% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-cnet.R
\name{adjust_cnet_set_relayout_gene}
\alias{adjust_cnet_set_relayout_gene}
\title{Adjust Set nodes then relayout Gene nodes}
\usage{
adjust_cnet_set_relayout_gene(
  g,
  nodes = NULL,
  x = 0,
  y = 0,
  use_grep = TRUE,
  do_reorder = TRUE,
  spread_labels = TRUE,
  repulse = 4,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} Cnet object}

\item{nodes}{\code{character} vector of one or more Set nodes.}

\item{x, y}{\code{numeric} values passed to \code{nudge_igraph_node()}.}

\item{use_grep}{\code{logical} passed to \code{nudge_igraph_node()}.}

\item{do_reorder}{\code{logical} indicating whether nodes should
be re-positioned within each subcluster by calling
\code{reorderIgraphNodes()}.}

\item{spread_labels}{\code{logical} indicating whether labels
should be re-oriented around each node by calling
\code{spread_igraph_labels()}.}

\item{repulse}{\code{numeric} value passed to \code{layout_with_qfr()}
to indicate the repulsion force. Typical values range between
\code{3} for loosely-packed nodes, and \code{5} or higher for more
closely-packed nodes.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to functions
\code{nudge_igraph_node()}, \code{layout_with_qfr()}, \code{spread_igraph_labels()},
\code{reorderIgraphNodes()}.}
}
\description{
Adjust Cnet Set nodes then relayout Gene nodes
}
\details{
This function operates on a Cnet \code{igraph} object,
distinguished by node attribute \code{"nodeType"} with
value \code{"Gene"} for Gene nodes, and \code{"Set"} for Set nodes.

This function is intended to help move a \code{Set} node
to improve visual spacing between nodes, then it
re-positions only the \code{Gene} nodes using \code{layout_with_qfr()},
keeping the \code{Set} nodes in fixed positions.

It calls \code{nudge_igraph_node()} using arguments \verb{nodes,x,y},
then defines the layout coordinates of Set nodes
as \code{constraints} that therefore are not allowed to change
when calling \code{layout_with_qfr()}. Note that Gene node
coordinates are allowed to change, even if Gene nodes
were included in \code{nodes}.
}
\seealso{
Other jam cnet utilities:
\code{\link[=adjust_cnet_nodeset]{adjust_cnet_nodeset()}},
\code{\link[=apply_nodeset_spacing]{apply_nodeset_spacing()}},
\code{\link[=bulk_cnet_adjustments]{bulk_cnet_adjustments()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=get_cnet_nodeset_vector]{get_cnet_nodeset_vector()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=make_cnet_test]{make_cnet_test()}},
\code{\link[=relayout_nodegroups]{relayout_nodegroups()}}
}
\concept{jam cnet utilities}
