% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-apply-cnet-direction.R
\name{apply_cnet_direction}
\alias{apply_cnet_direction}
\title{Apply Cnet border color by directionality}
\usage{
apply_cnet_direction(
  cnet,
  hitim = NULL,
  col = circlize::colorRamp2(breaks = c(-1, 0, 1), colors = c("blue", "grey80",
    "firebrick3")),
  col_l_max = 80,
  hide_solo_pie = TRUE,
  frame_blank = default_igraph_values()$vertex$frame.color,
  frame_lwd_blank = 0.2,
  border_lwd = 2,
  do_reorder = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cnet}{\code{igraph} object with node attribute \code{"pie.color"}
populated as a \code{list} of \code{character} vectors, named by
enrichment. The enrichment names should match \code{colnames(hitim)}.}

\item{hitim}{\code{numeric} matrix with values centered at zero for
no change, positive values \code{+1} for up-regulation, and
negative values \code{-1} for down-regulation. The values are
converted to color using color function \code{col}.}

\item{col}{\code{function} that takes \code{numeric} input and assigns
a color. The default assigns red for positive values, blue
for negative values, and white for zero, using
\code{colorjam::col_div_xf(1.2)}.}

\item{col_l_max}{\code{numeric} maximum HCL Lightness, default 80, for
output colors. For example, the middle color in \code{"RdBu_r"}
is nearly white, the \code{col_l_max} can be used to apply a darker grey.}

\item{hide_solo_pie}{\code{logical} default TRUE, whether a single-color
border for a multi-part pie node should only apply the color
to the overall node with 'frame.color', and not apply the color
to each pie wedge using 'pie.border'.
\itemize{
\item Default \code{hide_solo_pie=TRUE}: when all wedges (one or more)
have the same \code{pie.border} color, the 'pie.border' is defined
as \code{NA}, and 'frame.color' is assigned to this color.
The effect is to display the outline color and not each wedge.
Both 'pie.lwd' and 'frame.width' will be assigned \code{border_lwd},
and since 'pie.border' is \code{NA} it will not be rendered. Only the
'frame.color' will be rendered.
\item When \code{hide_solo_pie=FALSE} each pie wedge border color is assigned
to 'pie.border', \code{frame.color} will be assigned 'frame_blank',
and 'frame.width' will be assigned 'frame_lwd_blank' which is useful
for displaying a small outer frame for each node.
}}

\item{frame_blank}{\code{character} string to define the color used
for 'frame.color' when colors are defined in 'pie.border'.
The default uses the igraph defaults, currently 'black'.
In this case, the frame is drawn around the inner \code{pie.border}
colors, and only serves to add visual clarity. The frame border
can be blank \code{frame_blank="transparent"} or can be a thinner line,
controlled with \code{frame_lwd_blank=0.2}.}

\item{frame_lwd_blank}{\code{numeric} line width, default 0.2,
for nodes that have "blank" frame, which also means the
'pie.border' colors must also defined. In this case
the frame border can be invisible (\code{frame_lwd_blank=0}) or
a very thin line (default \code{frame_lwd_blank=0.2}) to surround the
inner borders drawn with 'pie.border'.}

\item{border_lwd}{\code{numeric} line width, default 2, used when a node
matches \code{rownames(hitim)}.
When the colors are applied to 'pie.color',
the border is defined with 'pie.lwd'.
When colors are applied to 'frame.color', the border is defined
with 'frame.width'.}

\item{do_reorder}{\code{logical} default FALSE, whether to reorder nodes
by node attributes such as color and border, by calling
\code{reorder_igraph_nodes()}.
When \code{do_reorder=TRUE}, other relevant arguments are passed
through \code{...} to \code{reorder_igraph_nodes()} in particular:
\itemize{
\item \code{colorV}: to control the expected order of colors.
It should be supplied if known upfront.
\item \code{sortAttributes}: to customize the default attribute sort order.
\item \code{nodeSortBy}: to customize the x-/y- axis arrangement.
\item \code{orderByAspect}: to enable x-/y- sorting by the aspect ratio of
nodes in the group. For example, tall-skinny node groups should sort by
y-axis first, short-wide node groups should sort by x-axis first.
}}

\item{...}{additional arguments are passed to \code{reorder_igraph_nodes()}
when \code{do_reorder=TRUE}.}
}
\description{
Apply Cnet border color by directionality
}
\details{
This function specifically requires the \code{"pie.color"}
node attribute of the \code{cnet} object \code{igraph} is populated
with a \code{list} of colors whose name is the enrichment,
and matches \code{colnames(hitim)} of the supplied hit
incidence matrix.

The node names \code{V(cnet)$name} also must match \code{rownames(hitim)},
otherwise the \code{"pie.border"} and \code{"coloredrect.border"} is assigned
its previous value.
In that case \code{"frame.color"} retains its original value.

This function may be run multiple times with different \code{hitim},
for example using \code{geneIMdirection} and \code{enrichIMdirection} in two steps.

When there are multiple colors for a given node, they are populated in
node attributes \code{"pie.border"} in \code{list} form. For these nodes,
the attribute \code{"frame.color"} is set to \code{NULL} so it is not displayed
on top of the pie wedge colors.

When there are no matching \code{rownames(hitim)}, or all colors
are identical, the \code{"pie.border"} is populated with \code{NULL},
so the pie wedges do not each show a color. Also for single-wedge
pie nodes, this process avoids drawing a small line at the top
of each node.
Instead the attribute \code{"frame.color"} is populated with the
one unique border color, so that only the outer border is colorized.
\subsection{Todo}{
\itemize{
\item Consider permitting input to be a simple igraph, then creating
pie.color, pie.border, pie.border.lwd attributes even if they
previously did not exist. It would probably optionally also use
\code{colorV} to colorize each component of multi-component nodes.
}
}
}
\seealso{
Other jam cnet igraph functions:
\code{\link[=rotate_igraph_layout]{rotate_igraph_layout()}}
}
\concept{jam cnet igraph functions}
