% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{apply_color_cap}
\alias{apply_color_cap}
\title{Apply color channel numeric range cap}
\usage{
apply_color_cap(x, channel = "l", range = c(0, 100), space = "hcl", ...)
}
\arguments{
\item{x}{vector colors, or list of color vectors. When input
is a \code{list}, the list is flattened, operations are performed,
then the result is split back into the original structure.}

\item{channel}{character channel recognized by \code{farver::set_channel()}.}

\item{range}{numeric range allowed for values returned by
\code{farver::get_channel()}.}

\item{space}{character name of a color space recognized by
\code{farver::set_channel()}.}

\item{...}{additional arguments are passed to \code{farver::get_channel()}
and \code{farver::set_channel()}.}
}
\description{
Apply color channel numeric range cap
}
\details{
This function extends \code{farver::set_channel()} by enforcing
a numeric range for any numeric channel accessible by
the \code{farver} package. Common use is to restrict the luminance
\code{"l"} channel of an \code{"hcl"} color from \code{c(0,100)} to \code{c(70,100)}
in order to make all colors brighter.
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
