% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-ashape.R
\name{ashape}
\alias{ashape}
\title{Alpha shape calculation}
\usage{
ashape(x, y = NULL, alpha, ...)
}
\arguments{
\item{x, y}{\code{numeric} vector with coordinate points.}

\item{alpha}{\code{numeric} with alpha threshold to use.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{ashape} object, which is a \code{list} containing:
\itemize{
\item edges: x,y coordinates of Delauney triangulation of the alpha-shape.
\item length: length of the alpha-shape.
\item alpha: value of alpha used.
\item alpha.extremes: \code{integer} index of points which were alpha-extremes.
\item delvor.obj: \code{delvor} object with Delauney/Voronoi supporting data.
\item x: x,y coordinates of input data
}
}
\description{
Alpha shape calculation for a set of points, and alpha threshold
}
\details{
This function is primarily intended to be called by \code{make_point_hull()},
since that function also iterates \code{alpha} values until it finds a
suitable, and successful, threshold.
}
\examples{
n <- 300
theta <- runif(n, 0, 2*pi)
r <- sqrt(runif(n, 0.25^2, 0.5^2))
x <- cbind(0.5+r*cos(theta), 0.5+r*sin(theta))
alpha <- 0.1
ashape.obj <- ashape(x, alpha=alpha)

plot(ashape.obj$x, asp=1)
segments(x0=ashape.obj$edges[, "x1"], x1=ashape.obj$edges[, "x2"],
   y0=ashape.obj$edges[, "y1"], y1=ashape.obj$edges[, "y2"])

}
\seealso{
Other jam utility functions:
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
