% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{avg_angles}
\alias{avg_angles}
\title{Average geometric angles}
\usage{
avg_angles(x, w = 1, maxAngle = 360, na.rm = TRUE, ...)
}
\description{
Average geometric angles
}
\examples{
x <- c(90, 180);
avg_angles(x);

}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
