% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-docs-MemPlotFolio.R
\docType{methods}
\name{Cnet-methods}
\alias{Cnet-methods}
\title{Cnet-methods for MemPlotFolio}
\arguments{
\item{x}{\code{MemPlotFolio} object}

\item{type}{\code{character} used for \code{CnetCollapsed()}}

\item{num}{\code{integer} number of exemplar pathways, used with
\code{CnetExemplar()}}

\item{cluster}{\code{character} name of pathway cluster which
matches an entry in \code{Clusters(x)}.}

\item{do_plot}{\code{logical} whether to render the igraph using
\code{jam_igraph()}.}

\item{legend_x, legend_y}{passed to \code{mem_legend()}. The default
\code{legend_x='bottomleft'}. Use \code{legend_x=FALSE} to hide the
legend.}

\item{main}{\code{character} with plot title.
\itemize{
\item For \code{CnetCluster()} if \code{ClusterLabels(x)} is defined, it
will use the corresponding label as plot title, otherwise
if it uses \code{names(Clusters(x))}.
To hide the title, use main=FALSE.
}}

\item{...}{additional arguments may be recognized.}
}
\value{
\code{igraph} Cnet object as defined for each method above.
}
\description{
Methods to create Concept Network (Cnet) plots
from a \code{MemPlotFolio} object.
}
\details{
The following methods are defined for \code{MemPlotFolio}:

\describe{
\item{\code{CnetCollapsed(x, type, do_plot, legend_x, legend_y)}}{
Returns \code{igraph} Cnet object with pathways collapsed
using \code{Clusters()}.
Argument \code{type} controls the pathway cluster label:
'title' uses the cluster title; 'set' uses the abbreviated
set names; 'cluster' uses \code{ClusterLabels()} if
defined.
Append \strong{\code{'2'}} suffix to hide gene labels, for example
\code{type='title2'} will display pathway cluster title, and
hide gene labels.
}
\item{\code{CnetExemplar(x, num, do_plot, legend_x, legend_y, main)}}{
Returns \code{igraph} Cnet object with exemplar pathways
for each cluster in \code{Clusters()}.
Argument 'num' defines the number of exemplar pathways
used for each cluster. It can be a single value, or
vector of values applied in order of \code{Clusters()}.
}
\item{\code{CnetCluster(x, cluster, do_plot, legend_x, legend_y, main)}}{
Returns \code{igraph} Cnet object foran individual cluster
defined in \code{Clusters()}.
Argument 'cluster' defines the cluster, and must match
an entry in \code{Clusters()} or be an \code{integer} index to
the cluster.
}
}
}
