% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{cnet2im}
\alias{cnet2im}
\title{Convert Cnet igraph to incidence matrix}
\usage{
cnet2im(g = NULL, df = NULL, ...)
}
\arguments{
\item{g}{igraph object containing Cnet data, specifically vertex
attribute name "nodeType" with values "Set" and "Gene", and
where "Set" nodes are only connected to "Gene" nodes.}

\item{df}{data.frame as optional input instead of \code{g}, usually
the result of a previous call to \code{cnet2df()}.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric matrix with colnames defined by \code{"Set"} node
names, and rownames defined by \code{"Gene"} node names.
}
\description{
Convert Cnet igraph to incidence matrix
}
\details{
This function takes igraph object containing "Cnet" data,
including vertex attribute \code{"nodeType"} with values \code{"Set"}
and \code{"Gene"}, and where \code{"Set"} nodes are only connected
to \code{"Gene"} nodes. It returns an incidence matrix whose
columns are "Set" node names and whose rows are "Gene"
node names.
}
\seealso{
Other jam conversion functions:
\code{\link[=cnet2df]{cnet2df()}},
\code{\link[=enrichDF2enrichResult]{enrichDF2enrichResult()}},
\code{\link[=enrichList2IM]{enrichList2IM()}},
\code{\link[=enrichList2df]{enrichList2df()}}
}
\concept{jam conversion functions}
