% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{cnetplot_internalJam}
\alias{cnetplot_internalJam}
\title{cnetplot internal function}
\usage{
cnetplot_internalJam(inputList, categorySize = "geneNum",
  showCategory = 5, pvalue = NULL, foldChange = NULL, fixed = TRUE,
  DE.foldChange = NULL, categoryColor = "#E5C494",
  geneColor = "#B3B3B3", colorRamp = "RdBu_r", ...)
}
\description{
cnetplot internal function
}
\details{
This function is intended to mimic the `DOSE:::cnetplot_internal()`
function to support `cnetplotJam()` customizations, including
not plotting the output, and including additional custom igraph
attributes.
}
\seealso{
Other jam igraph functions: \code{\link{cnetplotJam}},
  \code{\link{igraph2pieGraph}},
  \code{\link{layout_with_qfr}},
  \code{\link{shape.coloredrectangle.plot}}
}
\concept{jam igraph functions}
