% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-color-nodes-edges.R
\name{color_edges_by_nodes}
\alias{color_edges_by_nodes}
\title{Colorize igraph edges by nodes}
\usage{
color_edges_by_nodes(
  g,
  edge_alpha = NULL,
  Crange = c(0, 100),
  Lrange = c(0, 65),
  ...
)
}
\arguments{
\item{g}{\code{igraph} object that contains vertex node attribute \code{"color"}
as seen with \code{igraph::vertex_attr(g, "color")}.}

\item{edge_alpha}{\code{numeric} or \code{NULL}, where numeric value sets
the edge alpha transparency, where \code{edge_alpha=0} is completely
transparent, \code{edge_alpha=0.5} is 50\% transparent, and \code{edge_alpha=1}
is completely not transparent, and is opaque. When \code{edge_alpha=NULL}
the alpha values are supplied by \code{colorjam::blend_colors()}
which blends the two values.}

\item{...}{additional arguments are passed to \code{colorjam::blend_colors()}.}
}
\value{
\code{igraph} object with edge color attribute updated to
represent the result of blending node colors, seen by
\code{igraph::edge_attr(g)$color}.
}
\description{
Colorize igraph edges using node colors
}
\details{
This function colorizes edges by blending colors for the
nodes involved, by calling \code{colorjam::blend_colors()}.

The color for each node depends upon the node shape, so
the color or colors used to render each node shape will
be used for the edge. For example:
\itemize{
\item \code{shape="pie"} uses the average color from \code{V(g)$pie.color}
\item \code{shape="coloredrectangle"} uses the avereage color from
\code{V(g)$coloredrect.color}
\item everything else uses \code{V(g)$color}
}
}
\seealso{
Other jam igraph utilities:
\code{\link[=color_edges_by_nodegroups]{color_edges_by_nodegroups()}},
\code{\link[=color_nodes_by_nodegroups]{color_nodes_by_nodegroups()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=jam_igraph]{jam_igraph()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=subgraph_jam]{subgraph_jam()}}
}
\concept{jam igraph utilities}
