% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-color-nodes-edges.R
\name{color_nodes_by_nodegroups}
\alias{color_nodes_by_nodegroups}
\title{Color edges by nodegroups}
\usage{
color_nodes_by_nodegroups(
  g,
  nodegroups = NULL,
  nodegroup_colors = NULL,
  node_alpha = NULL,
  Crange = NULL,
  Lrange = NULL,
  color_attributes = c("color"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object that contains vertex node attribute \code{"color"}
as seen with \code{igraph::vertex_attr(g, "color")}.}

\item{nodegroups}{\code{list} or \code{communities} object that references
nodes in \code{g} and assigns one or more to nodegroups.
Default NULL calls \code{get_cnet_nodeset()}, passing '...'
for customization.}

\item{nodegroup_colors}{\code{character} or default NULL which assigns
categorical colors by \code{colorjam::rainbowJam()}.}

\item{node_alpha}{\code{numeric} of default NULL uses 1, with opacity
where 0 is transparent, 1 is completely solid.}

\item{Crange, Lrange}{\code{numeric} ranges passed to \code{colorjam::rainbowJam()}
to restrict the chroma and luminance of categorical colors.
Default NULL uses the defaults in \code{colorjam::rainbowJam()}.}

\item{color_attributes}{\code{character} vector of node attributes to
assign new colors, default 'color' assigns color to the node itself.
Note that this color is not seen when shape is 'pie' or 'jampie',
instead use 'pie.color'.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to\code{colorjam::rainbowJam()},
notably 'preset' which defines the color wheel used.}
}
\value{
\code{igraph} object with edge color attribute updated to
represent the result of blending node colors, seen by
\code{igraph::edge_attr(g)$color}.
}
\description{
Color edges by nodegroups
}
\examples{
cnet1 <- make_cnet_test()
igraph::vertex_attr(cnet1, "shape") <- "circle";
cnet2 <- color_nodes_by_nodegroups(cnet1);
# behold color
jam_igraph(cnet2)

# you can enable mark.groups which uses the same colorjam::rainbowJam()
jam_igraph(cnet2, mark.groups=TRUE)

}
\seealso{
Other jam igraph utilities:
\code{\link[=color_edges_by_nodegroups]{color_edges_by_nodegroups()}},
\code{\link[=color_edges_by_nodes]{color_edges_by_nodes()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=jam_igraph]{jam_igraph()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=subgraph_jam]{subgraph_jam()}}
}
\concept{jam igraph utilities}
