% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{colors_from_list}
\alias{colors_from_list}
\title{Ordered colors from a list of color vectors}
\usage{
colors_from_list(x, return_type = c("colors", "order"), verbose = FALSE, ...)
}
\arguments{
\item{x}{list of character vectors that contain valid R colors.}
}
\value{
character vector of unique colors in \code{x}
}
\description{
Ordered colors from a list of color vectors
}
\details{
This function takes a list of colors and returns the unique
order of colors based upon the order in vectors of the list.
It is mainly intended to be called by \code{reorderIgraphNodes()},
however the function is useful for inferring the proper order
of unique colors from a list of various subsets of colors.

The basic assumption is that there exists one true order of
unique colors, and that each vector in the list contains a
subset of those colors which is consistent with this
true order of colors.

The function uses only vectors that contain two or more
colors, and therefore requires that all unique colors are
present in the subset of vectos in the list where length >= 2.
It then uses vectors with two or more colors, calculates
the average observed rank for each color, then uses that
average rank to define the overall color order.

If not all unique colors are present in vectors with two or
more colors, the fallback sort uses \code{colorjam::sort_colors()}.
}
\seealso{
Other jam list functions:
\code{\link[=im2list]{im2list()}},
\code{\link[=imSigned2list]{imSigned2list()}},
\code{\link[=list2concordance]{list2concordance()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=list2imSigned]{list2imSigned()}}
}
\concept{jam list functions}
