% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-deconcat.R
\name{deconcat_df2}
\alias{deconcat_df2}
\title{Deconcatenate delimited column values in a data.frame}
\usage{
deconcat_df2(x, column, split = "[,; |/]+", blank = "", ...)
}
\arguments{
\item{x}{\code{data.frame} or compatible object}

\item{column}{\code{character} vector with one or more \code{colnames(x)}
that should be de-concatenated.}

\item{split}{\code{character} pattern used by \code{strsplit()} to split
multiple values in each column.}

\item{blank}{\code{character} string used to replace entries that
would otherwise be zero-length as returned by \code{strsplit()}.}

\item{...}{additional arguments are ignored.}
}
\description{
Deconcatenate delimited column values in a data.frame
}
\details{
This function deconcatenates delimited values in a column
of a \code{data.frame} by calling \code{strsplit()} on column values,
and repeating values in all other columns to match \code{lengths()}
following \code{strsplit()}.

This function includes a correction for cases where \code{strsplit()}
would otherwise return zero-length entries, and which would
otherwise be dropped from the output. From this function,
zero-length entries are replaced with \code{blank=""} so these
rows are not dropped from the output.
}
\examples{
df <- data.frame(one=c("AB", "BC", "AC"),
   two=c("a,b", "b,c", "a,c"));
deconcat_df2(df, column="two")

}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
