% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-bivariatecolor.R
\name{display_colorRamp2D}
\alias{display_colorRamp2D}
\title{Display colors from bivariate color function}
\usage{
display_colorRamp2D(
  col_fun,
  pretty.n = 12,
  xlab = "",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{col_fun}{\code{function} output from \code{colorRamp2D()}.}

\item{pretty.n}{\code{numeric} passed to \code{pretty()} to determine the
y-axis and x-axis label breaks, respectively, or
\code{NULL} to use breaks encoded in \code{col_fun}.}

\item{xlab, ylab, main}{\code{character} strings passed to \code{title()}.}

\item{...}{additional arguments are passed to \code{title()}.}
}
\description{
Display colors from bivariate color function
}
\details{
This function provides a very rudimentary method for displaying
a bivariate 2-D color matrix out from \code{colorRamp2D()}.
}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
