% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{drawEllipse}
\alias{drawEllipse}
\title{Draw ellipse}
\usage{
drawEllipse(
  x,
  y,
  a = 1,
  b = 1,
  angle = 0,
  segment = NULL,
  arc.only = TRUE,
  nv = 100,
  deg = TRUE,
  border = NULL,
  col = NA,
  lty = 1,
  lwd = 1,
  draw = TRUE,
  ...
)
}
\arguments{
\item{x, y}{\code{numeric} coordinates, where x can be a two-column numeric
matrix of x,y coordinates.}

\item{a, b}{\code{numeric} values indicating x- and y-axis radius, before
rotation if \code{angle} is non-zero.}

\item{angle}{\code{numeric} value indicating the rotation of ellipse.}

\item{segment}{NULL or \code{numeric} vector of two values indicating the
start and end angles for the ellipse, prior to rotation.}

\item{arc.only}{\code{logical} indicating whether to draw the ellipse
arc without connecting to the center of the ellipse. Set
\code{arc.only=FALSE} when segment does not include the full circle,
to draw only the wedge.}

\item{nv}{\code{numeric} the number of vertices around the center to draw.}

\item{deg}{\code{logical} indicating whether input \code{angle} and \code{segment}
values are in degrees, or \code{deg=FALSE} for radians.}

\item{border, col, lty, lwd}{arguments passed to \code{graphics::polygon()}.}

\item{draw}{\code{logical} indicating whether to draw the ellipse.}

\item{...}{additional arguments are passed to \code{graphics::polygon()}
when \code{draw=TRUE}.}
}
\value{
invisible list of x,y coordinates
}
\description{
Draw ellipse
}
\details{
This function draws an ellipse centered on the given coordinates,
rotated the given degrees relative to the center point, with give
x- and y-axis radius values.
}
\examples{
par("mar"=c(2, 2, 2, 2));
plot(NULL,
   type="n",
   xlim=c(-5, 20),
   ylim=c(-5, 18),
   ylab="", xlab="", bty="L",
   asp=1);
xy <- drawEllipse(
   x=c(1, 11, 11, 11),
   y=c(1, 11, 11, 11),
   a=c(5, 5, 5*1.5, 5),
   b=c(2, 2, 2*1.5, 2),
   angle=c(20, -15, -15, -15),
   segment=c(0, 360, 0, 120, 120, 240, 240, 360),
   arc.only=c(TRUE, FALSE, FALSE, TRUE),
   col=jamba::alpha2col(c("red", "gold", "dodgerblue", "darkorchid"), alpha=0.5),
   border=c("red", "gold", "dodgerblue", "darkorchid"),
   lwd=1,
   nv=99)
points(x=c(1, 11), y=c(1, 11), pch=20, cex=2)
jamba::drawLabels(x=c(12, 3, 13, 5),
   y=c(14, 10, 9, 2),
   labelCex=0.7,
   drawBox=FALSE,
   adjPreset=c("topright", "left", "bottomright", "top"),
   txt=c("0-120 degrees,\nangle=-15,\narc.only=TRUE",
      "120-240 degrees,\nangle=-15,\narc.only=TRUE,\nlarger radius",
      "240-360 degrees,\nangle=-15,\narc.only=FALSE",
      "angle=20"))

}
\seealso{
Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
