% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-enrichdf2er.R
\name{enrichDF2enrichResult}
\alias{enrichDF2enrichResult}
\title{Convert data.frame to enrichResult}
\usage{
enrichDF2enrichResult(
  enrichDF = NULL,
  pvalueCutoff = 1,
  pAdjustMethod = "none",
  keyColname = c("itemsetID", "ID", "Name", "Pathway"),
  pathGenes = "pathGenes",
  geneColname = c("geneNames", "geneID", "Gene", "Genes"),
  geneHits = "geneHits",
  geneRatioColname = c("GeneRatio", "^Ratio"),
  geneDelim = "[,/ ]+",
  geneSep = "/",
  pvalueColname = c("P.Value", "Pvalue", "Pval", "FDR", "adj.P.Val"),
  padjustColname = c("p.adjust", "p.adjusted", "padjust", "adjp", "adj.P.Val", "FDR"),
  qvalueColname = c("Q.Value", "Qvalue", "qval", "FDR"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  readable = NULL,
  msigdbGmtT = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichDF}{\code{data.frame} representing gene set enrichment
results.}

\item{pvalueCutoff}{\code{numeric} value range 0 to 1, to define the
P-value threshold for enrichment results to be considered
in downstream processing.}

\item{pAdjustMethod}{\code{character} string to define the P-value
adjustment method, or \code{"none"} for no additional adjustment.
See \code{stats::p.adjust()} for valid values.}

\item{keyColname}{\code{character} value of the \code{colnames(enrichDF)}
containing the unique row identifier. It can be a pathway_ID
or any uniquely identifying value.}

\item{pathGenes}{\code{character} or value of the \code{colnames(enrichDF)}
containing the number of genes in each pathway. This value will be
derived from \code{geneRatioColname} if needed.}

\item{geneColname}{\code{character} value of the \code{colnames(enrichDF)}
containing delimiited genes in each pathway.}

\item{geneHits}{\code{character} value of the \code{colnames(enrichDF)}
containing the integer count of the gene hits in each pathway.
This value will be derived from \code{geneRatioColname} if needed.}

\item{geneRatioColname}{\code{character} value of the \code{colnames(enrichDF)}
containing the character ratio of gene hits to pathway size,
in format "50/100". This value is used when either \code{"pathGenes"}
or \code{"geneHits"} are not supplied.}

\item{geneDelim}{\code{character} regular expression pattern used to separate
genes in the \code{pathGenes} column into a vector of character
values.}

\item{geneSep}{\code{character} string with separator to use for output
genes. Only \code{'/'} is supported for most \code{enrichResult} methods.}

\item{pvalueColname, padjustColname, qvalueColname}{\code{character} vector
to search \code{colnames(enrichDF)}, with enrichment P-value,
adjusted P-value, and Q-value, respectively.
\itemize{
\item The P-value is stored in 'pvalue' in the output object, to match
\code{enrichResult} convention.
\item When there is no adjusted P-value column matched, by default it
uses the P-value without change, since some downstream methods
require 'p.adjust' exists.
\item When there is no Q-value column matched, by default it uses
the adjusted P-value column value 'p.adjust' without change.
Note that the 'p.adjust' may also contain the value from 'pvalue'
as described above.
}}

\item{descriptionColname}{\code{character} vector
to search \code{colnames(enrichDF)}, for pathway or gene set description.}

\item{readable}{\code{logical} default NULL, sets the 'readable' flag for
the resulting \code{enrichResult} object.}

\item{msigdbGmtT}{optional 'GmtT' object (not currently implemented)}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{enrichResult} object.
}
\description{
Convert data.frame to enrichResult
}
\details{
This function takes a \code{data.frame} containing gene set enrichment
results, and converts it to a proper \verb{enrichResult`` object defined in }DOSE\verb{for R-4.5 and older, or}enrichit` for
R-4.6 and newer.

This object is supported by other functions in the \code{clusterProfiler}
suite of tools.
}
\seealso{
Other jam import functions:
\code{\link[=IPAlist_to_hits]{IPAlist_to_hits()}},
\code{\link[=importIPAenrichment]{importIPAenrichment()}}

Other jam conversion functions:
\code{\link[=cnet2df]{cnet2df()}},
\code{\link[=cnet2im]{cnet2im()}},
\code{\link[=enrichList2IM]{enrichList2IM()}},
\code{\link[=enrichList2df]{enrichList2df()}}
}
\concept{jam conversion functions}
\concept{jam import functions}
