% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{enrichList2geneHitList}
\alias{enrichList2geneHitList}
\title{Extract gene hit list from list of enrichResult}
\usage{
enrichList2geneHitList(
  enrichList,
  geneColname,
  geneDelim = "[,/ ]",
  make_unique = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichList}{\code{list} of \code{enrichResult} objects}

\item{geneColname}{\code{character} string with the column name containing
delimited gene identifiers.}

\item{geneDelim}{\code{character} regular expression used to split delimited
gene values. By default, \code{enrichResult} uses '/' forward slash
as delimiter, however the default here will split any space or
comma as well.}

\item{make_unique}{\code{logical} default TRUE, whether to return only unique
genes per set, or potentially multiple genes per set. Typically there
should only ever be one instance of a gene per set, but through a
variety of other mechanisms they may exist, for example if two gene
identifiers are resolved into the same gene symbol.}

\item{verbose}{\code{logical} whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of character vectors, containing the unique
set of genes involved in each enrichment.
}
\description{
Extract gene hit list from list of enrichResult
}
\details{
This function is mainly for internal use in multienrichjam,
it takes a list of \code{enrichResult} objects, and determines
the full set of genes involved in each \code{enrichResult}.

Note that genes are sorted using \code{jamba::mixedSort()} for
alpha-numeric sorting, based upon version sorting.

This function also works with \code{ComplexHeatmap::HeatmapList}
objects.
}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
