% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{enrichMapJam}
\alias{enrichMapJam}
\title{Create enrichMap igraph object from enrichResult, deprecated}
\usage{
enrichMapJam(
  x,
  n = 50,
  vertex.label.font = 1,
  vertex.label.cex = 1,
  keyColname = "ID",
  nodeLabel = c("Name", "Description", "ID"),
  descriptionColname = "Description",
  nodeLabelFunc = NULL,
  overlapThreshold = 0.2,
  msigdbGmtT = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{enrichResult} or \code{data.frame} containing
enrichment results, specifically expecting colnames to
contain one of \code{c("ID","Description","Name")}
to represent the node name, and \code{c("Description")} to represent
the description, if present.}

\item{n}{\code{numeric} value indicating the maximum number of nodes to
include in the final network.}

\item{vertex.label.font, vertex.label.cex}{\code{numeric} values
to define the default node label font and size.}

\item{keyColname, nodeLabel, descriptionColname}{\code{character} vectors
indicating the colname to use for the node name and label.}

\item{nodeLabelFunc}{\code{function}, default NULL uses
\code{fixSetLabels(x, width=30, ...)}. Use 'FALSE' to keep labels unchanged.
Note that the \code{igraph} node name remains the same as input,
and label is used as the visual label on the graph.
\itemize{
\item When providing a custom function, note that each node is called
individually and therefore the call is not vectorized.
}}

\item{overlapThreshold}{\code{numeric} value indicating the minimum
Jaccard overlap, where edges with lower values are deleted from
the \code{igraph} object.}

\item{msigdbGmtT}{not currently implemented.}

\item{...}{additional arguments are passed to \code{enrichDF2enrichResult()}
when the input \code{x} is a \code{data.frame}.}
}
\value{
\code{igraph} object with nodes representing each pathway,
sized based upon the number of genes involved in enrichment, and
colored based upon the \code{-log10(Pvalue)}
using \code{colorjam::col_linear_xf()}, a function that applies
a color gradient to a numeric range.
Each edge has attributes: \code{overlap} containing Jaccard overlap,
\code{overlap_count} with the number of genes in common between
the two nodes, and \code{overlap_max_pct} with the maximum percent
overlap between two nodes (overlap count)/(smaller node size).
}
\description{
Create enrichMap igraph object from enrichResult, deprecated,
use \code{mem2emap()}.
}
\details{
This function is a minor customization to \code{enrichplot::emapplot()},
which takes a single \code{enrichResult} and produces an Enrichment map
network. For the equivalent function using \code{Mem} as input, see
\code{mem2emap()}.

The major differences compared with \code{enrichplot}:
\itemize{
\item An \code{igraph} object is returned instead of \code{ggplot} object, since the
\code{igraph} object can be manipulated and is useful to review.
\item This function calculates overlap using \code{dist(...,method="binary")}
which is a much faster method for calculating the Jaccard overlap.
\item This function also calculates the overlap count, another helpful
measure for filtering network connections, for example to remove
links with only one gene, even if they overlap is above the
required threshold. Many spurious network connections are removed
with this filter, and it appears to be a helpful option.
}
}
\seealso{
Other jam deprecated functions:
\code{\link[=cnetplotJam]{cnetplotJam()}},
\code{\link[=grid_with_title]{grid_with_title()}},
\code{\link[=mem_multienrichplot]{mem_multienrichplot()}},
\code{\link[=with_qfr]{with_qfr()}}
}
\concept{jam deprecated functions}
