% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-subset-mem.R
\name{filter_mem_sets}
\alias{filter_mem_sets}
\title{Filter mem multienrichment object by Set names}
\usage{
filter_mem_sets(mem, includeSets = NULL, ...)
}
\arguments{
\item{mem}{\code{list} object returned by \code{multiEnrichMap()}}

\item{includeSets}{\code{character} vector of sets, matched with
\code{colnames(mem$memIM)}}

\item{...}{additional arguments are ignored}
}
\value{
\code{list} object of \code{mem} data content
}
\description{
Filter mem multienrichment object by Set names
}
\details{
This is intended to be an internal function. It simply takes a
\code{character} vector of set names, and subsets incidence matrix data
in the \code{mem} object. It performs no other filtering.
This function is called by \code{subset_mem()}, the recommended method
to subset a \code{mem} object.
}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{bulk_cnet_adjustments}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
