% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{find_colname}
\alias{find_colname}
\title{Find colname by character string or pattern matching}
\usage{
find_colname(
  pattern,
  x,
  max = 1,
  index = FALSE,
  require_non_na = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pattern}{character vector containing text strings or
regular expression patterns.}

\item{x}{input \code{data.frame} or other R object that
contains colnames.}

\item{max}{integer maximum number of results to return.}

\item{index}{logical indicating whether to return the column
index as an integer vector. When \code{index=FALSE} it returns
the matching \code{colnames(x)}; when \code{index=TRUE} it returns
the matching column numbers as an integer vector.}

\item{require_non_na}{logical indicating whether to require the
column to contain non-NA values, default is TRUE. The intent
of this function is to find colnames whose data will match
expectations, and when require_non_na is TRUE, this
function will continue until it finds a column with non-NA
values.}

\item{...}{additional arguments are passed to \code{jamba::provigrep()}.}
}
\value{
character vector with length \code{max}, or if no pattern
match is found it returns \code{NULL}. Also if there are no
\code{colnames(x)} then it returns \code{NULL}.
}
\description{
Find colname by character string or pattern matching
}
\details{
This function is intended to help find a colname given
a character vector of expected values or regular expression
patterns. By default it returns the first matching value,
but can return multiple if \code{max=Inf}.

If there are no \code{colnames(x)} then \code{NULL} is returned.

The order of operations:
\enumerate{
\item Match exact string.
\item Match exact string in case-insensitive manner.
\item Match the start of each string using \code{jamba::provigrep()}.
\item Match the end of each string using \code{jamba::provigrep()}.
\item Match each string using \code{jamba::provigrep()}.
}

The results from the first successful operation is returned.

When there are duplicate \code{colnames(x)} only the first
unique name is returned.
}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
