% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-get-cnet-nodeset.R
\name{get_cnet_nodeset}
\alias{get_cnet_nodeset}
\title{Get Cnet node set by connected Sets}
\usage{
get_cnet_nodeset(
  g,
  set_nodes = NULL,
  sep = ",",
  filter_set_only = TRUE,
  min_size = 1,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object specifically containing Cnet plot data,
with node attribute \code{"nodeType"} that has values \code{c("Gene", "Set")}.}

\item{set_nodes}{\code{character} vector, default NULL, used to subset
the returned data to nodes which are connected to these 'Set' nodes.
of one or more set names, which are
defined here as the names of the nodes with \code{nodeType="Set"}
to which each node with \code{nodeType="Gene"} is connected.
For example \code{set_nodes=c("A", "B")} will return all Gene nodes
that are connected only to Set nodes \code{"A"} and \code{"B"}. If
\code{set_nodes=NULL} then this function returns a \code{list} with
all set_nodes observed.}

\item{sep}{\code{character} string used as a delimited when combining
\code{set_nodes} into a name.}

\item{filter_set_only}{\code{logical} default TRUE, whether to exclude
Set nodes, with vertex attribute nodeType='Set', from being
members of nodegroups. In principle, 'Gene' nodes are typically
members of Cnet nodesets, and 'Set' nodes do not usually
share the same 'Gene' nodes with other 'Set' nodes, and therefore
are not included.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} named by nodeset, with \code{character} vectors of
node names in each nodeset. When \code{filter_set_only=TRUE} by default,
the nodes will not contain 'Set' nodes.
The \code{list} is named by nodes connected by nodes in each node group,
using \code{sep} as delimiter.
}
\description{
Get Cnet node set by connected Sets, using nodeType to
delineate the two types of nodes in the graph.
}
\details{
This function operates on a Cnet \code{igraph} object,
distinguished by node attribute \code{'nodeType'} with
values: \code{'Gene'} for Gene nodes, and \code{'Set'} for Set nodes.
This function returns Gene nodes that are connected
to the Sets given by argument \code{set_nodes}. It is
useful to identify which genes are connected to
Set \code{"A"} and \code{"B"} for example, in other words the
Gene nodes in a specific subcluster of the Cnet
\code{igraph}.

When \code{set_nodes} is \code{NULL}, this function returns a \code{list}
containing \code{character} vectors, where each vector represents
Gene nodes that are connected to each combination of
Set nodes. This option is useful for obtaining all
possible Gene subclusters.

This function is also called by \code{adjust_cnet_nodeset()} in
order to manipulate all nodes in a subcluster as a group,
for example calling \code{nudge_igraph_node()} on the whole set
of nodes.

Note that when input data do not contain vertex attribute 'nodeType',
data are still returned but cannot be filtered by 'Set' and 'Gene',
and so will contain all possible node-connection groupings.
}
\examples{
cnet1 <- make_cnet_test()
get_cnet_nodeset(cnet1)

get_cnet_nodeset(cnet1, c("SetB,SetD"))

}
\seealso{
Other jam igraph utilities:
\code{\link[=color_edges_by_nodegroups]{color_edges_by_nodegroups()}},
\code{\link[=color_edges_by_nodes]{color_edges_by_nodes()}},
\code{\link[=color_nodes_by_nodegroups]{color_nodes_by_nodegroups()}},
\code{\link[=jam_igraph]{jam_igraph()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=subgraph_jam]{subgraph_jam()}}

Other jam cnet utilities:
\code{\link[=adjust_cnet_nodeset]{adjust_cnet_nodeset()}},
\code{\link[=adjust_cnet_set_relayout_gene]{adjust_cnet_set_relayout_gene()}},
\code{\link[=apply_nodeset_spacing]{apply_nodeset_spacing()}},
\code{\link[=bulk_cnet_adjustments]{bulk_cnet_adjustments()}},
\code{\link[=get_cnet_nodeset_vector]{get_cnet_nodeset_vector()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=make_cnet_test]{make_cnet_test()}},
\code{\link[=relayout_nodegroups]{relayout_nodegroups()}}
}
\concept{jam cnet utilities}
\concept{jam igraph utilities}
