% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-get-cnet-nodeset.R
\name{get_cnet_nodeset_vector}
\alias{get_cnet_nodeset_vector}
\title{Get Cnet nodesets as a named vector}
\usage{
get_cnet_nodeset_vector(g, nodegroups = NULL, filter_set_only = TRUE, ...)
}
\arguments{
\item{g}{\code{igraph} Cnet object}

\item{nodegroups}{default is NULL, which calls \code{get_cnet_nodeset()}.
Also accepts \code{list} or \code{communities}.}

\item{filter_set_only}{\code{logical} default TRUE, passed to
\code{get_cnet_nodeset()} to determine whether 'Set' nodes are
assigned a nodeset.
\itemize{
\item When TRUE, 'Set' nodes are filtered out,
therefore their nodeset becomes NA.
\item When FALSE, 'Set' nodes are also assigned to a nodeset.
}}

\item{...}{additional arguments passed to \code{get_cnet_nodeset()}}
}
\value{
\code{character} vector named by node, whose value is the
name of the nodegroup. All nodes in \code{g} are returned in order,
with \code{NA} values used when there is no nodegroup assigned to
a particular node.
}
\description{
Get Cnet nodesets as a named vector, with nodegroups named by node
}
\details{
This function is a simple conversion of \code{list} output from
\code{get_cnet_nodeset()} that returns a \code{character} vector alternative.

It is easier to answer: "What nodegroup is this node in?"
}
\examples{
cnet <- make_cnet_test()
nsv <- get_cnet_nodeset_vector(cnet)
nsv

}
\seealso{
Other jam cnet utilities:
\code{\link[=adjust_cnet_nodeset]{adjust_cnet_nodeset()}},
\code{\link[=adjust_cnet_set_relayout_gene]{adjust_cnet_set_relayout_gene()}},
\code{\link[=apply_nodeset_spacing]{apply_nodeset_spacing()}},
\code{\link[=bulk_cnet_adjustments]{bulk_cnet_adjustments()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=make_cnet_test]{make_cnet_test()}},
\code{\link[=relayout_nodegroups]{relayout_nodegroups()}}
}
\concept{jam cnet utilities}
