% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-internals.R
\name{get_igraph_arrow_mode}
\alias{get_igraph_arrow_mode}
\title{Get igraph arrow mode}
\usage{
get_igraph_arrow_mode(graph, arrow.mode = NULL)
}
\description{
Get igraph arrow mode
}
\details{
This function mimics the internal function \code{igraph:::i.get.arrow.mode()}.
}
\seealso{
Other jam igraph internal functions:
\code{\link[=default_igraph_values]{default_igraph_values()}},
\code{\link[=jam_igraph_arrows]{jam_igraph_arrows()}},
\code{\link[=jam_plot_igraph]{jam_plot_igraph()}}
}
\concept{jam igraph internal functions}
