% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-gr2er.R
\name{gr2er}
\alias{gr2er}
\title{Convert gseaResult to enrichResult equivalent result content}
\usage{
gr2er(er, ...)
}
\arguments{
\item{er}{\code{gseaResult}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{gseaResult}
}
\description{
Convert gseaResult to enrichResult equivalent result content,
focused mainly on providing minimal result columns for downstream
work.
}
\details{
This function is intended for internal use, and is not a full
conversion of \code{gseaResult} to a proper \code{enrichResult}.
It simply makes minor modifications to colnames for convenience
within \code{multiEnrichMap()} and other supporting functions.

Specific changes:
\itemize{
\item Renames 'core_enrichment' to 'geneID' if 'geneID' does not exist.
\item Uses 'geneID' to create 'Count' column with number of genes.
}
}
\keyword{internal}
