% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{gsubs}
\alias{gsubs}
\title{Pattern replacement with multiple patterns}
\usage{
gsubs(pattern, replacement, x, ignore.case = TRUE,
  replaceMultiple = rep(TRUE, length(pattern)), ...)
}
\arguments{
\item{pattern}{character vector of patterns}

\item{replacement}{character vector of replacements}

\item{x}{character vector with input data to be curated}

\item{ignore.case}{logical indicating whether to perform
pattern matching in case-insensitive manner, where
\code{ignore.case=TRUE} will ignore the uppercase/lowercase
distinction.}

\item{...}{additional arguments are passed to \code{base::gsub()}
or \code{base::sub()}.}

\item{replace_multiple}{logical vector indicating whether to perform
global substitution, where \code{replace_multiple=FALSE} will
only replace the first occurrence of the pattern, using
\code{base::sub()}. Note that this vector can refer to individual
entries in \code{pattern}.}
}
\description{
Pattern replacement with multiple patterns
}
\details{
This function is a simple wrapper around \code{base::gsub()}
when considering a series of pattern-replacement
combinations. It applies each pattern match and replacement
in order and is therefore not vectorized.
}
