% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-handle-rowcol-splits.R
\name{handle_rowcol_splits}
\alias{handle_rowcol_splits}
\title{Handle row and column split parameters for gene-pathway data}
\usage{
handle_rowcol_splits(
  Mem,
  auto_split = TRUE,
  row_split,
  row_title,
  max_row_split = 12,
  cluster_rows,
  row_method,
  gene_im_weight = 0.5,
  column_split,
  column_title,
  max_column_split = 8,
  cluster_columns,
  column_method,
  enrich_im_weight = 0.3,
  trim_rows = TRUE,
  trim_columns = TRUE,
  p_cutoff = NULL,
  p_floor = NULL,
  seed = 123,
  verbose = FALSE,
  debug = FALSE,
  ...
)
}
\value{
\code{list} with
\itemize{
\item Mem
\item row_split
\item row_title
\item cluster_rows
\item column_split
\item column_title
\item cluster_columns
}
}
\description{
Handle row and column split parameters for gene-pathway data
}
\keyword{internal}
