% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{heatmap_column_order_deprecated}
\alias{heatmap_column_order_deprecated}
\title{Return Heatmap column order}
\usage{
heatmap_column_order_deprecated(hm)
}
\description{
Return Heatmap column order as a list of character vectors (deprecated)
}
\concept{jam deprecated functions

This function is deprecated, and is fully replaced
by \code{jamba::heatmap_column_order()}.

This function is a helpful utility to return the fully
qualified list of colnames in a \code{ComplexHeatmap::Heatmap}
object.}
