% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{heatmap_row_order_deprecated}
\alias{heatmap_row_order_deprecated}
\title{Return Heatmap row order (deprecated)}
\usage{
heatmap_row_order_deprecated(hm)
}
\description{
Return Heatmap row order as a list of character vectors
}
\concept{jam deprecated functions

This function is deprecated, and is fully replaced
by \code{jamba::heatmap_row_order()}.

This function is a helpful utility to return the fully
qualified list of rownames in a \code{ComplexHeatmap::Heatmap}
object.

This function also works with \code{ComplexHeatmap::HeatmapList}
objects.}
