% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{igraph2pieGraph}
\alias{igraph2pieGraph}
\title{Convert igraph to use pie node shapes}
\usage{
igraph2pieGraph(
  g,
  valueIM = NULL,
  valueIMcolors = NULL,
  colorV = NULL,
  updateLabels = FALSE,
  maxNchar = 62,
  backgroundColor = "white",
  seed = 123,
  defineLayout = FALSE,
  repulse = 3.6,
  removeNA = FALSE,
  NAvalues = c(NA, "transparent"),
  verbose = FALSE,
  ...
)
}
\description{
Convert igraph to use pie node shapes
}
\details{
This function converts an igraph to use pie node shapes,
where pie wedges are colored using values derived from a
numeric matrix \code{valueIM} or pre-defined in a character matrix
containing colors \code{valueIMcolors}.

Note that pie wedge sizes are equally-sized and do not vary
by score, instead the color intensity is applied to each pie
wedge.

Node names using \code{V(g)$name} matching \code{rownames(valueIMcolors)}
are colorized and the node shape is converted to pie. All other
nodes are not modified.

When \code{valueIMcolors} is not defined, it is derived from \code{valueIM}
using \code{colorjam::matrix2heatColors()}. In that case, \code{colorV} defines the color
used for numeric values in each column, and other options are passed
to \code{colorjam::matrix2heatColors()} via \code{...} arguments.
}
\seealso{
Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
