% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{im2list}
\alias{im2list}
\title{convert incidence matrix to list}
\usage{
im2list(x, empty = c(NA, "", 0, FALSE), ...)
}
\arguments{
\item{x}{\code{matrix} or equivalent object with \code{colnames(x)} indicating
list set names, and \code{rownames(x)} indicating list contents.}

\item{empty}{\code{character} vector of incidence matrix values that
should be considered "empty" and therefore do not indicate
the row in \code{x} is present for the given column in \code{x}.
All other items are considered to be present.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of \verb{character vectors}, where list names
are defined by \code{colnames(x)}, and list elements are vectors
that contain values from \code{rownames(x)}.
}
\description{
convert incidence matrix to list
}
\details{
This function converts an incidence \code{matrix}, or equivalent
\code{data.frame}, to a list. The \code{matrix} should contain either
numeric values such as \code{c(0, 1)}, or logical values such
as \code{c(TRUE,FALSE)}, otherwise values are considered either
zero == \code{FALSE}, or non-zero == \code{TRUE}.

The resulting list will be named by \code{colnames(x)} of the input,
and will contain members named by \code{rownames(x)} which are
either non-zero, or contain \code{TRUE}.

Values of \code{NA} are converted to zero \code{0} and therefore ignored.
}
\examples{
im <- matrix(c(0,1,-1,1,1,NA,-1,0,1),
   ncol=3,
   nrow=3,
   dimnames=list(letters[1:3], LETTERS[1:3]))
print(im);
# matrix input
im2list(im);

# data.frame
imdf <- data.frame(im);
print(imdf);
im2list(im);

# logical input
imtf <- (!im == 0);
print(imtf);
im2list(imtf);

}
\seealso{
Other jam list functions:
\code{\link[=colors_from_list]{colors_from_list()}},
\code{\link[=imSigned2list]{imSigned2list()}},
\code{\link[=list2concordance]{list2concordance()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=list2imSigned]{list2imSigned()}}
}
\concept{jam list functions}
