% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{importIPAenrichment}
\alias{importIPAenrichment}
\title{Import Ingenuity Pathway Analysis 'IPA' results}
\usage{
importIPAenrichment(
  ipaFile,
  headerGrep =
    "(^|\\t)((expr.|-log.|)p-value|Pvalue|Score($|\\t)|Symbol($|\\t)|Ratio($|\\t)|Consistency.Score|Master.Regulator($|\\t))",
  ipaNameGrep = c("Pathway", "Regulator$", "Regulators", "Regulator", "Disease",
    "Toxicity", "Category", "Categories", "Function", "Symbol$", "^ID$",
    "My.(Lists|Pathways)"),
  geneGrep = c("Molecules in Network", "Target molecules", "Molecules", "Symbol"),
  geneCurateFrom = c("[ ]*[(](complex|includes others)[)][ ]*", "^[, ]+|[, ]+$"),
  geneCurateTo = c("", ""),
  signColname = c("Expr Fold Change", "fold.*change", "log.*ratio", "log.*fold",
    "log.*fc", "lfc", "ratio", "fold", "fc"),
  signThreshold = 0,
  method = 1,
  sheet = 1,
  sep = "\\t",
  xlsxMultiSheet = TRUE,
  useXlsxSheetNames = FALSE,
  remove_blank_colnames = TRUE,
  convert_ipa_slash = TRUE,
  ipa_slash_sep = ":",
  revert_ipa_xref = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{ipaFile}{one of the four input types described above:
a character vector of text file names; a character vector of
Excel \code{.xlsx} file names; a list of \code{data.frame} objects.}

\item{headerGrep}{regular expression pattern used to
recognize header columns found in Ingenuity IPA enrichment data.}

\item{ipaNameGrep}{vector of regular expression patterns
used to recognize the name of the enriched entity,
for example the biological pathway, or network, or
disease category, etc.}

\item{geneGrep}{regular expression pattern used to recognize
the column containing genes, or the molecules tested for
enrichment which were found in the enriched entity.}

\item{geneCurateFrom, geneCurateTo}{vector of patterns and
replacements, respectively, used to curate values in
the gene column. These replacement rules are used to
ensure that genes are delimited consistently, with no
leading or trailing delimiters.}

\item{method}{integer value indicating the method used to
import data from a text file, where: \code{method=1} uses
\code{data.table::read.table()} and the \code{textConnection} argument;
\code{method=2} uses \code{readr::read_tsv()}. The motivation to use
\code{data.table::read.table()} is it performed better in the
presence of UTF-8 characters such as the alpha symbol.}

\item{sheet}{integer value used only when \code{ipaFile} is
a vector of Excel \code{.xlsx} files, and when the Excel
format includes multiple worksheets. This value will
extract enrichment data only from one worksheet from
each Excel file.}

\item{sep}{character string used when \code{ipaFile} is a vector
of text files, to split fields into columns. The default
will split fields by the tab character.}

\item{xlsxMultiSheet}{logical indicating whether input
Excel \code{.xlsx} files contain multiple worksheets.}

\item{useXlsxSheetNames}{logicl indicating whether to use the
Excel worksheet name for each imported enrichment table,
when importing from \code{.xlsx} files, and when
\code{xlsxMultiSheet=FALSE}. When \code{xlsxMultiSheet=TRUE} the
name is derived from the value matched using \code{ipaNameGrep},
because in this case, there are expected to me multiple
enrichment tables in one worksheet.}

\item{remove_blank_colnames}{\code{logical} indicating whether to drop
\code{colnames()} where all values are contained in \code{c(NA, "")}.
This option may be preferable \code{remove_blank_colnames=FALSE}
when all values in some column like \code{zScore} are \code{NA}, but
you would still like to retain the column for consistency
with other data. We found that IPA does not report \code{zScore}
values when there are only 4 or fewer genes involved in
each enrichment result.}

\item{convert_ipa_slash}{\code{logical} indicating whether to convert
IPA gene naming conventions, currently some genes are considered
one entity in the IPA system, for example \code{"HSPA1A/HSPA1B"} is
considered one gene, even though two Entrez gene entries
\code{"HSPA1A"} and \code{"HSPA1B"} can be represented. Regardless whether
one or both genes are provided to IPA, it considers it one
entity for the purpose of pathway enrichment hypergeometric testing.
Unfortunately, the forward slash \code{"/"} is also used by
\code{clusterProfiler} object \code{enrichResult} as gene delimiter, and
is hard-coded and cannot be changed. So it will automatically
consider \code{"HSPA1A/HSPA1B"} as two genes, causing a mismatch with
the IPA results.
When \code{convert_ipa_slash=TRUE} by default, it converts the
forward slash \code{"/"} to the value of argument \code{ipa_slash_sep}.}

\item{ipa_slash_sep}{\code{character} string used as a delimited when
\code{convert_ipa_slash=TRUE}, used to replace genes that contain
forward slash \code{"/"} to use another character.}

\item{revert_ipa_xref}{\code{logical} indicating whether to revert the
IPA gene symbols reported, which requires that the IPA data
contains a section \code{"Analysis Ready Molecules"}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of \code{data.frame} objects, where each \code{data.frame}
contains enrichment data for one of the Ingenuity IPA
enrichment tests.
}
\description{
Import Ingenuity Pathway Analysis 'IPA' results, by default
reverting IPA symbols to input values
}
\details{
This function parses Ingenuity Pathway Analysis ('IPA')
enrichment data into a \code{list} of \code{data.frame}
objects for downstream analysis.
Each \code{data.frame} represents the results of one Ingenuity IPA test,
however not all sections contain gene set enrichment results.
\subsection{Batch processing}{

When importing multiple files, argument \code{ipaFile} can be a vector
of '.xlsx' or '.txt' files. This workflow also calls
\code{IPAlist_to_hits()} to generate a gene hit matrix, stored
as \code{attr(ipalist, "geneHitIM")} to use in \code{multiEnrichMap()}.
}

\subsection{IPA Gene Xref Data}{

By default, the argument \code{revert_ipa_xref=TRUE} will convert the
IPA gene symbol values back to the original identifier.
In most cases* this behavior is desirable, with caveats:
\itemize{
\item When using platform data which
use a non-gene identifier, including microarray probesets, or
RefSeq transcript ID, or protein "UniProt" accession numbers,
it is recommended to use \code{revert_ipa_xref=FALSE}.
In these cases, the IPA gene symbol is expected to be
more user-friendly, and therefore more useful.
\itemize{
\item This option would be helpful to view the IPA gene symbols
as they appear in the IPA report - even if the symbols
sometimes do not match the input row identifiers.
}
\item It is helpful to use \code{revert_ipa_xref=TRUE} when the identifier
will also be used to compare to the source data, for example
if trying to make an expression heatmap of the genes involved
in enrichment results.
\itemize{
\item This option would convert IPA gene symbol back to Affymetrix
probeset ID, for example, if the probeset ID values were
used as the primary identifier for each measurement.
The probeset ID might be convenient to align with the input
data matrix.
\item The primary reason for this option is when providing
gene symbols as input to IPA, some will be renamed to IPA
preferred gene symbols, which would therefore be difficult
to match with the gene symbols provided to IPA.
}
}

In any case, the output \code{list} should contain an entry
"Analysis Ready Molecules" with the full IPA data table used
for the analysis. This \code{data.frame} will also contain any
statistical columns, if provided to IPA upfront.

See \code{IPAlist_to_hits()} for an automated way to create a gene hit
matrix from the 'Analysis Ready Molecules' returned by IPA.
}

\subsection{Motivation}{
\enumerate{
\item Separate multiple IPA enrichment tables.
\item Rename colnames to be consistent, compatible with
\code{enrichDF2enrichResult()}.
\item Revert IPA gene aliases to original user input, default but optional.
\item Generate \code{geneHitIM} hit matrix when processing multiple files.
}
}

\subsection{Input format}{
\enumerate{
\item \code{ipaFile} can be one or more text \code{.txt} files,
where the text file contains all IPA enrichment data in
tall format. This format is most common.
\item \code{ipaFile} can be one or more Excel \code{.xlsx} files,
which contains all IPA enrichment data in
one tall worksheet tab.
\item \code{ipaFile} can be one or more Excel \code{.xlsx} files,
where each type of IPA enrichment appears on a separate
Excel worksheet tab.
\item \code{ipaFile} can be a list of \code{data.frame} objects.
This option is intended when the IPA data has already
been imported into R as separate \code{data.frame} objects.
}
}

\subsection{Notes}{

When using \code{"Export All"} from 'IPA', the default text format
includes multiple enrichment tables concatenated together in one
file. Each enrichment table contains its own unique column
headers, with descriptive text in the line preceding the
column headers. This function is intended to separate the
enrichment tables into a list of \code{data.frame} objects, and
retain the descriptive text as names of the list.
}

\subsection{Troubleshooting}{
\itemize{
\item A common error occurs when reverting IPA gene symbols
to the original user-supplied identifier, by default
\code{revert_ipa_xref=TRUE}.
For errors during this step, consider \code{revert_ipa_xref=FALSE}
which will retain the gene symbol as recognized by IPA.
The downside of this approach is that it may be more difficult
to equate to the input identifier.
In that case look at the "Analysis Ready Molecules" \code{data.frame}
which should contain the
user-provided values as "ID"; the IPA recognized symbol as "Name",
and optionally a column "Symbol" which is edited by multienrichjam.
}
}
}
\examples{
ipaFile <- system.file(package="multienrichjam", "extdata",
   c("Newborns-IPA.txt", "OlderChildren-IPA.txt"));
ipalist <- importIPAenrichment(ipaFile)

}
\seealso{
Other jam import functions:
\code{\link[=IPAlist_to_hits]{IPAlist_to_hits()}},
\code{\link[=enrichDF2enrichResult]{enrichDF2enrichResult()}}
}
\concept{jam import functions}
