% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{imSigned2list}
\alias{imSigned2list}
\title{convert signed incidence matrix to list}
\usage{
imSigned2list(x, empty = c(NA, "", 0, FALSE), ...)
}
\arguments{
\item{x}{\code{matrix} or equivalent object with \code{colnames(x)} indicating
list set names, and \code{rownames(x)} indicating list contents.}

\item{empty}{\code{character} vector of incidence matrix values that
should be considered "empty" and therefore do not indicate
the row in \code{x} is present for the given column in \code{x}.
All other items are considered to be present, and are assigned
direction based upon the value in that cell of \code{x}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of named numeric vectors, where list names
are defined by \code{colnames(x)}, and vector names are derived
from \code{rownames(x)}. Values in each vector indicate the
signed direction, \code{c(-1,1)}.
}
\description{
convert signed incidence matrix to list
}
\details{
This function converts an signed incidence \code{matrix}
that contains positive and negative values, or equivalent
\code{data.frame}, to a list of named vectors containing values
\code{c(-1, 1)} to indicate signed direction.
The input \code{matrix} should contain numeric values where
positive and negative values indicate directionality.
When the input contains only logical values \code{c(TRUE,FALSE)}
the direction is assumed to be \code{+1} positive.

Values of \code{NA} are converted to zero \code{0} and therefore ignored.

Values that are \code{logical} with \code{TRUE} and \code{FALSE} are converted
to \code{numeric} before output.
}
\examples{
im <- matrix(c(0,1,-1,1,1,NA,-1,0,1),
   ncol=3,
   nrow=3,
   dimnames=list(letters[1:3], LETTERS[1:3]))
print(im);
# matrix input
im2list(im);
imSigned2list(im);
imSigned2list(im != 0);

}
\seealso{
Other jam list functions:
\code{\link[=colors_from_list]{colors_from_list()}},
\code{\link[=im2list]{im2list()}},
\code{\link[=list2concordance]{list2concordance()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=list2imSigned]{list2imSigned()}}
}
\concept{jam list functions}
