% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{imSigned2list_dep}
\alias{imSigned2list_dep}
\title{convert signed incidence matrix to list}
\usage{
imSigned2list_dep(x, verbose = FALSE, ...)
}
\value{
\code{list} of named numeric vectors, where list names
are defined by \code{colnames(x)}, and vector names are derived
from \code{rownames(x)}. Values in each vector indicate the
signed direction, \code{c(-1,1)}.
}
\description{
convert signed incidence matrix to list
}
\details{
This function converts an signed incidence \code{matrix}
that contains positive and negative values, or equivalent
\code{data.frame}, to a list of named vectors containing values
\code{c(-1, 1)} to indicate signed direction.
The input \code{matrix} should contain numeric values where
positive and negative values indicate directionality.
When the input contains only logical values \code{c(TRUE,FALSE)}
the direction is assumed to be \code{+1} positive.

Values of \code{NA} are converted to zero \code{0} and therefore ignored.

This function uses the \code{transactions} class from the \code{arules}
R package, which in our testing is substantially faster than
similar techniques from a variety of other R packages.
}
\seealso{
Other jam list functions: 
\code{\link{colors_from_list}()},
\code{\link{im2list}()},
\code{\link{im2list_dep}()},
\code{\link{imSigned2list}()},
\code{\link{list2concordance}()},
\code{\link{list2im}()},
\code{\link{list2imSigned}()}
}
\concept{jam list functions}
