% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Mem.R
\name{internal_EnrichmentMap}
\alias{internal_EnrichmentMap}
\title{EnrichmentMap internal function}
\usage{
internal_EnrichmentMap(
  x,
  do_plot = TRUE,
  legend_x = "bottomleft",
  legend_y = NULL,
  params = list(repulse = 3.5, width = 30, group = "default", mark.expand = 4,
    do_fixSetLabels = FALSE, do_legend = TRUE),
  ...
)
}
\arguments{
\item{x}{\code{Mem} or \code{MemPlotFolio} object.}

\item{do_plot}{\code{logical} whether to render the \code{igraph}, which
is done using \code{jam_igraph(Emap, ...)}.}

\item{legend_x, legend_y}{passed to \code{mem_legend()} to render
a color legend for the plot.}

\item{params}{\code{list} of parameters specific to EnrichmentMap:
\itemize{
\item \code{'repulse'}: \code{numeric} default 3.5 to define the initial
network layout.
\item \code{'width'}: \code{integer} character width to apply word-wrap
to node labels, and to community or nodegroup labels.
\item \code{'group'}: \code{character} string indicating the type of
node grouping to use:
\itemize{
\item \code{'default'}: will use, in order of preference:
if supplied a \code{MemPlotFolio} object, it will use
cluster labels if available, then cluster titles;
if supplied a \code{Mem} object it will use \code{igraph}
community detection and the corresponding top keywords.
\item \code{'clusters'}: requires \code{MemPlotFolio} input and uses
cluster titles.
\item \code{'cluster_labels'}: requires \code{MemPlotFolio} input and
cluster labels if available, otherwise cluster titles.
\item \code{'community'}: will use \code{igraph} community detection.
\item \code{'none'}: will not use node grouping.
}
\item \code{'mark.expand'}: \code{numeric} passed to \code{jam_igraph()} to define
node group expansion for the shaded hull around node groups,
default 4. Units are percentage of plot dimensions.
\item \code{'do_legend'}: \code{logical} default TRUE, whether to render
the color legend.
\item \code{'do_fixSetLabels'}: \code{logical} whether to apply \code{fixSetLabels()}
to adjust pathway/set nodes to be more user-friendly.
Default is FALSE.
When TRUE it calls \code{fixSetLabels()} with \code{width=params$width},
and when FALSE it calls \code{fixSetLabels()} with
\code{width=params$width} however it does no other adjustments except
to apply word-wrap. For pathways with underscores, it will
therefore not word-wrap at all, which may not be ideal.
\itemize{
\item Other parameters will be added to the function arguments
with default values when implemented.
}
}}

\item{...}{additional arguments are passed through to
\code{mem2emap()}, \code{jam_igraph()}, and \code{mem_legend()}.}
}
\value{
\code{igraph} Enrichment Map network is returned.
When \code{do_plot=TRUE} the network is drawn using
\code{jam_igraph(Emap, ...)}.
}
\description{
EnrichmentMap internal function
}
\keyword{internal}
