% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraphshapes.R
\name{jam_mypie}
\alias{jam_mypie}
\title{Vectorized mypie() function for igraph vertex pie polygons}
\usage{
jam_mypie(
  x,
  y,
  values,
  radius,
  edges = 200,
  col = NULL,
  angle = 45,
  density = NULL,
  border = NULL,
  frame.color = NULL,
  frame.lwd = NULL,
  lty = NULL,
  lwd = NULL,
  init.angle = 90,
  inner_pie_border = getOption("inner_pie_border", TRUE),
  inner_cex = 1.01,
  xy_max = NULL,
  overclose_polygons = FALSE,
  ...
)
}
\arguments{
\item{x, y}{\code{numeric} coordinate for the center of each \code{igraph} node.}

\item{values}{\code{numeric} vector of relative pie wedge sizes.}

\item{radius}{\code{numeric} radius of pie}

\item{edges}{\code{integer} number of edges to make a circle, default 200.}

\item{col}{\code{character} vector of R colors to fill each pie wedge,
in order, default NULL.}

\item{angle, density}{used to draw lines to fill each pie node, passed
to \code{polygon()}. Default angle=45, and density=NULL so it draws none.}

\item{border}{\code{character} vector of R colors for each pie wedge border,
default is NULL.}

\item{frame.color}{\code{character} R color used around the entire pie circle,
default is NULL.}

\item{frame.lwd}{\code{numeric} line width, default NULL uses \code{par('lwd')}
if a graphics device is open, otherwise uses 1.}

\item{lty}{\code{numeric} or \code{character} line type, default NULL.}

\item{lwd}{\code{numeric} line width, default NULL will use \code{par('lwd')}
if a graphics device is open, otherwise uses 1.}

\item{init.angle}{\code{numeric} angle in degrees (0 to 360) where \code{0} is the
top of the circle, proceeding clockwide.}

\item{inner_pie_border}{\code{logical} whether to apply \code{pie.border} colors
only along the inside of each pie wedge polygon, so that adjacent
colors will be seen beside each other without overlapping adjacent
borders. This method is currently in development.}

\item{overclose_polygons}{\code{logical} indicating whether to close polygon
coordinates with an extra couple points, to ensure the line join
\code{"ljoin"} is properly called when drawing each polygon.}

\item{...}{additional arguments are passed to \code{polygon()}}
}
\value{
\code{data.frame} with columns suitable for use in \code{polygon()}
after each column is expanded to vector form. Columns \code{x} and \code{y}
are stored \code{AsIs()} in \code{list} format, and should be combined
with one \code{NA} value between each \code{numeric} vector. The \code{NA} values
cause \code{polygon()} to draw a series of separate polygons in
one vectorized step, much faster than drawing a series of
polygons in an iterative loop.
}
\description{
Vectorized mypie() function for igraph vertex pie polygons
}
\details{
This is a function called internally by \code{shape.jampie.plot()},
not intended for direct use.

See \code{reorder_igraph_nodes()} for visual examples.

This function is a light rewrite of \code{igraph:::mypie()}, except
that this function determines polygon coordinates without
drawing them, instead returns the polygon coordinates to
the calling function \code{shape.jampie.plot()} which in turn
draws all polygons once using the vectorized approach
described for \code{graphics::polygon()}.

See \code{shape.jampie.plot()} for more detail.

To disable \code{pie.border} set to \code{NA} with \code{vertex.pie.border=NA}
or \code{V(g)[[2]]$pie.border <- NA}.

To disable \code{frame.color} set to \code{NA} with \code{vertex.frame.color=NA}
or \code{V(g)[2]$frame.color <- NA}.
}
\examples{
# See reorder_igraph_nodes() for visual examples.

}
\seealso{
Other jam igraph shapes:
\code{\link[=shape.coloredrectangle.plot]{shape.coloredrectangle.plot()}},
\code{\link[=shape.ellipse.clip]{shape.ellipse.clip()}},
\code{\link[=shape.ellipse.plot]{shape.ellipse.plot()}},
\code{\link[=shape.jampie.plot]{shape.jampie.plot()}}
}
\concept{jam igraph shapes}
