% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycat-ui-launch.R
\name{launch_shinycat}
\alias{launch_shinycat}
\title{Launch ShinyCAT: Cnet Adjustment Tool}
\usage{
launch_shinycat(g = NULL, envir = new.env(), ..., options = list(width = 1200))
}
\arguments{
\item{g}{\code{igraph} object, or can be NULL if the variable 'g' is
defined in 'envir'.}

\item{envir}{\code{environment} assigned to the R-shiny function space,
which may be useful when capturing the \code{igraph} object after
manipulation via the R-shiny app.}

\item{...}{additional arguments are ignored.}

\item{options}{\code{list} with additional settings, for example:
\itemize{
\item host: \code{character} string with hostname or IP address
to permit incoming connections.
Use '0.0.0.0' to accept all incoming host names.
\item port: \code{numerical} port to listen for incoming connections.
\item quiet: \code{logical} whether to
\item launch.browser: \code{logical} whether to launch a web browser, default
TRUE for interactive sessions.
\item display.mode: \code{character} string with display mode: 'normal' is
recommended; 'showcase' displays the code beside the app.
}}
}
\value{
\code{environment} invisibly, containing
\itemize{
\item 'g' the input \code{igraph}
\item 'adj_cnet' the output \code{igraph} after manipulation in the R-shiny app.
It also contains two attributes for reference:
\enumerate{
\item 'nodeset_adj': \code{data.frame} used to adjust nodesets.
\item 'node_adj': \code{data.frame} used to adjust nodes.
}
}
}
\description{
Launch ShinyCAT: Cnet Adjustment Tool
}
\details{
This function launches the R-shiny app 'ShinyCat' to manipulate
a Cnet \code{igraph} object, which is expected to contain node (vertex)
attribute 'nodeType' with values 'Gene' and 'Set'.

There are two ways to retain the resulting igraph object:
\enumerate{
\item Capture the output of this function, an \code{environment} described below.
\item Click "Adjustments" then "Save RData", which stores the 'igraph`
object as 'adj_cnet'.
}

The function returns an \code{environment} inside which are two objects:
\itemize{
\item 'g': the \code{igraph} data input
\item 'adj_cnet': the adjusted Cnet \code{igraph} output
}

Additionally, the 'adj_cnet' object contains two attributes:
\itemize{
\item 'nodeset_adj', 'node_adj': \code{data.frame} objects used by
\code{bulk_cnet_adjustments()} to convert 'g' into 'adj_cnet'.
}
}
\examples{
# create Cnet test data
g <- make_cnet_test();
cnetenv <- new.env();

# you must catch the output to use the resulting igraph object
# output_envir <- launch_shinycat();

}
\seealso{
Other jam igraph utilities:
\code{\link[=color_edges_by_nodegroups]{color_edges_by_nodegroups()}},
\code{\link[=color_edges_by_nodes]{color_edges_by_nodes()}},
\code{\link[=color_nodes_by_nodegroups]{color_nodes_by_nodegroups()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=jam_igraph]{jam_igraph()}},
\code{\link[=subgraph_jam]{subgraph_jam()}}

Other jam cnet utilities:
\code{\link[=adjust_cnet_nodeset]{adjust_cnet_nodeset()}},
\code{\link[=adjust_cnet_set_relayout_gene]{adjust_cnet_set_relayout_gene()}},
\code{\link[=apply_nodeset_spacing]{apply_nodeset_spacing()}},
\code{\link[=bulk_cnet_adjustments]{bulk_cnet_adjustments()}},
\code{\link[=get_cnet_nodeset]{get_cnet_nodeset()}},
\code{\link[=get_cnet_nodeset_vector]{get_cnet_nodeset_vector()}},
\code{\link[=make_cnet_test]{make_cnet_test()}},
\code{\link[=relayout_nodegroups]{relayout_nodegroups()}}

Other jam shiny functions:
\code{\link[=shinycat_server]{shinycat_server()}},
\code{\link[=shinycat_ui]{shinycat_ui()}}
}
\concept{jam cnet utilities}
\concept{jam igraph utilities}
\concept{jam shiny functions}
