% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{list2concordance}
\alias{list2concordance}
\title{Convert list to concordance matrix}
\usage{
list2concordance(x, naValue = NA, makeSigned = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{list} of named numerical vectors, where the sign (positive
or negative sign) indicates directionality, and is used to calculate
concordance, which is a measure of the agreement of the overall set
of directions shared between each pair of vectors.}

\item{naValue}{value passed to \code{jamba::rmNA()} used to replace any
\code{NaN} values in the output matrix. The \code{NaN} values result when a
pair of vectors has no shared non-zero named entry.}

\item{makeSigned}{logical indicating whether to force the vectors in
the input list \code{x} to contain only values \code{c(-1,0,1)}, by calling
\code{base::sign()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{list2imSigned()}.}
}
\description{
Convert list to concordance matrix
}
\details{
This function calculates pairwise concordance using
Kruskal concordance coefficient (ref) using the following equation:
\itemize{
\item (number_agree - number_disagree) / (total_shared)
}

The equation is applied to each pair of named vectors in the input
list \code{x}, and reflects the degree of agreement in direction (+ or -)
between shared named elements, with +1 being perfect concordance
(agreement), and -1 being perfect discordance (disagreement.) Values
of zero indicate equal agreement and disagreement, and therefore
reflect no concordance nor discordance. Values
of \code{NA} occur when no named entries are shared.

This function calls \code{list2imSigned()} to produce a signed
incidence matrix, which is then used with \code{base::crossprod()}
to calculate the full matrix of values.
}
\examples{
set.seed(123);
l123 <- lapply(jamba::nameVector(1:3), function(i){
   jamba::nameVector(
      sample(c(-1,-1,0,1), replace=TRUE, size=15),
      letters[1:15]
   )
});
list2concordance(l123);

# observe the signed incidence matrix
list2imSigned(l123);

}
\seealso{
Other jam list functions:
\code{\link[=colors_from_list]{colors_from_list()}},
\code{\link[=im2list]{im2list()}},
\code{\link[=imSigned2list]{imSigned2list()}},
\code{\link[=list2im]{list2im()}},
\code{\link[=list2imSigned]{list2imSigned()}}
}
\concept{jam list functions}
