% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{list2im}
\alias{list2im}
\title{convert list to incidence matrix}
\usage{
list2im(
  x,
  empty = 0,
  do_sparse = FALSE,
  keepCounts = FALSE,
  sort_rows = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{list} of item vectors, where vector values are item names}

\item{empty}{\code{numeric} used for missing or empty values, default 0.}

\item{do_sparse}{\code{logical} default FALSE, whether to return sparse Matrix.
When \code{do_sparse=TRUE}, it sets \code{keepCounts=FALSE} since a sparse Matrix
does not represent counts.}

\item{keepCounts}{\code{logical} default FALSE, whether to include the count
of any duplicated items.}

\item{sort_rows}{\code{logical} default FALSE, whether to sort rows using
\code{jamba::mixedOrder()}, or \code{function} can be supplied to use
another sort algorithm.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix with values 0 or 1 indicating absense or
presence of each item (row) in each set (column).
The sets are derived from \code{names(x)}.
}
\description{
convert list to incidence matrix
}
\details{
This function converts a list of vectors into an incidence matrix, where
the rows are the vector items and the columns are the list names.
}
\examples{
L1 <- list(A=c("C","A","B","A"),
   D=c("D","E","F","D"),
   A123=c(1:8,3,5),
   B=c("A1", "A2", "A10"),
   T=LETTERS[7:9]);
# Default behavior is to make items unique
list2im(L1);

# Option to report the counts
list2im(L1, keepCounts=TRUE);

# Option to sort item rows with jamba::mixedOrder()
list2im(L1, sort_rows=TRUE);

# Option to sort item rows with vanilla sort()
list2im(L1, sort_rows=sort);

}
\seealso{
Other jam list functions:
\code{\link[=colors_from_list]{colors_from_list()}},
\code{\link[=im2list]{im2list()}},
\code{\link[=imSigned2list]{imSigned2list()}},
\code{\link[=list2concordance]{list2concordance()}},
\code{\link[=list2imSigned]{list2imSigned()}}
}
\concept{jam list functions}
