% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{list2imSigned}
\alias{list2imSigned}
\title{convert list to directional incidence matrix}
\usage{
list2imSigned(x, emptyValue = NA, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{list} of named vectors, where the names are used to
identify each element, and become rownames in the output
incidence matrix. The vector values become values in the
incidence matrix.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{list2im()}.}
}
\value{
\code{numeric} matrix with rownames defined by vector names from
each vector in the input list. The colnames are defined by
names of the input list if they exist.
list. Values in the matrix are values from each vector.
}
\description{
convert list to directional incidence matrix
}
\details{
This function extends \code{list2im()} in that it stores the
value associated with each element in the list. As such, the input
format is a named vector, where the names of the vector are the items,
and the numeric values are the values to be stored in the
incidence matrix.

A common scenario is to generate a vector of genes, with values
\code{c(-1, 0, 1)} indicating the direction of gene expression changes,
named by the gene symbol. Each vector in the list represents one
statistical test. Here, \code{list2imSigned()} will convert this list
into a directional matrix representing the gene changes across the
comparisons.

Note that this function currently does not combine multiple values,
instead only the last occurring value is stored in the resulting
matrix. This decision is partly due to efficiency, and partly because
there are multiple possible methods to combine multiple values.
For example, taking the \code{mean(x)} for a given gene, which has
a value \code{1} and \code{-1} would result in \code{0} and might suggest the
gene is not a statistical hit. Instead, when multiple values
are anticipated per named vector entry, use functions in a
package like \code{data.table} or \code{dplyr} to apply a function to
combine values.
}
\examples{
L1 <- list(A=c("C","A","B","A"),
   D=c("D","E","F","D"),
   A123=c(1:8,3,5),
   T=LETTERS[7:9]);
L1;

# Convert each vector to a signed vector
set.seed(123);
L2 <- lapply(L1, function(i){
   i <- unique(i);
   jamba::nameVector(sample(c(-1,1), size=length(i), replace=TRUE), i);
});
L2;

# Convert to signed incidence matrix
list2imSigned(L2);

}
\seealso{
Other jam list functions:
\code{\link[=colors_from_list]{colors_from_list()}},
\code{\link[=im2list]{im2list()}},
\code{\link[=imSigned2list]{imSigned2list()}},
\code{\link[=list2concordance]{list2concordance()}},
\code{\link[=list2im]{list2im()}}
}
\concept{jam list functions}
