% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\name{list_to_Mem}
\alias{list_to_Mem}
\title{Coerce legacy multiEnrichMap list format to Mem object}
\usage{
list_to_Mem(x, ...)
}
\arguments{
\item{x}{\code{list} output from \code{multiEnrichMap()}}

\item{...}{additional arguments are ignored}
}
\description{
Coerce legacy multiEnrichMap list format to \code{Mem} S4 object
}
\examples{
jamsession::load_object("test_mem")
test_mem$thresholds <- list(p_cutoff=0.05, min_count=3)
Mem <- list_to_Mem(test_mem)

}
\seealso{
Other Mem: 
\code{\link{Mem-class}},
\code{\link{check_Mem}()}
}
\concept{Mem}
