% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-Mem.R
\docType{class}
\name{Mem-class}
\alias{Mem-class}
\alias{Mem}
\alias{[,Mem,ANY,ANY,ANY-method}
\alias{show,Mem-method}
\alias{Mem_to_list}
\alias{as.list,Mem-method}
\alias{names,Mem-method}
\alias{enrichments,Mem-method}
\alias{enrichments}
\alias{enrichments<-,Mem-method}
\alias{sets,Mem-method}
\alias{sets}
\alias{sets<-,Mem-method}
\alias{genes,Mem-method}
\alias{genes}
\alias{genes<-,Mem-method}
\alias{geneIM,Mem-method}
\alias{geneIM}
\alias{geneIMdirection,Mem-method}
\alias{geneIMdirection}
\alias{geneIMdirection<-,Mem-method}
\alias{geneIMcolors,Mem-method}
\alias{geneIMcolors}
\alias{geneIMcolors<-,Mem-method}
\alias{enrichList,Mem-method}
\alias{enrichList}
\alias{enrichIM,Mem-method}
\alias{enrichIM}
\alias{enrichIMcolors,Mem-method}
\alias{enrichIMcolors}
\alias{enrichIMcolors<-,Mem-method}
\alias{enrichIMdirection,Mem-method}
\alias{enrichIMdirection}
\alias{enrichIMdirection<-,Mem-method}
\alias{enrichIMgeneCount,Mem-method}
\alias{enrichIMgeneCount}
\alias{memIM,Mem-method}
\alias{memIM}
\alias{geneHitIM,Mem-method}
\alias{geneHitIM}
\alias{geneHitIM<-,Mem-method}
\alias{geneHitList,Mem-method}
\alias{geneHitList}
\alias{geneHitList<-,Mem-method}
\alias{headers,Mem-method}
\alias{colorV,Mem-method}
\alias{colorV}
\alias{colorV<-,Mem-method}
\alias{thresholds,Mem-method}
\alias{thresholds}
\alias{thresholds<-,Mem-method}
\alias{dim,Mem-method}
\alias{dimnames,Mem-method}
\alias{as}
\alias{coerce}
\alias{list_to_Mem}
\alias{updateObject,Mem-method}
\alias{geneInCategory,Mem-method}
\alias{geneInCategory}
\alias{setsByGene,Mem-method}
\alias{setsByGene}
\alias{EnrichmentMap,Mem-method}
\alias{EnrichmentMap}
\title{Mem S4 class, accessors, getters, and setters}
\usage{
\S4method{[}{Mem,ANY,ANY,ANY}(x, i, j, k, drop = FALSE)

\S4method{show}{Mem}(object)

Mem_to_list(x, ...)

\S4method{as.list}{Mem}(x)

\S4method{names}{Mem}(x)

\S4method{enrichments}{Mem}(x)

\S4method{enrichments}{Mem}(x) <- value

\S4method{sets}{Mem}(x, ...)

\S4method{sets}{Mem}(x) <- value

\S4method{genes}{Mem}(x, ...)

\S4method{genes}{Mem}(x) <- value

\S4method{geneIM}{Mem}(x, ...)

\S4method{geneIMdirection}{Mem}(x, ...)

\S4method{geneIMdirection}{Mem}(x) <- value

\S4method{geneIMcolors}{Mem}(x, ...)

\S4method{geneIMcolors}{Mem}(x) <- value

\S4method{enrichList}{Mem}(x, ...)

\S4method{enrichIM}{Mem}(x, ...)

\S4method{enrichIMcolors}{Mem}(x, ...)

\S4method{enrichIMcolors}{Mem}(x) <- value

\S4method{enrichIMdirection}{Mem}(x, ...)

\S4method{enrichIMdirection}{Mem}(x) <- value

\S4method{enrichIMgeneCount}{Mem}(x, ...)

\S4method{memIM}{Mem}(x, ...)

\S4method{geneHitIM}{Mem}(x, ...)

\S4method{geneHitIM}{Mem}(x) <- value

\S4method{geneHitList}{Mem}(x, ...)

\S4method{geneHitList}{Mem}(x) <- value

\S4method{headers}{Mem}(x)

\S4method{colorV}{Mem}(x, ...)

\S4method{colorV}{Mem}(x) <- value

\S4method{thresholds}{Mem}(x, ...)

\S4method{thresholds}{Mem}(x) <- value

\S4method{dim}{Mem}(x)

\S4method{dimnames}{Mem}(x)

list_to_Mem(mem)

\S4method{updateObject}{Mem}(object, ..., verbose = FALSE)

\S4method{geneInCategory}{Mem}(x)

\S4method{setsByGene}{Mem}(x)

\S4method{EnrichmentMap}{Mem}(
  x,
  do_plot,
  params = list(repulse = 3.5, width = 30, group = "default", mark.expand = 4, do_legend
    = TRUE),
  legend_x = "bottomleft",
  legend_y = NULL,
  ...
)
}
\arguments{
\item{x}{\code{Mem} object}

\item{i}{\code{character} or \code{integer} with gene names}

\item{j}{\code{character} or \code{integer} with pathway names}

\item{drop}{\code{logical} always set to FALSE for Mem objects.}

\item{...}{additional arguments are ignored}

\item{value}{\code{character} vector of names to assign to pathway gene sets}

\item{mem}{\code{list} output from \code{multiEnrichMap()}}
}
\value{
\code{Mem_to_list()} returns a \code{list} suitable for other mem functions.

\code{as.list(Mem)} returns a \code{list} suitable for other mem functions.

\code{names()} returns the \code{character} vector of enrichment names

\code{enrichments()} returns the \code{character} vector of enrichment names

\code{sets()} returns the \code{character} vector of pathway gene sets

\code{list_to_Mem()} returns a \code{Mem} S4 object, from 'list' or
'Mem' input.

\code{updateObject()} returns a \code{Mem} S4 object with the current
class version, and current valid class definition. It should be
able to update any serialized 'Mem' object from previous
class versions.

\code{geneInCategory()} returns a \code{list} named by pathway, containing
\code{character} vectors with genes in each pathway.
For the reciprocal, see \code{setsByGene()}.

\code{setsByGene()} returns a \code{list} named by gene,
containing \code{character} vectors with associated gene sets.
For the reciprocal, see \code{geneInCategory()}.
}
\description{
Mem class containing results from \code{multiEnrichMap()},
current class version "1.0.0".
}
\section{Functions}{
\itemize{
\item \code{x[i}: Subset a Mem object, \code{Mem[i, j, k]},
where \code{i} are genes, \code{j} are pathways, and \code{k} are enrichments.

\item \code{show(Mem)}: Show summary of a Mem object, dimensions defined by
genes, sets, enrichments, showing up to 5 entries of each.
Also includes a list of analysis parameters: topN, and thresholds.

\item \code{Mem_to_list()}: Coerce S4 \code{Mem} to legacy \code{list} mem format

\item \code{as.list(Mem)}: Coerce S4 \code{Mem} to legacy \code{list} mem format

\item \code{names(Mem)}: Names for the enrichment tests in a Mem object

\item \code{enrichments(Mem)}: Names for the enrichment tests in a Mem object

\item \code{enrichments(Mem) <- value}: Assign names for the enrichment tests in a Mem object

\item \code{sets(Mem)}: List pathway gene set names in a Mem object

\item \code{sets(Mem) <- value}: Assign pathway gene set names to a Mem object

\item \code{genes(Mem)}: List genes represented

\item \code{genes(Mem) <- value}: Assign gene names in a Mem object

\item \code{geneIM(Mem)}: The gene-enrichment matrix of genes represented in
enrichment tests.

\item \code{geneIMdirection(Mem)}: The matrix of genes tested versus enrichment tests, with directionality.

\item \code{geneIMdirection(Mem) <- value}: Assign directional matrix to the gene-enrichment matrix

\item \code{geneIMcolors(Mem)}: The matrix of colors indicating genes in each enrichment test

\item \code{geneIMcolors(Mem) <- value}: Assign a color matrix to the gene-enrichment matrix

\item \code{enrichList(Mem)}: The list of enrichResult data in an Mem object

\item \code{enrichIM(Mem)}: The pathway/P-value matrix

\item \code{enrichIMcolors(Mem)}: The pathway-enrichment directional color matrix

\item \code{enrichIMcolors(Mem) <- value}: Assign the pathway-enrichment directional color matrix

\item \code{enrichIMdirection(Mem)}: The pathway-enrichment directional score matrix

\item \code{enrichIMdirection(Mem) <- value}: Assign the pathway-enrichment directional score matrix

\item \code{enrichIMgeneCount(Mem)}: Pathway-enrichment gene count matrix, genes involved
in enrichment of each pathway, for each enrichment test.

\item \code{memIM(Mem)}: The gene-pathway incidence matrix

\item \code{geneHitIM(Mem)}: The matrix of genes tested for enrichment, including
genes not associated with enrichment results.

\item \code{geneHitIM(Mem) <- value}: Assign the matrix of genes tested in a Mem object

\item \code{geneHitList(Mem)}: The list of genes tested for enrichment, including
genes not associated with enrichment results.

\item \code{geneHitList(Mem) <- value}: Assign the list of genes tested for enrichment,
including genes not associated with enrichment results.

\item \code{headers(Mem)}: Enrichment data column headers associated to
enrichResult data in a Mem object

\item \code{colorV(Mem)}: List colors assigned to  represented

\item \code{colorV(Mem) <- value}: Assign colors to enrichments in a Mem object. If
supplied 'value' matches the number of 'enrichments(x)', colors are
assigned in the order supplied.
Otherwise, 'names(value)' must be present in 'enrichments(x)',
and values are assigned in matching order.

\item \code{thresholds(Mem)}: List thresholds defined

\item \code{thresholds(Mem) <- value}: thresholds defined in a Mem object.
Note that thresholds do not trigger any other updates to the
Mem object.

\item \code{dim(Mem)}: dimensions in order of genes, sets, and enrichments

\item \code{dimnames(Mem)}: dimension names for each genes, sets, and enrichments

\item \code{list_to_Mem()}: Convert legacy \code{list} mem format to S4 \code{Mem}

\item \code{updateObject(Mem)}: Confirm Mem object meets current class definition.

\item \code{geneInCategory(Mem)}: Genes in each category (pathway, gene set)

\item \code{setsByGene(Mem)}: Pathway gene sets associated with each gene

\item \code{EnrichmentMap(Mem)}: EnrichmentMap \code{igraph} network to connect
pathway gene sets based upon Jaccard overlap between each
pathway pair.
Note '...' arguments are passed to \code{jam_igraph()} and
\code{mem_legend()} when \code{do_plot=TRUE}.
Argument \code{'params'} is a \code{list} with additional arguments:
\itemize{
\item 'repulse': \code{numeric} repulse passed to \code{layout_with_qfr()}.
\item 'width': \code{integer} word-wrap width used for cluster labels.
\item 'group': \code{character} used to define the type of node group,
'clusters' uses pathway cluster title; 'cluster_labels' uses
cluster labels if defined; 'community' uses igraph community,
'none' display no node grouping.
\item 'mark.expand': passed to \code{jam_igraph()} to control the buffer
around the alpha hull shaded regions for node groups.
\item 'do_legend': \code{logical} whether to display the legend
}

}}
\examples{
# examples for Mem_to_list and as(Mem, "list")
data(Memtest)
mem1 <- Mem_to_list(Memtest)
jamba::sdim(mem1)

# list_to_Mem examples
data(Memtest)
mem <- as(Memtest, "list")
Mem <- list_to_Mem(mem)

setsByGene(Memtest)
}
\seealso{
Other Mem: 
\code{\link{Mem-data}},
\code{\link{Mem-slots}},
\code{\link{check_Mem}()}

Other jam Mem utilities: 
\code{\link{MemPlotFolio-class}},
\code{\link{fixSetLabels}()},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()}
}
\concept{Mem}
\concept{jam Mem utilities}
