% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multienrichjam-data.R
\docType{data}
\name{Mem-data}
\alias{Mem-data}
\alias{Memtest}
\alias{Reese_genes}
\title{Reese 2019 cohort multienrichment data}
\format{
\code{Memtest} is a \code{Mem} multi-enrichment object including two
enrichment tests 'Newborns' and 'OlderChildren'

\code{Reese_genes} is a \code{list} with names 'Newborns' and 'OlderChildren',
each with a \code{character} vector of gene symbols, named by ENTREZID,
representing genes associated with each cohort.
For example, the \code{names(Reese_genes$Newborns)} may be useful for
pathway enrichment methods that expect Entrez gene ID as input.
}
\source{
\href{https://doi.org/10.1016/j.jaci.2018.11.043}{Reese et al 2019}
doi:10.1016/j.jaci.2018.11.043,
\strong{Epigenome-wide meta-analysis of DNA methylation and childhood asthma}
}
\usage{
Memtest

Reese_genes
}
\description{
Data used for multi-enrichment analysis as described in
\href{https://doi.org/10.1016/j.jaci.2018.11.043}{Reese et al 2019}
doi:10.1016/j.jaci.2018.11.043,
\strong{Epigenome-wide meta-analysis of DNA methylation and childhood asthma}.
It is provided for convenient testing of multienrichjam.
}
\details{
Memtest is a \code{Mem} object with enrichment results
for 'Newborns' and 'OlderChildren'.

A minor change was made to published data:
\itemize{
\item Gene symbols were updated to current Entrez gene symbols
as of November 2025. This process changed 'INADL' to 'PATJ'
in the enrichment results.
\item Two additional genes were updated in 'geneHitIM' for genes
not present in the IPA pathway results.
\item The \code{Reese_genes} data were also updated for consistency,
and to ensure that the \code{names(Reese_genes)} using Entrez ID
would be consistent with the gene symbols shown.
Finally, the genes in \code{Reese_genes} should be usable for
other pathway enrichment tests.
}
}
\examples{
# Mem data for Reese 2019 cohorts
data(Memtest)
# enrichment names
names(Memtest)
# genes represented
genes(Memtest)
# gene sets, pathways represented
sets(Memtest)

# Reese 2019 genes by cohort
data(Reese_genes)
# Number of genes per cohort
lengths(Reese_genes)

# first 6 gene symbols for Newborns
head(Reese_genes$Newborns)

# first 6 Entrez gene IDs for OlderChildren
head(names(Reese_genes$OlderChildren))

}
\seealso{
Other Mem:
\code{\link{Mem-class}},
\code{\link{Mem-slots}},
\code{\link[=check_Mem]{check_Mem()}}
}
\concept{Mem}
\keyword{datasets}
