% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\docType{methods}
\name{[,Mem,ANY,ANY,ANY-method}
\alias{[,Mem,ANY,ANY,ANY-method}
\alias{some_vector}
\title{Subset Mem object}
\usage{
\S4method{[}{Mem,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

some_vector(
  x,
  maxToShow = 5,
  ellipsis = "...",
  ellipsisPos = c("middle", "start", "end"),
  quote = FALSE,
  sep = ", ",
  ...
)
}
\description{
Extract some elements of potentially long vector for display
}
