% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-mem2emap.R
\name{mem2emap}
\alias{mem2emap}
\title{Convert multiEnrichMap mem output to EnrichmentMap emap}
\usage{
mem2emap(
  mem,
  overlap = 0.2,
  p_cutoff = NULL,
  min_count = NULL,
  colorV = NULL,
  cluster_function = igraph::cluster_walktrap,
  cluster_list = NULL,
  num_keep_terms = 3,
  keep_terms_sep = ",\\n",
  repulse = 3.3,
  remove_singlets = FALSE,
  color_by_nodes = FALSE,
  size_by_genes = TRUE,
  median_size = 5,
  max_size = 25,
  apply_edge_width = TRUE,
  max_edge_width = 20,
  min_edge_width = 0.5,
  apply_edge_color = apply_edge_width,
  apply_direction = TRUE,
  direction_max = 2,
  direction_floor = 0.5,
  border_lwd = 3,
  seed = 123,
  spread_labels = TRUE,
  label_min_dist = 2.5,
  y_bias = 10,
  vertex.label.font = 2,
  use_layout_components = FALSE,
  use_shadowText = TRUE,
  do_express = FALSE,
  do_plot = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{Mem} output from \code{multiEnrichMap()}}

\item{overlap}{\code{numeric}, default 0.2, value between 0 and 1 with Jaccard
overlap coefficient required between any two pathways in order to
create a network edge connecting these two pathways.
\itemize{
\item Jaccard overlap coefficient is reciprocal of Jaccard distance,
using \code{(1 - Jaccard_distance)}, where Jaccard_distance is
calculated with \code{dist(x, method='binary')} with incidence matrix.
\item \code{overlap=0.2} is default, which specifies roughly 20\% overlap
in genes shared between two pathway nodes, relative to the union
of genes in those two pathway nodes.
\item \code{overlap=NULL} is a special case which uses the average between
median and mean Jaccard overlap, excluding the diagonal.
\item Note these genes must be involved in enrichment, and
therefore does not use all possible genes annotated to a pathway.
Therefore, connections are only created with enriched genes
are shared between pathways.
}}

\item{p_cutoff}{\code{numeric} threshold used for significant enrichment
P-value. The default NULL uses the value in the \code{mem} object.
This cutoff, together with \code{min_count}, define which nodes are
included, and the color fill for those nodes.}

\item{min_count}{\code{integer} threshold for minimum genes involved in
enrichment in order for a pathway to be considered significant
during this analysis. When NULL it uses the threshold in \code{mem}.}

\item{colorV}{\code{character} vector of R colors used for each enrichment.
Default NULL uses colors defined in \code{mem}.}

\item{cluster_function}{\code{function}, default \code{igraph::cluster_walktrap},
used to cluster \code{igraph} nodes in the resulting network graph.
It is used to enhance the visual summary.
\itemize{
\item This function is not used when \code{cluster_list} is supplied.
}}

\item{cluster_list}{\code{list} default NULL, optional list of pathway
clusters, containing \code{character} vectors of \code{igraph} node/vertex
names. It will be used instead of \code{cluster_function} when supplied.}

\item{num_keep_terms}{\code{integer}, default 3, number of text terms to
keep for each pathway cluster, used together with \code{cluster_function}.
Common words are removed, remaining terms are sorted by decreasing
occurrence, then used to summarize each cluster.}

\item{keep_terms_sep}{\code{character} string, default is comma-newline,
used to separate terms, used together with \code{num_keep_terms}.}

\item{repulse}{\code{numeric} value passed to \code{layout_with_qfr()},
default 3.3.
\itemize{
\item Use repulse 'FALSE', 'NULL', or '0' will skip the layout, as
is used internally by \code{mem_find_overlap()} to iterate numerous
overlaps without spending time on layout each iteration.
}}

\item{remove_singlets}{\code{logical}, default FALSE, whether to remove
pathway singlets which have no connections to other pathways.
\itemize{
\item Using TRUE will help simplify busy figures, at the expense of
being less complete.
}}

\item{color_by_nodes}{\code{logical}, default TRUE, whether to colorize pathway
clusters using node colors in each cluster.
Note that a mix of colors often turns brown, so this feature has
unpredictable benefit.}

\item{size_by_genes}{\code{logical} default TRUE, whether to size nodes
proportional to the number of genes involved in enrichment.}

\item{median_size}{\code{numeric} used when \code{size_by_genes=TRUE} to define
the median size, default is 5.}

\item{apply_edge_width}{\code{logical} default TRUE, whether to apply edge
width based upon the edge weight, which represents the Jaccard
overlap.}

\item{max_edge_width}{\code{numeric} maximum edge width, used when
\code{apply_edge_width=TRUE}. Default 25 is intended to represent
roughly the full size of a node.}

\item{min_edge_width}{\code{numeric} minimum edge width, used after
\code{max_edge_width} has been applied. The minimum is applied as a
noise floor, so that any edge at or below the \code{min_edge_width}
is set to this width. All edges above this width are left as-is.}

\item{apply_edge_color}{\code{logical} whether to apply edge
color based upon the edge weight, which represents the Jaccard
overlap. The default matches \code{apply_edge_width} so that they are
applied together.}

\item{apply_direction}{\code{logical} default TRUE, whether to apply direction
to the node border color, when supplied. Used with \code{direction_max}
and \code{direction_floor}.}

\item{direction_max}{\code{numeric}, default 2, indicating the directional
score at which the maximum color is applied. Typically when using
some derivative of z-score to represent directionality, values at or
above 2 are effectively "maximum".}

\item{direction_floor}{\code{numeric} default 0.5, the minimum directional score
for any color to be applied, below which values are considered
"non-directional". Typically when using some form of z-score, a value
at or below 0.5 may be considered "not directional."}

\item{border_lwd}{\code{numeric} line width of node borders, used to convey
any direction associated with enrichment.}

\item{seed}{\code{numeric} passed to \code{set.seed()} via the \code{layout_with_qfr()}
layout algorithm, default 123.}

\item{spread_labels}{\code{logical} default TRUE, whether to call
\code{spread_igraph_labels()}.}

\item{label_min_dist}{\code{numeric} default 2.5 defines the label distance
used in \code{spread_igraph_labels()}.}

\item{y_bias}{\code{numeric} default 10, passed to \code{spread_igraph_labels()}.}

\item{vertex.label.font}{\code{integer} to define the font style, default 2
sets bold font; alternatively use 1 for normal font.}

\item{use_layout_components}{\code{logical} default FALSE, whether to call
\code{igraph::layout_components()} to combine multiple network components,
rather than calling \code{layout_with_qfr()} on the overall graph object.}

\item{use_shadowText}{\code{logical} default TRUE, whether to render node
labels using \code{jamba::shadowText()}. This setting is also stored in
the graph attributes and is used by \code{jam_igraph()}.}

\item{do_express}{\code{logical} default FALSE, when TRUE it skips a number
of aesthetic steps. This option is intended mainly for
\code{mem_find_overlap()} to perform more rapid iterative evaluation
of overlap thresholds.}

\item{do_plot}{\code{logical}, default FALSE, whether to render the resulting
plot using \code{jam_igraph()}.}

\item{verbose}{\code{logical} whether to print verbose output.}

\item{...}{additional arguments are passed to \code{jam_igraph()} to customize
the network plot, used when \code{do_plot=TRUE}.}
}
\value{
\code{igraph} object with an Enrichment Map, nodes are pathways,
and edges are defined by the Jaccard overlap of genes involved
in enrichment of each pair of pathways.
\itemize{
\item Edges are defined only when the Jaccard overlap is at least
\code{overlap}.
\item Node sizes are proportional to the number of genes involved
in enrichment, with median node size defined by \code{median_size}.
\item By default, edge width and color is defined by the overlap,
ranging from 0 to 1. Graph attributes store the \code{max_edge_width}
with edge width at Jaccard overlap 1; and \code{edge_colramp} as
a function to convert weight to color, and can be used to
create a color legend.
}
}
\description{
Convert multiEnrichMap mem output to EnrichmentMap emap
}
\details{
This function is currently In development.

This function is intended to convert \code{mem} output from
\code{multiEnrichMap()} into an EnrichmentMap \code{igraph} format
which represents the statistical enrichment support from each
pathway enrichment.

This function can apply P-value thresholds using the input \code{mem},
or using a custom value.

A node community detection function is applied by default
which may help visualize subgroups of nodes.
When defined, the communities are used to define nodegroups
for edge bundling.

As of version 0.0.116.900, node size also scales the border
line width for 'frame.width' and 'pie.border.lwd', so that
the line width defined by \code{border_lwd} matches \code{median_size},
and is scaled proportionally by node size.
}
\examples{
emap <- mem2emap(Memtest, min_count=2, max_edge_width=5, overlap=0.33)
jam_igraph(emap)

emap2 <- relayout_nodegroups(emap, final_repulse=4, y_bias=10, label_min_dist=2.5)
jam_igraph(emap2)

emapB <- mem2emap(fixSetLabels(Memtest), min_count=3,
   overlap=0.35, median_size=10,
   apply_edge_width=TRUE, max_edge_width=15)
jam_igraph(emapB)

}
\seealso{
Other jam Mem utilities:
\code{\link{Mem-class}},
\code{\link{MemPlotFolio-class}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=mem2cnet]{mem2cnet()}}

Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam Mem utilities}
\concept{jam igraph functions}
