% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-MemPlotFolio.R
\docType{class}
\name{MemPlotFolio-class}
\alias{MemPlotFolio-class}
\alias{MemPlotFolio}
\alias{list_to_MemPlotFolio}
\alias{MemPlotFolio_to_list}
\alias{show,MemPlotFolio-method}
\alias{Clusters,MemPlotFolio-method}
\alias{GeneClusters,MemPlotFolio-method}
\alias{ClusterLabels,MemPlotFolio-method}
\alias{ClusterLabels}
\alias{ClusterLabels<-,MemPlotFolio-method}
\alias{ClusterData,MemPlotFolio-method}
\alias{ClusterData}
\alias{ClusterData<-,MemPlotFolio-method}
\alias{thresholds,MemPlotFolio-method}
\alias{thresholds}
\alias{metadata,MemPlotFolio-method}
\alias{metadata}
\alias{Caption,MemPlotFolio-method}
\alias{Caption}
\alias{CaptionLegendList,MemPlotFolio-method}
\alias{CaptionLegendList}
\alias{CaptionLegendList<-,MemPlotFolio-method}
\alias{EnrichmentHeatmap,MemPlotFolio-method}
\alias{EnrichmentHeatmap}
\alias{GenePathHeatmap,MemPlotFolio-method}
\alias{GenePathHeatmap}
\alias{CnetCollapsed,MemPlotFolio-method}
\alias{CnetCollapsed}
\alias{CnetExemplar,MemPlotFolio-method}
\alias{CnetExemplar}
\alias{CnetCluster,MemPlotFolio-method}
\alias{CnetCluster}
\alias{plot,MemPlotFolio,ANY-method}
\alias{EnrichmentMap,MemPlotFolio-method}
\alias{EnrichmentMap}
\title{Mem S4 class, accessors, getters, and setters}
\usage{
list_to_MemPlotFolio(mpf)

MemPlotFolio_to_list(x, ...)

\S4method{show}{MemPlotFolio}(object)

\S4method{Clusters}{MemPlotFolio}(x)

\S4method{GeneClusters}{MemPlotFolio}(x)

\S4method{ClusterLabels}{MemPlotFolio}(x)

\S4method{ClusterLabels}{MemPlotFolio}(x) <- value

\S4method{ClusterData}{MemPlotFolio}(x)

\S4method{ClusterData}{MemPlotFolio}(x) <- value

\S4method{thresholds}{MemPlotFolio}(x)

\S4method{metadata}{MemPlotFolio}(x)

\S4method{Caption}{MemPlotFolio}(x, ...)

\S4method{CaptionLegendList}{MemPlotFolio}(x, ...)

\S4method{CaptionLegendList}{MemPlotFolio}(x) <- value

\S4method{EnrichmentHeatmap}{MemPlotFolio}(x, do_plot, params = list(width = 30), ...)

\S4method{GenePathHeatmap}{MemPlotFolio}(x, do_plot, params = list(width = 30), ...)

\S4method{CnetCollapsed}{MemPlotFolio}(
  x,
  type,
  do_plot,
  params = list(width = 30, maxNchar = Inf, layout = NULL, rotate_degrees = 0),
  legend_x = "bottomleft",
  legend_y = NULL,
  ...
)

\S4method{CnetExemplar}{MemPlotFolio}(
  x,
  num,
  do_plot,
  params = list(width = 30, maxNchar = Inf, layout = NULL, rotate_degrees = 0),
  legend_x = "bottomleft",
  legend_y = NULL,
  main = NULL,
  ...
)

\S4method{CnetCluster}{MemPlotFolio}(
  x,
  cluster,
  do_plot,
  params = list(width = 30, maxNchar = Inf, layout = NULL, rotate_degrees = 0),
  main = NULL,
  legend_x = "bottomleft",
  legend_y = NULL,
  ...
)

\S4method{plot}{MemPlotFolio,ANY}(x, y, ...)

\S4method{EnrichmentMap}{MemPlotFolio}(
  x,
  do_plot = TRUE,
  params = list(repulse = 3.5, width = 30, group = "default", mark.expand = 4, do_legend
    = TRUE),
  legend_x = "bottomleft",
  legend_y = NULL,
  ...
)
}
\arguments{
\item{mpf}{\code{list} output from \code{mem_plot_folio()}}

\item{x}{\code{MemPlotFolio} object}

\item{...}{additional arguments passed to \code{plot_mpf()}}

\item{value}{\code{Legends} object from ComplexHeatmap package. Use NULL
to hide the caption legend list entries.}

\item{y}{ignored}
}
\value{
\code{list_to_MemPlotFolio()} returns a \code{MemPlotFolio} S4 object,
from 'list' or 'MemPlotFolio' input.

\code{MemPlotFolio_to_list()} returns a \code{list}

\code{Clusters(MemPlotFolio)} returns a \code{list} of \code{character} vectors,
named by cluster.

\code{GeneClusters(MemPlotFolio)} returns a \code{list} of
\code{character} vectors, named by cluster.

\code{ClusterLabels(MemPlotFolio)} returns a named \code{character} vector
of cluster labels, or \code{NULL} if none are defined.

\code{ClusterLabels<-(MemPlotFolio)} returns the updated \code{MemPlotFolio}
object with new cluster labels stored in \code{metadata$cluster_labels}.

\code{ClusterData(MemPlotFolio)} returns a named \code{list}
of cluster data, or \code{NULL} if none are defined.

\code{ClusterData<-(MemPlotFolio)} returns the updated \code{MemPlotFolio}
object with new cluster data stored in \code{metadata$cluster_data}.

\code{thresholds(MemPlotFolio)} returns a \code{list} of
thresholds used with \code{mem_plot_folio()}.

\code{metadata(MemPlotFolio)} returns a \code{list} of
metadata used with \code{mem_plot_folio()}.

\code{metadata(MemPlotFolio)} returns a \code{character} string
with caption summary used with \code{mem_plot_folio()}.
Multiple lines are delimited by newline characters.

\code{metadata(MemPlotFolio)} returns the caption summary
in the form of \code{ComplexHeatmap::Legends} suitable to \code{draw()}
as R grid graphics.

\code{CaptionLegendList<-(MemPlotFolio)} returns the updated \code{MemPlotFolio}
object with new caption legend list.

\code{EnrichmentHeatmap(MemPlotFolio)} returns a
\code{ComplexHeatmap::HeatmapList} when do_plot is TRUE (default),
in addition to rendering the heatmap. It returns
\code{ComplexHeatmap::Heatmap} when do_plot is FALSE, containing
enrichment P-values by enrichment, and pathway rows in clusters.
Optional arguments '...' include:
\itemize{
\item 'main' or 'column_title' for overall plot title;
\item 'column_title_gp' as \code{grid::gpar()} to adjust title font;
\item 'use_cluster_labels' \code{logical} whether to use \code{ClusterLabels()}
as row title entries.
}

\code{GenePathHeatmap(MemPlotFolio)} returns a
\code{ComplexHeatmap::HeatmapList} when do_plot is TRUE (default),
\code{ComplexHeatmap::Heatmap} when do_plot is FALSE, containing
the genes-pathways incidence matrix and associated caption.

\code{CnetCollapsed(MemPlotFolio)} returns an \code{igraph} object invisibly,
with Gene and Set nodes representing the collapsed pathway clusters.
The legend position can be adjusted using 'legend_x', 'legend_y' which
are passed to \code{mem_legend()} as 'x' and 'y'.

\code{CnetExemplar(MemPlotFolio)} returns an \code{igraph} object
with Gene and Set nodes for the 'num' number of exemplars per cluster.

\code{CnetCluster(MemPlotFolio)} returns an \code{igraph} object invisibly,
using all pathways in the cluster defined with argument \code{cluster}.
}
\description{
Mem class containing results from \code{multiEnrichMap()},
current class version "1.0.0".
}
\section{Functions}{
\itemize{
\item \code{list_to_MemPlotFolio()}: Convert legacy \code{list} to S4 \code{MemPlotFolio}

\item \code{MemPlotFolio_to_list()}: Coerce S4 \code{MemPlotFolio} to \code{list} format

\item \code{show(MemPlotFolio)}: Show summary of a MemPlotFolio object,
dimensions defined by genes, sets, enrichments.

\item \code{Clusters(MemPlotFolio)}: Returns the order of gene sets from
\code{MemPlotFolio} results as a \code{list} named by pathway cluster,
containing \code{character} vectors of pathway gene sets.

\item \code{GeneClusters(MemPlotFolio)}: Returns the order of genes from
\code{MemPlotFolio} results as a \code{list} named by gene cluster,
containing \code{character} vectors of genes.

\item \code{ClusterLabels(MemPlotFolio)}: Returns cluster labels stored in the
\code{metadata} slot of \code{MemPlotFolio}.

\item \code{ClusterLabels(MemPlotFolio) <- value}: Sets cluster labels in the \code{metadata} slot
of \code{MemPlotFolio}.

\item \code{ClusterData(MemPlotFolio)}: Returns cluster data stored in the
\code{metadata} slot of \code{MemPlotFolio}, intended to store further
information to describe each cluster.

\item \code{ClusterData(MemPlotFolio) <- value}: Sets cluster data in the \code{metadata} slot
of \code{MemPlotFolio}, intended to store further information to
describe each cluster. It should be the same length as \code{Clusters(x)}
and have no names, or have names equal to \code{names(Clusters(x))}.

\item \code{thresholds(MemPlotFolio)}: Returns the thresholds used with
\code{MemPlotFolio}.

\item \code{metadata(MemPlotFolio)}: Returns the metadata used with
\code{MemPlotFolio}.

\item \code{Caption(MemPlotFolio)}: Returns the caption summary for
\code{MemPlotFolio}.

\item \code{CaptionLegendList(MemPlotFolio)}: Returns the caption summary for
\code{MemPlotFolio} as \code{ComplexHeatmap::Legends}.

\item \code{CaptionLegendList(MemPlotFolio) <- value}: Sets define caption legend list to display
in \code{EnrichmentHeatmap()} and \code{GenePathHeatmap()}.

\item \code{EnrichmentHeatmap(MemPlotFolio)}: Draws the enrichment heatmap
from \code{MemPlotFolio} results.
Optional arguments '...' include:
\itemize{
\item 'main' or 'column_title' for overall plot title;
\item 'column_title_gp' as \code{grid::gpar()} to adjust title font;
\item 'use_cluster_labels' \code{logical} whether to use \code{ClusterLabels()}
as row title entries.
}

\item \code{GenePathHeatmap(MemPlotFolio)}: Draws the gene-pathway set heatmap
from \code{MemPlotFolio} results.

\item \code{CnetCollapsed(MemPlotFolio)}: Draws the Cnet collapsed
network from \code{MemPlotFolio} results. Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.
When 'type' is not defined: it uses  type='cluster' if
\code{ClusterLabels()} is defined, otherwise type='set';
then it appends '2' to the end if there are more than 1000 nodes.
Argument 'type' can be:
\itemize{
\item \code{type='title'} or \code{type=''} to use cluster title
\item \code{type='set'} to use abbreviated pathway names
\item \code{type='cluster'} to use cluster labels
\item \code{type='attribute'} to use any existing vertex attribute.
\item Append \strong{'2'} to the end to hide gene labels. For example \code{type='set2'}.
}

Recognized arguments in '...':
\itemize{
\item \code{'width'} to apply word-wrap to cluster labels, or vertex attribute.
\item \code{'maxNchar'} to set maximum string character length, passed to
\code{fixSetLabels()}.
\item \code{'layout'} passed to \code{set_igraph_layout(cnet, layout)} to update
the node layout.
\item \code{'rotate_degrees'} with \code{numeric} value. When set, it calls
\code{rotate_igraph_layout()} with defaults.
\item All other '...' arguments are passed to \code{jam_igraph()}.
}

The legend includes direction if encoded in the 'MemPlotFolio', but
can be forced with \code{do_directional=TRUE} or FALSE.

\item \code{CnetExemplar(MemPlotFolio)}: Draws the Cnet exemplar network
from \code{MemPlotFolio} results for 'num' exemplar per cluster.
Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.

The legend includes direction if encoded in the 'MemPlotFolio', but
can be forced with \code{do_directional=TRUE} or FALSE. Legend position
is controlled by legend_x,legend_y, default legend_x="bottomleft".

\item \code{CnetCluster(MemPlotFolio)}: Draws the Cnet network
from \code{MemPlotFolio} results for a specific cluster.
Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.

The legend includes direction if encoded in the 'MemPlotFolio', but
can be forced with \code{do_directional=TRUE} or FALSE.
Plot title 'main', default NULL uses the cluster label if present,
then cluster name. Use main=FALSE or main="" to hide the title.

\item \code{plot(x = MemPlotFolio, y = ANY)}: Plot a \code{MemPlotFolio} object calling
\code{plot_mpf()}

\item \code{EnrichmentMap(MemPlotFolio)}: EnrichmentMap \code{igraph} network to connect
pathway gene sets based upon Jaccard overlap between each
pathway pair.
Note '...' arguments are passed to \code{jam_igraph()} and
\code{mem_legend()} when \code{do_plot=TRUE}.
Argument \code{'params'} is a \code{list} with additional arguments:
'repulse', 'width', 'group', 'mark.expand', 'do_legend'.

}}
\examples{
# list_to_MemPlotFolio examples
data(Memtest)
mpf <- mem_plot_folio(Memtest, do_plot=FALSE,
   do_which=c(1, 2, 3, 4), returnType="list")
Mpf <- list_to_MemPlotFolio(mpf)

data(Memtest)
mpf <- mem_plot_folio(Memtest, do_plot=FALSE, returnType="list")
Mpf <- list_to_MemPlotFolio(mpf)

# enrichment heatmap
EnrichmentHeatmap(Mpf, column_title="Enrichment Heatmap")

# Gene-path heatmap
GenePathHeatmap(Mpf, column_title="Gene-Path Heatmap")

# Cnet collapsed sets
CnetCollapsed(Mpf, type="set", use_shadowText=TRUE, main="Cnet Collapsed Sets")

# Cnet exemplar plot
CnetExemplar(Mpf, num=2, use_shadowText=TRUE, main="Cnet Exemplars, num=2")

# Cnet cluster plot
CnetCluster(Mpf, cluster="B", use_shadowText=TRUE, main="Cnet Cluster 'B'")

}
\seealso{
Other jam Mem utilities:
\code{\link{Mem-class}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}}

Other MemPlotFolio:
\code{\link[=check_MemPlotFolio]{check_MemPlotFolio()}},
\code{\link[=score_gene_path_clusters]{score_gene_path_clusters()}}
}
\concept{MemPlotFolio}
\concept{jam Mem utilities}
