% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\docType{data}
\name{Memtest}
\alias{Memtest}
\title{Memtest S4 Mem object for testing in multienrichjam}
\format{
\code{Mem} object containing two enrichment tests
}
\usage{
Memtest
}
\description{
Mem object for testing in multienrichjam, with 'Newborns' and
'OlderChildren' described in
Reese et al 2019 \link{https://doi.org/10.1016/j.jaci.2018.11.043}.
}
\seealso{
Other Mem: 
\code{\link{Mem-class}},
\code{\link{check_Mem}()}
}
\concept{Mem}
\keyword{datasets}
