% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{mergeAllXY}
\alias{mergeAllXY}
\title{Merge a list of data.frames keeping all rows}
\usage{
mergeAllXY(...)
}
\description{
Merge a list of data.frames keeping all rows
}
\details{
This function is an extension to the `base::merge.data.frame()`
function, with the default arguments `all.x=TRUE, all.y=TRUE` in
order to maintain all rows of all input data.frames.
}
\seealso{
Other jam list functions: \code{\link{unnestList}}
}
\concept{jam list functions}
