% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multienrichjam-package.R
\docType{package}
\name{multienrichjam-package}
\alias{multienrichjam}
\alias{multienrichjam-package}
\title{Multi-enrichment analysis with multienrichjam}
\description{
Analysis of multiple gene set enrichment results, focused on pathway clustering driven by gene content, visualization of gene-pathway relationships, condensing redundant pathways to produce interpretable gene-pathway concept network graphs. Bipartite graphs render with edge bundling and efficient vectorized node rendering.
}
\details{
This package aims to enable analysis for one or multiple
pathway or functional enrichment datasets, with flexible
options for publication of data visualizations as figures.

The core workflow follows these steps:
\enumerate{
\item Import data for use.
\item Call \code{multiEnrichMap()} using relevant parameters.
\item Visualize with \code{mem_plot_folio()} for review.
\item Iterate then polish final figures.
}

The analysis produces:
\itemize{
\item \code{Mem-class}: multi-enrichment data used for visualization.
\item Data visualizations from \code{mem_plot_folio()} in \code{list} format.
}

Concept network (Cnet) plots can be customized using ShinyCat:
\itemize{
\item \code{launch_shinycat()} - interactive network adjustments
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmw86069.github.io/multienrichjam/}
  \item Report bugs at \url{http://github.com/jmw86069/multienrichjam/issues}
}

}
\author{
\strong{Maintainer}: James M. Ward \email{jmw86069@gmail.com} (\href{https://orcid.org/0000-0002-9510-2848}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item James M. Ward \email{jmw86069@gmail.com} (\href{https://orcid.org/0000-0002-9510-2848}{ORCID}) [copyright holder]
}

}
\keyword{internal}
