% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{multiEnrichMap}
\alias{multiEnrichMap}
\title{Prepare MultiEnrichMap data from enrichList}
\usage{
multiEnrichMap(
  enrichList,
  geneHitList = NULL,
  geneHitIM = NULL,
  colorV = NULL,
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  enrichLabels = NULL,
  subsetSets = NULL,
  overlapThreshold = deprecated(),
  p_cutoff = 0.05,
  cutoffRowMinP = deprecated(),
  enrichBaseline = -log10(p_cutoff),
  enrichLens = 0,
  enrichNumLimit = 4,
  nEM = 500,
  min_count = 3,
  topEnrichN = 20,
  topEnrichSources = c("gs_cat", "gs_subcat"),
  topEnrichCurateFrom = NULL,
  topEnrichCurateTo = NULL,
  topEnrichSourceSubset = NULL,
  topEnrichDescriptionGrep = NULL,
  topEnrichNameGrep = NULL,
  keyColname = c("ID", "Name", "pathway", "itemsetID", "Description"),
  nameColname = c("Name", "pathway", "Description", "itemsetID", "ID"),
  geneColname = c("geneID", "geneNames", "Genes"),
  countColname = c("gene_count", "count", "geneHits"),
  pvalueColname = c("padjust", "p.adjust", "adjp", "padj", "qvalue", "qval", "q.value",
    "pvalue", "p.value", "pval", "FDR"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  descriptionCurateFrom = c("^Genes annotated by the GO term "),
  descriptionCurateTo = c(""),
  directionColname = c("direction", "NES", "activation.z.{0,1}score", "z.{0,1}score"),
  direction_cutoff = 0,
  pathGenes = c("setSize", "pathGenes", "Count"),
  geneHits = c("Count", "geneHits", "gene_count"),
  geneDelim = "[,/ ]+",
  GmtTname = NULL,
  msigdbGmtT = NULL,
  returnType = c("Mem", "list"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichList}{\code{list} of \code{enrichResult} or \code{data.frame} objects.
\itemize{
\item The \code{names(enrichList)} are used in subsequent results.
\item Note that \code{data.frame} are converted to \code{enrichResult} using
\code{enrichDF2enrichResult(x, ...)} where the '...' ellipses are
used to recognize colnames.
\item Recommendation is to confirm each \code{data.frame} is properly
converted to \code{enrichResult} upfront.
}}

\item{geneHitList}{\code{list} of character vectors, or
\code{list} of \code{numeric} vectors whose names represent genes, or
or \code{NULL}. When \code{NULL} the gene hit list for each enrichment
result is inferred from the enrichment results themselves,
however this option may incompletely represent which genes
were statistical hits.
Note that \code{geneHitList} and \code{geneHitIM} serve the same purpose
and either can be supplied.}

\item{geneHitIM}{\code{numeric} matrix with gene rows, enrichment columns,
and \code{numeric} values indicating the presence and/or direction
of change for each gene.
Note that \code{geneHitList} and \code{geneHitIM} serve the same purpose
and either can be supplied.}

\item{colorV}{\code{character} vector of colors, length
equal to \code{length(enrichList)},
used to assign specific colors to each enrichment result.}

\item{nrow, ncol, byrow}{optional arguments used to customize
\code{igraph} node shape \code{"coloredrectangle"}, useful when the
number of \code{enrichList} results is larger than around 4. It
defines the number of columns and rows used for each node,
to display enrichment result colors, and whether to fill
colors by row when \code{byrow=TRUE}, or by column when \code{byrow=FALSE}.}

\item{enrichLabels}{\code{character} vector of enrichment labels to use,
as an optional alternative to \code{names(enrichList)}.}

\item{subsetSets}{\code{character} vector of optional set names to
use in the analysis, useful to analyze only a specific subset
of known pathways.}

\item{overlapThreshold}{\code{numeric} (deprecated), value between 0 and 1,
used to define the Enrichment Map, which is only created for legacy
output enabled with \code{returnType="list"}.
\itemize{
\item To create a Multi-Enrichment Map \code{igraph} object, see \code{mem2emap()}.
\item overlapThreshold is the Jaccard overlap score above which
two pathways will be linked in the resulting network.
}}

\item{p_cutoff}{\code{numeric} value between 0 and 1, default 0.05,
enrichment P-value required by at least one enrichment result to
be retained in downstream analyses.
This P-value can be reviewed in \code{enrichIM(Mem)} of the output,
which is a matrix of P-values by pathway and enrichment.
The column header assigned as P-value is stored in
\code{headers(Mem)$pvalueColname}.}

\item{cutoffRowMinP}{(deprecated in favor of 'p_cutoff').
When it is non-NULL, it will be used for backward compatibility.}

\item{enrichBaseline}{\code{numeric} value, default uses \code{-log10(p_cutoff)},
the -log10 P-value threshold required for the color gradient to assign
a color associated with enrichment.
\itemize{
\item In other words, P-values that do not meet this
threshold are not colored in the color gradient for enrichment
P-values.
\item To color all P-values use \code{enrichBaseline=0}.
}}

\item{enrichLens}{\code{numeric} value, default 0, indicating the "lens"
to enhance the intensity of color gradients. Numbers above 0 make
the color ramp more compressed, and more vivid at lower numeric values.}

\item{enrichNumLimit}{\code{numeric} value indicating the \code{-log10(P-value)}
above which each color gradient is considered the maximum color,
useful to apply a fixed threshold for each color gradient.}

\item{nEM}{\code{integer} (deprecated) maximum pathways to include in the
Enrichment Map, which is only created for legacy
output enabled with \code{returnType="list"}.
\itemize{
\item To create a Multi-Enrichment Map \code{igraph} object, see \code{mem2emap()}.
}}

\item{topEnrichN}{\code{integer} maximum rows to retain from each
\code{enrichResult} in 'enrichList', for each source when supplied.
Set \code{topEnrichN=0} or \code{topEnrichN=NULL} to retain all rows.
Only rows where the pathway met other filtering criteria in at
least one enrichment in 'enrichList' will be retained.}

\item{topEnrichSources, topEnrichCurateFrom, topEnrichCurateTo, topEnrichSourceSubset, topEnrichDescriptionGrep, topEnrichNameGrep}{arguments passed to \code{topEnrichListBySource()} when \code{topEnrichN}
is greater than \code{0}. The default values are used only when
input data matches these patterns.}

\item{keyColname, nameColname, geneColname, pvalueColname, descriptionColname}{\code{character} vector in each case with text strings or patterns to
use when matching or prioritizing colnames to assign to each type:
\itemize{
\item \code{key}: The primary unique key for each pathway
\item \code{name}: The short name for each pathway
\item \code{gene}: column with delimited gene symbols or identifiers tested
for enrichment of each pathway.
\item \code{pvalue}: column with enrichment P-value, typically prioritizing
either 'qvalue' or 'p.adjust'. Per \code{enrichResult} convention,
the selected column is renamed to 'pvalue'.
\item \code{description}: longer description associated with the pathway.
}

Each vector is passed to \code{find_colname()} to find
a suitable matching colname for each entry in \code{enrichList}.
That function prioritizes full colname matches, then leading or
trailing matches, then substring.}

\item{descriptionCurateFrom, descriptionCurateTo}{\code{character} vectors
with patterns and replacements, passed to \code{gsubs()}, intended to
help curate common descriptions to shorter, perhaps more
user-friendly labels. One example is removing the prefix
\code{"Genes annotated by the GO term "} from Gene Ontology pathways.
These label can be manually curated later in the \code{Mem-class}
methods, specifically \verb{sets()<-} will allows assignment of custom
names.}

\item{pathGenes, geneHits}{\code{character} values indicating the colnames
that contain the number of pathway genes, and the number of gene
hits, respectively. These values are optional, and not specifically
used by multienrichjam.}

\item{geneDelim}{\code{character} pattern used with \code{strsplit()} to
split multiple gene values into a list of vectors.
The default (required) delimiter for \code{enrichResult} objects
is \code{'/'}. A common alternative is \code{','} (comma-delimited).}

\item{verbose}{\code{logical} indicating whether to print verbose output.
For \code{verbose} to cascade to internal functions, use \code{verbose=2}.}

\item{...}{additional arguments are passed to various internal
functions.}
}
\value{
\code{list} object containing various result formats:
\itemize{
\item colorV: named vector of colors assigned to each enrichment,
where names match the names of each enrichment in \code{enrichList}.
}
}
\description{
Prepare MultiEnrichMap data from enrichList
}
\details{
This function performs most of the work of comparing multiple
enrichment results.
This function takes a list of \code{enrichResult} objects,
generates an overall pathway-gene incidence matrix, assembles
a pathway-to-Pvalue matrix, creates EnrichMap \code{igraph} network
objects, and CnetPlot \code{igraph} network objects. It also applies
node shapes and colors consistent with the colors used for
each enrichment result.

By default, each enrichment result table is subsetted for the
top \code{n=20} pathways sorted by pathway source, defined by
colnames \code{c("Source", "Category")}. For data without a source
column, the overall enrichment results are sorted to take the
top 20. Once the top 20 from each enrichment table are selected,
the overall set of pathways are used to retain these pathways
from all enrichment tables. In this way, a significant enrichment
result from one table will still be compared to a non-significant
result from another table.

The default values for \code{topEnrichN} and related arguments
are intended when using enrichment results from \code{MSigDB},
which has colnames \code{c("Source","Category")} and represents
100 or more combinations of sources and categories. The
default values will select the top 20 entries from the
canonical pathways, after curating the canonical pathway
categories to one \code{"CP"} source value.

To disable the top pathway filtering, set \code{topEnrichN=0}.

Colors can be defined for each enrichment result using the
argument \code{colorV}, otherwise colors are assigned using
\code{colorjam::rainbowJam()}.
}
\examples{
## See the Vignette for a full walkthrough example

}
\seealso{
Other multienrichjam core functions:
\code{\link[=jam_igraph]{jam_igraph()}},
\code{\link[=mem_plot_folio]{mem_plot_folio()}}
}
\concept{multienrichjam core functions}
