% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{nodegroups2communities}
\alias{nodegroups2communities}
\title{Convert nodegroups list to communities object}
\usage{
nodegroups2communities(nodegroups, algorithm = NULL, ...)
}
\arguments{
\item{algorithm}{\code{character} or \code{NULL}, indicating the name of the
community detection algorithm used.
\itemize{
\item When \code{algorithm} is defined, it is used instead of
\code{attr(nodegroups, "algorithm")}.
\item When \code{algorithm} is \code{NULL}, \code{attribute(nodegroups, "algorithm")} is
used if defined, otherwise \code{algorithm="nodegroups"}.
}}

\item{...}{additional arguments are ignored.}

\item{wc}{\code{communities} object as returned by \code{igraph} functions
such as \code{cluster_optimal()}, \code{cluster_walktrap()}, or
\code{cluster_leading_eigen()}.}
}
\value{
\code{community} object, which is essentially a \code{list} with
specific required elements:
\itemize{
\item \code{"membership"} - \code{integer} assignment of nodes to clusters
\item \code{"names"} - \code{character} list of node names
\item \code{"vcount"} - \code{integer} number of nodes
\item \code{"algorithm"} - \code{character} string with the name of the community
detection method used.
\item \code{"cluster_names"} - \code{character} labels associated with \code{membership}
index values. These names are not generated by \code{igraph} community
detection, and are therefore optional for use in most \code{igraph}
workflows. However, they are used in some \code{multienrichjam} functions,
specifically \code{make_point_hull()} which optionally displays a
label beside each node cluster during plotting.
}
}
\description{
Convert nodegroups list to communities object
}
\details{
Note that this function is "lossy", in that the output \code{communities}
object will not contain any supporting data specific to the
community detection algorithm originally used.
However, the output \code{communities} object will be accepted
by most \code{igraph} related functions that require
that object type as an input value.

The \code{names(nodegroups)} are used to define a new element in the
output \code{communities} object \code{"cluster_names"}, so the names
will be maintained in the data. Default \code{igraph} functions
do not use these names, but they are used by \code{multienrichjam}
for example by function \code{make_point_hull()} which uses these
names to label each cluster during plotting.
}
\seealso{
Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
