% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-internals.R
\name{parse_igraph_plot_params}
\alias{parse_igraph_plot_params}
\title{Parse igraph plot params}
\usage{
parse_igraph_plot_params(graph, params)
}
\description{
This function mimics the internal function \code{igraph:::i.parse.plot.params()}.
}
\seealso{
Other jam igraph internal functions: 
\code{\link{default_igraph_values}()},
\code{\link{get_igraph_arrow_mode}()},
\code{\link{jam_igraph_arrows}()}
}
\concept{jam igraph internal functions}
