% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-plot.R
\name{plot_layout_scale}
\alias{plot_layout_scale}
\title{Plot layout scale by percentage of coordinate range}
\usage{
plot_layout_scale(
  layout,
  grid_colors = c("grey80", "grey70"),
  grid_lty = c("dotted", "solid"),
  big_tick = 10,
  small_tick = 2.5,
  ...
)
}
\arguments{
\item{layout}{\code{matrix} or \code{data.frame} with at least two columns,
only the first two columns are used for the grid.}

\item{grid_colors}{\code{character} colors used for the small and big grid
lines, respectively.}

\item{grid_lty}{\code{integer} or \code{character} line type used for the
small and big grid lines, respectively.}

\item{big_tick, small_tick}{\code{numeric} values in percent, the step size
between big grid lines, and small grid lines, respectively.}

\item{...}{additional arguments are ignored.}
}
\description{
Plot layout scale by percentage of coordinate range
}
\details{
The \code{layout} argument supplied coordinates, and largest numeric
range of any column is used to define 100 percent scale.
A grey grid is drawn on a base R plot to indicate the big
and small steps across the range, using \code{big_tick} and \code{small_tick},
respectively.
}
\seealso{
Other jam plot functions:
\code{\link[=adjust_polygon_border]{adjust_polygon_border()}}
}
\concept{jam plot functions}
