% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-rotate.R
\name{rotate_coordinates}
\alias{rotate_coordinates}
\title{Rotate numeric coordinates}
\usage{
rotate_coordinates(
  x,
  degrees = 0,
  reflect = c("none", "x", "y", "z"),
  center = NULL,
  center_rule = c("median", "origin", "mean", "min", "max"),
  rotation_axes = c(1, 2),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{matrix} with 2 or more columns.}

\item{degrees}{numeric value indicating the degrees to rotate
layout coordinates, where 360 degrees is one complete rotation.}

\item{reflect}{\code{character} string indicating one or more axes
to reflect coordinates, which flips the position of coordinates
along that axis. It is usually called to flip x-axis or y-axis
coordinates, for example with \code{reflect="x"} or \code{reflect=1}.
Input is handled as follows:
\itemize{
\item if \code{reflect} contains \code{"none"}, then reflect is applied to
none of the coordinate axes, therefore the default
\code{reflect=c("none", "x", "y", "z")} will apply no reflection.
\item \code{character} input: \code{reflect} values are matched to \code{colnames(x)}.
When there are no \code{colnames(x)}, then \code{reflect} values of
\code{c("x", "y", "z")} are automatically recognized as columns
\code{c(1, 2, 3)} respectively.
\item \code{integer} input is treated as a vector of column index positions,
for example \code{reflect=c(2)} will reflect values on the second
coordinate column.
}}

\item{center}{\code{numeric} coordinates to use as the center, or
\code{center=NULL} to calculate the center using \code{center_rule}.}

\item{center_rule}{\code{character} string indicating which rule to
apply to determine the center coordinates when \code{center=NULL}.
Note that it has little effect on most downstream plotting
assuming the plot function adjusts x- and y-axis ranges to
the data range, but may modify the axis ranges as a result.
\itemize{
\item \code{"origin"} uses c(0, 0);
\item \code{"mean"} uses the mean of each axis;
\item \code{"median"} uses the median of each axis;
\item \code{"min"} uses the minimum of each axis;
\item \code{"max"} uses the max of each axis.
}}

\item{rotation_axes}{\code{integer} vector indicating which axis
coordinates to rotate, by default \code{c(1, 2)} uses the first
two axes in \code{x}. Note that \code{rotation_axes} must represent
columns present in x.}

\item{...}{additional arguments are ignored.}
}
\value{
\verb{numeric matrix} with the same number of columns as the
input \code{x}.
}
\description{
Rotate numeric coordinates, optionally after reflecting coordinates
along one or more coordinate axes.
}
\details{
This function rotates coordinates in two axes, by the angle
defined in \code{degrees}. It optionally reflects coordinates in
one or more axes, which occurs before rotation.

Note that the \code{reflect} is applied before \code{degrees}.

Rotation code kindly contributed by Don MacQueen to the \code{maptools}
package, and is reproduced here to avoid a dependency on \code{maptools}
and therefore the \code{sp} package.
}
\examples{
layout <- cbind(0:10, 0:10);
layout_rot50 <- rotate_coordinates(x=layout, degrees=50);
layout_rot40_ctrmean <- rotate_coordinates(x=layout, degrees=40, center_rule="mean");
layout_reflectx_ctrmean <- rotate_coordinates(x=layout, reflect="x", center_rule="mean");
plot(rbind(layout, layout_rot50, layout_rot40_ctrmean, layout_reflectx_ctrmean),
   col=rep(c("darkorchid", "darkorange1", "dodgerblue", "red4"), each=11),
   pch=rep(c(17, 20, 18, 17), each=11),
   cex=2);

}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
