% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraphshapes.R
\name{shape.ellipse.clip}
\alias{shape.ellipse.clip}
\title{clip function for igraph vertex shape ellipse}
\usage{
shape.ellipse.clip(coords, el, params, end = c("both", "from", "to"))
}
\description{
clip function for igraph vertex shape ellipse
}
\details{
This function defines the clipping function for custom igraph vertex
shape ellipse.
}
\seealso{
Other jam igraph shapes:
\code{\link[=jam_mypie]{jam_mypie()}},
\code{\link[=shape.coloredrectangle.plot]{shape.coloredrectangle.plot()}},
\code{\link[=shape.ellipse.plot]{shape.ellipse.plot()}},
\code{\link[=shape.jampie.plot]{shape.jampie.plot()}}
}
\concept{jam igraph shapes}
