% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraphshapes.R
\name{shape.ellipse.plot}
\alias{shape.ellipse.plot}
\title{plot function for igraph vertex shape ellipse}
\usage{
shape.ellipse.plot(coords, v = NULL, params)
}
\description{
plot function for igraph vertex shape ellipse
}
\details{
This function defines the plotting function for custom igraph vertex
shape ellipse.
}
\seealso{
Other jam igraph shapes:
\code{\link[=jam_mypie]{jam_mypie()}},
\code{\link[=shape.coloredrectangle.plot]{shape.coloredrectangle.plot()}},
\code{\link[=shape.ellipse.clip]{shape.ellipse.clip()}},
\code{\link[=shape.jampie.plot]{shape.jampie.plot()}}
}
\concept{jam igraph shapes}
