% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycat-server.R
\name{shinycat_server}
\alias{shinycat_server}
\title{R-shiny server function for shinycat}
\usage{
shinycat_server(input, output, session)
}
\arguments{
\item{input, output, session}{arguments used by \code{shiny::shinyApp()}}
}
\description{
R-shiny server function for shinycat, internal use by launch_shinycat()
}
\seealso{
Other jam shiny functions:
\code{\link[=launch_shinycat]{launch_shinycat()}},
\code{\link[=shinycat_ui]{shinycat_ui()}}
}
\concept{jam shiny functions}
