% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{sort_colors}
\alias{sort_colors}
\title{Sort colors}
\usage{
sort_colors(
  x,
  sort_by = c("h", "-c", "-l"),
  c_min = 4,
  grey_hue = 359,
  hue_offset = 0,
  ...
)
}
\description{
Sort colors
}
\examples{
x <- jamba::nameVector(colors());

## Basic color sort
c2 <- sort_colors(x);
jamba::showColors(c2, main="cmin=4");

## Increase filtering of unsaturated colors
c3 <- sort_colors(x, c_min=20);
jamba::showColors(c3, main="cmin=20");

## Increase filtering of unsaturated colors
c4 <- sort_colors(x, c_min=50);
jamba::showColors(c4, main="cmin=50");

}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
