% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-subset-mem.R
\name{subset_mem}
\alias{subset_mem}
\title{Subset mem multienrichment object}
\usage{
subset_mem(
  mem,
  includeSets = NULL,
  includeGenes = NULL,
  min_gene_ct = 1,
  min_set_ct = 1,
  p_cutoff = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{list} object as returned by \code{multiEnrichMap()}}

\item{includeSets}{\code{character} vector with specific sets to retain,
all other sets will be dropped.}

\item{includeGenes}{\code{character} vector with specific genes to retain,
all other genes will be dropped.}

\item{min_gene_ct}{\code{numeric} filter applied to genes representing the
minimum number of occurrences across sets in the \code{mem$memIM}
incidence matrix. The default value \code{min_gene_ct=1} effectively
requires a gene to be present in at least one set, which is useful
after filtering by \code{includeSets}.}

\item{min_set_ct}{\code{numeric} filter applied to sets representing the
minimum number of occurrences of genes in the \code{mem$memIM}
incidence matrix. The default value \code{min_set_ct=1} effectively
requires a set to contain at least one gene.}

\item{p_cutoff}{\code{numeric} optional enrichment P-value filter to apply
to \code{mem$enrichIM} enrichment P-values.
It is intended to apply optionally higher stringency by using a
lower \code{p_cutoff} than used by \code{multiEnrichMap()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} object of \code{mem} data
}
\description{
Subset mem multienrichment object
}
\details{
This function is intended to subset the incidence matrix data contained
in a \code{mem} object by heuristics. It does not update other data in
the \code{mem} object such as \code{enrichList} and \code{multiEnrichDF}, nor
any \code{igraph} objects. It is intended mainly to subset by sets (pathways),
or genes, then also subset other corresponding incidence matrix
data consistently.
}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{bulk_cnet_adjustments}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
