% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-with-ht-opts.R
\name{with_ht_opts}
\alias{with_ht_opts}
\alias{local_ht_opts}
\title{Withr mimic of with_options() for ComplexHeatmap options}
\usage{
with_ht_opts(new, code)

local_ht_opts(.new = list(), ..., .local_envir = parent.frame())
}
\arguments{
\item{new}{\code{list} named by heatmap option, with corresponding values
as described in \code{ComplexHeatmap::ht_opt()}.}

\item{code}{\code{any} valid R code to execute in the temporary environment.}
}
\value{
\code{any} The results of the evaluation of the \code{code} argument.
}
\description{
Withr mimic of with_options() for ComplexHeatmap options
}
\details{
This function uses the same mechanism as used by \code{withr::with_options()}
except focuses on options defined by \code{ComplexHeatmap::ht_opt()}.
}
\examples{
ComplexHeatmap::ht_opt("COLUMN_ANNO_PADDING")
#> 1mm
with_ht_opts(list(COLUMN_ANNO_PADDING=grid::unit(3, "mm")),
   ComplexHeatmap::ht_opt("COLUMN_ANNO_PADDING"))
#> 3mm
ComplexHeatmap::ht_opt("COLUMN_ANNO_PADDING")
#> 1mm

test_local <- function() {
   local_ht_opts(list(COLUMN_ANNO_PADDING=grid::unit(3, "mm")))
   print(ComplexHeatmap::ht_opt("COLUMN_ANNO_PADDING"))
}
test_local()
#> 3mm
ComplexHeatmap::ht_opt("COLUMN_ANNO_PADDING")
#> 1mm
}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
