% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{with_qfr}
\alias{with_qfr}
\title{Layout specification for Qgraph Fruchterman-Reingold, deprecated}
\usage{
with_qfr(..., repulse = 4)
}
\description{
Layout specification for Qgraph Fruchterman-Reingold, deprecated
}
\seealso{
Other jam deprecated functions:
\code{\link[=cnetplotJam]{cnetplotJam()}},
\code{\link[=enrichMapJam]{enrichMapJam()}},
\code{\link[=grid_with_title]{grid_with_title()}},
\code{\link[=mem_multienrichplot]{mem_multienrichplot()}}
}
\concept{jam deprecated functions}
