% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-direction.R
\name{add_pathway_direction}
\alias{add_pathway_direction}
\title{Add directionality to pathway enrichment}
\usage{
add_pathway_direction(
  x,
  gene_hits,
  delim = "/",
  dir_colname = "z-score",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{enrichResult} object as returned by \code{clusterProfiler::enricher()}}

\item{gene_hits}{\code{numeric} vector, with values \code{c(-1, 1)} to
indicate up or down, respectively, and whose \code{names(gene_hits)}
should match the gene symbols (or values) in \code{x@result$geneID}.
If no \code{names(gene_hits)} match values in the \code{geneID} column,
the direction score will ultimately be populated with \code{NA} values.}

\item{delim}{\code{character} string with delimiter used in the
\code{enrichResult} object, default: \code{"/"}}

\item{dir_colname}{\code{character} string indicating the column name to
create, default \code{"z-score"}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{enrichResult} object, with directional values populated
into the result column name defined by \code{dir_colname}.
When no genes in a gene set are present in \code{gene_hits} the
expected directional value is \code{NA}.
}
\description{
Add directionality to pathway enrichment
}
\details{
This function uses pathway enrichment data provided as an
\code{enrichResult} object. Data is added as a new column (default "z-score")
with \code{numeric} values indicating the direction of change.

It requires \code{gene_hits} with the direction of change for each gene,
which is used to determine the up- and down-regulated genes
in each gene set.

The default equation uses the IPA z-score of activation, since the
z-score has beneficial properties in downstream visualization functions.
It is symmetric around zero, and implies reasonable thresholds.

Note that when no genes involved in enrichment are present in the
\code{gene_hits} data, the resulting directional value is expected
to be \code{NA}, which will be treated as "non-directional" in downstream
visualization functions.
\subsection{IPA z-score activation calculation}{

The IPA Activation \code{"z-score"} uses the following formula:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{z <- (N_genes_up - N_genes_down) / sqrt(N_genes_up + N_genes_down)
}\if{html}{\out{</div>}}

Citation for IPA formula: \link{https://doi.org/10.1093/bioinformatics/btt703}
as referenced: \href{https://qiagen.my.salesforce-sites.com/KnowledgeBase/KnowledgeNavigatorPage?id=kA41i000000L5nQCAS&categoryName=BioX}{IPA FAQ - Statistical Calculations}
}

\subsection{Direction calculation (deprecated)}{

The deprecated calculation for \code{"direction"} is shown below:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{direction_1 <- (N_genes_up - N_genes_down) / (N_genes);
# then log2(1.2 + x) transformed
direction <- log2(1.2 + abs(direction_1)) * sign(direction_1)
}\if{html}{\out{</div>}}
}
}
\seealso{
Other jam enrichment functions: 
\code{\link{multiEnrichMap}()},
\code{\link{multienrichjam}()},
\code{\link{topEnrichBySource}()}
}
\concept{jam enrichment functions}
