% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-cnet.R
\name{adjust_cnet_nodeset}
\alias{adjust_cnet_nodeset}
\title{Adjust Cnet node set}
\usage{
adjust_cnet_nodeset(
  g,
  set_nodes = NULL,
  x = 0,
  y = 0,
  expand = 0,
  percent_spacing = NULL,
  rotate_degrees = 0,
  cnet_nodesets = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} Cnet plot, with vertex attribute \code{nodeType} that
contains values \code{"Gene"} and \code{"Set"}, as produced by \code{memIM2cnet()}.}

\item{set_nodes}{\code{character} vector of one or more set names, which are
defined here as the names of the nodes with \code{nodeType="Set"}
to which each node with \code{nodeType="Gene"} is connected.
For example \code{set_nodes=c("A", "B")} will return all Gene nodes
that are connected only to Set nodes \code{"A"} and \code{"B"}. Alternatively,
one can refer to a cluster by the node name of one member of the
cluster, as a convenience option.}

\item{x, y}{\code{numeric} indicating the relative amount to move nodes
in the node set, based upon the overall range of x and y values
in the layout coordinates.}

\item{expand}{\code{numeric} the relative amount to expand node coordinates,
where \code{expand=0} does no expansion, \code{expand=1} will expand coordinates
100\% of the distance to the center of the cluster, twice the original
distance; and \code{expand=-1} will expand coordinates to half the
distance to the center.}

\item{rotate_degrees}{\code{numeric} value indicating the rotation in
degrees to adjust the node coordinates, relative to the center of
the coordinates.}

\item{cnet_nodesets}{\code{list} optional to provide the known nodesets
upfront, in order to avoid re-determining the nodesets.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Adjust Cnet node set
}
\details{
This function is intended to help adjust a node set,
defined as a set of \code{"Gene"} nodes that all connect
to the same \code{"Set"} nodes. These nodes are usually clustered
together in a Cnet plot, and can be manipulated as a
group using this function. Use \code{get_cnet_nodeset(g)}
to get all node sets from a Cnet \code{igraph} object.

Specifically, this function can be used to expand,
rotate, and move the cluster of nodes in a Cnet
network layout. The purpose is usually to help the
visual clarity of the plot, and to reduce node label
overlaps.
}
\examples{
# the examples below are for internal data
# and will be rewritten to work with example data
if (1 == 2) {
g <- mem_dmjdm_gp_20_plots_jikl$cnet_collapsed_set;
set_nodes <- c("J", "L")
g2 <- adjust_cnet_nodeset(g, set_nodes=c("J", "L"), expand=0.5)

jam_igraph(g)
g2 <- adjust_cnet_nodeset(g, set_nodes=c("I", "K", "L"), x=0.01, y=-0.01, expand=0.2)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "J", "K", "L"), x=-0.08, y=-0.05, expand=0.2)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "L"), x=-0.05, y=0.05, expand=0.3)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "J", "L"), x=-0.02, y=0.03, expand=0.3)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "K"), x=0.03, y=0.09, expand=0.)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "J", "K"), x=0.02, y=-0.01, expand=0.)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("I", "J"), x=0.01, y=-0.00, expand=0.1)
g2 <- adjust_cnet_nodeset(g2, set_nodes=c("J"), x=-0.03, y=-0.02, expand=0.)
jam_igraph(g2)
}

}
\seealso{
Other jam cnet igraph functions: 
\code{\link{adjust_cnet_set_relayout_gene}()},
\code{\link{apply_nodeset_spacing}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{nudge_igraph_node}()},
\code{\link{rotate_igraph_layout}()}
}
\concept{jam cnet igraph functions}
