% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-apply-cnet-direction.R
\name{apply_cnet_direction}
\alias{apply_cnet_direction}
\title{Apply Cnet border color by directionality}
\usage{
apply_cnet_direction(
  cnet,
  hitim = NULL,
  col = colorjam::col_div_xf(1.2),
  col_l_max = 65,
  hide_solo_pie = TRUE,
  frame_blank = default_igraph_values()$vertex$frame.color,
  frame_lwd_blank = 0.2,
  border_lwd = 2,
  do_reorder = FALSE,
  ...
)
}
\arguments{
\item{cnet}{\code{igraph} object with node attribute \code{"pie.color"}
populated as a \code{list} of \code{character} vectors, named by
enrichment. The enrichment names should match \code{colnames(hitim)}.}

\item{hitim}{\code{numeric} matrix with values centered at zero for
no change, positive values \code{+1} for up-regulation, and
negative values \code{-1} for down-regulation. The values are
converted to color using color function \code{col}.}

\item{col}{\code{function} that takes \code{numeric} input and assigns
a color. The default assigns red for positive values, blue
for negative values, and white for zero, using
\code{colorjam::col_div_xf(1.2)}.}

\item{col_l_max}{\code{numeric} maximum HCL Lightness permitted for
output colors, for example because the middle color in \code{"RdBu_r"}
is nearly white, it may be preferable to apply a darker grey
color.}

\item{hide_solo_pie}{\code{logical} indicating whether a single-color
border should only be applied to \code{"frame.color"} and not to
individual \code{"pie.border"} entries.
\itemize{
\item When \code{hide_solo_pie=TRUE} and all wedges (one or more) will
have the same \code{pie.border} color, then \code{pie.border} is defined
as \code{NA}, and instead the \code{frame.color} is assigned to this color.
Both \code{pie.lwd} and \code{frame.lwd} will be assigned \code{border_lwd},
since \code{pie.border} is \code{NA} it will not be rendered, only the
\code{frame.color} will be rendered.
\item When \code{hide_solo_pie=FALSE} each pie wedge border color is assigned
to \code{pie.border}, \code{frame.color} will be assigned \code{frame_blank},
and \code{frame.lwd} will be assigned \code{frame_lwd_blank} which is useful
for displaying a small outer frame for each node.
}}

\item{frame_blank}{\code{character} string to define the color used
for \code{"frame.color"} when there are multiple colors in \code{"pie.border"}
and therefore there should be no visible \code{"frame.color"}.}

\item{do_reorder}{\code{logical} indicating whether to call
\code{reorder_igraph_nodes()} on the resulting \code{igraph}, so that
the border color can be used in the sort conditions.}

\item{...}{additional arguments are ignored.}
}
\description{
Apply Cnet border color by directionality
}
\details{
This function specifically requires the \code{"pie.color"}
node attribute of the \code{cnet} object \code{igraph} is populated
with a \code{list} of colors whose name is the enrichment,
and matches \code{colnames(hitim)} of the supplied hit
incidence matrix.

The node names \code{V(cnet)$name} also must match \code{rownames(hitim)},
otherwise the \code{"pie.border"} and \code{"coloredrect.border"} is assigned
its previous value.
In that case \code{"frame.color"} retains its original value.

This function may be run multiple times with different \code{hitim},
for example using \code{geneIMdirection} and \code{enrichIMdirection} in two steps.

When there are multiple colors for a given node, they are populated in
node attributes \code{"pie.border"} in \code{list} form. For these nodes,
the attribute \code{"frame.color"} is set to \code{NULL} so it is not displayed
on top of the pie wedge colors.

When there are no matching \code{rownames(hitim)}, or all colors
are identical, the \code{"pie.border"} is populated with \code{NULL},
so the pie wedges do not each show a color. Also for single-wedge
pie nodes, this process avoids drawing a small line at the top
of each node.
Instead the attribute \code{"frame.color"} is populated with the
one unique border color, so that only the outer border is colorized.
}
\seealso{
Other jam cnet igraph functions: 
\code{\link{adjust_cnet_nodeset}()},
\code{\link{adjust_cnet_set_relayout_gene}()},
\code{\link{apply_nodeset_spacing}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{make_cnet_test}()},
\code{\link{nudge_igraph_node}()},
\code{\link{rotate_igraph_layout}()}
}
\concept{jam cnet igraph functions}
