% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-apply-cnet-direction.R
\name{apply_cnet_direction}
\alias{apply_cnet_direction}
\title{Apply Cnet border color by directionality}
\usage{
apply_cnet_direction(
  cnet,
  hitim,
  col = colorjam::col_div_xf(1.2),
  hide_solo_pie = TRUE,
  frame_blank = "#FFFFFF00",
  do_reorder = FALSE,
  ...
)
}
\arguments{
\item{cnet}{\code{igraph} object with node attribute \code{"pie.color"}
populated as a \code{list} of \code{character} vectors, named by
enrichment. The enrichment names should match \code{colnames(hitim)}.}

\item{hitim}{\code{numeric} matrix with values centered at zero for
no change, positive values \code{+1} for up-regulation, and
negative values \code{-1} for down-regulation. The values are
converted to color using color function \code{col}.}

\item{col}{\code{function} that takes \code{numeric} input and assigns
a color. The default assigns red for positive values, blue
for negative values, and white for zero, using
\code{colorjam::col_div_xf(1.5)}.}

\item{hide_solo_pie}{\code{logical} indicating whether a single-color
border should only be applied to \code{"frame.color"} and not to
individual \code{"pie.border"} entries. When \code{hide_solo_pie=FALSE}
the border color is applied to each pie wedge, and if all
pie wedges are identical, it is also applied as \code{"frame.color"}.}

\item{frame_blank}{\code{character} string to define the color used
for \code{"frame.color"} when there are multiple colors in \code{"pie.border"}
and therefore there should be no visible \code{"frame.color"}.}

\item{...}{additional arguments are ignored.}
}
\description{
Apply Cnet border color by directionality
}
\details{
This function specifically requires the \code{"pie.color"}
node attribute of the \code{cnet} object \code{igraph} is populated
with a \code{list} of colors whose name is the enrichment,
and matches \code{colnames(hitim)} of the supplied hit
incidence matrix.

The node names \code{V(cnet)$name} also must match \code{rownames(hitim)},
otherwise the \code{"pie.border"} and \code{"coloredrect.border"} is assigned
its previous value.
In that case \code{"frame.color"} retains its original value.

This function may be run multiple times with different \code{hitim},
for example using \code{geneIMdirection} and \code{enrichIMdirection} in two steps.

When there are multiple colors for a given node, they are populated in
node attributes \code{"pie.border"} in \code{list} form. For these nodes,
the attribute \code{"frame.color"} is set to \code{NULL} so it is not displayed
on top of the pie wedge colors.

When there are no matching \code{rownames(hitim)}, or all colors
are identical, the \code{"pie.border"} is populated with \code{NULL},
so the pie wedges do not each show a color. Also for single-wedge
pie nodes, this process avoids drawing a small line at the top
of each node.
Instead the attribute \code{"frame.color"} is populated with the
one unique border color, so that only the outer border is colorized.
}
