% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-cnet.R
\name{apply_nodeset_spacing}
\alias{apply_nodeset_spacing}
\title{Apply minimum node spacing for each Cnet node set}
\usage{
apply_nodeset_spacing(cnet, percent_spacing = 3, apply_negative = FALSE, ...)
}
\arguments{
\item{cnet}{\code{igraph} object that contains attribute
\code{"nodeType"} and values \code{nodeType=c("Gene", "Set")},
consistent with use of \code{get_cnet_nodesets()} and
\code{adjust_cnet_nodeset()}.}
}
\value{
\code{igraph} Cnet object, whose nodes are adjusted
according to the arguments given.
}
\description{
Apply minimum node spacing for each Cnet node set
}
\details{
This function automates the process of spacing nodesets
in a Cnet plot by calling \code{adjust_cnet_nodeset()} on
all nodesets whose median nearest node distance is
above or below a threshold.

By default nodes must be at least \code{percent_spacing}
distance apart, relative to the maximum x-axis and
y-axis layout coordinate range. Roughly speaking,
nodes are clearly visible with distance between nodes
of at least 3 percent of the coordinate range.

Nodes are not compressed down to this distance unless
using the argument \code{apply_negative=TRUE}.

Note that this function does not detect when two
clusters overlap. For other specific layout adjustments
use \code{adjust_cnet_nodeset()}. In future, this function
may attempt to re-position nodesets relative to
each other, but that process tends to be iterative
and is not guaranteed to have a solution.
}
\seealso{
Other jam cnet igraph functions: 
\code{\link{adjust_cnet_nodeset}()},
\code{\link{adjust_cnet_set_relayout_gene}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{nudge_igraph_node}()},
\code{\link{rotate_igraph_layout}()}
}
\concept{jam cnet igraph functions}
