% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{avg_angles}
\alias{avg_angles}
\title{Average geometric angles}
\usage{
avg_angles(x, w = 1, maxAngle = 360, na.rm = TRUE, ...)
}
\description{
Average geometric angles
}
\examples{
x <- c(90, 180);
avg_angles(x);

}
\seealso{
Other jam utility functions: 
\code{\link{avg_colors_by_list}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{isColorBlank}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
