% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{avg_angles}
\alias{avg_angles}
\title{Average geometric angles}
\usage{
avg_angles(x, w = 1, maxAngle = 360, na.rm = TRUE, ...)
}
\description{
Average geometric angles
}
\examples{
x <- c(90, 180);
avg_angles(x);

}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{deconcat_df2}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{sort_colors_deprecated}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
