% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-bulk-cnet-adjustments.R
\name{bulk_cnet_adjustments}
\alias{bulk_cnet_adjustments}
\title{Bulk Cnet plot adjustments}
\usage{
bulk_cnet_adjustments(
  g,
  nodeset_adj,
  node_adj,
  apply_negative = TRUE,
  nodesets = NULL,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object expected to be a Cnet object with vertex attribute
'nodeType' containing values 'Gene' and 'Set'.}

\item{nodeset_adj}{\code{data.frame} recognizing five colnames:
\itemize{
\item nodeset: character name of each nodeset
\item x,y: numeric x,y coordinate adjustments to add to existing coordinates
\item rotate_degrees: numeric angle in degrees to rotate existing nodes
\item percent_spacing: absolute spacing to enforce for nodes in the nodeset.
}}

\item{node_adj}{\code{data.frame} recognizing five colnames:
\itemize{
\item node: \code{character} name of each node
\item x,y: \code{numeric} x,y coordinate adjustments to add to existing coordinates
\item label.dist: \code{numeric} value to add to existing label distance.
\item label.degree: \code{numeric} angle in degrees to add to the existing angle.
}}

\item{apply_negative}{\code{logical} whether to allow shrinking nodeset spacing.}

\item{nodesets}{\code{list} as output from \code{get_cnet_nodesets()}, or any
list named by nodeset, containing \code{character} vector of node names.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{igraph} object after applying adjustments.
}
\description{
Bulk Cnet plot adjustments
}
\details{
This function applies a series of adjustments to nodesets and nodes in
a Cnet \code{igraph} object. Nodesets are adjusted first, then nodes
are adjusted afterward. This order is important because the x,y
coordinates may be altered by nodeset first, before the node adjustment
is applied.

All x,y coordinate adjustments are performed using scaled units,
using the max x- or y-range for the plot layout.
All Cnet \code{igraph} plots are assumed to use \code{asp=1} to enforce
aspect ratio 1, in order to maintain the intended visual distances
used by the layout algorithms.

Therefore it is possible to expand a plot layout such that the
overall scale is wider than before, thereby affecting the scaled unit
space.

In any event, this function is intended to replicated the effect done
by \code{launch_shinycat()}.
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
