% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_Mem}
\alias{check_Mem}
\title{Check Mem object}
\usage{
check_Mem(object)
}
\arguments{
\item{object}{\code{Mem} object}
}
\description{
Check whether a Mem object is valid.
}
\details{
It requires:
\itemize{
\item \code{slotName(object)}: \verb{"memIM", "enrichIM", "geneIM", "enrichList"}
\item \code{object@memIM} to inherit \code{"matrix"}
\item \code{object@enrichIM} to inherit \code{"matrix"}
\item \code{object@geneIM} to inherit \code{"matrix"}
\item \code{colnames(object@memIM)} to equal \code{rownames(object@enrichIM)}
\item \code{rownames(object@memIM)} to equal \code{colnames(object@geneIM)}
\item \code{names(enrichList)} to equal \code{colnames(object@enrichIM)}
}

General guidance for Mem objects:
\itemize{
\item The aim is to support all user-facing tasks using S4 methods,
to prevent the need to access slots directly.
}
}
\seealso{
Other Mem:
\code{\link{Mem-class}},
\code{\link{Mem-data}},
\code{\link{Mem-slots}}
}
\concept{Mem}
