% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{check_MemPlotFolio}
\alias{check_MemPlotFolio}
\title{Check MemPlotFolio object}
\usage{
check_MemPlotFolio(object)
}
\arguments{
\item{object}{\code{MemPlotFolio} object}
}
\value{
\code{logical} whether the object is valid.
}
\description{
Check whether a MemPlotFolio object is valid.
}
\details{
It requires:
\itemize{
\item slots: 'enrichment_hm', 'gp_hm', 'clusters_mem',
'cnet_clusters', 'cnet_exemplars', 'cnet_clusters'
}

General guidance for MemPlotFolio objects:
\itemize{
\item The aim is to collect all parameters and associated data
together so that the various types of plots are all using
consistent underlying data and settings.
}
}
\seealso{
Other MemPlotFolio:
\code{\link{MemPlotFolio-class}},
\code{\link[=score_gene_path_clusters]{score_gene_path_clusters()}}
}
\concept{MemPlotFolio}
