% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{cnet2df}
\alias{cnet2df}
\title{Summarize Cnet igraph as a data.frame}
\usage{
cnet2df(g, getNeighbors = TRUE, checkSubsets = getNeighbors, ...)
}
\arguments{
\item{g}{igraph object containing Cnet data, specifically vertex
attribute name "nodeType" with values "Set" and "Gene", and
where "Set" nodes are only connected to "Gene" nodes.}

\item{getNeighbors}{logical indicating whether to include
the connected neighbor node names.}

\item{checkSubsets}{logical indicating whether to test "Set"
nodes to determine if the neighbors are all represented by
another "Set" node.}

\item{...}{additional arguments are ignored.}
}
\value{
data.frame with the node name, label, degree (number of
edges), membership (based upon connected component), and
if \code{getNeighbors=TRUE} it includes comma-delimited names
of neighboring nodes.
}
\description{
Summarize Cnet igraph as a data.frame
}
\details{
This function provides a data.frame summary of an igraph object
containing "Cnet" data, including vertex attribute \code{"nodeType"}
with values \code{"Set"} and \code{"Gene"}, and where \code{"Set"} nodes are
only connected to \code{"Gene"} nodes.

The data.frame is intended to provide a convenient method for
subsetting nodes, typically based upon a connected cluster,
or the minimum number of edges per node. For example, filter
for the connected component containing a node of interest, or
filter for \code{"Set"} nodes with more than one \code{"Gene"}.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}

Other jam conversion functions: 
\code{\link{cnet2im}()},
\code{\link{enrichDF2enrichResult}()},
\code{\link{enrichList2IM}()},
\code{\link{enrichList2df}()},
\code{\link{enrichMapJam}()}
}
\concept{jam conversion functions}
\concept{jam igraph functions}
