% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{cnet2im}
\alias{cnet2im}
\title{Convert Cnet igraph to incidence matrix}
\usage{
cnet2im(g = NULL, df = NULL, ...)
}
\arguments{
\item{g}{igraph object containing Cnet data, specifically vertex
attribute name "nodeType" with values "Set" and "Gene", and
where "Set" nodes are only connected to "Gene" nodes.}

\item{df}{data.frame as optional input instead of \code{g}, usually
the result of a previous call to \code{cnet2df()}.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric matrix with colnames defined by \code{"Set"} node
names, and rownames defined by \code{"Gene"} node names.
}
\description{
Convert Cnet igraph to incidence matrix
}
\details{
This function takes igraph object containing "Cnet" data,
including vertex attribute \code{"nodeType"} with values \code{"Set"}
and \code{"Gene"}, and where \code{"Set"} nodes are only connected
to \code{"Gene"} nodes. It returns an incidence matrix whose
columns are "Set" node names and whose rows are "Gene"
node names.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}

Other jam conversion functions: 
\code{\link{cnet2df}()},
\code{\link{enrichDF2enrichResult}()},
\code{\link{enrichList2IM}()},
\code{\link{enrichList2df}()},
\code{\link{enrichMapJam}()}
}
\concept{jam conversion functions}
\concept{jam igraph functions}
