% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{cnetplotJam}
\alias{cnetplotJam}
\title{Create a cnetplot igraph object}
\usage{
cnetplotJam(
  x,
  showCategory = 5,
  categorySize = "geneNum",
  nodeLabel = c("Name", "Description", "ID"),
  foldChange = NULL,
  fixed = TRUE,
  doPlot = FALSE,
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  normalizeGeneSize = TRUE,
  labelCex = 0.45,
  colorSub = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{enrichResults object}

\item{showCategory}{integer number of categories to include in the
resulting Cnet plot.}

\item{categorySize}{character value indicating how to size the pathway
nodes, where \code{"geneNum"} sizes nodes by the number of genes in that
pathway, and \code{"pvalue"} sizes nodes by the enrichment P-value using
the format \code{-log10(pvalue)}.}

\item{nodeLabel}{character value indicating which colname in
\code{as.data.frame(x)} to use as a node label. Depending upon the source
of data, there may be alternative colnames that are more suitable
as node labels.}

\item{foldChange}{numeric vector named by gene, or NULL. When supplied,
the vector names must use the same nomenclature as the \code{x} input
object, which can be inspected with \code{print(head(x@gene))}.}

\item{fixed}{optional argument passed to \code{netplot}.}

\item{doPlot}{logical indicating whether to plot the result.}

\item{categoryColor, geneColor}{character color, used to colorize
category (pathway) nodes, or gene nodes, respectively.}

\item{normalizeGeneSize}{logical indicating whether to re-scale the
gene node sizes so the mean gene node size is no larger than the
median category node size. This option is intended to help gene
and category node sizes to be in relative proportion.}

\item{labelCex}{numeric value to re-scale label text in the igraph
object, and is applied directly to the igraph object.}

\item{colorSub}{character vector of valid R colors, whose names
are compared to node names, for example \code{V(g)$name \%in\% names(colorSub)}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Create a cnetplot igraph object
}
\details{
The purpose of this function is to mimic the steps in \code{DOSE:::cnetplot()}
except not plot the output, and provide some customizations.

This function calls \code{cnetplot_internalJam()}, which among other things
adds a node attribute to the resulting \code{igraph}, \code{"nodeType"},
where \code{nodeType="Gene"} identifies gene nodes, and \code{nodeType="Set"}
identifies pathway/gene set nodes.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
