% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{collapse_mem_clusters}
\alias{collapse_mem_clusters}
\title{Collapse Multienrichment clusters}
\usage{
collapse_mem_clusters(
  mem,
  clusters,
  choose = NULL,
  per_cluster = Inf,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  return_type = c("cnet", "mem"),
  max_labels = 4,
  max_nchar_labels = 25,
  include_cluster_title = TRUE,
  verbose = FALSE,
  ...
)
}
\value{
By default \code{return_type="cnet"} and this function returns
an augmented Cnet plot. The labels of each cluster are defined
by the input \code{names(clusters)}, however an \code{igraph} attribute
\code{"set_labels"} includes an abbreviated label of the top ranked
sets for each cluster. This label is probably the closest thing
to summarizing the composition of each cluster.
}
\description{
Collapse Multienrichment clusters
}
\details{
This function is similar to \code{rank_mem_clusters()} in that it
starts with \code{mem} results from \code{multiEnrichMap()} and a list
of \code{clusters} of pathways/sets. Instead of ranking and choosing
exemplars from each clusters, it simply collapses each cluster
into one super-set that contains union of all genes.

It does also run \code{rank_mem_clusters()} in the event one would
want to collapse only the top \code{per_cluster} entries for each
cluster, but the default is to include them all.
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
