% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{collapse_mem_clusters}
\alias{collapse_mem_clusters}
\title{Collapse Multienrichment clusters}
\usage{
collapse_mem_clusters(
  mem,
  clusters,
  choose = NULL,
  per_cluster = Inf,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  return_type = c("cnet", "Mem", "mem"),
  max_labels = 4,
  max_nchar_labels = 25,
  include_cluster_title = TRUE,
  cluster_color_min_fraction = 0.5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{Mem} or legacy \code{list} mem object}

\item{clusters}{\code{list} named by cluster name, containing \code{character}
vectors all of whose values should be present in \code{sets(mem)}.}

\item{choose}{\code{character} default NULL, passed to \code{rank_mem_clusters()}
to choose a subset of clusters to return. Values must be found in
\code{names(clusters)}.}

\item{per_cluster}{\code{integer} default 'Inf' with the maximum sets
to include in each cluster. Use an integer like \code{per_cluster=1}
to choose one "exemplar" per cluster, ranked according
to \code{byCols} which is described in \code{rank_mem_clusters()}.}

\item{byCols}{\code{character} vector of summary values to use when
sorting sets within each cluster, passed to and described
in \code{rank_mem_clusters()}.}

\item{return_type}{\code{character} default "cnet" with the output format:
\itemize{
\item "cnet" returns an \code{igraph} Cnet plot.
\item "Mem" returns a \code{Mem} object after collapsing by cluster.
\item "mem" returns legacy \code{list} mem output.
}}

\item{max_labels}{\code{integer} default 4, the number of pathway labels
to include in each cluster, ordered according to 'byCols'.}

\item{max_nchar_labels}{\code{integer} default 25, max character length
for each pathway label.}

\item{include_cluster_title}{\code{logical} default TRUE, whether to include
the cluster title as a prefix.
For example if the cluster name is 'A' the new title would be:
'A: pathway 1; pathway 2; pathway 3; pathway 4'.}

\item{cluster_color_min_fraction}{\code{numeric} value default 0.5,
the fraction of pathways in a cluster which has "non blank"
assigned color in \code{enrichIMcolors()} in order for the color
to be assigned to the resulting node, in the case of Cnet
output with \code{return_type="cnet"}.}

\item{verbose}{\code{logical} whether to print verbose output.}

\item{...}{additional arguments are passed to \code{cnet2mem()} which
includes thresholds used by \code{apply_cnet_direction()} such as
\code{direction_col_fn} (color function applied to enrichIMdirection),
\code{gene_direction_col_fn} (color function applied to geneIMdirection),
and others.
To avoid applying directional color to Set nodes (nodeType='Set')
use \code{direction_cutoff=100} and \code{direction_max=101}, which will
consider any enrichIMdirection value at or below 100 to be
non-directional.}
}
\value{
By default \code{return_type="cnet"} and this function returns
an augmented Cnet plot. The labels of each cluster are defined
by the input \code{names(clusters)}, however an \code{igraph} attribute
\code{"set_labels"} includes an abbreviated label of the top ranked
sets for each cluster. This label is probably the closest thing
to summarizing the composition of each cluster.
}
\description{
Collapse Multienrichment clusters
}
\details{
This function is similar to \code{rank_mem_clusters()} in that it
starts with \code{mem} results from \code{multiEnrichMap()} and a list
of \code{clusters} of pathways/sets. Instead of ranking and choosing
exemplars from each clusters, it simply collapses each cluster
into one super-set that contains union of all genes.

It does also run \code{rank_mem_clusters()} in the event one would
want to collapse only the top \code{per_cluster} entries for each
cluster, but the default is to include them all.
}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{colorRamp2D}()},
\code{\link{curateIPAcolnames}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
