% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-color-nodes-edges.R
\name{color_edges_by_nodegroups}
\alias{color_edges_by_nodegroups}
\title{Color edges by nodegroups}
\usage{
color_edges_by_nodegroups(
  g,
  nodegroups,
  nodegroup_colors = NULL,
  edge_alpha = NULL,
  Crange = c(60, 100),
  Lrange = c(45, 85),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object that contains vertex node attribute \code{"color"}
as seen with \code{igraph::vertex_attr(g, "color")}.}

\item{nodegroups}{\code{list} or \code{communities} object that references
nodes in \code{g} and assigns one or more to nodegroups.}

\item{edge_alpha}{\code{numeric} or \code{NULL}, where numeric value sets
the edge alpha transparency, where \code{edge_alpha=0} is completely
transparent, \code{edge_alpha=0.5} is 50\% transparent, and \code{edge_alpha=1}
is completely not transparent, and is opaque. When \code{edge_alpha=NULL}
the alpha values are supplied by \code{colorjam::blend_colors()}
which blends the two values.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{colorjam::blend_colors()}.}
}
\value{
\code{igraph} object with edge color attribute updated to
represent the result of blending node colors, seen by
\code{igraph::edge_attr(g)$color}.
}
\description{
Color edges by nodegroups
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{label_communities}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{mem2emap}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
