% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-color-nodes-edges.R
\name{color_nodes_by_nodegroups}
\alias{color_nodes_by_nodegroups}
\title{Color edges by nodegroups}
\usage{
color_nodes_by_nodegroups(
  g,
  nodegroups,
  nodegroup_colors = NULL,
  node_alpha = NULL,
  Crange = c(60, 100),
  Lrange = c(45, 85),
  color_attributes = c("color"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object that contains vertex node attribute \code{"color"}
as seen with \code{igraph::vertex_attr(g, "color")}.}

\item{nodegroups}{\code{list} or \code{communities} object that references
nodes in \code{g} and assigns one or more to nodegroups.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{colorjam::blend_colors()}.}

\item{edge_alpha}{\code{numeric} or \code{NULL}, where numeric value sets
the edge alpha transparency, where \code{edge_alpha=0} is completely
transparent, \code{edge_alpha=0.5} is 50\% transparent, and \code{edge_alpha=1}
is completely not transparent, and is opaque. When \code{edge_alpha=NULL}
the alpha values are supplied by \code{colorjam::blend_colors()}
which blends the two values.}
}
\value{
\code{igraph} object with edge color attribute updated to
represent the result of blending node colors, seen by
\code{igraph::edge_attr(g)$color}.
}
\description{
Color edges by nodegroups
}
\seealso{
Other jam igraph utilities: 
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{jam_igraph}()},
\code{\link{launch_shinycat}()},
\code{\link{subgraph_jam}()}
}
\concept{jam igraph utilities}
