% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{communities2nodegroups}
\alias{communities2nodegroups}
\title{Convert communities object to nodegroups list format}
\usage{
communities2nodegroups(wc, sep = ",", ...)
}
\arguments{
\item{wc}{\code{communities} object as returned by \code{igraph} functions
such as \code{cluster_optimal()}, \code{cluster_walktrap()}, or
\code{cluster_leading_eigen()}.
\itemize{
\item Alternatively, a \code{list} object with 'membership' and 'names'.
\item Alternatively, a \code{list} object which mimics the output of this
function, in which case the input \code{wc} is returned as-is.
}}

\item{sep}{\code{character} string, default ',' (comma), used when
cluster_names are defined as a \code{list} with vector of names
for each cluster. Passed to \code{jamba::cPaste()}.}

\item{...}{additional arguments are passed to \code{jamba::cPaste()} when
relevant.}
}
\value{
\code{list} of \code{character} vectors, where each vector contains
names of \code{igraph} nodes. When \code{algorithm} is defined in the
input object, it is included as an attribute of the output \code{list},
accessible with \code{attr(out, "algorithm")}.

When optional value \code{"cluster_names"} is present in the \code{communities}
object, they are used to define the output \code{list} names.
}
\description{
Convert communities object to nodegroups list format
}
\details{
Note that this function is "lossy", in that the output \code{list}
does not contain all the information necessary to reconstitute
the input \code{communities} object in detail. However, the output
\code{list} can be converted to a \code{communities} object that will
be accepted by most \code{igraph} related functions that require
that object type as an input value.

Alternatively, this function can be used to confirm the input
is already appropriate nodegroups \code{list} format, by supplying
\code{list} format upfront.
}
\seealso{
Other jam igraph functions:
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=igraph2pieGraph]{igraph2pieGraph()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
