% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{communities2nodegroups}
\alias{communities2nodegroups}
\title{Convert communities object to nodegroups list format}
\usage{
communities2nodegroups(wc, ...)
}
\arguments{
\item{wc}{\code{communities} object as returned by \code{igraph} functions
such as \code{cluster_optimal()}, \code{cluster_walktrap()}, or
\code{cluster_leading_eigen()}.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of \code{character} vectors, where each vector contains
names of \code{igraph} nodes. When \code{algorithm} is defined in the
input object, it is included as an attribute of the output \code{list},
accessible with \code{attr(out, "algorithm")}.

When optional value \code{"cluster_names"} is present in the \code{communities}
object, they are used to define the output \code{list} names.
}
\description{
Convert communities object to nodegroups list format
}
\details{
Note that this function is "lossy", in that the output \code{list}
does not contain all the information necessary to reconstitute
the input \code{communities} object in detail. However, the output
\code{list} can be converted to a \code{communities} object that will
be accepted by most \code{igraph} related functions that require
that object type as an input value.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{label_communities}()},
\code{\link{layout_with_qfr}()},
\code{\link{layout_with_qfrf}()},
\code{\link{mem2emap}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
