% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{curateIPAcolnames}
\alias{curateIPAcolnames}
\title{Curate Ingenuity IPA colnames}
\usage{
curateIPAcolnames(
  jDF,
  ipaNameGrep = c("^Name$", "^ID$", "Canonical Pathways", "Upstream Regulator",
    "Diseases or Functions Annotation", "Diseases . Functions", "My Lists",
    "Ingenuity Toxicity Lists", "My Pathways"),
  geneGrep = c("Molecules in Network", "Target molecules", "Molecules", "Symbol"),
  geneCurateFrom = c(" [(](complex|includes others)[)]", "^[,]+|[,]+$"),
  geneCurateTo = c("", ""),
  convert_ipa_slash = TRUE,
  ipa_slash_sep = ":",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{jDF}{data.frame from one Ingenuity IPA enrichment test.}

\item{ipaNameGrep}{vector of regular expression patterns
used to recognize the name of the enriched entity,
for example the biological pathway, or network, or
disease category, etc.}

\item{geneGrep}{regular expression pattern used to recognize
the column containing genes, or the molecules tested for
enrichment which were found in the enriched entity.}

\item{geneCurateFrom, geneCurateTo}{vector of patterns and
replacements, respectively, used to curate values in
the gene column. These replacement rules are used to
ensure that genes are delimited consistently, with no
leading or trailing delimiters.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Curate Ingenuity IPA colnames
}
\details{
This function is intended to help curate colnames observed
in Ingenuity IPA enrichment data. The IPA enrichment data
includes multiple types of enrichment tests, each with slightly
different column headers. This function is intended to
make the colnames more consistent.

This function will rename the first recognized gene colname
to \code{"geneNames"} for consistency with downstream analyses.

The values in the recognized gene colname are curated
using \verb{geneCurateFrom,geneCurateTo} for multiple
pattern-replacement substitutions. This mechanism is
used to ensure consistent delimiters and values used
for each enrichment table.

Any colname matching \code{"-log.*p.value"} is considered
-log10 P-value, and is converted to normal P-values
for consistency with downstream analyses.

Any recognized P-value column is renamed to \code{"P-value"}
for consistency with downstream analyses.

When the recognized P-value column contains a range,
for example \code{"0.00017-0.0023"}, the lower P-value is
chosen. In that case, the higher P-value is stored in
a new column \code{"max P-value"}. P-value ranges
are reported in the disease category analysis by
Ingenuity IPA, after collating individual pathways
by disease category and storing the range of enrichment
P-values.
}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
