% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-deconcat.R
\name{deconcat_df2}
\alias{deconcat_df2}
\title{Deconcatenate delimited column values in a data.frame}
\usage{
deconcat_df2(x, column, split = "[,; |/]+", blank = "", ...)
}
\arguments{
\item{x}{\code{data.frame} or compatible object}

\item{column}{\code{character} vector with one or more \code{colnames(x)}
that should be de-concatenated.}

\item{split}{\code{character} pattern used by \code{strsplit()} to split
multiple values in each column.}

\item{blank}{\code{character} string used to replace entries that
would otherwise be zero-length as returned by \code{strsplit()}.}

\item{...}{additional arguments are ignored.}
}
\description{
Deconcatenate delimited column values in a data.frame
}
\details{
This function deconcatenates delimited values in a column
of a \code{data.frame} by calling \code{strsplit()} on column values,
and repeating values in all other columns to match \code{lengths()}
following \code{strsplit()}.

This function includes a correction for cases where \code{strsplit()}
would otherwise return zero-length entries, and which would
otherwise be dropped from the output. From this function,
zero-length entries are replaced with \code{blank=""} so these
rows are not dropped from the output.
}
\examples{
df <- data.frame(one=c("AB", "BC", "AC"),
   two=c("a,b", "b,c", "a,c"));
deconcat_df2(df, column="two")

}
