% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-internals.R
\name{default_igraph_values}
\alias{default_igraph_values}
\title{Default igraph parameter values}
\usage{
default_igraph_values()
}
\value{
\code{list} of default igraph plotting and data parameters,
including \code{"plot"}, \code{"vertex"}, and \code{"edge"}.
}
\description{
Default igraph parameter values
}
\seealso{
Other jam igraph internal functions: 
\code{\link{get_igraph_arrow_mode}()},
\code{\link{jam_igraph_arrows}()},
\code{\link{parse_igraph_plot_params}()}
}
\concept{jam igraph internal functions}
