% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-bivariatecolor.R
\name{display_colorRamp2D}
\alias{display_colorRamp2D}
\title{Display colors from bivariate color function}
\usage{
display_colorRamp2D(
  col_fun,
  pretty.n = 12,
  xlab = "",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{col_fun}{\code{function} output from \code{colorRamp2D()}.}

\item{pretty.n}{\code{numeric} passed to \code{pretty()} to determine the
y-axis and x-axis label breaks, respectively, or
\code{NULL} to use breaks encoded in \code{col_fun}.}

\item{xlab, ylab, main}{\code{character} strings passed to \code{title()}.}

\item{...}{additional arguments are passed to \code{title()}.}
}
\description{
Display colors from bivariate color function
}
\details{
This function provides a very rudimentary method for displaying
a bivariate 2-D color matrix out from \code{colorRamp2D()}.
}
