% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamgraph-edgebundle.R
\name{edge_bundle_bipartite}
\alias{edge_bundle_bipartite}
\title{Bundle edges in a bipartite graph}
\usage{
edge_bundle_bipartite(g, type = "nodeType", ...)
}
\description{
Bundle edges in a bipartite graph
}
