% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-enrichdf2er.R
\name{enrichDF2enrichResult}
\alias{enrichDF2enrichResult}
\title{Convert data.frame to enrichResult}
\usage{
enrichDF2enrichResult(
  enrichDF = NULL,
  pvalueCutoff = 1,
  pAdjustMethod = "none",
  keyColname = c("itemsetID", "ID", "Name", "Pathway"),
  pathGenes = "pathGenes",
  geneColname = c("geneNames", "geneID", "Gene", "Genes"),
  geneHits = "geneHits",
  geneRatioColname = c("GeneRatio", "^Ratio"),
  geneDelim = "[,/ ]+",
  geneSep = ",",
  pvalueColname = c("P.Value", "Pvalue", "FDR", "adj.P.Val"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  msigdbGmtT = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichDF}{\code{data.frame} representing gene set enrichment
results.}

\item{pvalueCutoff}{\code{numeric} value range 0 to 1, to define the
P-value threshold for enrichment results to be considered
in downstream processing.}

\item{pAdjustMethod}{\code{character} string to define the P-value
adjustment method, or \code{"none"} for no additional adjustment.
See \code{stats::p.adjust()} for valid values.}

\item{keyColname}{\code{character} value of the \code{colname(enrichDF)}
containing the unique row identifier. It can be a pathway_ID
or any uniquely identifying value.}

\item{pathGenes}{\code{character} or value of the \code{colname(enrichDF)}
containing the number of genes in each pathway. This value will be
derived from \code{geneRatioColname} if needed.}

\item{geneColname}{\code{character} value of the \code{colname(enrichDF)}
containing delimiited genes in each pathway.}

\item{geneHits}{\code{character} value of the \code{colname(enrichDF)}
containing the integer count of the gene hits in each pathway.
This value will be derived from \code{geneRatioColname} if needed.}

\item{geneRatioColname}{\code{character} value of the \code{colname(enrichDF)}
containing the character ratio of gene hits to pathway size,
in format "50/100". This value is used when either \code{"pathGenes"}
or \code{"geneHits"} are not supplied.}

\item{geneDelim}{\code{character} regular expression pattern used to separate
genes in the \code{pathGenes} column into a vector of character
values.}

\item{pvalueColname}{\code{character} value of the \code{colname(enrichDF)}
containing enrichment P-values to use in downstream processing.}

\item{msigdbGmtT}{optional gmtT object, a representation of
\code{arules::transactions-class}. (Not currently implemented.)}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Convert data.frame to enrichResult
}
\details{
This function takes a data.frame containing gene set enrichment
results, and converts it to a proper \code{enrichResult} object,
compatible with methods provided by the \code{clusterProfiler}
package.
}
\seealso{
Other jam conversion functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{enrichList2IM}()},
\code{\link{enrichList2df}()},
\code{\link{enrichMapJam}()}
}
\concept{jam conversion functions}
