% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-enrichlist2df.R
\name{enrichList2df}
\alias{enrichList2df}
\title{Convert enrichList to data.frame}
\usage{
enrichList2df(
  enrichList,
  keyColname = c("ID", "Name", "pathway", "itemsetID", "Description"),
  geneColname = c("geneID", "geneNames", "Genes"),
  geneCountColname = c("gene_count", "Count", "geneCount", "geneHits"),
  pvalueColname = c("padjust", "p.adjust", "adjp", "padj", "qvalue", "qval", "q.value",
    "pvalue", "p.value", "pval", "FDR"),
  pvalueFloor = 1e-200,
  msigdbGmtT = NULL,
  geneDelim = "[,/ ]+",
  geneSep = ",",
  verbose = FALSE,
  debug = 0,
  ...
)
}
\arguments{
\item{enrichList}{\code{list} of \code{enrichResult} or \code{data.frame} objects.}

\item{keyColname, geneColname, geneCountColname, pvalueColname}{\code{character} string indicating each column to use, or a vector
of column names in order of priority.}

\item{pvalueFloor}{\code{numeric} lowest permitted P-value, intended to
prevent using actual zero \code{0} which may occur sometimes when
upstream processing rounds the value.}

\item{msigdbGmtT}{not currently implemented.}

\item{geneDelim}{\code{character} pattern used to split genes.}

\item{geneSep}{\code{character} string used to join genes, default \code{","}
uses comma-delimited values.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{debug}{\code{integer} used for debugging, with values \verb{0, 1, 2, 3}
recognized. Value \code{0}, default, performs no debug.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} with combined results across all \code{enrichList}.
}
\description{
Convert enrichList to a single data.frame, merging results
}
\details{
This function takes one or more enrichment results, and
combines results into one \code{data.frame}.
\itemize{
\item Genes are combined, reporting the unique genes in sorted order.
\item \code{numeric} columns report the lowest or highest value, as appropriate:
\itemize{
\item P-value columns report the lowest observed value, in \code{pvalueColname}
\item Count columns report the highest observed value, except the
primary count column reports the number of unique genes.
}
}
}
\seealso{
Other jam conversion functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{enrichDF2enrichResult}()},
\code{\link{enrichList2IM}()}
}
\concept{jam conversion functions}
