% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{enrichList2df}
\alias{enrichList2df}
\title{Convert enrichList to data.frame}
\usage{
enrichList2df(enrichList, keyColname = "itemsetID",
  geneColname = "geneNames", geneCountColname = "geneCount",
  pvalueColname = "P-value", pvalueFloor = 1e-200, msigdbGmtT = NULL,
  verbose = FALSE, debug = 0, ...)
}
\description{
Convert enrichList to data.frame
}
