% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{enrichList2geneHitList}
\alias{enrichList2geneHitList}
\title{Extract gene hit list from list of enrichResult}
\usage{
enrichList2geneHitList(
  enrichList,
  geneColname,
  geneDelim,
  make_unique = TRUE,
  ...
)
}
\value{
\code{list} of character vectors, containing the unique
set of genes involved in each enrichment.
}
\description{
Extract gene hit list from list of enrichResult
}
\details{
This function is mainly for internal use in multienrichjam,
it takes a list of \code{enrichResult} objects, and determines
the full set of genes involved in each \code{enrichResult}.
}
