% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{enrichList2geneHitList}
\alias{enrichList2geneHitList}
\title{Extract gene hit list from list of enrichResult}
\usage{
enrichList2geneHitList(
  enrichList,
  geneColname,
  geneDelim,
  make_unique = TRUE,
  ...
)
}
\value{
\code{list} of character vectors, containing the unique
set of genes involved in each enrichment.
}
\description{
Extract gene hit list from list of enrichResult
}
\details{
This function is mainly for internal use in multienrichjam,
it takes a list of \code{enrichResult} objects, and determines
the full set of genes involved in each \code{enrichResult}.

This function also works with \code{ComplexHeatmap::HeatmapList}
objects.
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
