% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{enrichList2IM}
\alias{enrichList2IM}
\title{Convert enrichList to IM incidence matrix}
\usage{
enrichList2IM(enrichList, GmtT = NULL, addAnnotations = TRUE,
  keyColname = "Name", valueColname = "P-value", emptyValue = 1,
  verbose = FALSE, ...)
}
\description{
Convert enrichList to IM incidence matrix
}
\seealso{
Other jam conversion functions: \code{\link{enrichDF2enrichResult}},
  \code{\link{multiEnrichMap}}
}
\concept{jam conversion functions}
