% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{find_colname}
\alias{find_colname}
\title{Find colname by character string or pattern matching}
\usage{
find_colname(
  pattern,
  x,
  max = 1,
  index = FALSE,
  require_non_na = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pattern}{character vector containing text strings or
regular expression patterns.}

\item{x}{input \code{data.frame} or other R object that
contains colnames.}

\item{max}{integer maximum number of results to return.}

\item{index}{logical indicating whether to return the column
index as an integer vector. When \code{index=FALSE} it returns
the matching \code{colnames(x)}; when \code{index=TRUE} it returns
the matching column numbers as an integer vector.}

\item{require_non_na}{logical indicating whether to require the
column to contain non-NA values, default is TRUE. The intent
of this function is to find colnames whose data will match
expectations, and when require_non_na is TRUE, this
function will continue until it finds a column with non-NA
values.}

\item{...}{additional arguments are passed to \code{jamba::provigrep()}.}
}
\value{
character vector with length \code{max}, or if no pattern
match is found it returns \code{NULL}. Also if there are no
\code{colnames(x)} then it returns \code{NULL}.
}
\description{
Find colname by character string or pattern matching
}
\details{
This function is intended to help find a colname given
a character vector of expected values or regular expression
patterns. By default it returns the first matching value,
but can return multiple if \code{max=Inf}.

If there are no \code{colnames(x)} then \code{NULL} is returned.

The order of operations:
\enumerate{
\item Match exact string.
\item Match exact string in case-insensitive manner.
\item Match the start of each string using \code{jamba::provigrep()}.
\item Match the end of each string using \code{jamba::provigrep()}.
\item Match each string using \code{jamba::provigrep()}.
}

The results from the first successful operation is returned.

When there are duplicate \code{colnames(x)} only the first
unique name is returned.
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
