% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-find-enrich-colnames.R
\name{find_enrich_colnames}
\alias{find_enrich_colnames}
\title{Find enrichment colnames}
\usage{
find_enrich_colnames(
  x,
  keyColname = c("ID", "Name", "pathway", "itemsetID", "Description"),
  nameColname = c("Name", "pathway", "Description", "itemsetID", "ID"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  geneColname = c("geneID", "geneNames", "Genes"),
  countColname = c("gene_count", "count", "geneHits"),
  geneRatioColname = c("GeneRatio", "^Ratio"),
  pvalueColname = c("padjust", "p.adjust", "adjp", "padj", "qvalue", "qval", "q.value",
    "pvalue", "p.value", "pval", "FDR"),
  directionColname = c("direction", "NES", "activation.z.{0,1}score", "z.{0,1}score"),
  pathGenes = c("setSize", "pathGenes", "Count"),
  geneHits = c("Count", "geneHits", "gene_count"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame}, \code{enrichList}, \code{Mem}, or \code{list} of \code{data.frame}
objects.}

\item{keyColname}{\code{character} default 'ID' indicating the primary
identifier for each set. This column may be a numeric identifier.}

\item{nameColname}{\code{character} default 'Name', with the set name,
typically a short name for each set.}

\item{descriptionColname}{\code{character} default 'Description' with the
longer set description. It will use \code{nameColname} or \code{keyColname}
as needed.}

\item{geneColname}{\code{character} default 'geneID' containing delimited
genes associated with each enrichment result.}

\item{countColname}{\code{character} default 'Count' with the number of
genes in the \code{geneColname} column. It will be calculated as needed.}

\item{geneRatioColname}{\code{character} default 'GeneRatio' with the
numeric (decimal) ratio of test genes to pathway genes, or
a character indication in the form '6/24' with 'tested/pathway'
gene counts.}

\item{pvalueColname}{\code{character} default 'padjust' with the best available
column to use for statistical significance of enrichment.}

\item{directionColname}{\code{character} default looks for matching entries
in order:
\itemize{
\item 'direction': some directional score with positive and negative
scores, reasonably scaled so that -2 and +2 are meaningful values,
but the exact values are not critical to this analysis.
\item 'NES': the GSEA normalized enrichment score),
\item 'activation z-score': provided by Ingenuity Pathway Analysis
\item 'z-score' or 'zScore': provided alternative by some tools. This
score may be removedin future, since clusterProfiler provides it as
a supplement to the hypergeometric enrichment, and it only
contains positive values, therefore it is not directional.
}}

\item{pathGenes}{\code{character} default 'setSize' indicating the number
of genes in each set as tested for enrichment. This number is not
always reported, however it is not used by 'multienrichjam', but
is required by some \code{clusterProfiler} functions.}

\item{geneHits}{\code{character} default 'Count' indicating the number of
genes in the \code{geneColname} column. It will be calculated as needed.}

\item{verbose}{\code{logical} default FALSE, whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} of recognized colnames named by the type
of column, using \code{NA} for any column types not found.

\code{list} named by each column name argument, with one \code{character}
value or \code{NULL} in each entry.
}
\description{
Find enrichment colnames
}
\examples{
newborn_txt <- system.file("extdata",
   "Newborns-IPA.txt",
   package="multienrichjam");
ipa_dfs <- importIPAenrichment(newborn_txt);
find_enrich_colnames(ipa_dfs[[1]])

er <- enrichDF2enrichResult(ipa_dfs[[1]])
find_enrich_colnames(er)

}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}},
\code{\link[=xyAngle]{xyAngle()}}
}
\concept{jam utility functions}
