% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-find-enrich-colnames.R
\name{find_enrich_colnames}
\alias{find_enrich_colnames}
\title{Find enrichment colnames}
\usage{
find_enrich_colnames(
  x,
  keyColname = c("ID", "Name", "pathway", "itemsetID", "Description"),
  nameColname = c("Name", "pathway", "Description", "itemsetID", "ID"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  geneColname = c("geneID", "geneNames", "Genes"),
  countColname = c("gene_count", "count", "geneHits"),
  geneRatioColname = c("GeneRatio", "^Ratio"),
  pvalueColname = c("padjust", "p.adjust", "adjp", "padj", "qvalue", "qval", "q.value",
    "pvalue", "p.value", "pval", "FDR"),
  directionColname = c("activation.z.{0,1}score", "z.{0,1}score"),
  pathGenes = c("setSize", "pathGenes", "Count"),
  geneHits = c("Count", "geneHits", "gene_count"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame}, \code{enrichList}, \code{Mem}, or \code{list} of \code{data.frame}
objects.}

\item{keyColname}{\code{character} default 'ID' indicating the primary
identifier for each set. This column may be a numeric identifier.}

\item{nameColname}{\code{character} default 'Name', with the set name,
typically a short name for each set.}

\item{descriptionColname}{\code{character} default 'Description' with the
longer set description. It will use \code{nameColname} or \code{keyColname}
as needed.}

\item{geneColname}{\code{character} default 'geneID' containing delimited
genes associated with each enrichment result.}

\item{countColname}{\code{character} default 'Count' with the number of
genes in the \code{geneColname} column. It will be calculated as needed.}

\item{geneRatioColname}{\code{character} default 'GeneRatio' with the
numeric (decimal) ratio of test genes to pathway genes, or
a character indication in the form '6/24' with 'tested/pathway'
gene counts.}

\item{pvalueColname}{\code{character} default 'padjust' with the best available
column to use for statistical significance of enrichment.}

\item{directionColname}{\code{character} default 'zscore' with a directional
score, typically a z-score or some other reasonably scaled
numeric value where the sign indicates directionality, with '+'
meaning activated and '-' meaning suppressed.}

\item{pathGenes}{\code{character} default 'setSize' indicating the number
of genes in each set as tested for enrichment. This number is not
always reported, however it is not used by 'multienrichjam', but
is required by some \code{clusterProfiler} functions.}

\item{geneHits}{\code{character} default 'Count' indicating the number of
genes in the \code{geneColname} column. It will be calculated as needed.}

\item{verbose}{\code{logical} default FALSE, whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} of recognized colnames named by the type
of column, using \code{NA} for any column types not found.

\code{list} named by each column name argument, with one \code{character}
value or \code{NULL} in each entry.
}
\description{
Find enrichment colnames
}
\examples{
newborn_txt <- system.file("extdata",
   "Newborns-IPA.txt",
   package="multienrichjam");
ipa_dfs <- importIPAenrichment(newborn_txt);
find_enrich_colnames(ipa_dfs[[1]])

er <- enrichDF2enrichResult(ipa_dfs[[1]])
find_enrich_colnames(er)

}
