% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-fix-set-labels.R,
%   R/multienrichjam-data.R
\docType{data}
\name{fixSetLabels}
\alias{fixSetLabels}
\alias{words}
\alias{abbrev}
\title{Fix Set or pathway labels for legibility}
\format{
\code{words} is a \code{data.frame} with colnames 'from' and 'to',
for pathway word Perl-compatible regular expression pattern and
replacement. It is used as default in \code{fixSetLabels(..., words_from)}.

\code{abbrev} is a \code{data.frame} with colnames 'from' and 'to',
for pathway word Perl-compatible regular expression pattern and
replacement. It is used as default in \code{fixSetLabels(..., abbrev_from)}.
}
\usage{
fixSetLabels(
  x,
  wrap = TRUE,
  width = 40,
  maxNchar = Inf,
  suffix = "...",
  nodeType = c("Set", "Gene", "any"),
  do_abbreviations = TRUE,
  adjustCase = TRUE,
  lowercaseAll = TRUE,
  removeGrep = "^(KEGG|PID|REACTOME|BIOCARTA|NABA|SA|SIG|ST|WP|HALLMARK)[_. ]",
  words_from = NULL,
  words_to = NULL,
  add_from = NULL,
  add_to = NULL,
  abbrev_from = NULL,
  abbrev_to = NULL,
  ...
)

words

abbrev
}
\arguments{
\item{x}{any of the following objects:
\itemize{
\item \code{character} vector
\item \code{igraph} object. The \code{igraph::V(g)$name} attribute is used as input,
and the resulting label is then stored as \code{V(g)$label}.
When \code{nodeType} is also defined, and nodes have attribute 'nodeType',
only nodes with that attribute value will be edited.
The default is \code{nodeType="Set"}.
\item \code{Mem} object. The \code{sets()} are adjusted by this function.
}}

\item{wrap}{\code{logical} indicating whether to apply word wrap, based upon
the supplied \code{width} argument.}

\item{width}{integer value used when \code{wrap=TRUE}, it is sent to
\code{base::strwrap()}.}

\item{maxNchar}{\code{numeric} value or \code{Inf} to limit the maximum characters
allowed for each string. This option is preferred when \code{wrap=TRUE}
is not feasible, for example heatmap labels. When \code{NULL} or \code{Inf}
no limit is applied. See \code{base::nchar()}.}

\item{suffix}{\code{character} value, default \code{"..."}, used
when \code{maxNchar} is below \code{Inf}.
When a string is shortened to \code{maxNchar}, the \code{suffix} helps indicate
that there was additional text.}

\item{nodeType}{\code{character} string ussed when \code{x} is \code{igraph},
to limit changes to nodes by attribute values in \code{"nodeType"}.
Use \code{"any"} or \code{NULL} to affect all nodes.}

\item{do_abbreviations}{\code{logical}, default TRUE, whether to apply
\verb{abbrev_from,abbrev_to}. These patterns are intended specifically
to help shorten a long phrase, possibly removing words, or using
common abbreviations.}

\item{adjustCase}{\code{logical}, default TRUE,  indicating whether to
adjust the uppercase and lowercase lettering by calling
\code{jamba::ucfirst()}. The default sets all characters to lowercase,
then applies uppercase to the first letter of each word.}

\item{lowercaseAll}{\code{logical} used only when \code{adjustCase=TRUE},
passed to \code{jamba::ucfirst()}}

\item{removeGrep}{\code{character} regular expression pattern used to remove
patterns from the resulting label.
\itemize{
\item The default removes common canonical pathway source prefix terms
use in MSigDB data, for example KEGG, BIOCARTA, PID, etc.
Use \code{""} or \code{NULL} to skip this step.
\item Multiple values can be defined, they are applied in order.
}}

\item{words_from, words_to}{\code{character} default NULL uses internal data
\code{words} with pattern and replacement.
\itemize{
\item Input \code{words_from} can be a two-column \code{data.frame} expected to have
'from' and 'to' in order.
\item Supplied as vectors, the 'words_from' are regular expression patterns,
replaced with 'words_to' in order, applied case-sensitive.
It does use Perl regular expression in \code{base::gsub()}, which is useful
to use with 'backslash-b' to enforce a word boundary for example.
}}

\item{add_from, add_to}{\code{character} vectors used in addition to
\code{words_from},\code{words_to}.
\itemize{
\item 'add_from' can be supplied as a two-column \code{data.frame} as described
for 'words_from'.
\item These values are applied after \code{words_from},\code{words_to}, so that
user-defined replacements have priority.
}}

\item{abbrev_from, abbrev_to}{\code{character} default NULL uses internal data
\code{abbrev} with pattern and replacement.
Intended to apply a specific abbreviation, and only applied when
\code{do_abbreviations=TRUE}.
\itemize{
\item 'abbrev_from' can be supplied as a \code{data.frame} as described for
'words_from'.
\item The abbreviations are "opinionated" in that they may remove words
or shorten common phrases which do not seem critical to
understanding the meaning of most biological pathways.
}

Examples:
\itemize{
\item "Extracellular Matrix" becomes "ECM"
\item "Mitochondrial" becomes "Mito"
\item " Pathway" at the end of a phrase is removed, as it is not
required to understand the rest of the label.
\item "Signaling by " at the start of a phrase is removed, as
it also is not typically necessary to understand the label.
}}

\item{...}{additional arguments are passed to \code{jamba::ucfirst(x, ...)},
for example \code{firstWordOnly=TRUE} will capitalize only the first word.}
}
\value{
object whose class matches input 'x':
\itemize{
\item \code{character} vector
\item \code{igraph} object with vertex 'label' updated from 'name'
\item \code{Mem} object with updated \code{sets()}
}
}
\description{
Fix Set or pathway labels for legibility
}
\details{
This function is a convenient wrapper for several steps that edit
gene set and pathways labels to be slightly more legible.
It operates on:
\itemize{
\item \code{character} vector, returning \code{character} vector
\item \code{igraph} object, where is uses 'name' to update 'label'
\item \code{Mem} object, where it updates \code{sets(Mem)}
}

The arguments have extensive default values encoded, which are
represented in data \code{multienrichjam::words} for basic word replacement,
and \code{multienrichjam::abbrev} for abbreviations used only
when \code{do_abbreviations=TRUE}.

Summary of typical changes:
\itemize{
\item The vast majority of changes are custom biological terms which
are expected to have certain capitalization, for example
'Mapk' is usually written 'MAPK'.
\item Some changes are motivated to fix common artifacts in public data,
for example \code{'PI3kakt'} refers to \code{'PI3K/AKT'}.
\item To use your own replacements, supply \code{words_from} as a two-column
\code{data.frame}, or two vectors \code{words_from} and \code{words_to}.
\item To add custom effects to default, supply \code{abbrev_from} as a two-column
\code{data.frame}, or use two vectors \code{abbrev_from} and \code{abbrev_to}.
}

For \code{igraph} input, the vertex 'name' is used as the starting point.
To revert changes, use \code{igraph::V(x)$label <- igraph::V(x)$name}.

For \code{Mem} input, the \code{sets(x)} are updated, with no immediate way
to revert changes. It may become useful to do so in future, however.
}
\examples{
x <- c("KEGG_INSULIN_SIGNALING_PATHWAY",
   "KEGG_T_CELL_RECEPTOR_SIGNALING_PATHWAY",
   "KEGG_NEUROTROPHIN_SIGNALING_PATHWAY");
fixSetLabels(x);
fixSetLabels(x, do_abbreviations=FALSE);

jamba::nullPlot();
jamba::drawLabels(txt=x,
   preset=c("top", "center", "bottom"));

}
\seealso{
Other jam Mem utilities: 
\code{\link{Mem-class}},
\code{\link{MemPlotFolio-class}},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()}

Other jam igraph functions: 
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{label_communities}()},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()}
}
\concept{jam Mem utilities}
\concept{jam igraph functions}
\keyword{datasets}
