% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-cnet.R
\name{get_cnet_nodeset}
\alias{get_cnet_nodeset}
\title{Get Cnet node set by connected Sets}
\usage{
get_cnet_nodeset(g, set_nodes = NULL, sep = ",", ...)
}
\arguments{
\item{g}{\code{igraph} object specifically containing Cnet plot data,
with node attribute \code{"nodeType"} that has values \code{c("Gene", "Set")}.}

\item{set_nodes}{\code{character} vector of one or more set names, which are
defined here as the names of the nodes with \code{nodeType="Set"}
to which each node with \code{nodeType="Gene"} is connected.
For example \code{set_nodes=c("A", "B")} will return all Gene nodes
that are connected only to Set nodes \code{"A"} and \code{"B"}. If
\code{set_nodes=NULL} then this function returns a \code{list} with
all set_nodes observed.}

\item{sep}{\code{character} string used as a delimited when combining
\code{set_nodes} into a name.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{character} vector when argument \code{set_nodes} is supplied
as a \code{character} vector; when \code{set_nodes} is \code{NULL},
this function returns a \code{list} of \code{character} vectors whose
names represent the Set node names connected to each
Gene node, where multiple Set node names are
delimited by \code{sep}.
}
\description{
Get Cnet node set by connected Sets
}
\details{
This function operates on a Cnet \code{igraph} object,
distinguished by node attribute \code{"nodeType"} with
value \code{"Gene"} for Gene nodes, and \code{"Set"} for Set nodes.
This function returns Gene nodes that are connected
to the Sets given by argument \code{set_nodes}. It is
useful to identify which genes are connected to
Set \code{"A"} and \code{"B"} for example, in other words the
Gene nodes in a specific subcluster of the Cnet
\code{igraph}.

When \code{set_nodes} is \code{NULL}, this function returns a \code{list}
containing \code{character} vectors, where each vector represents
Gene nodes that are connected to each combination of
Set nodes. This option is useful for obtaining all
possible Gene subclusters.

This function is also called by \code{adjust_cnet_nodeset()} in
order to manipulate all nodes in a subcluster as a group,
for example calling \code{nudge_igraph_node()} on the whole set
of nodes.
}
\seealso{
Other jam cnet igraph functions: 
\code{\link{adjust_cnet_nodeset}()},
\code{\link{adjust_cnet_set_relayout_gene}()},
\code{\link{apply_nodeset_spacing}()},
\code{\link{nudge_igraph_node}()},
\code{\link{rotate_igraph_layout}()}
}
\concept{jam cnet igraph functions}
