% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-points-to-ahull.R
\name{get_hull_data}
\alias{get_hull_data}
\title{Get data for alpha hull (internal)}
\usage{
get_hull_data(
  x,
  verbose = FALSE,
  hull_method = "alphahull",
  buffer = NULL,
  alpha = NULL,
  expand = 0.1,
  ...
)
}
\arguments{
\item{x}{\code{numeric} matrix with x,y coordinates}

\item{verbose}{\code{logical} indicating whether to print verbose output}

\item{hull_method}{\code{character} string with the preferred hull method.}

\item{alpha}{\code{numeric} passed to \code{alphahull:ashape()}}

\item{expand}{\code{numeric} used to define \code{alpha} based upon coordinate range.}

\item{...}{additional arguments are ignored.}
}
\description{
Get data for alpha hull (internal)
}
