% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-get-igraph-layout.R
\name{get_igraph_layout}
\alias{get_igraph_layout}
\title{Obtain or create layout for igraph object}
\usage{
get_igraph_layout(g, layout = NULL, make_circular = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{g}{\code{igraph} object}

\item{layout}{one of:
\itemize{
\item \code{numeric} matrix of layout coordinates, with \code{nrow(layout)}
equal to the number of nodes \code{igraph::vcount(g)}.
\item \code{function} that takes \code{igraph} input, and returns \code{numeric} matrix
of layout coordinates.
\item \code{NULL}: to use the layout embedded in \code{g} using
\code{igraph::graph_attr(g, "layout")} if it exists. If it does not exist,
then \code{make_circular=TRUE} causes a new layout to be created with
arbitrary circular coordinates; or \code{make_circular=FALSE} will return
\code{NULL}.
}}

\item{make_circular}{\code{logical} indicating whether to create a makeshift
layout when the input data does not already contain layout coordinates,
and when \code{layout} is not supplied.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix of x,y coordinates with nrow equal
to the number of nodes in the input, from \code{igraph::vcount(g)}.
Note that \code{rownames()} are defined to match node names
\code{igraph::V(g)$name}, unlike default \code{igraph} layouts.
}
\description{
Obtain or create layout for igraph object
}
\details{
This function is a simple helper function intended to retrieve
the node layout for an \code{igraph} object. When \code{layout} is supplied
as a \code{function} it is used to define a specific layout matrix.
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
