% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{grid_with_title}
\alias{grid_with_title}
\title{Draw Heatmap with title and subtitle using grid viewports}
\usage{
grid_with_title(
  object,
  title = NULL,
  title_fontsize = 18,
  title_just = "centre",
  caption = NULL,
  caption_fontsize = 12,
  caption_just = "left",
  caption_x = NULL,
  ...
)
}
\arguments{
\item{object}{an object with class \code{"Heatmap"}, \code{"gTree"}, or
any object where \code{methods::hasMethod("draw", class(object))}
is \code{TRUE}.}

\item{title}{character string used as title. When \code{NULL} or
\code{nchar(title)==0} then no title is displayed.}

\item{title_fontsize}{numeric value indicating the font size,
with units \code{"points"}.}

\item{title_just}{character string indicating the justification
(alignment) of the title.}

\item{caption, caption_fontsize, caption_just}{arguments equivalent
to the title_* arguments.}

\item{caption_x}{numeric value in grid units specifying where to
position the caption text. By default when \code{caption_just} is \code{"left"}
the \code{caption_x} is defined by \code{grid::unit(0.2, "npc")}, which
positions the caption at the left side (20\% from the left edge)
of the plot, with text proceeding to the right of that point.}

\item{...}{additional arguments are ignored.}
}
\value{
The byproduct of this function is to draw a grid visualization
that includes a title and subtitle, then the \code{object} in the center.
}
\description{
This function is intended to help make it easier to wrap a
heatmap from \code{ComplexHeatmap::Heatmap()} inside the relevant
grid viewports in order to display one large title at the
top of the resulting visualization, and optionally one
subtitle at the bottom of the visualization.
}
\details{
The input can be one heatmap (\code{Heatmap} object),
or object of class \code{"gTree"}, or any object with a method
\code{draw()} associated with it, detected by
\code{methods::hasMethod("draw", class(object))}.

A good example of \code{"gTree"} object is the result of
calling \code{draw()} inside \code{grid::grid.grabExpr()}, for example:

\code{grid::grid.grabExpr(ComplexHeatmap::draw(...))}

The \code{ComplexHeatmap::draw()} function has extended capability
for arranging one or more heatmaps and associated annotations.
}
\seealso{
Other jam plot functions: 
\code{\link{jam_igraph}()},
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()},
\code{\link{mem_plot_folio}()}
}
\concept{jam plot functions}
