% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{gsubs_remove}
\alias{gsubs_remove}
\title{Pattern replacement with multiple patterns}
\usage{
gsubs_remove(
  pattern,
  replacement,
  x,
  ignore.case = TRUE,
  replaceMultiple = rep(TRUE, length(pattern)),
  ...
)
}
\arguments{
\item{pattern}{character vector of patterns}

\item{replacement}{character vector of replacements}

\item{x}{character vector with input data to be curated}

\item{ignore.case}{logical indicating whether to perform
pattern matching in case-insensitive manner, where
\code{ignore.case=TRUE} will ignore the uppercase/lowercase
distinction.}

\item{...}{additional arguments are passed to \code{base::gsub()}
or \code{base::sub()}.}

\item{replace_multiple}{logical vector indicating whether to perform
global substitution, where \code{replace_multiple=FALSE} will
only replace the first occurrence of the pattern, using
\code{base::sub()}. Note that this vector can refer to individual
entries in \code{pattern}.}
}
\description{
Pattern replacement with multiple patterns
}
\details{
This function is a simple wrapper around \code{base::gsub()}
when considering a series of pattern-replacement
combinations. It applies each pattern match and replacement
in order and is therefore not vectorized.
}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{bulk_cnet_adjustments}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
