% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-plot.R
\name{handle_igraph_param_list}
\alias{handle_igraph_param_list}
\title{Handle igraph attribute parameter list}
\usage{
handle_igraph_param_list(
  x,
  attr,
  factor_l,
  i_values = NULL,
  attr_type = c("node", "vertex", "edge"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{igraph} object}

\item{attr}{\code{character} name of the attribute to update in \code{x}.}

\item{factor_l}{\code{list} or \code{numeric} vector or \code{function}:
\itemize{
\item \code{list} of \code{numeric} vectors where \code{names(factor_l)}
correspond to attribute names, and the names of numeric vectors
are attribute values The attribute names and attribute values are
used to match relevant entities of type \code{attr_type}.
For matching entities, attribute values are used as defined by
attribute name \code{attr}, and are multiplied by the matching
numeric value in \code{factor_l}.
\item \code{numeric} vector which is directly multiplied by \code{i_values}
to produce an adjusted output vector \code{i_values}.
\item \code{function} which is used to modify \code{i_values} by calling
\code{factor_l(i_values)} to produce adjusted output \code{i_values}.
}}

\item{i_values}{\code{vector} of attribute values that represent the current
attribute values in \code{x} for the attribute \code{attr}.}

\item{attr_type}{\code{character} string indicating the type of entity
being adjusted in \code{x}:
\itemize{
\item \code{"node"} or \code{"vertex"} refers to \code{igraph::vertex_attr()}
\item \code{"edge"} refers to \code{igraph::edge_attr()}
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{vector} of attribute values representing \code{attr}.
}
\description{
Handle igraph attribute parameter list, internal function for \code{jam_igraph()}
}
\details{
This mechanism is intended to help update \code{igraph} attributes
in bulk operations by the attribute values associated with
nodes or edges. Most commonly, the argument \code{factor_l} is
multiplied by numeric attributes to scale attribute values,
for example label font size, or node size.

For example:\preformatted{handle_igraph_param_list(x,
   attr="size",
   factor_l=list(nodeType=c(Gene=1, Set=2)),
   i_values=rep(1, igraph::vcount(x)),
   attr_type="node")
}

This function call will match node attribute \code{nodeType},
the size of nodes with attribute value
\code{nodeType="Set"} are multiplied \code{size * 2},
\code{nodeType="Gene"} are multiplied \code{size * 1}.
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
