% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{heatmap_row_order}
\alias{heatmap_row_order}
\title{Return Heatmap row order}
\usage{
heatmap_row_order(hm)
}
\description{
Return Heatmap row order as a list of character vectors
}
\details{
This function is a helpful utility to return the fully
qualified list of rownames in a \code{ComplexHeatmap::Heatmap}
object.

This function also works with \code{ComplexHeatmap::HeatmapList}
objects.
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{deconcat_df2}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{heatmap_column_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
