% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-highlights.R
\name{highlight_edges_by_node}
\alias{highlight_edges_by_node}
\title{Highlight edges connected to a node or nodes}
\usage{
highlight_edges_by_node(
  g,
  node = NULL,
  highlight_color = "royalblue3",
  highlight_width = NULL,
  highlight_cex = 2,
  nonhighlight_alpha = NULL,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object}

\item{node}{\code{character} node name, or \code{integer} node index.}

\item{highlight_color}{\code{character} color used for highlighting,
default 'royalblue'.}

\item{highlight_width}{\code{numeric} default NULL, set a fixed edge width.}

\item{highlight_cex}{\code{numeric} default 2, multiplies the edge width
by this expansion factor.}

\item{nonhighlight_alpha}{\code{numeric} default NULL, applies an optional
alpha transparency to non-highlighted edges. Values must be
between 0 (transparent) and 1 (opaque), where a value 0.5 is
recommended when used.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{igraph} object after adjusting edge attributes.
}
\description{
Highlight edges connected to a node or nodes
}
\details{
This function highlights edges connected to one of more nodes
by applying a color, and adjusting the edge width.

Note that when edge attributes 'width' and 'color' are not yet
defined in the \code{igraph} object, they will be defined and stored
into the object by using \code{default_igraph_values()}.
For example: \code{default_igraph_values()$edge$width} and
\code{default_igraph_values()$edge$color}.
}
\examples{
cnet1 <- make_cnet_test(seed=1);
# highlight edges by node
cnet1h <- highlight_edges_by_node(cnet1, "CA", nonhighlight_alpha=0.5)
# plot the original
jam_igraph(cnet1)
# plot the highlighted variant
jam_igraph(cnet1h, label_dist_factor=3)

# highlight edges using nodeset
nodesets <- get_cnet_nodeset(cnet1)
cnet1hns <- highlight_edges_by_node(cnet1,
   nodesets[[grep("CA", nodesets)]],
   nonhighlight_alpha=0.5)
jam_igraph(cnet1hns, node_factor=2, label_factor=2, label_dist_factor=3,
   mark.groups=unname(nodesets[grep("CA", nodesets)]))

}
\seealso{
Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
