% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{igraph2pieGraph}
\alias{igraph2pieGraph}
\title{Convert igraph to use pie node shapes}
\usage{
igraph2pieGraph(g, valueIM = NULL, valueIMcolors = NULL,
  colorV = NULL, updateLabels = FALSE, maxNchar = 62,
  backgroundColor = "white", seed = 123, defineLayout = FALSE,
  repulse = 3.6, removeNA = FALSE, NAvalues = c(NA, "transparent"),
  verbose = FALSE, ...)
}
\description{
Convert igraph to use pie node shapes
}
\details{
This function converts an igraph to use pie node shapes,
where pie wedges are colored using values derived from a
numeric matrix `valueIM` or pre-defined in a character matrix
containing colors `valueIMcolors`.

Note that pie wedge sizes are equally-sized and do not vary
by score, instead the color intensity is applied to each pie
wedge.

Node names using `V(g)$name` matching `rownames(valueIMcolors)`
are colorized and the node shape is converted to pie. All other
nodes are not modified.

When `valueIMcolors` is not defined, it is derived from `valueIM`
using `colorjam::matrix2heatColors()`. In that case, `colorV` defines the color
used for numeric values in each column, and other options are passed
to `colorjam::matrix2heatColors()` via `...` arguments.
}
\seealso{
Other jam igraph functions: \code{\link{cnetplotJam}},
  \code{\link{cnetplot_internalJam}},
  \code{\link{layout_with_qfr}},
  \code{\link{shape.coloredrectangle.plot}}
}
\concept{jam igraph functions}
