% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{importIPAenrichment}
\alias{importIPAenrichment}
\title{Import Ingenuity IPA enrichment results}
\usage{
importIPAenrichment(ipaFile,
  headerGrep = "(^|\\t)((expr.|-log.|)p-value|Pvalue|Score\\t|Symbol\\t|Ratio\\t|Consistency.Score|Master.Regulator\\t)",
  ipaNameGrep = c("Pathway", "Regulator$", "Regulators", "Regulator",
  "Disease", "Toxicity", "Category", "Categories", "Function", "Symbol$",
  "^ID$", "My.(Lists|Pathways)"), geneGrep = c("Molecules in Network",
  "Molecules"), geneCurateFrom = c("^[,]+|[,]+$"),
  geneCurateTo = c(""), method = 1, sheet = 1, sep = "\\t",
  xlsxMultiSheet = TRUE, useXlsxSheetNames = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{ipaFile}{one of the four input types described above:
a character vector of text file names; a character vector of
Excel \code{.xlsx} file names; a list of \code{data.frame} objects.}

\item{headerGrep}{regular expression pattern used to
recognize header columns found in Ingenuity IPA enrichment data.}

\item{ipaNameGrep}{vector of regular expression patterns
used to recognize the name of the enriched entity,
for example the biological pathway, or network, or
disease category, etc.}

\item{geneGrep}{regular expression pattern used to recognize
the column containing genes, or the molecules tested for
enrichment which were found in the enriched entity.}

\item{geneCurateFrom, geneCurateTo}{vector of patterns and
replacements, respectively, used to curate values in
the gene column. These replacement rules are used to
ensure that genes are delimited consistently, with no
leading or trailing delimiters.}

\item{method}{integer value indicating the method used to
import data from a text file, where: \code{method=1} uses
\code{data.table::read.table()} and the \code{textConnection} argument;
\code{method=2} uses \code{readr::read_tsv()}. The motivation to use
\code{data.table::read.table()} is it performed better in the
presence of UTF-8 characters such as the alpha symbol.}

\item{sheet}{integer value used only when \code{ipaFile} is
a vector of Excel \code{.xlsx} files, and when the Excel
format includes multiple worksheets. This value will
extract enrichment data only from one worksheet from
each Excel file.}

\item{sep}{character string used when \code{ipaFile} is a vector
of text files, to split fields into columns. The default
will split fields by the tab character.}

\item{xlsxMultiSheet}{logical indicating whether input
Excel \code{.xlsx} files contain multiple worksheets.}

\item{useXlsxSheetNames}{logicl indicating whether to use the
Excel worksheet name for each imported enrichment table,
when importing from \code{.xlsx} files, and when
\code{xlsxMultiSheet=FALSE}. When \code{xlsxMultiSheet=TRUE} the
name is derived from the value matched using \code{ipaNameGrep},
because in this case, there are expected to me multiple
enrichment tables in one worksheet.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
list of \code{data.frame} objects, where each \code{data.frame}
contains enrichment data for one of the Ingenuity IPA
enrichment tests.
}
\description{
Import Ingenuity IPA enrichment results
}
\details{
This function parses Ingenuity IPA enrichment data into
a form usable as a list of  enrichment \code{data.frame}
objects for downstream analysis. Each \code{data.frame}
will represent the results of one Ingenuity IPA
enrichment test.

The input data can be one of four forms:
\enumerate{
\item \code{ipaFile} can be a text \code{.txt} file,
where the text file contains all IPA enrichment data in
tall format. This format is most common.
\item \code{ipaFile} can be an Excel \code{.xlsx} file,
which contains all IPA enrichment data in
one tall worksheet tab.
\item \code{ipaFile} can be an Excel \code{.xlsx} file,
where each type of IPA enrichment appears on a separate
Excel worksheet tab.
\item \code{ipaFile} can be a list of \code{data.frame} objects.
This option is intended when the IPA data has already
been imported into R as separate \code{data.frame} objects.
}

The basic motivation for this function is two-fold:
\enumerate{
\item Separate multiple IPA enrichment tables.
\item Rename colnames to be consistent.
}

When using \code{"Export All"} from IPA, the default text format
includes multiple enrichment tables concatenated together in one
file. Each enrichment table contains its own unique column
headers, with descriptive text in the line preceding the
column headers. This function is intended to separate the
enrichment tables into a list of \code{data.frame} objects, and
retain the descriptive text as names of the list.
}
\concept{jam import functions}
