% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-import.R
\name{IPAlist_to_hits}
\alias{IPAlist_to_hits}
\title{Convert IPA list to a gene hit list or matrix}
\usage{
IPAlist_to_hits(
  IPAlist,
  signColname = c("Expr Fold Change", "fold.*change", "log.*ratio", "log.*fold",
    "log.*fc", "lfc", "ratio", "fold", "fc"),
  signThreshold = 0,
  geneColname = c("name", "symbol", "ID"),
  emptyValue = 0,
  returnType = c("list", "matrix"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{IPAlist}{\code{list} of output from \code{importIPAenrichment()}}

\item{signColname}{\code{character} vector of patterns to match the colname
in the 'Analysis Ready Molecules' IPA xref table which should
indicate the direction of change where present.}

\item{signThreshold}{\code{numeric} minimum absolute threshold, default 0,
for a value in 'signColname' to be considered a non-zero change.}

\item{geneColname}{\code{character} vector of patterns to match the colname
in the 'Analysis Ready Molecules' IPA xref table which should
indicate the gene involved. This column is validated by matching
with the 'Canonical Pathways' geneID or geneNames column to ensure
all values involved in pathway enrichment are also present in the
xref column.}

\item{emptyValue}{\code{numeric} default 0, passed to \code{list2imSigned()} to
use when \code{returnType="matrix"} in the matrix for non-hits.}

\item{returnType}{\code{character} string
\itemize{
\item 'list' (default): returns a signed gene list, with \code{integer} vectors
named by gene.
\item 'matrix': returns a numeric \code{matrix} with gene rownames.
}}

\item{...}{additional arguments are passed to internal functions}
}
\value{
\code{list} by default, or \code{matrix} with \code{returnType="matrix"}.
}
\description{
Convert IPA list to a gene hit list or matrix
}
\details{
Given a \code{list} of IPA data produced by \code{importIPAenrichment()}
produce a single \code{list} or incidence \code{matrix} representing
all gene hits tested by IPA.
}
\seealso{
Other jam import functions: 
\code{\link{enrichDF2enrichResult}()},
\code{\link{importIPAenrichment}()}
}
\concept{jam import functions}
