% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{isColorBlank}
\alias{isColorBlank}
\title{Determine if colors are blank colors}
\usage{
isColorBlank(
  x,
  c_max = 7,
  l_min = 95,
  alpha_max = 0.1,
  blankColor = c("#FFFFFF", "#FFFFFFFF", "transparent"),
  ...
)
}
\arguments{
\item{x}{character vector of R colors.}

\item{c_max}{maximum chroma as determined by HCL color space, in
range of no color 0 to maximum color 100.}

\item{l_min}{numeric minimum luminance required for a color to be
considered blank, combined with the \code{c_max} argument. This
threshold prevents grey colors from being considered blank,
unless their luminance is above this threshold.}

\item{alpha_max}{numeric value indicating the alpha transparency
below which a color is considered blank, in range of fully
transparent 0, to fully non-transparent 1.}

\item{blankColor}{character vector of R colors directly matched to
the input \code{x} vector. The value \code{"transparent"} is useful here,
because it is not easily converted to HCL color space.}

\item{...}{additional arguments are ignored.}
}
\description{
Determine if colors are blank colors
}
\details{
This function takes a vector of colors and determines if each color
is considered a "blank color", based upon direct match and the
color chroma saturation and luminance. For example, extremely pale
colors from \code{colorjam::vals2colorLevels()} may be considered "blank" if the
color saturation is extremely low. Similarly, colors with
extremely high alpha transparency may be considered "blank".
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
