% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{jam_igraph}
\alias{jam_igraph}
\title{Jam wrapper to plot igraph}
\usage{
jam_igraph(
  x,
  ...,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  node_factor = 1,
  edge_factor = 1,
  label_factor = 1,
  label_dist_factor = 1,
  expand = 0.03,
  rescale = FALSE
)
}
\arguments{
\item{x}{\code{igraph} object to be plotted}

\item{...}{additional arguments are passed to \code{igraph::plot.igraph()}}

\item{xlim, ylim}{default x and y axis limits}

\item{expand}{numeric value used to expand the x and y axis ranges,
where \code{0.03} expands each size \verb{3\%}.}

\item{rescale}{logical indicating whether to rescale the layout
coordinates to \code{c(-1, 1)}. When \code{rescale=FALSE} the original
layout coordinates are used as-is without change.}
}
\description{
Jam wrapper to plot igraph
}
\details{
This function is a lightweight wrapper around \code{igraph::plot.igraph()}
intended to handle \code{rescale=FALSE} properly, which is not done
in the former function (as of January 2020). The desired outcome
is for the \code{xlim} and \code{ylim} defaults to be scaled according to the
\code{igraph} layout. Similarly, the \code{vertex.size} and \code{vertex.label.dist}
parameters should also be scaled proportionally.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{shape.coloredrectangle.plot}()},
\code{\link{shape.ellipse.plot}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}

Other jam plot functions: 
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()}
}
\concept{jam igraph functions}
\concept{jam plot functions}
