% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-igraph-internals.R
\name{jam_igraph_arrows}
\alias{jam_igraph_arrows}
\title{Render igraph arrows}
\usage{
jam_igraph_arrows(
  x1,
  y1,
  x2,
  y2,
  code = 2,
  size = 1,
  width = 1,
  open = TRUE,
  sh.adj = 0.1,
  sh.lwd = 1,
  sh.col = if (is.R()) par("fg") else 1,
  sh.lty = 1,
  h.col = sh.col,
  h.col.bo = sh.col,
  h.lwd = sh.lwd,
  h.lty = sh.lty,
  arrows_only = FALSE,
  curved = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x1, y1, x2, y2}{\code{numeric} coordinates for initial and final x and y
coordinates.}

\item{code}{\code{integer} indicating the position of arrow:
\itemize{
\item \code{code=1} arrow is positioned on the line end
\item \code{code=2} arrow is positioned on the line start
\item \code{code=3} arrow is positioned on both ends of the line
}}

\item{size}{\code{numeric} scaled size of the arrow head, which is applied to
both the length and width of the arrow head.}

\item{width}{\code{numeric} scalar for the arrow head width, which is only
applied to the relative arrow width.}

\item{open}{\code{logical} indicating whether the arrow head should be a filled
polygon, otherwise only the outer "V" lines are drawn.}

\item{sh.adj}{\code{numeric} adjustment for segment length, where:
\itemize{
\item \code{sh.adj=0}  will extend the edge line (underneath the arrow head)
to the end of the line
\item \code{sh.adj=1} will extend the edge line only to the base of the arrow head
\item \code{sh.adj=1.1} will leave a gap approximately 10\% the arrow head length,
between the edge line and the start of the arrow head.
}}

\item{sh.lwd}{\code{numeric} line width of main segment edge line}

\item{sh.col}{\code{character} color of main segment edge line}

\item{sh.lty}{\code{numeric} line type of main segment edge line}

\item{h.col, h.col.bo}{\code{character} arrow head color and arrow head border
color, respectively.}

\item{h.lwd}{\code{numeric} arrow head line width}

\item{h.lty}{\code{numeric} arrow head line type}

\item{arrows_only}{\code{logical} indicating whether to draw only arrows,
or \code{arrows_only=FALSE} to draw arrows and edge lines.}

\item{curved}{\code{logical} indicating whether to draw curved edges}

\item{verbose}{\code{logical} indicating whether to print verbose output.}
}
\description{
Render igraph arrows
}
\details{
This function is a mimic of the internal \code{igraph:::igraph.Arrows()}
which is not permitted to be called directly for CRAN-approved
R packages.
}
\examples{
plot(NULL, xlim=c(-3, 3), ylim=c(-4, 4), type="n", xlab="", ylab="", bty="n")
jam_igraph_arrows(-2, 3, 2, 3, code=1, open=FALSE, sh.col="blue", sh.lwd=2)
jam_igraph_arrows(-2, 2, 2, 2, code=2, open=FALSE, sh.col="red", sh.lwd=2)
jam_igraph_arrows(-2, 1, 2, 1, code=3, open=FALSE, sh.col="gold", sh.lwd=2)
jam_igraph_arrows(-2, 0, 2, 0, code=3, arrows_only=TRUE, open=FALSE, sh.col="purple4", sh.lwd=2)

jam_igraph_arrows(-2, -1, 2, -1, code=1, open=FALSE, sh.col="blue", h.col="#FF000055", sh.lwd=2, size=2, sh.adj=0.1)
jam_igraph_arrows(-2, -2, 2, -2, code=1, open=FALSE, sh.col="blue", h.col="#FF000055", sh.lwd=2, size=2, sh.adj=1.1)
jam_igraph_arrows(-2, -3, 2, -3, code=2, open=FALSE, sh.col="blue", h.col="#FF000055", sh.lwd=2, size=2, sh.adj=1.1)
jam_igraph_arrows(-2, -4, 2, -4, code=3, open=FALSE, sh.col="blue", h.col="#FF000055", sh.lwd=2, size=2, sh.adj=1.1)
text(x=rep(0, 8), y=seq(from=3, to=-4)+0.2,
   labels=c("code=1",
      "code=2",
      "code=3",
      "code=3, arrows_only=TRUE",
      "code=1, size=2, sh.adj=0.1",
      "code=1, size=2, sh.adj=1.1",
      "code=2, size=2, sh.adj=1.1",
      "code=3, size=2, sh.adj=1.1"))

}
\seealso{
Other jam igraph internal functions: 
\code{\link{default_igraph_values}()},
\code{\link{get_igraph_arrow_mode}()},
\code{\link{parse_igraph_plot_params}()}
}
\concept{jam igraph internal functions}
