% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{jam_plot_igraph}
\alias{jam_plot_igraph}
\title{Jam igraph vectorized plot function}
\usage{
jam_plot_igraph(
  x,
  axes = FALSE,
  add = FALSE,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  mark.groups = list(),
  mark.shape = 1/2,
  mark.col = colorjam::rainbowJam(length(mark.groups), alpha = 0.3),
  mark.border = colorjam::rainbowJam(length(mark.groups), alpha = 1),
  mark.expand = 15,
  pie_to_jampie = TRUE,
  ...
)
}
\arguments{
\item{x}{The graph to plot.}

\item{axes}{Logical, whether to plot axes, defaults to FALSE.}

\item{add}{Logical scalar, whether to add the plot to the current device, or
delete the device's current contents first.}

\item{xlim}{The limits for the horizontal axis, it is unlikely that you want
to modify this.}

\item{ylim}{The limits for the vertical axis, it is unlikely that you want
to modify this.}

\item{mark.groups}{A list of vertex id vectors. It is interpreted as a set
of vertex groups. Each vertex group is highlighted, by plotting a colored
smoothed polygon around and \dQuote{under} it. See the arguments below to
control the look of the polygons.}

\item{mark.shape}{A numeric scalar or vector. Controls the smoothness of the
vertex group marking polygons. This is basically the \sQuote{shape}
parameter of the \code{\link[graphics]{xspline}} function, its possible
values are between -1 and 1. If it is a vector, then a different value is
used for the different vertex groups.}

\item{mark.col}{A scalar or vector giving the colors of marking the
polygons, in any format accepted by \code{\link[graphics]{xspline}}; e.g.
numeric color ids, symbolic color names, or colors in RGB.}

\item{mark.border}{A scalar or vector giving the colors of the borders of
the vertex group marking polygons. If it is \code{NA}, then no border is
drawn.}

\item{mark.expand}{A numeric scalar or vector, the size of the border around
the marked vertex groups. It is in the same units as the vertex sizes. If a
vector is given, then different values are used for the different vertex
groups.}

\item{pie_to_jampie}{logical indicating whether to convert
vertex shape \code{"pie"} to \code{"jampie"} in order to use vectorized
plotting.}

\item{...}{Additional plotting parameters. See \link[igraph]{igraph.plotting} for
the complete list.}
}
\description{
Jam igraph vectorized plot function
}
\details{
This function is a complete copy of \code{igraph:::plot.igraph()} with
two changes to enable vectorized plotting in these situations:
\enumerate{
\item When there are multiple vertex \code{"shape"} attributes, the
base plot function draws each individual \code{igraph} vertex one by one.
The \code{jam_plot_igraph()} plots nodes of each shape in a group. For
fairly large \code{igraph} objects (about 1000 nodes), this change
is substantially faster.
\item When there are multiple font families, the default plot function
draws each label one by one. The \code{jam_plot_igraph()} draws
labels in groups of font family. This situation is very rare, however
when used the speed improvement is substantial.
}

Note that this function is not called by default, and is only called
by \code{multienrichjam::jam_igraph()}.

Smaller changes: \code{mark.col} and \code{mark.border} default colors
now call \code{colorjam::rainbowJam()} instead of \code{grDevices::rainbow()}.

All arguments are documented in \code{igraph::plot.igraph()}.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
