% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{label_communities}
\alias{label_communities}
\title{Assign labels to igraph communities}
\usage{
label_communities(
  wc,
  labels = NULL,
  add_catchwords = NULL,
  num_keep_terms = 3,
  keep_terms_sep = ",\\n",
  ...
)
}
\arguments{
\item{wc}{\code{communities} object, or \code{list} in form of nodegroups,
which is a \code{list} of \code{character} vectors that contain \code{igraph}
node names.}

\item{labels}{\code{character} vector of optional labels to assign directly
to community clusters. When not defined, the auto-detection method
is used.}

\item{add_catchwords}{\code{character} of optional words to include as
catchwords, to be excluded from use in the final label.}

\item{num_keep_terms}{\code{integer} maximum number of terms to be included
in the final output label, when auto-detection is used.}

\item{keep_terms_sep}{\code{character} string used as a delimited to separate
each term when multiple terms are concatenated together to form
the cluster label.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{communities} or \code{list} format matching the input \code{wc} format.
\itemize{
\item When \code{communities} is input, additional value \code{cluster_names}
will contain a \code{character} vector of names corresponding to each
integer index in \code{wc$membership}.
\item When \code{nodegroups} is input, the \code{list} names will be a \code{character}
vector of cluster labels.
}
}
\description{
Assign labels to igraph communities
}
\seealso{
Other jam igraph functions: 
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()}
}
\concept{jam igraph functions}
