% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-layout-communities.R
\name{layout_communities}
\alias{layout_communities}
\title{Layout igraph communities}
\usage{
layout_communities(g, layout = layout_with_qfrf(repulse = 3), ...)
}
\arguments{
\item{g}{\code{igraph} with graph attributes that include either
'mark.groups' or 'nodegroups'.}

\item{layout}{\code{function} to apply to each igraph community.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{igraph}
}
\description{
Layout igraph communities, applies layout to each node community
}
\details{
This function mimics \code{igraph::layout_components()}, in that it
splits the input \code{igraph} into a \code{list}, applies layout to
each graph, then merges the layouts together.
}
