% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{layout_with_qfrf}
\alias{layout_with_qfrf}
\title{igraph layout function using qgraph Fruchterman-Reingold}
\usage{
layout_with_qfrf(repulse = 3.5, seed = 123, ...)
}
\arguments{
\item{repulse}{numeric value typically between 3 and 5, passed
to \code{layout_with_qfr()}, which in turn is passed to
\code{qgraph::qgraph.layout.fruchtermanreingold()}.}

\item{seed}{numeric value used to set the R random seed, in order
to make layouts consistent.}

\item{...}{additional arguments are passed to \code{layout_with_qfr()}.}
}
\value{
function used to calculate layout coordinates of
an \code{igraph} object.
}
\description{
igraph layout function using qgraph Fruchterman-Reingold
}
\details{
This function returns a layout function, which can be convenient
when calling \code{igraph::plot.igraph()}, in order to set
layout parameters in the same call.
}
\seealso{
Other jam igraph layouts:
\code{\link[=layout_communities]{layout_communities()}},
\code{\link[=layout_with_qfr]{layout_with_qfr()}},
\code{\link[=relayout_with_qfr]{relayout_with_qfr()}},
\code{\link[=removeIgraphSinglets]{removeIgraphSinglets()}},
\code{\link[=reorderIgraphNodes]{reorderIgraphNodes()}},
\code{\link[=rotate_igraph_layout]{rotate_igraph_layout()}},
\code{\link[=spread_igraph_labels]{spread_igraph_labels()}}
}
\concept{jam igraph layouts}
