% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{layout_with_qfrf}
\alias{layout_with_qfrf}
\title{igraph layout function using qgraph Fruchterman-Reingold}
\usage{
layout_with_qfrf(repulse = 3.1, seed = 123, ...)
}
\description{
igraph layout function using qgraph Fruchterman-Reingold
}
