% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{layout_with_qfrf}
\alias{layout_with_qfrf}
\title{igraph layout function using qgraph Fruchterman-Reingold}
\usage{
layout_with_qfrf(repulse = 3.5, seed = 123, ...)
}
\arguments{
\item{repulse}{numeric value typically between 3 and 5, passed
to \code{layout_with_qfr()}, which in turn is passed to
\code{qgraph::qgraph.layout.fruchtermanreingold()}.}

\item{seed}{numeric value used to set the R random seed, in order
to make layouts consistent.}

\item{...}{additional arguments are passed to \code{layout_with_qfr()}.}
}
\value{
function used to calculate layout coordinates of
an \code{igraph} object.
}
\description{
igraph layout function using qgraph Fruchterman-Reingold
}
\details{
This function returns a layout function, which can be convenient
when calling \code{igraph::plot.igraph()}, in order to set
layout parameters in the same call.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{label_communities}()},
\code{\link{layout_with_qfr}()},
\code{\link{mem2emap}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
