% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{list2im}
\alias{list2im}
\title{convert list to incidence matrix}
\usage{
list2im(x, keepCounts = FALSE, emptyValue = 0, verbose = FALSE, ...)
}
\arguments{
\item{x}{list of vectors}

\item{keepCounts}{boolean indicating whether to return values indicating
the number of occurrences of each item.}

\item{emptyValue}{any single value that should be used for blank entries,
by default zero \code{0}. Use \code{emptyValue=NA} to return \code{NA} for missing
entries.}

\item{verbose}{boolean indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric matrix whose rownames were vector items of the input list,
and whole colnames were list names.
}
\description{
convert list to incidence matrix
}
\details{
This function converts a list of vectors into an incidence matrix, where
the rows are the vector items and the columns are the list names.
It uses an object from the \code{arules} package called
\code{arules::transactions} which offers highly efficient methods
for interconverting from list to matrix. The
\code{\link[arules]{transactions}} class is itself an enhanced data matrix,
which stores data using sparse matrix object type from the
\code{\link{Matrix}} package, but also associates a \code{data.frame} to both
the rows and columns of the matrix to offer additional row and column
annotation, as needed.

Performance benchmarks showed high speed of converting a list to a matrix,
but also that the resulting matrix was substantially smaller (5-20 times)
then comparable methods producing a data matrix.

When argument \code{keepCounts=TRUE}, the method of applying counts only
updates entries with multiple instances, which helps make this step
relatively fast.
}
\examples{
L1 <- list(A=c("C","A","B","A"),
   D=c("D","E","F","D"),
   A123=c(1:8,3,5),
   T=LETTERS[7:9]);
# Default behavior is to make items unique
list2im(L1);

# Option to report the counts
list2im(L1, keepCounts=TRUE);

}
\seealso{
Other jam list functions: 
\code{\link{colors_from_list}()},
\code{\link{im2list}()},
\code{\link{imSigned2list}()},
\code{\link{list2concordance}()},
\code{\link{list2imSigned}()}
}
\concept{jam list functions}
