% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-sets.r
\name{list2imSigned}
\alias{list2imSigned}
\title{convert list to directional incidence matrix}
\usage{
list2imSigned(x, makeUnique = TRUE, doTimer = FALSE, verbose = FALSE,
  ...)
}
\value{
numeric matrix with rownames defined by names from each vector
   in the input list, and colnames defined by the names of the input
   list. Values in the matrix are values from each vector.
}
\description{
convert list to directional incidence matrix
}
\details{
This function extends \code{\link{list2im}} in that it stores the
value associated with each element in the list. As such, the input
format is a named vector, where the names of the vector are the items,
and the numeric values are the values to be stored in the
incidence matrix.

A common scenario is to generate a vector of genes, with values
\code{c(-1, 0, 1)} indicating the direction of gene expression changes,
named by the gene symbol. Each vector in the list represents one
statistical test. Here, \code{list2imSigned()} will convert this list
into a directional matrix representing the gene changes across the
comparisons.
}
\examples{
L1 <- list(A=c("C","A","B","A"),
   D=c("D","E","F","D"),
   A123=c(1:8,3,5),
   T=LETTERS[7:9]);
L1;

# Convert each vector to a signed vector
set.seed(123);
L2 <- lapply(L1, function(i){
   i <- unique(i);
   nameVector(sample(c(-1,1), size=length(i), replace=TRUE), i);
});
L2;

# Convert to signed incidence matrix
list2imSigned(L2);

}
