% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-bivariatecolor.R
\name{make_legend_bivariate}
\alias{make_legend_bivariate}
\title{Display colors from bivariate color function}
\usage{
make_legend_bivariate(
  col_fun,
  pretty.n = 5,
  name = "bivariate",
  xlab = "",
  ylab = "",
  title = "",
  border = TRUE,
  digits = 3,
  grid_height = grid::unit(5, "mm"),
  grid_width = grid_height,
  row_breaks = NULL,
  column_breaks = NULL,
  row_gap = grid::unit(0, "mm"),
  column_gap = grid::unit(0, "mm"),
  ...
)
}
\description{
Display colors from bivariate color function
}
\examples{
mcolor <- matrix(ncol=3,
   c("seashell", "salmon1", "firebrick3",
      "gray99", "lightgoldenrod1", "gold",
      "aliceblue", "skyblue", "dodgerblue3"));
row_breaks <- c(0, 0.5, 1);
column_breaks <- c(-1, 0, 1);
rownames(mcolor) <- row_breaks;
colnames(mcolor) <- column_breaks;
jamba::imageByColors(mcolor);

col_fun <- colorRamp2D(column_breaks=column_breaks,
   row_breaks=row_breaks,
   mcolor=mcolor)
lgds <- make_legend_bivariate(col_fun,
   ylab="-log10pvalue",
   xlab="z-score",
   pretty.n=5);
jamba::nullPlot(doBoxes=FALSE);
ComplexHeatmap::draw(lgds)

lgds <- make_legend_bivariate(col_fun,
   ylab="-log10pvalue",
   xlab="z-score",
   pretty.n=NULL);
jamba::nullPlot(doBoxes=FALSE);
ComplexHeatmap::draw(lgds)

lgds <- make_legend_bivariate(col_fun,
   ylab="-log10pvalue",
   xlab="z-score",
   column_breaks=c(-1, -0.5,  0, 0.5, 1),
   row_breaks=c(0, 0.25, 0.5, 0.75, 1),
   column_gap=grid::unit(1, "mm"),
   row_gap=grid::unit(1, "mm"),
   pretty.n=5);
jamba::nullPlot(doBoxes=FALSE);
ComplexHeatmap::draw(lgds)

}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
