% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-points-to-ahull.R
\name{make_point_hull}
\alias{make_point_hull}
\title{Make alpha hull from points}
\usage{
make_point_hull(
  x,
  expand = 0.05,
  buffer = NULL,
  alpha = NULL,
  seed = 124,
  col = "#FF000033",
  border = "#FF0000FF",
  lwd = 2,
  lty = 1,
  max_iterations = 100,
  do_plot = FALSE,
  add = FALSE,
  hull_method = c("default", "ahull", "alphahull", "igraph", "sf", "chull"),
  smooth = TRUE,
  shape = 1/2,
  label = NULL,
  label.cex = 1,
  label.x.nudge = 0,
  label.y.nudge = 0,
  label_preset = NULL,
  label_adj_preset = label_preset,
  min_points = 1,
  xy_range = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{numeric} matrix with 2 columns that contains the
coordinate of each point.}

\item{expand}{\code{numeric} value, default 0.05, the buffer width around
each point, scaled based upon the total range of coordinates,
used only when \code{buffer} is not supplied.}

\item{buffer}{\code{numeric} value indicating the absolute buffer width
around each point. This value is used if provided, otherwise
\code{expand} is used to derive a value for \code{buffer}.}

\item{alpha}{\code{numeric} value passed to \code{ashape()} when
hull_method is \code{"alphahull"}. This value determines the level of
detail of the resulting hull.}

\item{seed}{\code{numeric} seed used with \code{set.seed()} to define reproducible
behavior.}

\item{lwd, lty}{line width and line type parameters, respectively.}

\item{max_iterations}{\code{integer} number of attempts to call
\code{ashape()} with varying values of \code{alpha}. Each iteration
checks to confirm the resulting polygon includes all input points.}

\item{do_plot}{\code{logical} indicating whether to plot the polygon
output.}

\item{add}{\code{logical} used when \code{do_plot=TRUE} to indicate whether
the hull should be drawn onto an existing plot device, or whether
to open a new plot prior to drawing the hull.}

\item{hull_method}{\code{character} string indicating the hull method to use:
\itemize{
\item \code{"default"} - will use \code{"alphahull"} if the \code{alphahull} R package
is available.
\item \code{"alphahull"} - use \code{ashape()} which is the preferred
method, in fact the only available option that will allow a concave
shape in the output.
\item \code{"igraph"} - calls hidden function \code{igraph:::convex_hull()} as
used when drawing \code{mark.groups} around grouped nodes.
\item \code{"sf"} - calls \code{sf::st_convex_hull()}, with same effective output
as \code{"igraph"}.
\item \code{"chull"} - calls \code{grDevices::chull()}, again with same effective
output as \code{"igraph"}, but with benefit of not incurring additional
R package dependencies.
}}

\item{smooth}{\code{logical} indicating whether to smooth the final polygon
using \code{graphics::xspline()}.}

\item{label_preset}{\code{character} (default \code{NULL}) indicating the side
to place a label, when \code{label} is provided.
Recognized values: \verb{"bottom", "top", "left", "right"}.
When \code{NULL} it detects the offset from the plot center.}

\item{label_adj_preset}{\code{character} (default label_preset) indicating
the label adjustment relative to the position of the label. In
most cases it should equal \code{label_preset}.}

\item{min_points}{\code{integer} minimum points to use, default 1 will
create a point hull even around only 1 point. To require at least
3 points, use \code{min_points=3}.}

\item{xy_range}{\code{numeric} range, default NULL, to define the plot range,
when coordinates in \code{x} are not sufficient to describe the span
of the plot. It assumes 1:1 aspect ratio.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}

\item{color, border}{\code{character} colors used when \code{do_plot=TRUE} to draw
the resulting hull polygon.}
}
\value{
\code{numeric} matrix with polygon coordinates, where
each polygon is separated by one row that contains \code{NA}
values. This output is sufficient for vectorized plotting
in base R graphics using \code{graphics::polygon()}.
}
\description{
Make alpha hull from points
}
\details{
This function makes an alpha hull around points, calling
\code{ashape()} then piecing together the somewhat
random set of outer edges into a coherent polygon.
}
\examples{
set.seed(12)
n <- 22
xy <- cbind(x=sample(seq_len(n), size=n, replace=TRUE),
   y=sample(seq_len(n), size=n, replace=TRUE));
xy <- rbind(xy, xy[1, , drop=FALSE])
x4 <- sf::st_multipoint(xy)

plot(x4, col="red", pch=20, cex=3,
   main="hull_method='ahull'")
phxy <- make_point_hull(x=xy,
   do_plot=TRUE,
   label="ahull",
   min_points=24,
   add=TRUE, xpd=TRUE)
   
# test single-point hull
phxy1 <- make_point_hull(x=head(xy, 1),
   do_plot=TRUE,
   xy_range=par("usr"),
   label="ahull,\nsinglet",
   col="#4169E144",
   border="royalblue",
   add=TRUE, xpd=TRUE)

# test doublet hull
phxy2 <- make_point_hull(x=xy[c(5, 22), ],
   do_plot=TRUE,
   xy_range=par("usr"),
   label="ahull,\nsinglet",
   col="#A020F055",
   border="#A020F0",
   add=TRUE, xpd=TRUE)

# test triplet hull
phxy3 <- make_point_hull(x=xy[c(11, 18), ],
   do_plot=TRUE,
   xy_range=par("usr"),
   label="ahull,\nsinglet",
   col="#A020F055",
   border="#A020F0",
   add=TRUE, xpd=TRUE)

plot(x4, col="red", pch=20, cex=3,
   main="hull_method='chull'")
phxy2 <- make_point_hull(x=xy, expand=0.05, do_plot=TRUE,
   add=TRUE, verbose=TRUE, xpd=TRUE, hull_method="chull")

}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{curateIPAcolnames}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{summarize_node_spacing}()},
\code{\link{with_ht_opts}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
