% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\docType{class}
\name{Mem-class}
\alias{Mem-class}
\alias{Mem}
\alias{[,Mem,ANY,ANY,ANY-method}
\alias{list_to_Mem}
\alias{as}
\alias{coerce}
\alias{Mem_to_list}
\alias{enrichments,Mem-method}
\alias{names,Mem-method}
\alias{enrichments<-,Mem-method}
\alias{sets,Mem-method}
\alias{sets<-,Mem-method}
\alias{genes,Mem-method}
\alias{genes<-,Mem-method}
\alias{geneIM,Mem-method}
\alias{geneIMdirection,Mem-method}
\alias{geneIMcolors,Mem-method}
\alias{enrichList,Mem-method}
\alias{enrichIM,Mem-method}
\alias{enrichIMcolors,Mem-method}
\alias{enrichIMdirection,Mem-method}
\alias{enrichIMgeneCount,Mem-method}
\alias{memIM,Mem-method}
\alias{geneHitIM,Mem-method}
\alias{headers,Mem-method}
\alias{colorV,Mem-method}
\alias{colorV<-,Mem-method}
\alias{thresholds,Mem-method}
\alias{dim,Mem-method}
\alias{dimnames,Mem-method}
\title{Mem class}
\usage{
\S4method{[}{Mem,ANY,ANY,ANY}(x, i, j, k, drop = FALSE)

list_to_Mem(mem)

Mem_to_list(x, ...)

\S4method{enrichments}{Mem}(x)

\S4method{names}{Mem}(x)

\S4method{enrichments}{Mem}(x) <- value

\S4method{sets}{Mem}(x, ...)

\S4method{sets}{Mem}(x) <- value

\S4method{genes}{Mem}(x, ...)

\S4method{genes}{Mem}(x) <- value

\S4method{geneIM}{Mem}(x, ...)

\S4method{geneIMdirection}{Mem}(x, ...)

\S4method{geneIMcolors}{Mem}(x, ...)

\S4method{enrichList}{Mem}(x, ...)

\S4method{enrichIM}{Mem}(x, ...)

\S4method{enrichIMcolors}{Mem}(x, ...)

\S4method{enrichIMdirection}{Mem}(x, ...)

\S4method{enrichIMgeneCount}{Mem}(x, ...)

\S4method{memIM}{Mem}(x, ...)

\S4method{geneHitIM}{Mem}(x, ...)

\S4method{headers}{Mem}(x)

\S4method{colorV}{Mem}(x, ...)

\S4method{colorV}{Mem}(x) <- value

\S4method{thresholds}{Mem}(x, ...)

\S4method{dim}{Mem}(x)

\S4method{dimnames}{Mem}(x)
}
\arguments{
\item{x}{\code{Mem} object}

\item{i}{\code{character} or \code{integer} with gene names}

\item{j}{\code{character} or \code{integer} with pathway names}

\item{drop}{\code{logical} always set to FALSE for Mem objects.}

\item{mem}{\code{list} output from \code{multiEnrichMap()}}

\item{...}{additional arguments are ignored}

\item{value}{\code{character} vector of names to assign to pathway gene sets}
}
\value{
\code{list_to_Mem()} returns a \code{Mem} S4 object, from 'list' or 'Mem' input.

\code{Mem_to_list()} returns a \code{list} suitable for ther mem functions.
}
\description{
Mem class containing results from \code{multiEnrichMap()}.
}
\details{
\subsection{Slots}{
\itemize{
\item \strong{enrichList}: \code{list} of \code{enrichResults} named by enrichment test.
\item \strong{enrichLabels}: \code{character} vector with custom enrichment labels,
named by enrichment test.
\item \strong{colorV}: \code{character} vector of R colors, named by enrichment test.
\item \strong{geneHitList}: \code{list} of vectors named by enrichment test.
Vectors are either:\if{html}{\out{<br>}}
\code{character} vector of genes tested as hits for enrichment, or\if{html}{\out{<br>}}
\code{integer} vector named by genes with values indicating
the directionality, using \code{1} or \code{-1}.
\item \strong{geneHitIM}: \code{numeric} matrix with enrichment tests as colnames, and
genes as rownames.\if{html}{\out{<br>}}
Values are \code{1} or \code{0} to indicate gene hits tested,\if{html}{\out{<br>}}
optionally \code{1} or \code{-1} to indicate directionality.
\item \strong{memIM}: \code{numeric} matrix with gene sets / pathways as colnames, and
genes as rownames.\if{html}{\out{<br>}}
Values are \code{1} or \code{0} indicating genes present in each pathway.
\item \strong{geneIM}: \code{numeric} matrix with enrichment test as colnames, and genes
as rownames.\if{html}{\out{<br>}}
This matrix differs from 'geneHitIM' in that it only includes
genes also in 'memIM', and in that order.
\item \strong{enrichIM}: \code{numeric} matrix with enrichment test as colnames, and
gene sets / pathways as rownames.\if{html}{\out{<br>}}
Matrix values are enrichment P-values,
using the appropriate adjusted P-value or FDR to represent significance.
\item \strong{multiEnrichDF}: \code{data.frame} with single combined enrichment table.\if{html}{\out{<br>}}
Data include union of genes involved in each pathway, and
lowest P-value observed across all enrichment tests.
\item \strong{multiEnrichResult}: \code{enrichResult} equivalent of 'multiEnrichDF'.\if{html}{\out{<br>}}
It is intended to facilitate use in \code{enrichplot} functions.
\item \strong{thresholds}: \code{list} of thresholds used when calling \code{multiEnrichMap()}.
\item \strong{headers}: \code{list} to associate actual colnames (headers) in 'enrichList'
with the conceptual names used in multienrichjam.
\item \strong{geneIMcolors}: \code{character} matrix with enrichment test colnames, and
gene rownames.\if{html}{\out{<br>}}
Values are colors assigned using 'colorV'.
\item \strong{geneIMdirection}: \code{numeric} matrix equivalent to 'geneIM' with values
\code{1} and \code{-1} indicating directionality.\if{html}{\out{<br>}}
When no directionality is provided, all values will be \code{1} or \code{0}.
\item \strong{enrichIMcolors}: \code{character} matrix with enrichment test colnames, and
gene set / pathway rownames.\if{html}{\out{<br>}}
Values are color gradients using the
appropriate P-value threshold used with \code{multiEnrichMap()}.
\item \strong{enrichIMdirection}: \code{numeric} matrix equivalent to 'enrichIM' with values
indicating directional score.\if{html}{\out{<br>}}
Directional scores may include a IPA z-score or other directional
score where available.\if{html}{\out{<br>}}
When not available, all values are \code{1}.
\item \strong{enrichIMgeneCount}: \code{integer} matrix equivalent to 'enrichIM'.\if{html}{\out{<br>}}
Values are the number of genes involved in the enrichment of each pathway.
}
}
}
\section{Functions}{
\itemize{
\item \code{x[i}: Subset a Mem object, \code{Mem[i, j, k]},
where \code{i} are genes, \code{j} are pathways, and \code{k} are enrichments.

\item \code{list_to_Mem()}: Convert legacy \code{list} mem format to S4 \code{Mem}

\item \code{Mem_to_list()}: Coerce S4 \code{Mem} to legacy \code{list} mem format

\item \code{enrichments(Mem)}: Names for the enrichment tests in a Mem object

\item \code{names(Mem)}: Names for the enrichment tests in a Mem object

\item \code{enrichments(Mem) <- value}: Assign names for the enrichment tests in a Mem object

\item \code{sets(Mem)}: List pathway gene set names in a Mem object

\item \code{sets(Mem) <- value}: Assign pathway gene set names to a Mem object

\item \code{genes(Mem)}: List genes represented

\item \code{genes(Mem) <- value}: Assign gene names in a Mem object

\item \code{geneIM(Mem)}: The gene-enrichment matrix of genes represented in
enrichment tests.

\item \code{geneIMdirection(Mem)}: The matrix of genes tested versus enrichment tests, with directionality.

\item \code{geneIMcolors(Mem)}: The matrix of colors indicating genes in each enrichment test

\item \code{enrichList(Mem)}: The list of enrichResult data in an Mem object

\item \code{enrichIM(Mem)}: The pathway/P-value matrix

\item \code{enrichIMcolors(Mem)}: The pathway/directional enrichment color matrix

\item \code{enrichIMdirection(Mem)}: The pathway-enrichment directional score matrix

\item \code{enrichIMgeneCount(Mem)}: Pathway-enrichment gene count matrix, genes involved
in enrichment of each pathway, for each enrichment test.

\item \code{memIM(Mem)}: The gene-pathway incidence matrix

\item \code{geneHitIM(Mem)}: The matrix of genes tested for enrichment, including
genes not associated with enrichment results.

\item \code{headers(Mem)}: Enrichment data column headers associated to
enrichResult data in a Mem object

\item \code{colorV(Mem)}: List colors assigned to  represented

\item \code{colorV(Mem) <- value}: Assign colors to enrichments in a Mem object. If
supplied 'value' matches the number of 'enrichments(x)', colors are
assigned in the order supplied.
Otherwise, 'names(value)' must be present in 'enrichments(x)',
and values are assigned in matching order.

\item \code{thresholds(Mem)}: List thresholds defined

\item \code{dim(Mem)}: dimensions in order of genes, sets, and enrichments

\item \code{dimnames(Mem)}: dimension names for each genes, sets, and enrichments

}}
\examples{
# See vignettes for a full walk-through

# list_to_Mem examples
data(Memtest)
mem <- as(Memtest, "list")
Mem <- list_to_Mem(mem)

# examples for Mem_to_list and as(Mem, "list")
data(Memtest)
mem1 <- Mem_to_list(Memtest)
mem2 <- as(Memtest, "list")
identical(mem1, mem2)

}
\seealso{
Other Mem: 
\code{\link{Memtest}},
\code{\link{check_Mem}()}
}
\concept{Mem}
