% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\docType{class}
\name{Mem-class}
\alias{Mem-class}
\title{Mem class}
\description{
Mem class containing results from \code{multiEnrichMap()}.
}
\details{
\subsection{Slots}{
\itemize{
\item enrichList
\item enrichLabels
\item colorV
\item geneHitList
\item geneHitIM
\item memIM
\item geneIM
\item enrichIM
\item multiEnrichDF
\item multiEnrichResult
\item thresholds
\item colnames
\item geneIMcolors
\item geneIMdirection
\item enrichIMcolors
\item enrichIMdirection
\item enrichIMgeneCount
}
}
}
\examples{
# See vignettes for a full walk-through

}
\seealso{
Other Mem: 
\code{\link{check_Mem}()},
\code{\link{list_to_Mem}()}
}
\concept{Mem}
