% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{Mem-slots}
\alias{Mem-slots}
\title{Mem description of slots}
\description{
Mem description of slots, for reference, slots may change at any time
}
\details{
Note that the Mem S4 object is intended to be used via the accessor
functions as described in \code{Mem-class}.

However, for reference, and in case \emph{really necessary} the slots are
described in more detail below. These slots may change at any time,
as the 'Mem' S4 object evolves after more usage, so relying upon
specific slots to be stable may not be ideal.

The descriptions below should be edited whenever the 'Mem' S4 slot
structure changes, but be prepared there may sometimes be some delay
even under the best of intentions.
}
\section{Enrichment data input}{
\itemize{
\item \strong{enrichList}: \code{list} of \code{enrichResults} named by enrichment test.
This list represents the input enrichment data.
\item \strong{enrichLabels}: \code{character} vector with custom enrichment labels,
named by enrichment test.
\item \strong{geneHitList}: \code{list} of vectors named by enrichment test.
Vectors are either:\if{html}{\out{<br>}}
\code{character} vector of genes tested as hits for enrichment, or\if{html}{\out{<br>}}
\code{integer} vector named by genes with values indicating
the directionality, using \code{1} or \code{-1}.
\item \strong{geneHitIM}: \code{numeric} matrix with enrichment tests as colnames, and
genes as rownames.\if{html}{\out{<br>}}
Values are \code{1} or \code{0} to indicate gene hits tested,\if{html}{\out{<br>}}
optionally \code{1} or \code{-1} to indicate directionality.
}
}

\section{Incidence matrix data}{
\itemize{
\item \strong{memIM}: \code{numeric} matrix with gene sets / pathways as colnames, and
genes as rownames.\if{html}{\out{<br>}}
Values are \code{1} or \code{0} indicating genes present in each pathway.
\item \strong{geneIM}: \code{numeric} matrix with enrichment test as colnames, and genes
as rownames.\if{html}{\out{<br>}}
This matrix differs from 'geneHitIM' in that it only includes
genes also in 'memIM', and in that order.
\item \strong{enrichIM}: \code{numeric} matrix with enrichment test as colnames, and
gene sets / pathways as rownames.\if{html}{\out{<br>}}
Matrix values are enrichment P-values,
using the appropriate adjusted P-value or FDR to represent significance.
}
}

\section{Combined enrichment data}{
Enrichment data are combined into singular objects, mainly intended
to be used with \code{clusterProfiler} related functions.
Also the \code{data.frame} could be used as a supplemental table.
\itemize{
\item \strong{multiEnrichDF}: \code{data.frame} with single combined enrichment table.\if{html}{\out{<br>}}
Data include union of genes involved in each pathway, and
lowest P-value observed across all enrichment tests.
\item \strong{multiEnrichResult}: \code{enrichResult} equivalent of 'multiEnrichDF'.\if{html}{\out{<br>}}
It is intended to facilitate use in \code{enrichplot} functions.
}
\subsection{Analysis details}{
\itemize{
\item \strong{colorV}: \code{character} vector of R colors, named by enrichment test.
\item \strong{thresholds}: \code{list} of thresholds used when calling \code{multiEnrichMap()}.
\itemize{
\item \strong{p_cutoff}: \code{numeric} P-value threshold for significant enrichment.
\item \strong{min_count}: \code{integer} number of genes involved in a pathway
for the pathway to be considered significant.
}
\item \strong{headers}: \code{list} to associate actual colnames (headers) in 'enrichList'
with the conceptual names used in multienrichjam.
}
}

\subsection{Additional incidence matrix data}{
\itemize{
\item \strong{geneIMcolors}: \code{character} matrix with enrichment test colnames, and
gene rownames.\if{html}{\out{<br>}}
Values are colors assigned using 'colorV'.
\item \strong{geneIMdirection}: \code{numeric} matrix equivalent to 'geneIM' with values
\code{1} and \code{-1} indicating directionality.\if{html}{\out{<br>}}
When no directionality is provided, all values will be \code{1} or \code{0}.
\item \strong{enrichIMcolors}: \code{character} matrix with enrichment test colnames, and
gene set / pathway rownames.\if{html}{\out{<br>}}
Values are color gradients using the
appropriate P-value threshold used with \code{multiEnrichMap()}.
\item \strong{enrichIMdirection}: \code{numeric} matrix equivalent to 'enrichIM' with values
indicating directional score.\if{html}{\out{<br>}}
Directional scores may include a IPA z-score or other directional
score where available.\if{html}{\out{<br>}}
When not available, all values are \code{1}.
\item \strong{enrichIMgeneCount}: \code{integer} matrix equivalent to 'enrichIM'.\if{html}{\out{<br>}}
Values are the number of genes involved in the enrichment of each pathway.
}
}
}

\seealso{
Other Mem: 
\code{\link{Mem-class}},
\code{\link{Mem-data}},
\code{\link{check_Mem}()}
}
\concept{Mem}
