% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-mem2cnet.R
\name{mem2cnet}
\alias{mem2cnet}
\alias{memIM2cnet}
\title{Convert MultiEnrichment incidence matrix to Cnet plot}
\usage{
mem2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  forceColors = FALSE,
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 12,
  geneSize = 6,
  categoryCex = 1,
  geneCex = 0.8,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col_fn = NULL,
  direction_cutoff = NULL,
  direction_max = 2,
  gene_direction_cutoff = 0,
  gene_direction_max = 1.2,
  gene_direction_col_fn = NULL,
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  remove_singlet_genes = TRUE,
  do_reorder = TRUE,
  spread_labels = FALSE,
  vertex.label.font = 2,
  use_shadowText = TRUE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)

memIM2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  forceColors = FALSE,
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 12,
  geneSize = 6,
  categoryCex = 1,
  geneCex = 0.8,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col_fn = NULL,
  direction_cutoff = NULL,
  direction_max = 2,
  gene_direction_cutoff = 0,
  gene_direction_max = 1.2,
  gene_direction_col_fn = NULL,
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  remove_singlet_genes = TRUE,
  do_reorder = TRUE,
  spread_labels = FALSE,
  vertex.label.font = 2,
  use_shadowText = TRUE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{memIM}{one of the following:
\itemize{
\item \code{Mem} S4 object, preferred. In this case, the arguments regarding
'geneIM' and 'enrichIM' data are taken directly from 'mem'.
\item legacy \code{list} mem object, for backward compatibility,
\item \code{numeric} matrix in the form of a \code{memIM} gene-pathway incidence
matrix. In this case, other arguments involving \code{geneIM} and \code{enrichIM}
matrices are required for correct behavior of this function.
When \code{mem} format is supplied, relevant arguments which are empty will
use corresponding data from \code{mem}, for example \code{geneIM}, \code{geneIMcolors},
\code{enrichIM}, \code{enrichIMcolors}.
}}

\item{categoryShape, geneShape}{\code{character} string with node shape,
default 'pie' uses pie nodes.
Note that 'pie' shapes with only one segment are converted to
'circle' for convenience.
When using \code{jam_igraph()} to plot, it already treats single-segment
pie nodes as circle anyway, and vectorizes pie node plotting to
make the overall rendering time substantially faster.}

\item{forceColors}{\code{logical} default FALSE, whether to force node colors
to use \code{categoryColor} and \code{geneColor} arguments, otherwise colors
are defined using \code{enrichIMcolors()} and \code{geneIMcolors()},
respectively.
Enabling this option will skip all pie and directional styling.}

\item{categoryColor, geneColor}{\code{character} R color for default node colors,
used when \code{geneIMcolors},\code{enrichIMcolors} is not supplied, respectively.}

\item{categoryLabelColor, geneLabelColor}{\code{character} R color used as
default node label color.}

\item{categorySize, geneSize}{\code{numeric} default node size.}

\item{categoryCex, geneCex}{\code{numeric} adjustment to default node label
font size.}

\item{frame_darkFactor}{\code{numeric} passed to \code{jamba::makeColorDarker()}
so the frame color is slightly darker than the node fill color.}

\item{geneIM, geneIMcolors, geneIMdirection, enrichIM, enrichIMcolors, enrichIMdirection}{\code{matrix} data used only when the input 'mem' is not an \code{Mem} or legacy
\code{list} mem object which already contains these data.
When input 'mem' is supplied as a \code{matrix}, these value enable
this function to operate on nearly any custom data.
\itemize{
\item \code{geneIMcolors} is used to define gene (row) node colors
\item \code{enrichIMcolors} is used to define set (column) node colors
\item \code{geneIMdirection},\code{enrichIMdirection} is used to define optional
border colors defined by the direction, where -1 is down,
0 is no change, and +1 is up.
Use \code{direction_col_fn} to define a custom color function, however the
default uses the reversed "RdBu" Brewer color ramp with blue (down),
white (no change) and red (up).
}}

\item{coloredrect_nrow, coloredrect_ncol, coloredrect_byrow}{arguments
used when \code{geneShape="coloredrectangle"}, to define layout and
placement of colors across columns in \code{geneIMcolors}. By default,
one row of colors is used.}

\item{colorV}{\code{character} optional vector of R colors, taken from 'mem'
for \code{Mem} or legacy \code{list} mem objects.
When supplied, it should be a vector named to match
\code{colnames(enrichIM)}.
When defined, these colors override \code{enrichIMcolors}.}

\item{direction_col_fn}{\code{function} used to colorize 'Set' node borders
via enrichIMdirection. The default uses \code{colorjam::col_div_xf()}
which applies reverse Brewer "RdBu" for blue (down), white (no change),
and red (up) with maximum color at \code{2.0}.
When not supplied, \code{direction_cutoff} and \code{direction_max} are used
with \verb{colorjam::col_div_xf().}}

\item{direction_cutoff}{\code{numeric} default is taken from \code{memIM} when
provided as \code{Mem}, typical default is \code{1.0} to require directional
z-score at least \code{1.0} in order to apply any directional border
color.}

\item{direction_max}{\code{numeric} default 2, the numeric value to apply
maximum color from the color gradient function.}

\item{gene_direction_cutoff}{\code{numeric} default 0, the minimum value
in geneIMdirection to apply up/down color. Using 0 will colorize
any non-zero color.}

\item{gene_direction_max}{\code{numeric} default 1.2, the numeric value to apply
maximum color from the color gradient function. Typical values
are \code{-1} and \code{1}, so using \code{1.2} will apply nearly the maximum color.}

\item{gene_direction_col_fn}{\code{function} used to colorize 'Gene' node borders
via enrichIMdirection. The default uses \code{colorjam::col_div_xf()}
which applies reverse Brewer "RdBu" for blue (down), white (no change),
and red (up) with maximum color at \code{1.2}.
*When not supplied, \code{gene_direction_cutoff} and \code{gene_direction_max}
are used with \verb{colorjam::col_div_xf().}}

\item{hide_solo_pie}{\code{logical} default TRUE, passed to
\code{apply_cnet_direction()} to determine whether to display border
only as one outer frame color when all colors are identical.
When FALSE, all pie wedges are individually colored.}

\item{remove_blanks}{\code{logical} default TRUE, whether to remove blank
color subsections from each node, using \code{removeIgraphBlanks()}.
This argument is useful for 'pie', 'jampie', or 'coloredrectangle'
node shapes, so they will only indicate the relevant color.}

\item{remove_singlet_genes}{\code{logical} default TRUE, whether to remove
singlet genes, which are genes (rows) not represented in any pathway
gene sets (columns).}

\item{do_reorder}{\code{logical} default TRUE used to re-order nodes
in "equivalent positions" by border color, color, etc. by calling
\code{reorder_igraph_nodes()}. Equivalent nodes are common in Cnet
(bipartite) networks, for example numerous 'Gene' nodes may
be connected to the same 2 pathway 'Set' nodes, and therefore
the node positions are interchangeable. Sorting by node attributes
(notably color, then alphabetically by label) helps visual
review.}

\item{spread_labels}{\code{logical} default FALSE, whether to spread node
labels away from incoming edges, also adding label distance via
vertex attribute 'label.dist'.
This step calls \code{spread_igraph_labels()}, and can be customized
further by passing arguments through '...' ellipses.}

\item{vertex.label.font}{\code{integer} default 2, used to set default font
to use bold face (font=2) for labels. Use 1 for normal font.}

\item{use_shadowText}{\code{logical} default TRUE, applied to graph
attributes to apply shadowText by default when plotting with
\code{jam_igraph()}.}

\item{repulse}{\code{numeric} value passed to \code{layout_with_qfr()}
when either \code{do_reorder} or \code{spread_labels} is TRUE. Otherwise
there is no layout applied.
The 'repulse' value effectively defines node spacing, usually
with values between 3 and 4, which 3 giving broad spacing,
and 4 giving very close spacing of nodes in equivalent network
positions, and broad spacing otherwise. Higher values tend
to "clump" nodes closer together.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to downstream functions:
\itemize{
\item when \code{remove_blanks=TRUE} it is passed to \code{removeIgraphBlanks()}
\item when \code{spread_labels=TRUE} it is passed to \code{spread_igraph_labels()}
}}
}
\value{
\code{igraph} object with Concept network data, containing
pathways connected to genes. Each node has attribute \code{"nodeType"}
of either \code{"Set"} or \code{"Gene"}.
}
\description{
Convert MultiEnrichment incidence matrix to Cnet plot
}
\details{
This function takes \code{Mem} S4 object, or legacy \code{list} mem data,
or even bare \code{matrix} data, and produces a Concept network (Cnet)
in the form of an \code{igraph} object.

Cnet network data are a form of bipartite graph,
where every node is either 'Gene' or 'Set', and where nodes
are only connected across \verb{'Gene' <--> 'Set'}.
See \code{igraph::graph_from_biadjacency_matrix()}.
A Cnet plot has particularly useful characteristic that
Set (pathway) nodes tend to be connected to very many Gene nodes,
creating an imbalance which is useful when defining a visual layout.

A vertex attribute \code{'nodeType'} stores the type of node: 'Gene' or 'Set'.

The \code{memIM} defines the network, and non-zero value constitutes
an edge (connection) between row (Gene) and column (Set).

Other data are associated with nodes when available.
\enumerate{
\item \code{geneIM}, \code{geneIMcolors}, \code{geneIMdirection}:
define 'Gene' node fill and border colors.
\item \code{enrichIM}, \code{enrichIMcolors}, \code{enrichIMdirection}:
define 'Set' node fill and border colors.
}

Everything else helps customize the output, and can be used as-is.
\subsection{Generic Cnet}{

The \code{memIM} can be supplied as an \code{matrix} of \code{integer} values, with
no other data supplied, and this is sufficient to create a Cnet plot.
Therefore, any incidence matrix can be used to create a network.

If no other data are provided, nodes are colored by
\code{categoryColor} (Set) and \code{geneColor} (Gene), and shapes defined
by \code{categoryShape} and \code{geneShape}.

Be aware that \code{remove_singlet_genes=TRUE} default will hide
rows (Genes) which have no connection to any columns (Sets).
This default ensures that an incidence matrix can be subset
for columns (Sets) of interest without also filtering rows.

However, columns (Sets) are not removed, because they typically are
not empty in the most common workflows.
Singlet columns, Set nodes with no connected Gene nodes, can be
filtered by calling \code{removeIgraphSinglets()}.
}
}
\examples{
use_sets <- c("eNOS Signaling",
   "Growth Hormone Signaling",
   "mTOR Signaling")
jam_igraph(mem2cnet(Memtest[, use_sets, ]),
   use_shadowText=TRUE)

}
\seealso{
Other jam Mem utilities:
\code{\link{Mem-class}},
\code{\link{MemPlotFolio-class}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=mem2emap]{mem2emap()}}

Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=rectifyPiegraph]{rectifyPiegraph()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam Mem utilities}
\concept{jam igraph functions}
