% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-mem2cnet.R
\name{mem2cnet}
\alias{mem2cnet}
\alias{memIM2cnet}
\title{Convert MultiEnrichment incidence matrix to Cnet plot}
\usage{
mem2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 12,
  geneSize = 6,
  categoryCex = 1,
  geneCex = 0.8,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col = colorjam::col_div_xf(1.2),
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  remove_singlet_genes = TRUE,
  spread_labels = FALSE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)

memIM2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 12,
  geneSize = 6,
  categoryCex = 1,
  geneCex = 0.8,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col = colorjam::col_div_xf(1.2),
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  remove_singlet_genes = TRUE,
  spread_labels = FALSE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{categoryShape, geneShape}{\code{character} string indicating the
preferred default node shape. In general, \code{pie} shapes with one segment
are converted to \code{circle} for simplicity during plot functions,
although when using \code{jam_igraph()} to plot, it will treat single-segment
pie nodes as circle anyway, and vectorizes pie node plotting to
make the overall rendering time substantially faster.}

\item{categoryColor, geneColor}{\code{character} R color for default node colors,
used when \code{geneIMcolors},\code{enrichIMcolors} is not supplied, respectively.}

\item{categoryLabelColor, geneLabelColor}{\code{character} R color used as
default node label color.}

\item{categorySize, geneSize}{\code{numeric} default node size.}

\item{categoryCex, geneCex}{\code{numeric} adjustment to default node label
font size.}

\item{frame_darkFactor}{\code{numeric} passed to \code{jamba::makeColorDarker()}
so the frame color is slightly darker than the node fill color.}

\item{geneIM, geneIMcolors, geneIMdirection, enrichIM, enrichIMcolors, enrichIMdirection}{\code{matrix} data used only when the input 'mem' is not an \code{Mem} or legacy
\code{list} mem object which already contains these data.
When input 'mem' is supplied as a \code{matrix}, these value enable
this function to operate on nearly any custom data.
\itemize{
\item \code{geneIMcolors} is used to define gene (row) node colors
\item \code{enrichIMcolors} is used to define set (column) node colors
\item \code{geneIMdirection},\code{enrichIMdirection} is used to define optional
border colors defined by the direction, where -1 is down,
0 is no change, and +1 is up.
Use \code{direction_col} to define a custom color function, however the
default uses the reversed "RdBu" Brewer color ramp with blue (down),
white (no change) and red (up).
}}

\item{coloredrect_nrow, coloredrect_ncol, coloredrect_byrow}{arguments
used when \code{geneShape="coloredrectangle"}, to define layout and
placement of colors across columns in \code{geneIMcolors}. By default,
one row of colors is used.}

\item{colorV}{\code{character} optional vector of R colors, taken from 'mem'
for \code{Mem} or legacy \code{list} mem objects.
When supplied, it should be a vector named to match
\code{colnames(enrichIM)}.
When defined, these colors override \code{enrichIMcolors}.}

\item{direction_col}{\code{function} used to colorize node borders based
upon directionality. The default uses \code{colorjam::col_div_xf()}
which applies reverse Brewer "RdBu" for blue (down), white (no change),
and red (up) with threshold \code{1.2}.}

\item{hide_solo_pie}{\code{logical} default TRUE, passed to
\code{apply_cnet_direction()} to determine whether to display border
only as one outer frame color when all colors are identical.
When FALSE, all pie wedges are individually colored.}

\item{remove_blanks}{\code{logical} default TRUE, whether to remove blank
color subsections from each node, using \code{removeIgraphBlanks()}.
This argument is useful for 'pie', 'jampie', or 'coloredrectangle'
node shapes, so they will only indicate the relevant color.}

\item{remove_singlet_genes}{\code{logical} default TRUE, whether to remove
singlet genes, which are genes (rows) not represented in any pathway
gene sets (columns).}

\item{spread_labels}{\code{logical} default FALSE, whether to spread node
labels away from incoming edges, also adding label distance via
vertex attribute 'label.dist'.
This step calls \code{spread_igraph_labels()}, and can be customized
further by passing arguments through '...' ellipses.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to downstream functions:
\itemize{
\item when \code{remove_blanks=TRUE} it is passed to \code{removeIgraphBlanks()}
\item when \code{spread_labels=TRUE} it is passed to \code{spread_igraph_labels()}
}}

\item{mem}{one of the following:
\itemize{
\item \code{Mem} S4 object, preferred. In this case, the arguments regarding
'geneIM' and 'enrichIM' data are taken directly from 'mem'.
\item legacy \code{list} mem object, for backward compatibility,
\item \code{numeric} matrix in the form of a \code{memIM} gene-pathway incidence
matrix. In this case, other arguments involving \code{geneIM} and \code{enrichIM}
matrices are required for correct behavior of this function.
When \code{mem} format is supplied, relevant arguments which are empty will
use corresponding data from \code{mem}, for example \code{geneIM}, \code{geneIMcolors},
\code{enrichIM}, \code{enrichIMcolors}.
}}
}
\value{
\code{igraph} object with Concept network data, containing
pathways connected to genes. Each node has attribute \code{"nodeType"}
of either \code{"Set"} or \code{"Gene"}.
}
\description{
Convert MultiEnrichment incidence matrix to Cnet plot
}
\details{
This function takes \code{list} output from \code{multiEnrichMap()}
and uses three elements to create a Cnet plot:
\enumerate{
\item \code{"memIM"} gene-pathway incidence matrix
\item \code{"geneIM"} the gene incidence matrix
\item \code{"enrichIM"} the pathway enrichment matrix
}
}
\examples{
use_sets <- c("eNOS Signaling",
   "Growth Hormone Signaling",
   "mTOR Signaling")
jam_igraph(mem2cnet(Memtest[, use_sets, ]),
   use_shadowText=TRUE)

}
\seealso{
Other jam Mem utilities: 
\code{\link{Mem-class}},
\code{\link{fixSetLabels}()},
\code{\link{mem2emap}()}

Other jam igraph functions: 
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{label_communities}()},
\code{\link{mem2emap}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()}
}
\concept{jam Mem utilities}
\concept{jam igraph functions}
