% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot-mpf.R
\name{mem_do_newpage}
\alias{mem_do_newpage}
\title{Determine whether Mem should call newpage}
\usage{
mem_do_newpage(x, ...)
}
\arguments{
\item{x}{not currently implemented, it may be used
in future.}

\item{...}{additional arguments are ignored.}
}
\description{
Determine whether Mem should call newpage
}
\details{
\subsection{Rules}{
\itemize{
\item Running outside Positron, then yes. End.
\item If running inside knitr, yes. End.
\item Otherwise no.
}
}
}
\keyword{internal}
