% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_enrichment_heatmap}
\alias{mem_enrichment_heatmap}
\title{MultiEnrichment Heatmap of enrichment P-values}
\usage{
mem_enrichment_heatmap(
  mem,
  style = c("dotplot_inverted", "dotplot", "heatmap"),
  apply_direction = FALSE,
  p_cutoff = NULL,
  min_count = 1,
  p_floor = 1e-10,
  point_size_factor = 1,
  point_size_max = 8,
  point_size_min = 2,
  row_method = "euclidean",
  column_method = "euclidean",
  name = "-log10P",
  row_dend_reorder = TRUE,
  row_dend_width = grid::unit(30, "mm"),
  row_fontsize = NULL,
  row_cex = 1,
  row_split = NULL,
  row_gap = grid::unit(2, "mm"),
  cluster_rows = TRUE,
  column_fontsize = NULL,
  column_cex = 1,
  cluster_columns = FALSE,
  sets = NULL,
  color_by_column = FALSE,
  cex.axis = 1,
  lens = 3,
  cexCellnote = 1,
  column_title = NULL,
  row_names_max_width = grid::unit(300, "mm"),
  column_names_max_height = grid::unit(300, "mm"),
  heatmap_legend_param = NULL,
  hm_cell_size = NULL,
  legend_height = grid::unit(6, "cm"),
  legend_cex = 1,
  direction_cutoff = 0,
  gene_count_max = NULL,
  top_annotation = NULL,
  outline = TRUE,
  show_enrich = NULL,
  use_raster = FALSE,
  do_plot = TRUE,
  ...
)
}
\arguments{
\item{mem}{\code{Mem} or legacy \code{list} mem created by \code{multiEnrichMap()}.}

\item{style}{\code{character} string indicating the style of heatmap:
\code{"heatmap"} produces a regular heatmap, shaded by \code{log10(Pvalue)};
\code{"dotplot"} produces a dotplot, where the dot size is proportional
to the number of genes. See function description for details on
how to include the point size legend beside the heatmap.
The main benefit of using "dotplot" style is that it also indicates
the relative number of genes involved in the enrichment.}

\item{apply_direction}{\code{logical}, default FALSE, whether to define
a bivariate color scheme which uses \code{mem$enrichIMdirection}
when defined. The color scheme is intended to indicate both the
directional strength (usually with some type of z-score) and
the statistical enrichment (usually with the enrichment P-value
or adjusted P-value).}

\item{p_cutoff}{\code{numeric} value of the enrichment P-value cutoff,
by default this value is obtained from \code{mem$p_cutoff} to be
consistent with the original \code{multiEnrichMap()} analysis.
P-values above \code{p_cutoff} are not colored, and are therefore white.
This behavior is intended to indicate pathways with P-value above
this threshold did not meet the threshold, instead of
pathways with similar P-values displaying with similar color.}

\item{min_count}{\code{numeric} number of genes required for a pathway
to be considered dysregulated.}

\item{p_floor}{\code{numeric} minimum P-value used for the color gradient.
P-values below this floor are colored with the maximum color gradient.
This value is intended to be used in cases where one enrichment
P-value is very low (e.g. 1e-36) to prevent all other P-values from
being colored pale red-white and not be noticeable.}

\item{point_size_factor}{\code{numeric} used to adjust the legend point size,
since the heatmap point size is dependent upon the number of rows,
the legend may require some manual adjustment to make sure the
legend matches the heatmap.}

\item{point_size_min, point_size_max}{\code{numeric} values which define
the minimum and maximum point sizes, effectively defining the range
permitted when \code{style="dotplot"}.}

\item{row_method}{\code{character} string of the distance method
to use for row and column clustering. The clustering is performed
by \code{amap::hcluster()}.}

\item{name}{\code{character} value passed to \code{ComplexHeatmap::Heatmap()},
used as a label above the heatmap color legend.}

\item{row_dend_reorder}{\code{logical} indicating whether to reorder the
row dendrogram using the method described in
\code{ComplexHeatmap::Heatmap()}. The end result is minor reshuffling of
leaf nodes on the dendrogram based upon mean signal in each row,
which can sometimes look cleaner.}

\item{row_fontsize, column_fontsize}{optional \code{numeric} arguments passed to
\code{ComplexHeatmap::Heatmap()} to size row and column labels.}

\item{cluster_columns}{\code{logical} indicating whether to cluster heatmap
columns, by default columns are not clustered.}

\item{sets}{\code{character} vector of sets (pathways) to include in the heatmap,
all other sets will be excluded.}

\item{color_by_column}{\code{logical} indicating whether to colorize the
heatmap using \code{mem$colorV} as defined for each comparison. This
option is currently experimental, and produces a base R heatmap
using \code{jamba::imageByColors()}.}

\item{cex.axis}{\code{numeric} adjustment for axis labels, passed to
\code{jamba::imageByColors()} only when \code{color_by_column=TRUE}.}

\item{lens}{\code{numeric} value used in color gradients, defining the extent
the color gradient is enhanced in the mid-ranges (positive \code{lens}),
or diminished in the mid-ranges (negative \code{lens}). There is no
quantitative standard measure for color gradient changes, so this
option is intended to help adjust and improve the visual perception
of the current data.}

\item{cexCellnote}{\code{numeric} character expansion value used only
when \code{color_by_column=TRUE}, used to adjust the P-value label size
inside each heatmap cell.}

\item{column_title}{optional \code{character} string with title to display
above the heatmap.}

\item{row_names_max_width, column_names_max_height, heatmap_legend_param}{arguments passed to \code{ComplexHeatmap::Heatmap()} and provided here
for convenience.}

\item{hm_cell_size}{\code{grid::unit} or \code{numeric}, default \code{NULL}, to define
an optional fixed heatmap cell size, useful to define absolute
square heatmap cells. When \code{numeric} it is interpreted as
"mm" units. Note that the heatmap total height is determined by
the number of cells, and the total row gaps defined by the
number of row gaps with \code{row_split} multiplied by \code{row_gap}.}

\item{legend_height}{\code{grid::unit}, default 6 cm (60 mm), to define
the absolute height of the color gradient in the color key.
This value is only used when \code{heatmap_legend_param} is not defined.}

\item{legend_cex}{\code{numeric} default 1, used to scale the legend
fontsize relative to the default fontsize \code{10}.
This value is only used when \code{heatmap_legend_param} is not defined.}

\item{top_annotation}{\code{HeatmapAnnotation} as produced by
\code{ComplexHeatmap::HeatmapAnnotation()} or \code{NULL}, used to display
customized annotation at the top of the heatmap. The order of
columns must match the order of columns in the data displayed
in the heatmap.}

\item{outline}{\code{logical} default TRUE, whether to draw an outline
for each heatmap cell. Note: The outline is not drawn for
\code{style="dotplot"} which already adds lines through the middle
of each cell, not the border of each cell.}

\item{show_enrich}{\code{numeric} default NULL, indicating which of the
enrichment metrics to show as a label in each cell.
When only one type is shown, there is no prefix, but for multiple
types, a prefix is shown for each. The metrics in order include:
\enumerate{
\item "-log10P"
\item "direction"
\item "z-score"
\item "number of genes"
}}

\item{use_raster}{\code{logical} passed to \code{ComplexHeatmap::Heatmap()}
indicating whether to rasterize the heatmap output, used when
\code{style="heatmap"}. Rasterization is not relevant to dotplot output
since dotplot is drawn using an individual circle in each heatmap cell.}

\item{do_plot}{\code{logical} indicating whether to display the plot with
\code{ComplexHeatmap::draw()} or \code{jamba::imageByColors()} as relevant.
The underlying data is returned invisibly.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for customization.}
}
\description{
MultiEnrichment Heatmap of enrichment P-values
}
\details{
Note: It is recommended to call \code{mem_plot_folio()} with \code{do_which=1}
in order to utilize the gene-pathway content during clustering,
which is more effective at clustering similar pathways by gene
content. Otherwise pathways are clustered using only the
\code{-log10(p)} enrichment P-value.

This function is a lightweight wrapper to \code{ComplexHeatmap::Heatmap()}
intended to visualize the enrichment P-values from multiple
enrichment results. The P-value threshold is used to colorize
every cell whose P-value meets the threshold, while all other
cells are therefore white.
\subsection{Drawing Dotplot with Point Legend}{

The \code{style} argument controls whether a heatmap or dotplot is
created.
\itemize{
\item \code{style="dotplot"}: each heatmap cell is not filled, and the color
is drawn as a circle with size proportional to the number of
genes involved in enrichment. A separate point legend is returned
as an attribute of the heatmap object.
\item \code{style="dotplot_inverted"}: each heatmap cell is filled, and
a circle is drawn with size proportional to the number of
genes involved in enrichment. A separate point legend is returned
as an attribute of the heatmap object.
}

To draw the dotplot heatmap including the point legend,
use this command:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{ComplexHeatmap::draw(hm,
   annotation_legend_list=attr(hm, "annotation_legend_list"))
}\if{html}{\out{</div>}}

Generally, the clustering
using the gene-pathway incidence matrix is more effective at
representing biologically-driven pathway clusters.
}
}
\seealso{
Other custom plot functions: 
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{plot_mpf}()}
}
\concept{custom plot functions}
