% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_enrichment_heatmap}
\alias{mem_enrichment_heatmap}
\title{MultiEnrichment Heatmap of enrichment P-values}
\usage{
mem_enrichment_heatmap(
  mem,
  p_cutoff = 0.05,
  p_floor = 1e-06,
  row_method = "euclidean",
  name = "-log10P",
  row_dend_reorder = TRUE,
  row_fontsize = 8,
  column_fontsize = 12,
  cluster_columns = FALSE,
  sets = NULL,
  color_by_column = FALSE,
  cex.axis = 1,
  column_title = NULL,
  ...
)
}
\arguments{
\item{mem}{\code{list} object created by \code{multiEnrichMap()}. Specifically
the object is expected to contain \code{enrichIM}.}

\item{p_cutoff}{numeric value of the enrichment P-value cutoff,
above which P-values are not colored, and are therefore white.
The enrichment P-values are displayed as an annotated heatmap
at the top of the main heatmap. Any cell that has a color meets
at least the minimum P-value threshold.}

\item{p_floor}{numeric minimum P-value used for the color gradient.
P-values below this floor are colored with the maximum color gradient.}

\item{row_method}{character string of the distance method
to use for row and column clustering. The clustering is performed
by \code{amap::hcluster()}.}

\item{name}{character value passed to \code{ComplexHeatmap::Heatmap()},
used as a label above the heatmap color legend.}

\item{row_dend_reorder}{logical indicating whether to reorder the
row dendrogram using the method described in
\code{ComplexHeatmap::Heatmap()}. The end result is minor reshuffling of
leaf nodes on the dendrogram based upon mean signal in each row,
which can sometimes look cleaner.}

\item{row_fontsize, column_fontsize}{arguments passed to
\code{ComplexHeatmap::Heatmap()} to size row and column labels.}

\item{cluster_columns}{logical indicating whether to cluster heatmap
columns, by default columns are not clustered.}

\item{sets}{character vector of sets (pathways) to include in the heatmap,
all other sets will be excluded.}

\item{color_by_column}{logical indicating whether to colorize the
heatmap using \code{mem$colorV} as defined for each comparison. This
option is currently experimental, and produces a base R heatmap
using \code{jamba::imageByColors()}.}

\item{cex.axis}{sent to \code{jamba::imageByColors()} only when
\code{color_by_column=TRUE}.}

\item{column_title}{optional character string with title to display
above the heatmap.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for customization.}
}
\description{
MultiEnrichment Heatmap of enrichment P-values
}
\details{
This function is a lightweight wrapper to \code{ComplexHeatmap::Heatmap()}
intended to visualize the enrichment P-values from multiple
enrichment results. The P-value threshold is used to colorize
every cell whose P-value meets the threshold, while all other
cells are therefore white.
}
\seealso{
Other jam plot functions: 
\code{\link{jam_igraph}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()}
}
\concept{jam plot functions}
