% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_enrichment_heatmap}
\alias{mem_enrichment_heatmap}
\title{MultiEnrichment Heatmap of enrichment P-values}
\usage{
mem_enrichment_heatmap(
  mem,
  style = c("dotplot", "heatmap"),
  p_cutoff = mem$p_cutoff,
  min_count = 1,
  p_floor = 1e-06,
  point_size_factor = 1,
  point_size_max = 8,
  point_size_min = 1,
  row_method = "euclidean",
  column_method = "euclidean",
  name = "-log10P",
  row_dend_reorder = TRUE,
  row_dend_width = grid::unit(30, "mm"),
  row_fontsize = NULL,
  row_cex = 1,
  row_split = NULL,
  cluster_rows = TRUE,
  column_fontsize = NULL,
  column_cex = 1,
  cluster_columns = FALSE,
  sets = NULL,
  color_by_column = FALSE,
  cex.axis = 1,
  lens = 3,
  cexCellnote = 0,
  column_title = NULL,
  row_names_max_width = grid::unit(12, "cm"),
  column_names_max_height = grid::unit(12, "cm"),
  heatmap_legend_param = NULL,
  legend_height = grid::unit(6, "cm"),
  apply_direction = FALSE,
  direction_cutoff = 0,
  gene_count_max = NULL,
  top_annotation = NULL,
  show = NULL,
  use_raster = FALSE,
  do_plot = TRUE,
  ...
)
}
\arguments{
\item{mem}{\code{list} object created by \code{multiEnrichMap()}. Specifically
the object is expected to contain \code{enrichIM}.}

\item{style}{\code{character} string indicating the style of heatmap:
\code{"heatmap"} produces a regular heatmap, shaded by \code{log10(Pvalue)};
\code{"dotplot"} produces a dotplot, where the dot size is proportional
to the number of genes. See function description for details on
how to include the point size legend beside the heatmap.
The main benefit of using "dotplot" style is that it also indicates
the relative number of genes involved in the enrichment.}

\item{p_cutoff}{\code{numeric} value of the enrichment P-value cutoff,
by default this value is obtained from \code{mem$p_cutoff} to be
consistent with the original \code{multiEnrichMap()} analysis.
P-values above \code{p_cutoff} are not colored, and are therefore white.
This behavior is intended to indicate pathways with P-value above
this threshold did not meet the threshold, instead of
pathways with similar P-values displaying with similar color.}

\item{min_count}{\code{numeric} number of genes required for a pathway
to be considered dysregulated.}

\item{p_floor}{\code{numeric} minimum P-value used for the color gradient.
P-values below this floor are colored with the maximum color gradient.
This value is intended to be used in cases where one enrichment
P-value is very low (e.g. 1e-36) to prevent all other P-values from
being colored pale red-white and not be noticeable.}

\item{point_size_factor}{\code{numeric} used to adjust the legend point size,
since the heatmap point size is dependent upon the number of rows,
the legend may require some manual adjustment to make sure the
legend matches the heatmap.}

\item{row_method}{character string of the distance method
to use for row and column clustering. The clustering is performed
by \code{amap::hcluster()}.}

\item{name}{character value passed to \code{ComplexHeatmap::Heatmap()},
used as a label above the heatmap color legend.}

\item{row_dend_reorder}{logical indicating whether to reorder the
row dendrogram using the method described in
\code{ComplexHeatmap::Heatmap()}. The end result is minor reshuffling of
leaf nodes on the dendrogram based upon mean signal in each row,
which can sometimes look cleaner.}

\item{row_fontsize, column_fontsize}{arguments passed to
\code{ComplexHeatmap::Heatmap()} to size row and column labels.}

\item{cluster_columns}{logical indicating whether to cluster heatmap
columns, by default columns are not clustered.}

\item{sets}{character vector of sets (pathways) to include in the heatmap,
all other sets will be excluded.}

\item{color_by_column}{logical indicating whether to colorize the
heatmap using \code{mem$colorV} as defined for each comparison. This
option is currently experimental, and produces a base R heatmap
using \code{jamba::imageByColors()}.}

\item{cex.axis}{sent to \code{jamba::imageByColors()} only when
\code{color_by_column=TRUE}.}

\item{lens}{\code{numeric} value used in color gradients, to define whether
the color gradient should be enhanced in the mid-ranges (positive \code{lens}),
or diminished in the mid-ranges (negative \code{lens}).}

\item{cexCellnote}{numeric character expansion value used only
when \code{color_by_column=TRUE}, used to adjust the P-value label size
inside each heatmap cell.}

\item{column_title}{optional character string with title to display
above the heatmap.}

\item{row_names_max_width, column_names_max_height, heatmap_legend_param}{arguments passed to \code{ComplexHeatmap::Heatmap()}.}

\item{top_annotation}{\code{HeatmapAnnotation} as produced by
\code{ComplexHeatmap::HeatmapAnnotation()} or \code{NULL}, used to display
annotation at the top of the heatmap.}

\item{use_raster}{\code{logical} passed to \code{ComplexHeatmap::Heatmap()}
indicating whether to rasterize the heatmap output, used when
\code{style="heatmap"}, since dotplot output is always drawn with
individual circles in each heatmap cell.}

\item{do_plot}{\code{logical} indicating whether to display the plot with
\code{ComplexHeatmap::draw()} or \code{jamba::imageByColors()} as relevant.
The underlying data is returned invisibly.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for customization.}
}
\description{
MultiEnrichment Heatmap of enrichment P-values
}
\details{
This function is a lightweight wrapper to \code{ComplexHeatmap::Heatmap()}
intended to visualize the enrichment P-values from multiple
enrichment results. The P-value threshold is used to colorize
every cell whose P-value meets the threshold, while all other
cells are therefore white.

The \code{style} argument controls whether a heatmap or dotplot is
created. When \code{style="dotplot"} the cells are colored as usual
but are drawn as circles sized proportional to the number of
genes involved in enrichment. Because \code{ComplexHeatmap::Heatmap()}
is used for this step, a separate point legend is returned
as an attribute of the heatmap object.

To draw the dotplot heatmap including the point legend,
use this form:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{ComplexHeatmap::draw(hm,
   annotation_legend_list=attr(hm, "annotation_legend_list"))
}\if{html}{\out{</div>}}
}
\seealso{
Other jam plot functions: 
\code{\link{adjust_polygon_border}()},
\code{\link{grid_with_title}()},
\code{\link{jam_igraph}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()},
\code{\link{mem_plot_folio}()},
\code{\link{plot_layout_scale}()}
}
\concept{jam plot functions}
