% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{mem_find_overlap}
\alias{mem_find_overlap}
\title{Find best overlap threshold for EnrichMap}
\usage{
mem_find_overlap(
  mem,
  overlap_range = c(0.1, 0.99),
  overlap_count = 2,
  node_fraction = 0.5,
  max_cutoff = 0.4,
  debug = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{list} output from \code{multiEnrichMap()}}

\item{overlap_range}{numeric range of Jaccard overlap values}

\item{overlap_count}{numeric value passed to \code{mem_multienrichplot()}
which is used to filter the multienrichmap by Jaccard overlap
and by overlap_count.}

\item{max_cutoff}{numeric value between 0 and 1, to define the
maximum fraction of nodes in the largest connected component,
compared to the total number of non-singlet nodes.}

\item{debug}{logical indicating whether to return full debug
data, which is used internally to determine the best overlap
cutoff to use.}
}
\description{
Find best overlap threshold for EnrichMap
}
\details{
This function implements a straightforward approach to determine
a reasonable Jaccard overlap threshold for EnrichMap data.
It finds the overlap threshold at which the first connected
component is no more than \code{max_cutoff} fraction of the whole
network. This fraction is defined as the number of nodes in the
largest connected component, divided by the total number of
non-singlet nodes. When all nodes are connected, this fraction == 1.

We found empirically that a \code{max_cutoff=0.4}, the point at which the
largest connected component contains no more than 40\% of all nodes,
seems to be a reasonably good place to start.
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{bulk_cnet_adjustments}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
