% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_gene_path_heatmap}
\alias{mem_gene_path_heatmap}
\title{MultiEnrichment Heatmap of Genes and Pathways}
\usage{
mem_gene_path_heatmap(
  mem,
  genes = NULL,
  sets = NULL,
  min_gene_ct = 2,
  min_set_ct = 3,
  column_fontsize = NULL,
  row_fontsize = NULL,
  row_method = "binary",
  column_method = "binary",
  cluster_columns = NULL,
  cluster_rows = NULL,
  name = "gene_ct",
  p_cutoff = 0.05,
  row_split = NULL,
  column_split = NULL,
  auto_split = TRUE,
  column_title = NULL,
  row_title = NULL,
  row_title_rot = 90,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{list} object created by \code{multiEnrichMap()}. Specifically
the object is expected to contain \code{colorV}, \code{enrichIM},
\code{memIM}, \code{geneIM}.}

\item{genes}{character vector of genes to include in the heatmap,
all other genes will be excluded.}

\item{sets}{character vector of sets (pathways) to include in the heatmap,
all other sets will be excluded.}

\item{min_gene_ct}{minimum number of occurrences of each gene
across the pathways, all other genes are excluded.}

\item{min_set_ct}{minimum number of genes required for each set,
all other sets are excluded.}

\item{column_fontsize, row_fontsize}{arguments passed to
\code{ComplexHeatmap::Heatmap()} to size column and row labels.}

\item{row_method, column_method}{character string of the distance method
to use for row and column clustering. The clustering is performed
by \code{amap::hcluster()}.}

\item{name}{character value passed to \code{ComplexHeatmap::Heatmap()},
used as a label above the heatmap color legend.}

\item{p_cutoff}{numeric value of the enrichment P-value cutoff,
above which P-values are not colored, and are therefore white.
The enrichment P-values are displayed as an annotated heatmap
at the top of the main heatmap. Any cell that has a color meets
at least the minimum P-value threshold.}

\item{column_title}{optional character string with title to display
above the heatmap.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for customization.}
}
\description{
MultiEnrichment Heatmap of Genes and Pathways
}
