% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_gene_path_heatmap}
\alias{mem_gene_path_heatmap}
\title{MultiEnrichment Heatmap of Genes and Pathways}
\usage{
mem_gene_path_heatmap(
  mem,
  genes = NULL,
  sets = NULL,
  min_gene_ct = 1,
  min_set_ct = 1,
  min_set_ct_each = 4,
  column_fontsize = NULL,
  column_cex = 1,
  row_fontsize = NULL,
  row_cex = 1,
  row_method = "binary",
  column_method = "binary",
  enrich_im_weight = 0.3,
  gene_im_weight = 0.5,
  cluster_columns = NULL,
  cluster_rows = NULL,
  name = NULL,
  p_cutoff = mem$p_cutoff,
  p_floor = 1e-06,
  row_split = NULL,
  column_split = NULL,
  auto_split = TRUE,
  column_title = LETTERS,
  row_title = letters,
  row_title_rot = 90,
  colorize_by_gene = FALSE,
  na_col = "white",
  seed = 123,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{list} object created by \code{multiEnrichMap()}. Specifically
the object is expected to contain \code{colorV}, \code{enrichIM},
\code{memIM}, \code{geneIM}.}

\item{genes}{character vector of genes to include in the heatmap,
all other genes will be excluded.}

\item{sets}{character vector of sets (pathways) to include in the heatmap,
all other sets will be excluded.}

\item{min_gene_ct}{minimum number of occurrences of each gene
across the pathways, all other genes are excluded.}

\item{min_set_ct}{minimum number of genes required for each set,
all other sets are excluded.}

\item{min_set_ct_each}{minimum number of genes required for each set,
required for at least one enrichment test.}

\item{column_fontsize, row_fontsize}{arguments passed to
\code{ComplexHeatmap::Heatmap()} to size column and row labels.}

\item{row_method, column_method}{character string of the distance method
to use for row and column clustering. The clustering is performed
by \code{amap::hcluster()}.}

\item{enrich_im_weight}{\code{numeric} value between 0 and 1 indicating
the relative weight of enrichment \verb{-log10 P-value} and gene
incidence matrix when combined prior to clustering the columns
in the resulting heatmap. When \code{enrich_im_weight=0} then
the data is scaled using zero weight for enrichment and therefore
only uses the gene incidence matrix; when \code{enrich_im_weight=1}
the gene incidence matrix is scaled using zero weight and
therefore clustering only uses the \verb{-log10 P-value} values.}

\item{name}{character value passed to \code{ComplexHeatmap::Heatmap()},
used as a label above the heatmap color legend.}

\item{p_cutoff}{numeric value of the enrichment P-value cutoff,
above which P-values are not colored, and are therefore white.
The enrichment P-values are displayed as an annotated heatmap
at the top of the main heatmap. Any cell that has a color meets
at least the minimum P-value threshold. This value by default
is taken from input \code{mem}, using \code{mem$p_cutoff}, for
consistency with the input multienrichment analysis.}

\item{column_split, row_split}{optional arguments passed to
\code{ComplexHeatmap::Heatmap()} to split the heatmap by columns
or rows, respectively. Note that when \code{row_split} is \code{NULL}
and \code{auto_split=TRUE}, it will determine an appropriate number
of clusters based upon the number of rows. To turn off row split,
use \code{row_split=NULL} or \code{row_split=0} or \code{row_split=1};
likewise for \code{column_split}.}

\item{column_title}{optional character string with title to display
above the heatmap.}

\item{row_title}{optional character string with title to display
beside the heatmap. Note when \code{row_split} is defined, the
\code{row_title} is applied to each heatmap section.}

\item{row_title_rot}{\code{numeric} value indicating the rotation of
\code{row_title} text, where \code{0} is not rotated, and \code{90} is rotated
90 degrees.}

\item{colorize_by_gene}{\code{logical} indicating whether to color the
main heatmap body using the colors from \code{geneIM} which represents
each enrichment in which a given gene is involved. Colors are
blended using \code{avg_colors_by_list()}, using colors from
\code{mem$colorV}, applied to \code{mem$geneIM}.}

\item{na_col}{\code{character} string indicating the color to use for
NA or missing values. Typically this argument is only used
when \code{colorize_by_gene=TRUE}, where entries with no color are
recognized as \code{NA} by \code{ComplexHeatmap::Heatmap()}.}

\item{seed}{\code{numeric} value passed to \code{set.seed()} to allow
reproducible results, typically with clustering operations.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for customization.}
}
\description{
MultiEnrichment Heatmap of Genes and Pathways
}
\details{
This function takes the \code{mem} list output from
\code{multiEnrichMap()} and creates a gene-by-pathway incidence
matrix heatmap, using \code{ComplexHeatmap::Heatmap()}.
It uses three basic sources of data to annotate the heatmap:
\enumerate{
\item \code{mem$memIM} the gene-set incidence matrix
\item \code{mem$geneIM} the gene incidence matrix by dataset
\item \code{mem$enrichIM} the pathway enrichment P-value matrix by dataset
}

It will try to estimate a reasonable number of column and row
splits in the dendrogram, based solely upon the number of
columns and rows. These guesses can be controlled with argument
\code{column_split} and \code{row_split}, respectively.

When pathways are filtered by \code{min_gene_ct}, \code{min_set_ct},
and \code{min_set_ct_each}, the order of operations is as follows:
\enumerate{
\item \code{min_set_ct_each}, \code{min_set_ct} - these filters are applied
before filtering genes, in order to ensure all genes are present
from the start.
\item \code{min_gene_ct} - genes are filtered after pathway filtering,
in order to remove pathways which were not deemed "significant"
based upon the required number of genes. Only after those pathways
are removed can the number of occurrences of each gene be judged
appropriately.
}
}
\seealso{
Other jam plot functions: 
\code{\link{grid_with_title}()},
\code{\link{jam_igraph}()},
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()},
\code{\link{mem_plot_folio}()}
}
\concept{jam plot functions}
