% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_legend}
\alias{mem_legend}
\title{MultiEnrichMap color legend}
\usage{
mem_legend(
  mem,
  x = "bottomleft",
  y = NULL,
  bg = "#FFFFFF00",
  box.col = "transparent",
  title = "Color Key",
  cex = 0.8,
  ncol = 1,
  pch = 21,
  pt.cex = 2,
  pt.lwd = 2,
  inset = 0,
  do_directional = NULL,
  directional_column = c("same", "added-bottom", "added-top"),
  directional_colors = c(`up-regulated` = "firebrick3", `no change` = "grey80",
    `down-regulated` = "blue"),
  ...
)
}
\arguments{
\item{mem}{\code{list} object output from \code{multiEnrichMap()}, specifically
expected to contain element \code{"colorV"}.}

\item{x, y, bg, box.col, title, cex, ncol, pch, pt.cex, pt.lwd, inset}{arguments
are passed to \code{graphics::legend()}.
Note \code{pt.lwd} is mostly relevant with \code{do_direction=TRUE}, which
adds open circles to the legend, whose line width has default
\code{pt.lwd=2}.}

\item{do_directional}{\code{logical} indicating whether to include
directional colors defined in \code{directional_colors}, indicated only
as the border color of nodes.}

\item{directional_column}{\code{character} indicating how to add the
directional colors to columns of legend, with two options:
\itemize{
\item "same": appends \code{directional_colors} to legend colors using
the defined \code{ncol} number of columns.
\item "added-bottom": appends \code{directional_colors} as a new column
so the resulting legend with have \code{ncol+1} columns.
In this case, intervening empty rows are filled with blank space,
and the \code{directional_colors} are shown in the bottom-most rows in the
far right column of the legend.
\item "added-top": appends \code{directional_colors} as a new column
so the resulting legend with have \code{ncol+1} columns.
In this case, intervening empty rows are filled with blank space,
and the \code{directional_colors} are shown in the top-most rows in the
far right column of the legend.
}}

\item{directional_colors}{\code{character} vector of R colors, named by
the label to be shown in the legend, displayed in order (top to bottom,
left to right) they appear in this vector.
To remove the entry \code{"no change"}, supply a new vector:
\code{directional_colors=c(up.regulated="firebrick3", down.regulated="blue")}}

\item{...}{additional arguments are passed to \code{legend()}.}
}
\description{
MultiEnrichMap color legend
}
\details{
This function is a simple wrapper around \code{legend()} to add
a color key to a plot, typically for \code{igraph} plots.
}
\seealso{
Other custom plot functions: 
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{plot_mpf}()}
}
\concept{custom plot functions}
