% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_legend}
\alias{mem_legend}
\title{MultiEnrichMap color legend}
\usage{
mem_legend(
  mem,
  x = "bottomleft",
  y = NULL,
  bg = "#FFFFFF99",
  box.col = "transparent",
  title = "Color Key",
  cex = 0.8,
  ncol = 1,
  pch = 21,
  pt.cex = 2,
  ...
)
}
\arguments{
\item{mem}{\code{list} object output from \code{multiEnrichMap()}, specifically
expected to contain element \code{"colorV"}.}

\item{x, y, bg, box.col, title, cex, ncol, pch, pt.cex}{arguments passed
to \code{legend()}.}

\item{...}{additional arguments are passed to \code{legend()}.}
}
\description{
MultiEnrichMap color legend
}
\details{
This function is a simple wrapper around \code{legend()} to add
a color key to a plot, typically for \code{igraph} plots.
}
\seealso{
Other jam plot functions: 
\code{\link{jam_igraph}()},
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_multienrichplot}()}
}
\concept{jam plot functions}
