% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot.R
\name{mem_plot_folio}
\alias{mem_plot_folio}
\title{Multienrichment folio of summary plots}
\usage{
mem_plot_folio(
  mem,
  do_which = NULL,
  p_cutoff = NULL,
  p_floor = 1e-06,
  main = "",
  use_raster = TRUE,
  min_gene_ct = 2,
  min_set_ct = 2,
  min_set_ct_each = 4,
  column_method = "euclidean",
  row_method = "euclidean",
  exemplar_range = c(1, 2, 3),
  pathway_column_split = NULL,
  pathway_row_split = NULL,
  edge_color = NULL,
  cex.main = 2,
  cex.sub = 1.5,
  max_labels = 4,
  max_nchar_labels = 25,
  include_cluster_title = TRUE,
  repulse = 4,
  use_shadowText = FALSE,
  color_by_column = FALSE,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mem}{\code{list} object created by \code{multiEnrichMap()}. Specifically
the object is expected to contain \code{colorV}, \code{enrichIM},
\code{memIM}, \code{geneIM}.}

\item{do_which}{integer vector of plots to produce. When \code{do_which}
is \code{NULL}, then all plots are produced. This argument is intended
to help produce one plot from a folio, therefore each plot is referred
by the number of the plot, in order.}

\item{p_cutoff}{numeric value indicating the enrichment P-value threshold
used for \code{multiEnrichMap()}, but when \code{NULL} this value is taken
from the \code{mem} input, or \code{0.05} is used by default.}

\item{p_floor}{numeric value indicating the lowest enrichment P-value
used in the color gradient on the Enrichment Heatmap.}

\item{main}{character string used as a title on Cnet plots.}

\item{use_raster}{logical indicating whether to use raster heatmaps,
passed to \code{ComplexHeatmap::Heatmap()}.}

\item{min_gene_ct, min_set_ct}{integer values passed to
\code{mem_gene_path_heatmap()}. The \code{min_gene_ct} requires each set
to contain \code{min_gene_ct} genes, and \code{min_set_ct} requires each gene
to be present in at least \code{min_set_ct} sets.}

\item{min_set_ct_each}{minimum number of genes required for each set,
required for at least one enrichment test.}

\item{column_method, row_method}{arguments passed to
\code{ComplexHeatmap::Heatmap()} which indicate the distance method used
to cluster columns and rows, respectively.}

\item{exemplar_range}{integer vector (or \code{NULL}) used to create Cnet
exemplar plots, using this many exemplars per cluster.}

\item{pathway_column_split}{integer value passed as \code{column_split}
to \code{mem_gene_path_heatmap()}, indicating the number of pathway
clusters to create in the gene-pathway heatmap.}

\item{cex.main, cex.sub}{numeric values passed to \code{title()} which
size the default title and sub-title in Cnet plots.}

\item{color_by_column}{\code{logical} indicating whether to colorize
the enrichment heatmap columns using \code{colorV} in the input \code{mem}.
This argument is only relevant when \code{do_which} include \code{1}.}

\item{byCols}{\code{character} vector describing how to sort the
pathways within Cnet clusters. This argument is passed
to \code{rank_mem_clusters()}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to downstream functions.
Notably, \code{sets} is passed to \code{mem_gene_path_heatmap()} which
allows one to define a specific subset of sets to use in the
gene-pathway heatmap.}
}
\value{
\code{list} is returned via invisible, which contains each
relevant object.
}
\description{
Multienrichment folio of summary plots
}
\details{
This function is intended to create multiple summary plots
from a Multienrichment result, the \code{list} output from
\code{multiEnrichMap()}.

The plots created:
\enumerate{
\item Heatmap using enrichment P-values using \code{mem_enrichment_heatmap()}
\item Gene-Pathway Heatmap using \code{mem_gene_path_heatmap()}
\item Cnet plot using Gene-Pathway Heatmap collapsed clusters
\item Cnet plot using Gene-Pathway Heatmap cluster examplars (n per cluster)
\item Cnet plots one per cluster
}
}
\seealso{
Other jam plot functions: 
\code{\link{grid_with_title}()},
\code{\link{jam_igraph}()},
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()},
\code{\link{mem_multienrichplot}()}
}
\concept{jam plot functions}
