% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-mem-plot-folio.R
\name{mem_plot_folio}
\alias{mem_plot_folio}
\alias{prepare_folio}
\title{Multienrichment folio of summary plots}
\usage{
mem_plot_folio(
  mem,
  do_which = NULL,
  mpf = NULL,
  p_cutoff = NULL,
  p_floor = 1e-10,
  main = "",
  use_raster = FALSE,
  min_gene_ct = 1,
  min_set_ct = 1,
  min_set_ct_each = NULL,
  column_method = "euclidean",
  cluster_columns = NULL,
  row_method = "euclidean",
  cluster_rows = NULL,
  exemplar_range = c(1, 2, 3),
  pathway_column_split = NULL,
  pathway_column_title = LETTERS,
  gene_row_split = NULL,
  gene_row_title = letters,
  edge_color = NULL,
  cex.main = 2,
  cex.sub = 1.5,
  row_cex = 1,
  column_cex = 1,
  max_labels = 4,
  max_nchar_labels = 25,
  include_cluster_title = TRUE,
  repulse = 4,
  use_shadowText = FALSE,
  color_by_column = FALSE,
  style = "dotplot_inverted",
  enrich_im_weight = 0.3,
  gene_im_weight = 0.5,
  colorize_by_gene = TRUE,
  cluster_color_min_fraction = 0.4,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  edge_bundling = "connections",
  apply_direction = NULL,
  rotate_heatmap = FALSE,
  row_anno_padding = grid::unit(3, "mm"),
  column_anno_padding = grid::unit(3, "mm"),
  returnType = c("MemPlotFolio", "list"),
  do_plot = TRUE,
  verbose = FALSE,
  ...
)

prepare_folio(..., do_plot = FALSE)
}
\arguments{
\item{mem}{\code{Mem} or \code{list} object from \code{multiEnrichMap()}.}

\item{do_which}{\code{integer} vector of plots to produce.
When \code{do_which} is \code{NULL}, then all plots are produced.
This argument is intended to produce only a subset of plots.}

\item{mpf}{\code{MemPlotFolio}, default NULL, used only to re-apply the same
settings as another \code{MemPlotFolio}.
Note: When supplied, all values in this object \code{thresholds(mpf)}
are used, and no corresponding function arguments are used.}

\item{p_cutoff}{\code{numeric} value, default NULL is taken from \code{mem},
indicating the enrichment P-value threshold.}

\item{p_floor}{\code{numeric} with the lowest enrichment P-value
used in the color gradient on the Enrichment Heatmap.
The purpose is to prevent very low P-values from shifting the
color gradient too far from the \code{p_cutoff} causing those colors
to be pale and nearly white.}

\item{main}{\code{character} string used as a title on Cnet plots.}

\item{use_raster}{\code{logical} default FALSE, deprecated,
whether to use raster heatmaps, passed to \code{ComplexHeatmap::Heatmap()}.
\itemize{
\item Note that \code{use_raster=TRUE} may produce visual artifacts especially
with argument \code{colorize_by_gene=TRUE} in \code{mem_gene_path_heatmap()}.
Changing this argument is no longer supported
}}

\item{min_gene_ct, min_set_ct}{\code{integer} values passed to
\code{mem_gene_path_heatmap()}. The \code{min_gene_ct} requires each set
to contain \code{min_gene_ct} genes, and \code{min_set_ct} requires each gene
to be present in at least \code{min_set_ct} sets.}

\item{min_set_ct_each}{\code{integer}, default NULL, minimum genes per set
in at least one enrichment test.
\itemize{
\item Default NULL uses \code{thresholds(mem)$min_count} to use the same
criteria.
\item The distinction from \code{min_set_ct} is that this threshold
requires this number of genes in one enrichment,
while \code{min_set_ct} applies the threshold to the
combined multi-enrichment data.
}}

\item{column_method, row_method}{\code{character} arguments passed to
\code{ComplexHeatmap::Heatmap()} which indicate the distance method used
to cluster columns and rows, respectively.}

\item{cluster_columns, cluster_rows}{\code{logical}, default NULL, whether to
cluster columns (Sets) and rows (Genes), respectively.
When NULL it uses default clustering with \code{amap::hcluster()} and
applies \code{column_method} or \code{row_method}, respectively.}

\item{exemplar_range}{\code{integer} vector (or \code{NULL}) used to create Cnet
exemplar plots, using this many exemplars per cluster.}

\item{pathway_column_split, gene_row_split}{\code{integer} value passed
as \code{column_split} and \code{row_split}, respectively, to
\code{mem_gene_path_heatmap()}, indicating the number of pathway
clusters, and gene clusters, to create in the gene-pathway heatmap.
When either value is \code{NULL} then auto-split logic is used.}

\item{pathway_column_title, gene_row_title}{\code{character} vectors
passed to \code{mem_gene_path_heatmap()} as \code{column_title} and
\code{row_title}, respectively. When one value is supplied, it is
displayed and centered across all the respective splits. When
multiple values are supplied, values are used to the number
of splits, and recycled as needed. In that case, repeated
values are made unique by \code{jamba::makeNames()}.}

\item{cex.main, cex.sub}{\code{numeric} values passed to \code{title()} which
size the default title and sub-title in Cnet plots.}

\item{row_cex, column_cex}{\code{numeric} character expansion factor, used
to adjust the relative size of row and column labels,
respectively. A value of \code{1.1} will make row font size 10\%
larger.}

\item{color_by_column}{\code{logical} indicating whether to colorize
the enrichment heatmap columns using \code{colorV} in the input \code{mem}.
This argument is only relevant when \code{do_which} include \code{1}.}

\item{enrich_im_weight, gene_im_weight}{\code{numeric} value between 0 and
1, passed to \code{mem_gene_path_heatmap()}, used to apply relative
weight to clustering columns and rows, respectively, when
combining the gene-pathway incidence matrix with either column
enrichment P-values, or row gene incidence matrix data.}

\item{colorize_by_gene}{\code{logical} passed to \code{mem_gene_path_heatmap()}
indicating whether the heatmap body for the gene-pathway heatmap
will be colorized using the enrichment colors for each gene.}

\item{cluster_color_min_fraction}{\code{numeric} value passed to
\code{collapse_mem_clusters()} used to determine which enrichment
colors to associate with each Cnet cluster.}

\item{byCols}{\code{character} vector describing how to sort the
pathways within Cnet clusters. This argument is passed
to \code{rank_mem_clusters()}.}

\item{edge_bundling}{\code{character} string passed to \code{jam_igraph()}
to control edge bundling. The default \code{edge_bundling="connections"}
will bundle Cnet plot edges for genes that share the same pathway
connections.}

\item{apply_direction}{\code{logical} or \code{NULL} indicating whether to
indicate directionality in the \code{mem_enrichment_heatmap()} which is
the first plot in the series. The default \code{apply_direction=NULL}
will auto-detect whether there is directionality present in the
data, and will set \code{apply_direction=TRUE} only when there are non-NA
values that differ from zero.}

\item{rotate_heatmap}{\code{logical} passed to \code{mem_gene_path_heatmap()()}
and only this function, default \code{FALSE}. It indicates whether to
rotate the heatmap to have gene columns and pathway rows.
If you find most people tilt their head to read the pathways,
it might be preferable.}

\item{row_anno_padding, column_anno_padding}{\code{grid::unit} or \code{numeric}
which will be converted to "mm" units. These values control the
space between the heatmap body and row/column annotations,
respectively, only relevant for \code{mem_gene_path_heatmap()()}.
The value is only applied during \code{draw()} and cannot be
defined in the \code{Heatmap} object itself, which is why it is
included here and not \code{mem_gene_path_heatmap()()}.}

\item{do_plot}{\code{logical} indicating whether to render each plot.
Default \code{TRUE} for \code{mem_plot_folio()}, and \code{FALSE} for \code{prepare_folio()}.
In either case, plot data are created and returned, but \code{do_plot=TRUE}
will draw each plot on a unique page, suitable for use with
PDF output and \code{onefile=TRUE} for example.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to downstream functions.
Some useful examples:
\itemize{
\item \code{sets} is passed to \code{mem_gene_path_heatmap()} which
allows one to define a specific subset of sets to use in the
gene-pathway heatmap.
\item \code{cell_size} is passed to \code{mem_enrichment_heatmap()} with the
option to define square cell size in the heatmap dotplot. However,
the resulting heatmap will be at least \code{ncol * cell_height} width,
and \code{nrow * cell_size[2]} height, in addition to the heights of
the title and column labels, and widths of the color key and
dendrogram.
}}
}
\value{
\code{MemPlotFolio} object using \code{invisible()}, containing each
plot object enabled by the argument \code{do_which}.
The \code{MemPlotFolio-class} data are accessible using common functions:
\itemize{
\item \code{EnrichmentHeatmap()}
\item \code{GenePathHeatmap()}
\item \code{CnetCollapsed()}
\item \code{CnetExemplar()}
\item \code{CnetCluster()}
\item \code{Clusters()}
\item \code{GeneClusters()}
\item \code{thresholds()}
\item \code{metadata()}
\item \code{Caption()}
\item \code{CaptionLegendList()}
}
}
\description{
Multienrichment folio of summary plots
}
\details{
\code{prepare_folio()} and \code{mem_plot_folio()} both prepare data
visualizations from \code{Mem} data input.
However, \code{prepare_folio()} does not render figures,
while \code{mem_plot_folio()} does render each resulting figure.

Multiple figures can be added to a single PDF file
using \code{pdf(file, onefile=TRUE)} or \code{cairo_pdf(filename, onefile=TRUE)}.

The data are returned as \code{MemPlotFolio-class} which can be used to
create figures.

Pathways are clustered using the gene-pathway incidence matrix,
then used to define pathway clusters.
This step can be customized by supplying \code{pathway_column_split}
as a \code{list} of \code{character} vectors containing
pathway (set) names.
\subsection{Pathway Clustering}{

Pathways are hierarchical clustered by default using \code{amap::hcluster()},
with \code{column_method="euclidean"}.
The resulting hclust/dendrogram is split using \code{pathway_column_split}
with an \code{integer} number of sub-clusters. The default (NULL) will
determine a default based upon the total Sets in the analysis, and
is intended to be customized by the analyst.

The data to be clustered include the \code{memIM()} matrix of Genes (rows)
and Sets (columns), itself an incidence matrix. It could therefore
be clustered using \code{column_method="binary"}. However, the default
behavior is to append \code{-log10()} values from the \code{enrichIM} matrix,
in order to cluster both the incidence matrix, and the enrichment
P-values together. The matrix data are weighted relative to one another
using \code{enrich_im_weight=0.3}. Note that the enrichment P-value matrix
applies the \code{p_cutoff} such that values that do not meet the threshold
are considered '1' and become '0' with \code{-log10()} transformation.

In order to cluster the gene-pathway incidence matrix with no
influence of the enrichment P-value matrix, use \code{enrich_im_weight=0}.
In this case it often works well to use \code{column_method="binary"}.

Caveat:
Clustering is itself "imperfect": The results are not definitive,
and there is generally no one "true" answer.
However, the output is mainly intended to help organize
information already present in the data, and not to declare
ground truth.
As such, the results are considered stable for the methods and
parameters used, and the interpretation is performed in that context.

The clustering methods and parameters are included in the legend
alongside the \code{GenePathHeatmap()} and the \code{EnrichmentHeatmap()}
for clarity.

Our experience is that clusters do not need to be perfect to be
useful, informative, and valid.
\subsection{Custom Cluster Function}{

It is also possible to supply a custom function to perform clustering,
by using argument \code{cluster_columns} which is passed via '...'
to the underlying function \code{mem_gene_path_heatmap()}. This function
should take a \code{matrix} of \code{numeric} values as input, and return
either hclust or dendrogram output, suitable for use with \code{cutree()}.

An interesting custom function is \code{cluster::daisy(x, method='gower')}
whose output can be converted to \code{hclust}. The 'gower' method accepts
mixed input types such as ordinal, numeric, signed, or other values.

When supplying a custom function via \code{cluster_columns}, we recommend
defining a label using argument \code{column_method}, for example
\code{custom_method='gower'}. This label will appear in the legend of
the resulting visualizations.

The default \code{amap::hcluster()} was chosen because it is fast, includes
both the \code{dist()} and \code{hclust()} steps in one function, and
includes novel distance metrics such as \code{method='correlation'}
which have proven useful in other contexts.
}

\subsection{Pathway Grouping}{

Pathways can be grouped manually by defining \code{pathway_column_split}
as a \code{list} of \code{character} vectors, where those vectors match
pathway names returned by \code{sets(Mem)} - which are also colnames
of the \code{memIM()} incidence matrix.
This step may also use a subset of sets, for example including only
pathways of relevance to the downstream analysis.
}

}

\subsection{Gene Clustering}{

Genes in the Gene-Pathway Heatmap are clustered using similar logic
as described for Pathways. Note that gene clusters are not used
directly in other analysis steps (yet).

However, \code{score_gene_path_clusters()} offers some metric to identify
"hot spots" where a majority of genes in a gene cluster are
represented in the majority of pathways in a pathway cluster.
These "hot spots" are subject to further study, as potential
hubs for interpreting core components across similar biological
pathways.

Custom gene groups can be supplied with \code{gene_row_split} as
a \code{list} of \code{character} vectors whose values match \code{genes(Mem)}.
}

\subsection{Enrichment Heatmap}{

The Gene-Pathway Heatmap is used to define the pathway clusters
(or pathway groups), and therefore the order of pathways shown
in the Gene-Pathway Heatmap.

The same pathway order is used with the Enrichment Heatmap,
for consistency.

An important point is that the Enrichment Heatmap does not
cluster nor order pathways using enrichment P-values,
and instead re-uses the dendrogram (if relevant) from the
Gene-Pathway Heatmap.

Enrichment P-values are not specific to the genes
contained in each pathway, and therefore clustering by enrichment
P-value can be misleading by grouping pathways together which
are not related by the underlying biological data.
}

\subsection{Cnet Collapsed Plot}{

The pathways in each pathway cluster (or pathway group) are collapsed
into one virtual pathway, then used to create a Concept network (Cnet)
plot. This Cnet plot is intended to balance the motivation to
show complete pathway enrichment data, with the motivation to
reduce redundant information.

The pathway clusters (or pathway groups) are the key components
of the Cnet Collapsed output. Therefore adjusting the clustering
options \code{enrich_im_weight} or \code{column_method} are the most common
ways to influence and optimize the outcome.
}

\subsection{Cnet Exemplar Plot}{

There are two general paradigms for Concept networks (Cnets).
\enumerate{
\item Pathways in clusters: \code{CnetCollapsed()}
\item Exemplar Pathways per cluster: \code{CnetExemplar()}
}

The Cnet Exemplar plot includes one pathway for each pathway cluster
\code{num=1} but may include two \code{num=2} or three \code{num=3} pathways
per cluster if relevant. The exemplar pathway is intended as a
representative of the cluster, which can help simplify the number
of genes displayed, while also focusing the meaning of those genes
to the pathway shown.

This option may be preferred when pathway clusters are not visibly
distinctive, for example when one pathway cluster does not appear
to have a common set of genes shared with other pathways in the same
cluster.

The alternative occurs when pathways in each cluster do share
many of the same genes, and in this case the \code{CnetCollapsed()}
plot may be more effective.

Finally, a custom Cnet exemplar plot may be warranted when
displaying a few pathways with particular relevance to a
research study or experiment. The steps are described in the
README.Rmd and README html page for the multienrichjam package:
\url{https://jmw86069.github.io/multienrichjam}
}

\subsection{Cnet Cluster Plot}{

The last set of Concept network (Cnet) plots include all pathways
in each individual pathway cluster. These plots are useful for
pathway clusters which may have different functional sub-components,
or when pathways in a cluster do not appear to be very cohesive.

The Cnet Cluster plot helps display the relationship of pathways within
one pathway cluster, which may help provide the basis for interpreting
the results.
}

\subsection{Order of Plots}{

When using \code{mem_plot_folio()} plots are produced in a specific order.
The argument \code{do_which} may also be useful to focus the output
only on a particular plot, thereby skipping the preparatory
steps used for other plot types.
\enumerate{
\item \strong{Enrichment Heatmap} (Plot #1), \code{EnrichmentHeatmap()}.
Detailed arguments are described in \code{mem_enrichment_heatmap()}.
\item \strong{Gene-Pathway Heatmap} (Plot #2) \code{GenePathHeatmap()}.
Detailed arguments are described in \code{mem_gene_path_heatmap()}.
\item \strong{Cnet Cluster Plot} (Plots #3,#4,#5) \code{CnetCollapsed()}.
Detailed arguments are described in \code{collapse_mem_clusters()} and
\code{mem2cnet()}.
\itemize{
\item For \code{mem_plot_folio()} three styles are produced,
with different node labeling strategies.
\item Plot #3 uses pathway cluster titles.
\item Plot #4 uses pathway names.
\item Plot #5 uses pathway names, and hides gene labels.
}
\item \strong{Cnet Exemplar Plots} (Plots #6,#7,#8) \code{CnetExemplar()}.
The number of exemplars uses \code{exemplar_range=c(1, 2, 3)}.
\item \strong{Cnet Cluster Plots} (Plots #9,#10,#11,etc.) \code{CnetCluster()}.
Detailed arguments are described in \code{mem2cnet()}.
}
}

\subsection{Final Points}{

To define your own pathway cluster labels, define \code{pathway_column_title}
as a vector with length equal to \code{pathway_column_split}. These labels
become network node labels in subsequent plots, and in the
resulting \code{igraph} object.
}
}
\examples{
data(Memtest)
Mpf <- prepare_folio(Memtest)
GenePathHeatmap(Mpf, column_anno_padding=grid::unit(3, "mm"))

}
\seealso{
Other multienrichjam core functions: 
\code{\link{jam_igraph}()},
\code{\link{multiEnrichMap}()}
}
\concept{multienrichjam core functions}
