% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-mem2cnet.R
\name{memIM2cnet}
\alias{memIM2cnet}
\alias{mem2cnet}
\title{Convert MultiEnrichment incidence matrix to Cnet plot}
\usage{
memIM2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 6,
  geneSize = 3,
  categoryCex = 0.8,
  geneCex = 0.5,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col = colorjam::col_div_xf(1.2),
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  spread_labels = FALSE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)

mem2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = "darkblue",
  geneLabelColor = "grey25",
  categorySize = 6,
  geneSize = 3,
  categoryCex = 0.8,
  geneCex = 0.5,
  frame_darkFactor = NULL,
  geneIM = NULL,
  geneIMcolors = NULL,
  geneIMdirection = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  enrichIMdirection = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  colorV = NULL,
  direction_col = colorjam::col_div_xf(1.2),
  hide_solo_pie = TRUE,
  remove_blanks = TRUE,
  spread_labels = FALSE,
  repulse = 3.5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{memIM}{\code{numeric} matrix, or \code{mem} output in \code{list} format. When
\code{mem} format is supplied, relevant arguments which are empty will
use corresponding data from \code{mem}, for example \code{geneIM}, \code{geneIMcolors},
\code{enrichIM}, \code{enrichIMcolors}.}

\item{categoryShape, geneShape}{\code{character} string indicating the
preferred default node shape. In general, \code{pie} shapes with one segment
are converted to \code{circle} for simplicity during plot functions,
although when using \code{jam_igraph()} to plot, it will treat single-segment
pie nodes as circle anyway, and vectorizes pie node plotting to
make the overall rendering time substantially faster.}

\item{categoryColor, geneColor}{\code{character} R color for default node colors,
used when \code{geneIMcolors},\code{enrichIMcolors} is not supplied, respectively.}

\item{categoryLabelColor, geneLabelColor}{\code{character} R color used as
default node label color.}

\item{categorySize, geneSize}{\code{numeric} default node size.}

\item{categoryCex, geneCex}{\code{numeric} adjustment to default node label
font size.}

\item{frame_darkFactor}{\code{numeric} passed to \code{jamba::makeColorDarker()}
so the frame color is slightly darker than the node fill color.}

\item{geneIM, geneIMcolors, geneIMdirection, enrichIM, enrichIMcolors, enrichIMdirection}{\code{matrix} objects typically associated with \code{mem} output from
\code{multiEnrichMap()}, however they are optional so this function
can be applied broadly to any incidence matrix.
\itemize{
\item \code{geneIMcolors} is used to define gene (row) node colors
\item \code{enrichIMcolors} is used to define set (column) node colors
\item \code{geneIMdirection},\code{enrichIMdirection} is used to define optional
border colors defined by the direction, where -1 is down, 0 is no change, and
+1 is up. Use \code{direction_col} to define a custom color function.
}}

\item{coloredrect_nrow, coloredrect_ncol, coloredrect_byrow}{arguments
used when \code{geneShape="coloredrectangle"}, to define layout and
placement of colors across columns in \code{geneIMcolors}. By default,
one row of colors is used.}

\item{colorV}{\code{character} optional vector of R colors, named by enrichment
names that appear in \code{colnames(enrichIM)} when supplied. When
defined, these colors override those defined in \code{enrichIMcolors}.}

\item{remove_blanks}{\code{logical} indicating whether to remove blank
subsections from each node, by calling \code{removeIgraphBlanks()}.}

\item{spread_labels}{\code{logical} indicating whether to call
\code{spread_igraph_labels()} to orient labels radially away from incoming
edges.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to downstream functions:
\itemize{
\item when \code{remove_blanks=TRUE} it is passed to \code{removeIgraphBlanks()}
\item when \code{spread_labels=TRUE} it is passed to \code{spread_igraph_labels()}
}}
}
\value{
\code{igraph} object with Concept network data, containing
pathways connected to genes. Each node has attribute \code{"nodeType"}
of either \code{"Set"} or \code{"Gene"}.
}
\description{
Convert MultiEnrichment incidence matrix to Cnet plot
}
\details{
This function takes \code{list} output from \code{multiEnrichMap()}
and uses three elements to create a Cnet plot:
\enumerate{
\item \code{"memIM"} gene-pathway incidence matrix
\item \code{"geneIM"} the gene incidence matrix
\item \code{"enrichIM"} the pathway enrichment matrix
}
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{label_communities}()},
\code{\link{layout_with_qfr}()},
\code{\link{layout_with_qfrf}()},
\code{\link{mem2emap}()},
\code{\link{mem_multienrichplot}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{sync_igraph_communities}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
