% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{memIM2cnet}
\alias{memIM2cnet}
\title{Convert MultiEnrichment incidence matrix to Cnet plot}
\usage{
memIM2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = c("blue3"),
  geneLabelColor = "grey40",
  categorySize = 5,
  geneSize = 2.5,
  categoryCex = 0.9,
  geneCex = 0.7,
  frame_darkFactor = 1.4,
  geneIM = NULL,
  geneIMcolors = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  ...
)
}
\description{
Convert MultiEnrichment incidence matrix to Cnet plot
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{shape.coloredrectangle.plot}()},
\code{\link{shape.ellipse.plot}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subsetCnetIgraph}()}
}
\concept{jam igraph functions}
