% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-MemPlotFolio.R
\docType{class}
\name{MemPlotFolio-class}
\alias{MemPlotFolio-class}
\alias{MemPlotFolio}
\alias{list_to_MemPlotFolio}
\alias{MemPlotFolio_to_list}
\alias{show,MemPlotFolio-method}
\alias{Clusters,MemPlotFolio-method}
\alias{GeneClusters,MemPlotFolio-method}
\alias{thresholds,MemPlotFolio-method}
\alias{metadata,MemPlotFolio-method}
\alias{Caption,MemPlotFolio-method}
\alias{CaptionLegendList,MemPlotFolio-method}
\alias{EnrichmentHeatmap,MemPlotFolio-method}
\alias{GenePathHeatmap,MemPlotFolio-method}
\alias{CnetCollapsed,MemPlotFolio-method}
\alias{CnetExemplar,MemPlotFolio-method}
\alias{CnetCluster,MemPlotFolio-method}
\title{Mem S4 class, accessors, getters, and setters}
\usage{
list_to_MemPlotFolio(mpf)

MemPlotFolio_to_list(x, ...)

\S4method{show}{MemPlotFolio}(object)

\S4method{Clusters}{MemPlotFolio}(x)

\S4method{GeneClusters}{MemPlotFolio}(x)

\S4method{thresholds}{MemPlotFolio}(x)

\S4method{metadata}{MemPlotFolio}(x)

\S4method{Caption}{MemPlotFolio}(x, ...)

\S4method{CaptionLegendList}{MemPlotFolio}(x, ...)

\S4method{EnrichmentHeatmap}{MemPlotFolio}(x, do_plot, ...)

\S4method{GenePathHeatmap}{MemPlotFolio}(x, do_plot, ...)

\S4method{CnetCollapsed}{MemPlotFolio}(x, type, do_plot, ...)

\S4method{CnetExemplar}{MemPlotFolio}(x, num, do_plot, ...)

\S4method{CnetCluster}{MemPlotFolio}(x, cluster, do_plot, ...)
}
\arguments{
\item{mpf}{\code{list} output from \code{mem_plot_folio()}}

\item{x}{\code{MemPlotFolio} object}

\item{...}{additional arguments are ignored}
}
\value{
\code{list_to_MemPlotFolio()} returns a \code{MemPlotFolio} S4 object,
from 'list' or 'MemPlotFolio' input.

\code{MemPlotFolio_to_list()} returns a \code{list}

\code{Clusters(MemPlotFolio)} returns a \code{list} of \code{character} vectors,
named by cluster.

\code{GeneClusters(MemPlotFolio)} returns a \code{list} of
\code{character} vectors, named by cluster.

\code{thresholds(MemPlotFolio)} returns a \code{list} of
thresholds used with \code{mem_plot_folio()}.

\code{metadata(MemPlotFolio)} returns a \code{list} of
metadata used with \code{mem_plot_folio()}.

\code{metadata(MemPlotFolio)} returns a \code{character} string
with caption summary used with \code{mem_plot_folio()}.
Multiple lines are delimited by newline characters.

\code{metadata(MemPlotFolio)} returns the caption summary
in the form of \code{ComplexHeatmap::Legends} suitable to \code{draw()}
as R grid graphics.

\code{EnrichmentHeatmap(MemPlotFolio)} returns a
\code{ComplexHeatmap::HeatmapList} when do_plot is TRUE (default),
\code{ComplexHeatmap::Heatmap} when do_plot is FALSE, containing
enrichment P-values by enrichment, and pathway rows in clusters.

\code{GenePathHeatmap(MemPlotFolio)} returns a
\code{ComplexHeatmap::HeatmapList} when do_plot is TRUE (default),
\code{ComplexHeatmap::Heatmap} when do_plot is FALSE, containing
the genes-pathways incidence matrix and associated caption.

\code{CnetCollapsed(MemPlotFolio)} returns an \code{igraph} object invisibly,
with Gene and Set nodes representing the collapsed pathway clusters.

\code{CnetExemplar(MemPlotFolio)} returns an \code{igraph} object
with Gene and Set nodes for the 'num' number of exemplars per cluster.

\code{CnetCluster(MemPlotFolio)} returns an \code{igraph} object invisibly,
using all pathways in the cluster defined with argument \code{cluster}.
}
\description{
Mem class containing results from \code{multiEnrichMap()},
current class version "1.0.0".
}
\section{Functions}{
\itemize{
\item \code{list_to_MemPlotFolio()}: Convert legacy \code{list} to S4 \code{MemPlotFolio}

\item \code{MemPlotFolio_to_list()}: Coerce S4 \code{MemPlotFolio} to \code{list} format

\item \code{show(MemPlotFolio)}: Show summary of a MemPlotFolio object,
dimensions defined by genes, sets, enrichments.

\item \code{Clusters(MemPlotFolio)}: Returns the order of gene sets from
\code{MemPlotFolio} results as a \code{list} named by pathway cluster,
containing \code{character} vectors of pathway gene sets.

\item \code{GeneClusters(MemPlotFolio)}: Returns the order of genes from
\code{MemPlotFolio} results as a \code{list} named by gene cluster,
containing \code{character} vectors of genes.

\item \code{thresholds(MemPlotFolio)}: Returns the thresholds used with
\code{MemPlotFolio}.

\item \code{metadata(MemPlotFolio)}: Returns the metadata used with
\code{MemPlotFolio}.

\item \code{Caption(MemPlotFolio)}: Returns the caption summary for
\code{MemPlotFolio}.

\item \code{CaptionLegendList(MemPlotFolio)}: Returns the caption summary for
\code{MemPlotFolio} as \code{ComplexHeatmap::Legends}.

\item \code{EnrichmentHeatmap(MemPlotFolio)}: Draws the enrichment heatmap
from \code{MemPlotFolio} results.

\item \code{GenePathHeatmap(MemPlotFolio)}: Draws the gene-pathway set heatmap
from \code{MemPlotFolio} results.

\item \code{CnetCollapsed(MemPlotFolio)}: Draws the Cnet collapsed
network from \code{MemPlotFolio} results. Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.
Argument \code{type} can be:
\itemize{
\item \code{type=''} (default) to use cluster title
\item \code{type='set'} to use abbreviated pathway names
\item \code{type='set2'} to use abbreviated pathway names, with hidden gene labels
}

\item \code{CnetExemplar(MemPlotFolio)}: Draws the Cnet exemplar network
from \code{MemPlotFolio} results for 'num' exemplar per cluster.
Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.

\item \code{CnetCluster(MemPlotFolio)}: Draws the Cnet network
from \code{MemPlotFolio} results for a specific cluster.
Note that '...' arguments are
passed to \code{jam_igraph()} and \code{mem_legend()} when \code{do_plot=TRUE}.

}}
\examples{
# list_to_MemPlotFolio examples
data(Memtest)
mpf <- mem_plot_folio(Memtest, do_plot=FALSE,
   do_which=c(1, 2, 3, 4), returnType="list")
Mpf <- list_to_MemPlotFolio(mpf)

data(Memtest)
mpf <- mem_plot_folio(Memtest, do_plot=FALSE, returnType="list")
Mpf <- list_to_MemPlotFolio(mpf)

# enrichment heatmap
EnrichmentHeatmap(Mpf, column_title="Enrichment Heatmap")

# Gene-path heatmap
GenePathHeatmap(Mpf, column_title="Gene-Path Heatmap")

# Cnet collapsed sets
CnetCollapsed(Mpf, type="set", use_shadowText=TRUE, main="Cnet Collapsed Sets")

# Cnet exemplar plot
CnetExemplar(Mpf, num=2, use_shadowText=TRUE, main="Cnet Exemplars, num=2")

# Cnet cluster plot
CnetCluster(Mpf, cluster="B", use_shadowText=TRUE, main="Cnet Cluster 'B'")

}
\seealso{
Other jam Mem utilities: 
\code{\link{Mem-class}},
\code{\link{fixSetLabels}()},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()}

Other MemPlotFolio: 
\code{\link{check_MemPlotFolio}()},
\code{\link{score_gene_path_clusters}()}
}
\concept{MemPlotFolio}
\concept{jam Mem utilities}
