% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multienrichjam-data.R
\docType{data}
\name{Memtest}
\alias{Memtest}
\alias{Reese_genes}
\title{Reese 2019 cohort multienrichment data}
\format{
Memtest is a \code{Mem} multi-enrichment object including two
enrichment tests 'Newborns' and 'OlderChildren'

\code{Reese_genes} is a \code{list} with names 'Newborns' and 'OlderChildren',
with a \code{character} vectors with genes associated with each cohort.
}
\source{
Reese et al 2019 \link{https://doi.org/10.1016/j.jaci.2018.11.043},
\strong{Epigenome-wide meta-analysis of DNA methylation and childhood asthma}
}
\usage{
Memtest

Reese_genes
}
\description{
Data used for multi-enrichment analysis as described in
Reese et al 2019 \link{https://doi.org/10.1016/j.jaci.2018.11.043},
\strong{Epigenome-wide meta-analysis of DNA methylation and childhood asthma}.
It is provided for convenient testing of multienrichjam.
}
\details{
Memtest is a \code{Mem} object with enrichment results
for 'Newborns' and 'OlderChildren' as described in
}
\examples{
# Mem data for Reese 2019 cohorts
data(Memtest)
# enrichment names
names(Memtest)
# genes represented
genes(Memtest)
# gene sets, pathways represented
sets(Memtest)

# Reese 2019 genes by cohort
data(Reese_genes)
# Number of genes per cohort
lengths(Reese_genes)

}
\seealso{
Other Mem: 
\code{\link{Mem-class}},
\code{\link{check_Mem}()}
}
\concept{Mem}
\keyword{datasets}
