% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{multiEnrichMap}
\alias{multiEnrichMap}
\title{Prepare MultiEnrichMap data from enrichList}
\usage{
multiEnrichMap(enrichList, geneHitList = NULL, colorV = NULL,
  nrow = NULL, ncol = NULL, byrow = FALSE, enrichLabels = NULL,
  subsetSets = NULL, overlapThreshold = 0.1, cutoffRowMinP = 0.05,
  enrichBaseline = -log10(cutoffRowMinP), enrichLens = 0,
  enrichNumLimit = 4, nEM = 500, topEnrichN = 15,
  topEnrichSources = c("Category", "Source"),
  topEnrichCurateFrom = c("CP:.*"), topEnrichCurateTo = c("CP"),
  topEnrichSourceSubset = c("C2_CP", "C5_BP", "C5_CC", "C5_MF"),
  topEnrichDescriptionGrep = NULL, topEnrichNameGrep = NULL,
  keyColname = c("ID", "itemsetID"), nameColname = c("ID", "Name",
  "itemsetID"), geneColname = c("geneID", "geneNames", "Genes"),
  pvalueColname = c("P-value", "pvalue", "Pval"),
  descriptionColname = c("Description", "Name", "Pathway"),
  descriptionCurateFrom = c("^Genes annotated by the GO term "),
  descriptionCurateTo = c(""), pathGenes = c("setSize", "pathGenes"),
  geneHits = c("Count", "geneHits"), geneDelim = "[,/ ]+",
  GmtTname = NULL, msigdbGmtT = NULL, verbose = FALSE, ...)
}
\description{
Prepare MultiEnrichMap data from enrichList
}
\seealso{
Other jam enrichment functions: \code{\link{topEnrichBySource}}
}
\concept{jam enrichment functions}
