% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{order_colors}
\alias{order_colors}
\title{Order colors}
\usage{
order_colors(
  x,
  sort_by = c("H", "-C", "-L"),
  c_min = 4,
  grey_hue = 359,
  hue_offset = 0,
  ...
)
}
\description{
Order colors
}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{deconcat_df2}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{sort_colors_deprecated}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
