% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{order_colors}
\alias{order_colors}
\title{Order colors}
\usage{
order_colors(x, sort_by = c("H", "-L", "-C"), ...)
}
\arguments{
\item{x}{\code{character} vector of colors}

\item{sort_by}{\code{character} vector with color colnames used for sorting,
passed as argument 'byCols' to \code{colorjam::sort_colors()}.}

\item{...}{additional arguments are passed to \code{colorjam::sort_colors()}.}
}
\description{
Order colors
}
\examples{
set.seed(123);
x <- sample(grep("royal|^golden", colors(), value=TRUE))
x_sorted <- x[order_colors(x)];
jamba::showColors(list(input=x, sorted=x_sorted))

}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{curateIPAcolnames}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{summarize_node_spacing}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
