% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-plot-mpf.R
\name{plot_mpf}
\alias{plot_mpf}
\title{Plot function for MemPlotFolio objects}
\usage{
plot_mpf(
  Mpf,
  plot_which = c("all", "EnrichmentHeatmap", "eh", "GenePathHeatmap", "gp",
    "CnetCollapsed", "cc", "CnetExemplar", "ce", "CnetCluster", "c", "EnrichmentMap",
    "em"),
  do_md_tabs = FALSE,
  md_tab_open = NULL,
  md_tab_suffix = NULL,
  md_tab_close = NULL,
  md_tab_level = 2,
  md_title = "Multi-Enrichment Folio",
  cc_type = NULL,
  ce_num = NULL,
  c_cluster = NULL,
  em_group = NULL,
  Mem = NULL,
  params = list(em = list(repulse = 3.5, node_factor = 1, width = 30), cc = list(repulse
    = 3.5, node_factor = 1), ce = list(node_factor = 1), c = list(node_factor = 1)),
  do_newpage = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{Mpf}{\code{MemPlotFolio} object as returned by
\code{prepare_folio()} or \code{mem_plot_folio()}.}

\item{plot_which}{\code{character} vector with one or more plots,
default 'all' renders all plots defined in 'Mpf'.
Plots will be created in the order provided, except for 'all'
which follows the default order.
The following values are recognized:
\itemize{
\item 'all': All available plot types are included.
\item 'EnrichmentHeatmap' or 'eh': \code{EnrichmentHeatmap()}
\item 'GenePathHeatmap' or 'gp': \code{GenePathHeatmap()}
\item 'CnetCollapsed' or 'cc': \code{CnetCollapsed()}
\item 'CnetExemplar' or 'ce': \code{CnetExemplar()}
\item 'CnetCluster' or 'c': \code{CnetCluster()}
\item 'EnrichmentMap' or 'em': \code{mem2emap()}. Note this plot
requires either 'Mem' argument, or 'Mem' being included
in the \code{Mpf@metadata$Mem}.
}}

\item{do_md_tabs}{\code{logical} default FALSE, whether to print
markdown-compatible tab headers before rendering each
plot, done using \code{cat()} to STDOUT.
To enable specific Rmarkdown or Quarto (Qmd) style,
provide a \code{character} string, which will override all
other checks, and will define 'md_tab_open' and 'md_tab_close'.
\itemize{
\item 'Rmd' or 'rmd' to enable Rmarkdown '{.tabset}' style
\item 'Qmd' or 'qmd' or 'Quarto' or 'quarto' to enable
Quarto '::: {.panel-tabset}' style, with footer ':::'.
}}

\item{md_tab_open}{\code{character} string to define the tab open
style. When NULL, it will auto-detect an appropriate style.
\itemize{
\item Rmarkdown with HTML output: \code{''}
\item Quarto with HTML output: \code{'\\\\n\\\\n::: {.panel-tabset}\\\\n\\\\n'}.
\item All others: \code{''}.
}}

\item{md_tab_suffix}{\code{character} string to define the tab suffix
style. When NULL, it will auto-detect an appropriate style.
\itemize{
\item Rmarkdown with HTML output: \code{' {.tabset}\\\\n\\\\n'}
\item All others: \code{''}.
}}

\item{md_tab_close}{\code{character} string to define the tab close
style. When NULL, it will auto-detect an appropriate style.
\itemize{
\item Rmarkdown with HTML output: \code{''}
\item Quarto with HTML output: \code{'\\\\n\\\\n:::\\\\n\\\\n'}.
\item All others: \code{''}
}

In Quarto, to keep the tabset open, define \code{md_tab_close=''}.}

\item{md_tab_level}{\code{integer} heading level to begin, default 2
uses heading '##'.}

\item{md_title}{\code{character} default 'Multi-Enrichment Folio' will
print a title heading using \code{md_tab_level}, then create the tabset
underneath at a lower heading.
\itemize{
\item Use md_title=FALSE or md_title=NULL to suppress this header.
}}

\item{cc_type, ce_num, c_cluster}{passed to corresponding functions,
and supports multiple values.
\itemize{
\item cc_type: passed to CnetCollapsed(Mpf, type=cc_type).
When NULL it will use 'cluster' if \code{ClusterLabels(Mpf)} is
available, otherwise 'set'. If more than 500 nodes, it appends
'2' to the end, which hides gene node labels.
\item ce_num: passed to CnetExemplar(Mpf, num=ce_num). When NULL
it will use ce_num=1.
\item c_cluster: passed to CnetCluster(Mpf, cluster=c_cluster).
When NULL it will iterate each cluster in Mpf.
}}

\item{em_group}{\code{character} default NULL uses
community detection in \code{mem2emap()}.
\itemize{
\item 'community': uses community detection in \code{mem2emap()}.
\item 'clusters': uses \code{Clusters(Mpf)}.
\item 'cluster_labels': uses \code{ClusterLabels(Mpf)}.
}}

\item{Mem}{\code{Mem} default NULL, only included for backwards support of
Mpf objects which do not have 'Mem' included in metadata.
Experimental.}

\item{params}{\code{list} named by shorthand plot type
('eh','gp','em', 'cc', 'ce', 'c') each containing a list of
optional parameters relevant to each plot.
In each case, parameters in \code{params} are passed to the
internal function as arguments.For example,
\code{em=list(repulse=3.5)} is passed as
\code{EnrichmentMap(..., repulse=3.5)}.
Limited functionality currently, however check here for
recognized arguments as they become available.
\itemize{
\item 'em' EnrichmentMap:
\itemize{
\item repulse: apply via \code{layout_with_qfr(g, repulse=repulse)},
default 3.5.
\item width: \code{integer} word-wrap character width, default 30.
Word-wrap is applied using \code{fixSetLabels()}
however it uses \code{do_abbreviations=FALSE} and \code{removeGrep=NULL}
so the primary effect is to apply word-wrap, and capitalization.
}
}}

\item{do_newpage}{\code{logical} default NULL sets TRUE when knitr is running,
whether to call \code{grid::grid.newpage()} after certain plots to
encourage Rmd and Qmd to recognize end of a plot for the purpose
of placing figures inside the correct tabset regions. Experimental.
In some cases, knitr appears to embed a figure only when it
detects that figure is "complete", and sometimes it does so after
the next markdown tabset has been printed to STDOUT via \code{cat()}.
Using \code{grid::grid.newpage()} appears to force knitr to recognize
a new plot has started, therefore embeds the previous figure image.
Also, knitr skips blank plots, so calling newpage, then having
another plot call a newpage is not problematic.}

\item{verbose}{\code{logical} whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} plot data, named by type, returned invisibly.
}
\description{
Plot function for MemPlotFolio objects
}
\details{
Main purpose is to permit independent plotting
for MemPlotFolio objects, which should allow custom
settings for things like Rmd or Qmd tabs, or for
multi-page PDF output.

Note argument \code{plot_which} takes \code{character} vector,
to be distinct from \code{plot_which} which takes \code{integer}
as used historically by \code{mem_plot_folio()}.

Todo:
\itemize{
\item Future work will allow customization of each plot.
}
\subsection{Markdown tab output}{

When \code{do_md_tabs=TRUE} the default is to autodetect
appropriate values for \code{md_tab_open} and \code{md_tab_close}.
These values are defined if running in knitr context,
with HTML output, using recognized input file type
'Rmd', 'rmd', 'Qmd', 'qmd'. In all other cases, it sets
tabs to empty character ''.

You can set \code{do_md_tabs='Rmd'} to force Rmarkdown tabset style,
or \code{do_md_tabs='Qmd'} to force Quarto tabset style.
In either case, the tabs are defined using defaults,
where: Rmd uses '{.tabset}' and no footer; and
Qmd uses '::: {.panel-tabset}' and ':::' footer.

Setting \code{do_md_tabs=TRUE} then all other values \code{''} blank
will print markdown-style headings for each plot type,
without adding tabset tags. This output may be preferred
for a "long form" report where each plot has its own
distinct header, which may also appear in a
table of contents.

Secondary goals:
\itemize{
\item Decouple all the sub-plot customizations
so they can be applied without re-creating the object.
\item Permit subsetting the Mpf data as relevant, perhaps by
\code{Clusters()} as a drill-down technique.
}

Active development:
\itemize{
\item Each plot will support more customization options over time,
particularly to adjust Cnet network layouts, and other
custom labeling options.
\item Notably, support for \code{ClusterLabels()} is propagating
through various plots, and \code{ClusterData()} may provide
interactive capability, for example detailed mouse-over
or event-driven data for each pathway or gene cluster.
}
}
}
\seealso{
Other custom plot functions: 
\code{\link{mem_enrichment_heatmap}()},
\code{\link{mem_gene_path_heatmap}()},
\code{\link{mem_legend}()}
}
\concept{custom plot functions}
