% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-utils.R
\name{rank_mem_clusters}
\alias{rank_mem_clusters}
\title{Rank Multienrichment clusters}
\usage{
rank_mem_clusters(
  mem,
  clusters,
  choose = NULL,
  per_cluster = Inf,
  byCols = c("minp_rank", "composite_rank", "gene_count_rank"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mem}{\code{Mem} or legacy \code{list} mem output from \code{multiEnrichMap()}}

\item{clusters}{\code{list} containing set names, that must match
\code{colnames(mem$memIM)} and \code{rownames(mem$enrichIM)}.}

\item{choose}{\code{character} vector with optional subset of clusters
to return, matching \code{names(clusters)};
or \code{integer} vector referring to \code{clusters} by index position.
When \code{choose} is \code{NULL}, all clusters are returned.}

\item{per_cluster}{\code{integer} vector with the number of entries
to return for each cluster. Values will be recycled to the
length of the clusters to be returned, defined by \code{choose}
or by \code{length(clusters)} when \code{choose} is \code{NULL}.}

\item{byCols}{\code{character} vector used to sort the resulting
\code{data.frame} within each cluster. This argument is passed
directly to \code{jamba::mixedSortDF()}. Default 'minp_rank'.
Recognized columns:
\itemize{
\item '\strong{minp_rank}': the lowest P-value for all enrichments.
\item '\strong{composite_rank}': a composite score which sorts by the
order of magnitude of the lowest P-value, then by number
of genes.
\item '\strong{gene_count_rank}': the set with the most genes.
}

Note that any column order can be reversed using "-" prefix,
for example '-gene_count_rank' would return the set with the
lowest gene count.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} sorted by the criteria defined by \code{byCols},
with colname \code{"set"} to indicate the pathway/set name.
It includes additional columns which may be useful in
filtering or sorting.
\itemize{
\item 'gene_count': the total number of genes involved in enrichment
\item 'minp': the lowest enrichment P-value
\item 'gene_count_rank': the rank by descending gene count per cluster.
\item 'minp_rank': the rank by ascending minimum P-value.
\item 'composite_rank': the rank of the composite score, ascending.
The composite rank uses the order of magnitude of minimum P-value,
then descending number of genes.
}
}
\description{
Rank Multienrichment clusters
}
\details{
This function takes \code{list} output from \code{multiEnrichMap()}, and
a \code{list} of clusters, and returns a \code{data.frame} that contains
several rank order metrics. It is intended to be used with
column clusters following \code{mem_gene_path_heatmap()},
see examples.

The argument \code{per_cluster} is intended to make it convenient
to pick the top exemplar pathways, especially when argument
\code{byCols} is defined so that it sorts by the rank columns.

The argument \code{choose} is intended to make it easy to retrieve
pathways from specific clusters.
}
\examples{
## Start with mem
# mem <- multiEnrichMap(...);
# gp_hm <- mem_gene_path_heatmap(mem, column_split=4);
## Retrieve clusters from the Heatmap output, there should be 4 clusters
# clusters <- heatmap_column_order(gp_hm)
# clusters_df <- rank_mem_clusters(mem, clusters)

}
\seealso{
Other jam utility functions: 
\code{\link{ashape}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{curateIPAcolnames}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{find_enrich_colnames}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rotate_coordinates}()},
\code{\link{summarize_node_spacing}()},
\code{\link{with_ht_opts}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
