% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{rectifyPiegraph}
\alias{rectifyPiegraph}
\title{Convert pie igraph node shapes to coloredrectangle}
\usage{
rectifyPiegraph(
  g,
  nrow = 2,
  ncol = 5,
  byrow = TRUE,
  whichNodes = seq_len(igraph::vcount(g)),
  ...
)
}
\arguments{
\item{g}{igraph object, expected to contain one or more
nodes with shape \code{"pie"}.}

\item{nrow, ncol}{integer values indicating the default
number of rows and columns to use when displaying
the colors for each node.}

\item{byrow}{logical indicating whether each vector
of node colors should fill the nrow,ncol matrix
by each row, similar to how values are filled
in \code{base::matrix()} with argument \code{byrow}.}

\item{whichNodes}{integer vector of nodes in \code{g}
which should be considered. Only nodes with shape
\code{"pie"} will be converted which are also within
the \code{whichNodes} vector. By default, all nodes
are converted, but \code{whichNodes} allows converting
only a subset of nodes.}

\item{...}{additional arguments are ignored.}
}
\value{
igraph object where node shapes were changed
from \code{"pie"} to \code{"coloredrectangle"}.
}
\description{
Convert pie igraph node shapes to coloredrectangle
}
\details{
This function simply converts an igraph network with \code{"pie"}
node shapes, to use the \code{"coloredrectangle"} node shape
provided by the multienrichjam package.

In the process, it transfers related node attributes:
\itemize{
\item \code{"pie.color"} are copied to \code{"coloredrect.color"}
\item \code{"pie.names"} are copied to \code{"coloredrect.names"}. The
\code{"coloredrect.names"} can be used to label a color key.
\item \code{"size"} is converted to \code{"size2"} after applying
\code{sqrt(size) * 1.5}. The \code{"size2"} value is used to
define the size of coloredrectangle nodes.
}
}
\seealso{
Other jam igraph functions:
\code{\link[=communities2nodegroups]{communities2nodegroups()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=edge_bundle_bipartite]{edge_bundle_bipartite()}},
\code{\link[=edge_bundle_nodegroups]{edge_bundle_nodegroups()}},
\code{\link[=fixSetLabels]{fixSetLabels()}},
\code{\link[=flip_edges]{flip_edges()}},
\code{\link[=get_bipartite_nodeset]{get_bipartite_nodeset()}},
\code{\link[=highlight_edges_by_node]{highlight_edges_by_node()}},
\code{\link[=igraph2pieGraph]{igraph2pieGraph()}},
\code{\link[=label_communities]{label_communities()}},
\code{\link[=mem2cnet]{mem2cnet()}},
\code{\link[=mem2emap]{mem2emap()}},
\code{\link[=nodegroups2communities]{nodegroups2communities()}},
\code{\link[=removeIgraphBlanks]{removeIgraphBlanks()}},
\code{\link[=subsetCnetIgraph]{subsetCnetIgraph()}},
\code{\link[=subset_igraph_components]{subset_igraph_components()}},
\code{\link[=sync_igraph_communities]{sync_igraph_communities()}}
}
\concept{jam igraph functions}
