% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{rectifyPiegraph}
\alias{rectifyPiegraph}
\title{Convert pie igraph node shapes to coloredrectangle}
\usage{
rectifyPiegraph(
  g,
  nrow = 2,
  ncol = 5,
  byrow = TRUE,
  whichNodes = seq_len(vcount(g)),
  ...
)
}
\arguments{
\item{g}{igraph object, expected to contain one or more
nodes with shape \code{"pie"}.}

\item{nrow, ncol}{integer values indicating the default
number of rows and columns to use when displaying
the colors for each node.}

\item{byrow}{logical indicating whether each vector
of node colors should fill the nrow,ncol matrix
by each row, similar to how values are filled
in \code{base::matrix()} with argument \code{byrow}.}

\item{whichNodes}{integer vector of nodes in \code{g}
which should be considered. Only nodes with shape
\code{"pie"} will be converted which are also within
the \code{whichNodes} vector. By default, all nodes
are converted, but \code{whichNodes} allows converting
only a subset of nodes.}

\item{...}{additional arguments are ignored.}
}
\value{
igraph object where node shapes were changed
from \code{"pie"} to \code{"coloredrectangle"}.
}
\description{
Convert pie igraph node shapes to coloredrectangle
}
\details{
This function simply converts an igraph network with \code{"pie"}
node shapes, to use the \code{"coloredrectangle"} node shape
provided by the multienrichjam package.

In the process, it transfers related node attributes:
\itemize{
\item \code{"pie.color"} are copied to \code{"coloredrect.color"}
\item \code{"pie.names"} are copied to \code{"coloredrect.names"}. The
\code{"coloredrect.names"} can be used to label a color key.
\item \code{"size"} is converted to \code{"size2"} after applying
\code{sqrt(size) * 1.5}. The \code{"size2"} value is used to
define the size of coloredrectangle nodes.
}
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
