% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamgraph-relayout-nodegroups.R
\name{relayout_nodegroups}
\alias{relayout_nodegroups}
\title{Relayout each nodegroup in a bipartite (Cnet) graph, experimental}
\usage{
relayout_nodegroups(
  cnet,
  repulse = 3.5,
  fix_set_nodes = TRUE,
  spread_labels = TRUE,
  nodegroups = NULL,
  add_edges = TRUE,
  edge_factor = 2,
  do_final_relayout = NULL,
  final_repulse = 3.5,
  ...
)
}
\arguments{
\item{cnet}{\code{igraph} object with node layout already defined}

\item{repulse}{\code{numeric} default 3.5, passed to \code{relayout_with_qfr()}}

\item{fix_set_nodes}{\code{logical} default TRUE, whether to fix all
nodes with nodeType=='Set' to prevent them from moving.}

\item{spread_labels}{\code{logical} default TRUE, whether to apply
\code{spread_igraph_labels()} after the relayout iterations are complete.}

\item{nodegroups}{\code{list} default NULL, intended to pass a custom
set of nodegroups. When NULL it uses \code{get_cnet_nodeset()}.}

\item{...}{additional arguments are passed to
\code{spread_igraph_labels()}}
}
\value{
\code{igraph} object with updated layout.
}
\description{
Relayout each nodegroup in a bipartite (Cnet) graph, experimental
}
\details{
This function iteratively re-applies a layout function
to each nodegroup in a graph, constraining the position of
all other nodes for each iteration.
Currently the layout uses \code{relayout_with_qfr()}, in future it
may use any layout function.

The purpose is to "encourage" nodes in a nodegroup to become
bundled together.

Strategy:
\itemize{
\item Each nodegroup is isolated, and \code{relayout_with_qfr()} is called
on nodes in each nodegroup, while constraining all other nodes so
they cannot move.
\item In theory, using the same \code{repulse}, the nodes would not move at all.
Changing the repulse force could encourage nodes to stay together.
\item By default \code{add_edges=TRUE} which adds phantom edges to connect all
nodes in a node group.
\itemize{
\item The edge weight is scaled down by the number
of nodes.
\item These phantom edges are intended to help 'encourage' the nodegroup
nodes to group together.
\item Otherwise, most layout algorithms are only focused on specific
edge forces, and not secondary forces which are common in Cnet plots.
\item For example, nodes in a nodegroup all share
the same network connections, however they are not otherwise
attracted to each other in a network layout. In absence of any
repulsive force, they would all be co-located. But with some
repulsive force, they are repelled from each other, and sometimes
end up radially positioned around the plot, and not grouped together.
\item The phantom edges add some minimal force for nodes to be
grouped closer together, and are removed once the layout is complete.
}
\item As a final polishing step, \code{do_final_relayout=TRUE} enables a final
round of global node layout, with \code{final_repulse}.
\itemize{
\item We observed that sometimes the nodegroups are too clumped,
in a big circular "ball" due to the phantom edge process above.
The final relayout is helpful to allow nodes to space out somewhat.
\item It may be helpful to pass \code{niter} to control the number of
layout iterations in this final step. The default is 500.
}
}
\subsection{Todo}{

Bonus points:
\itemize{
\item Determine if each nodegroup is "split" around another nodegroup.
If so, iterate that nodegroup using varying 'repulse' or other
strategies to attempt to minimize the issue.
}
}
}
\examples{
cnet <- make_cnet_test();
ns <- get_cnet_nodeset(cnet)
# mark.groups: highlight just one nodegroup
# nodegroups: enables the edge_bundling to work properly
jam_igraph(cnet, mark.groups=ns["SetA,SetB"], nodegroups=ns,
   main="One nodegroup is split")

cnet2 <- relayout_nodegroups(cnet, nodegroups=ns["SetA,SetB"])
jam_igraph(cnet2, mark.groups=ns["SetA,SetB"], nodegroups=ns,
   main="This nodegroup is 'encouraged' to re-group")

cnet2 <- relayout_nodegroups(cnet, nodegroups=ns)
jam_igraph(cnet2, mark.groups=ns["SetA,SetB"], nodegroups=ns,
   main="Re-layout across all nodegroups")

}
\seealso{
Other jam cnet utilities: 
\code{\link{adjust_cnet_nodeset}()},
\code{\link{adjust_cnet_set_relayout_gene}()},
\code{\link{apply_nodeset_spacing}()},
\code{\link{bulk_cnet_adjustments}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{get_cnet_nodeset_vector}()},
\code{\link{launch_shinycat}()},
\code{\link{make_cnet_test}()}
}
\concept{jam cnet utilities}
