% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{relayout_with_qfr}
\alias{relayout_with_qfr}
\title{igraph re-layout using qgraph Fruchterman-Reingold}
\usage{
relayout_with_qfr(
  g,
  repulse = 3.5,
  spread_labels = TRUE,
  seed = 123,
  init = NULL,
  constrain = NULL,
  constraints = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object}

\item{repulse}{\code{numeric} exponent power used to scale the radius
effect around each vertex.
The default is slightly higher than the cube of
the number of vertices, but as the number of vertices increases,
values from 3.5 to 4 and higher are more effective for layout.}

\item{spread_labels}{\code{logical} indicating whether to call
\code{spread_igraph_labels()}, which places node labels at an angle offset
from the node, in order to improve default label positions.}

\item{seed}{\code{numeric} to set consistent random seed.}

\item{init}{optional \code{numeric} matrix with initial positions.
Default NULL will use existing layout coordinates if defined,
otherwise it uses NULL.}

\item{constrain}{\code{character} optional vector of node names that should
be constrained. This argument is a convenient shortcut for defining
\code{constraints}, which is a layout coordinate matrix with \code{NA} values
on each row where the coordinate is free to move, and \code{numeric}
values where the coordinate is fixed. For graph \code{g} that contains layout
in \code{igraph::graph_attr(g, "layout")}, the \code{init} can be defined with
this layout, then \code{constraints} is defined using \code{constrain}.}

\item{constraints}{optional two-column matrix with the coordinates of
nodes which should not be modified, and NA values for nodes where
the position can be modified.}

\item{verbose}{\code{logical} whether to print verbose output.}

\item{...}{additional arguments are passed to \code{layout_with_qfr()} and
\code{spread_igraph_labels()} as needed.}
}
\value{
\code{igraph} object, with layout coordinates stored in
graph attribute \code{"layout"}, accessible for example with
\code{graph$layout} or \code{graph_attr(graph, "layout")}.
When \code{spread_labels=TRUE},
\code{V(g)$label.degree} and \code{V(g)$label.dist} are updated
by calling \code{spread_igraph_labels()}.
}
\description{
igraph re-layout using qgraph Fruchterman-Reingold
}
\details{
This function extends \code{layout_with_qfr()} by applying the layout
to the \code{igraph} object itself, while also calling
\code{spread_igraph_labels()} to adjust label positions accordingly.

The main benefit to using this function is to update the layout
and node label positions in one step,
while also returning the \code{igraph} object ready to be plotted as-is.
}
\seealso{
Other jam igraph layouts:
\code{\link[=layout_communities]{layout_communities()}},
\code{\link[=layout_with_qfr]{layout_with_qfr()}},
\code{\link[=layout_with_qfrf]{layout_with_qfrf()}},
\code{\link[=removeIgraphSinglets]{removeIgraphSinglets()}},
\code{\link[=reorderIgraphNodes]{reorderIgraphNodes()}},
\code{\link[=rotate_igraph_layout]{rotate_igraph_layout()}},
\code{\link[=spread_igraph_labels]{spread_igraph_labels()}}
}
\concept{jam igraph layouts}
