% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{removeIgraphSinglets}
\alias{removeIgraphSinglets}
\title{Remove igraph singlet nodes}
\usage{
removeIgraphSinglets(g, min_degree = 1, ...)
}
\arguments{
\item{g}{igraph object}

\item{min_degree}{numeric threshold with the minimum number of
connections, also known as the "degree", required for each node.}

\item{...}{additional arguments are ignored.}
}
\description{
Remove igraph singlet nodes
}
\details{
This function is a lightweight method to remove igraph nodes with
no connections. In fact, the \code{min_degree} can be used to require
a minimum number of connections, but the intended use is to remove
the singlet nodes that have no connections.
}
\seealso{
Other jam igraph layouts:
\code{\link[=layout_communities]{layout_communities()}},
\code{\link[=layout_with_qfr]{layout_with_qfr()}},
\code{\link[=layout_with_qfrf]{layout_with_qfrf()}},
\code{\link[=relayout_with_qfr]{relayout_with_qfr()}},
\code{\link[=reorderIgraphNodes]{reorderIgraphNodes()}},
\code{\link[=rotate_igraph_layout]{rotate_igraph_layout()}},
\code{\link[=spread_igraph_labels]{spread_igraph_labels()}}
}
\concept{jam igraph layouts}
