% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{removeIgraphSinglets}
\alias{removeIgraphSinglets}
\title{Remove igraph singlet nodes}
\usage{
removeIgraphSinglets(g, min_degree = 1, ...)
}
\arguments{
\item{g}{igraph object}

\item{min_degree}{numeric threshold with the minimum number of
connections, also known as the "degree", required for each node.}

\item{...}{additional arguments are ignored.}
}
\description{
Remove igraph singlet nodes
}
\details{
This function is a lightweight method to remove igraph nodes with
no connections. In fact, the \code{min_degree} can be used to require
a minimum number of connections, but the intended use is to remove
the singlet nodes that have no connections.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
