% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{reorderIgraphNodes}
\alias{reorderIgraphNodes}
\title{Re-order igraph nodes}
\usage{
reorderIgraphNodes(
  g,
  sortAttributes = c("color", "coloredrect.color"),
  nodeSortBy = c("x", "y"),
  layout = function(...) layout_with_qfr(..., repulse = 3.5),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{igraph object}

\item{sortAttributes}{character vector of node attribute
names, to be applied in order when sorting nodes.}

\item{nodeSortBy}{character vector containing \code{"x"} and
\code{"y"} indicating the primary axis used to sort nodes.}

\item{layout}{numeric matrix of node coordinates, or
function used to produce layout coordinates. When layout
is \code{NULL}, this function tries to use node attributes
\code{"x"} and \code{"y"}. If those attributes do not exist,
the \code{layout_with_qfr()} is used with default \code{repulse=3.5}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
igraph with nodes positioned to order
nodes by color. The layout coordinates are stored in
the node attributes, accessible with \code{V(g)$x} and
\code{V(g)$y}. When there are not multiple nodes sharing
the same neighbors, the original igraph object is
returned, with the addition of layout coordinates.
}
\description{
Re-order igraph nodes
}
\details{
This function takes an igraph and a layout in the
form of coordinates, or a function used to produce
coordinates. It repositions nodes within equivalent
positions, ordering nodes by color along either the
\code{"x"} or \code{"y"} direction.

Equivalent node positions are those with the same
neighboring nodes. For example if node \code{"A"} and
node \code{"B"} both have neighbors \code{c("D", "E", "F")}
then nodes \code{"A"} and \code{"B"} are considered equivalent,
and will be reordered by their color.

This function is particularly effective with concept
network (Cnet) graphs, where multiple terms may
be connnected to the same concept. For MultiEnrichmap,
it typically works when multiple genes are connected
to the same pathways. When this happens, the genes
are sorted to group the colors.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{memIM2cnet}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{shape.coloredrectangle.plot}()},
\code{\link{shape.ellipse.plot}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subsetCnetIgraph}()}
}
\concept{jam igraph functions}
