% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-rotate.R
\name{rotate_coordinates}
\alias{rotate_coordinates}
\title{Rotate numeric coordinates}
\usage{
rotate_coordinates(
  x,
  degrees = 0,
  reflect = c("none", "x", "y", "z"),
  center = NULL,
  center_rule = c("origin", "mean", "median", "min", "max"),
  rotation_axes = c(1, 2),
  ...
)
}
\arguments{
\item{x}{\code{matrix} with 2 or more columns.}

\item{degrees}{numeric value indicating the degrees to rotate
layout coordinates, where 360 degrees is one complete rotation.}

\item{reflect}{\code{character} string indicating one or more axes
to reflect coordinates; \code{"none"} to reflect no axis; or
\code{integer} vector of column index positions. When there
are no \code{colnames(x)}, then \code{c("x","y","z")} are recognized
as columns \code{c(1, 2, 3)} respectively. Otherwise, when
\code{reflect} is a \code{character} vector it is matched to
\code{colnames(x)}.}

\item{center}{\code{numeric} coordinates to use as the center, or
\code{center=NULL} to calculate the center using \code{center_rule}.}

\item{center_rule}{\code{character} string indicating which rule to
apply to determine the center coordinates, when \code{center=NULL}:
\code{"origin"} uses c(0, 0); \code{"mean"} uses the mean of each axis;
\code{"median"} uses the median of each axis; \code{"min"} uses the minimum
of each axis; \code{"max"} uses the max of each axis.}

\item{rotation_axes}{\code{integer} vector indicating which axis
coordinates to rotate, by default \code{c(1, 2)} uses the first
two axes in \code{x}. Note that \code{rotation_axes} must represent
columns present in x.}

\item{...}{additional arguments are passed to
\code{multienrichjam::spread_igraph_labels()} and
\code{multienrichjam::reorderIgraphNodes()}.}
}
\value{
\verb{numeric matrix} with the same number of columns as the
input \code{x}.
}
\description{
Rotate numeric coordinates
}
\details{
This function rotates coordinates in two axes, by the angle
defined in \code{degrees}. It optionally reflects coordinates in
one or more axes.

Note that the \code{reflect} is applied before \code{degrees}.

Rotation code kindly contributed by Don MacQueen to the \code{maptools}
package, and is reproduced here to avoid a dependency on \code{maptools}
and therefore the \code{sp} package.
}
\examples{
layout <- cbind(0:10, 0:10);
layout_rot50 <- rotate_coordinates(x=layout, degrees=50);
layout_rot40_ctrmean <- rotate_coordinates(x=layout, degrees=40, center_rule="mean");
layout_reflectx_ctrmean <- rotate_coordinates(x=layout, reflect="x", center_rule="mean");
plot(rbind(layout, layout_rot50, layout_rot40_ctrmean, layout_reflectx_ctrmean),
   col=rep(c("darkorchid", "darkorange1", "dodgerblue", "red4"), each=11),
   pch=rep(c(17, 20, 18, 17), each=11),
   cex=2);

}
\seealso{
Other jam utility functions: 
\code{\link{apply_color_cap}()},
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{deconcat_df2}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{heatmap_column_order}()},
\code{\link{heatmap_row_order}()},
\code{\link{isColorBlank}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{sort_colors}()},
\code{\link{subgraph_jam}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
