% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{shape.coloredrectangle.plot}
\alias{shape.coloredrectangle.plot}
\title{plot function for igraph vertex shape coloredrectangle}
\usage{
shape.coloredrectangle.plot(coords, v = NULL, params)
}
\arguments{
\item{coords}{two-column numeric matrix with x- and y-coordinates,
respectively.}

\item{v}{optional ids of the vertices to plot. It should match
the number of rows in the \code{coords} argument.}

\item{params}{a function object that can be called to query
vertex/edge/plot graphical parameters. The first argument of
the function is \code{'vertex'}, \code{'edge'} or \code{'plot'} to decide
the type of the parameter, the second is a character string
giving the name of the parameter.}
}
\value{
invisible \code{list} of \code{data.frame} objects which were
used to draw the rectangle objects.
However the purpose of this function is the by-product that it
draws rectangles onto an igraph graph.
}
\description{
plot function for igraph vertex shape coloredrectangle
}
\details{
This function defines the plotting function for custom igraph vertex
shape coloredrectangle. The coloredrectangle shape is described as:

\itemize{
\item{a vertex drawn as a rectangle, filled with squares which are
each individually colored.}
\item{the squares are arrayed into a number of columns and rows,
which are defined for each vertex using \code{coloredrect.ncol} and
\code{coloredrect.nrow}, respectively.}
\item{the vector of colors is arrayed as values of a matrix, therefore
\code{coloredrect.byrow} is logical to indicate whether colors should fill
the matrix by row, similar to how \code{byrow=TRUE} is used.}
\item{the colors for each vertex are defined by
\code{coloredrect.color}. When this value does not exist, this function
will attempt to use values from \code{pie.color} if they exist.}
\item{the size of the rectangle is defined by \code{size2}, which is
typically the height of a \code{rectangle} node. The intent was to use
a size parameter which is already convenient, but which does not
conflict with the size used for vertex shapes such as \code{"circle"}.}
\item{when a vertex has 3 columns, and 2 rows, and only 5 colors,
the colors are not cycled to fill the complete node. Instead, the
last color is \code{"transparent"} to render no color in that square.}
\item{the frame around each node is described with \code{frame.color} and
\code{frame.lwd} which draws one rectangular border around the full
vertex. The border around all coloredrect vertices is drawn before the
squares are drawn for all coloredrect vertices, so that the border will
not overlap the squares. It has the visible effect of showing vertex
borders behind the vertex colors, while allowing for vectorized drawing,
which is substantially more efficient than per-vertex rendering.
The \code{graphics::symbols()} function is used to render rectangles,
it accepts only one value for \code{lwd} and \code{lty}, so rectangles
are split by \code{lwd} and \code{lty} and rendered in order. As a
result, if each vertex frame had a different lwd,lty combination, the
rendering would effectively be non-vectorized. For large numbers of
vertices and distinct lwd values, it would be preferred to reduce the
lwd values to limited significant digits (e.g. with
\code{signif(...,digits=2)}). That said, line width is not an optimal
way to convey a quantitative measurement.}
\item{a colored border can be drawn around each square in the vertex,
defined by \code{coloredrect.border} and \code{coloredrect.lwd},
however it is recommended to use
this color only as a visible break, for example the default is
\code{"grey30"} which draws a simple border. Note that squares are
rendered after the vertex frame color, so each square color border will
be drawn over (on top of) the frame border.}
\item{when \code{coloredrect.ncol} does not exist, it will attempt
to use two rows of colors if \code{coloredrect.color} contains
two or more values.}
\item{when \code{coloredrect.nrow} does not exist, it will use a value
based upon \code{coloredrect.ncol} and the length of
\code{coloredrect.color}.}
}

Currently, the plotting of coloredrectangle vertices does not define
clipping, which means edges are drawn to the center of each vertex,
and the coloredrectangle vertex shapes are drawn on top of the edges.
Transparent nodes will therefore show edges beneath them.
}
\seealso{
\code{igraph::shapes()}

Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{drawEllipse}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{igraph2pieGraph}()},
\code{\link{layout_with_qfrf}()},
\code{\link{layout_with_qfr}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{shape.ellipse.plot}()},
\code{\link{subsetCnetIgraph}()}
}
\concept{jam igraph functions}
