% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{shape.ellipse.plot}
\alias{shape.ellipse.plot}
\title{plot function for igraph vertex shape ellipse}
\usage{
shape.ellipse.plot(coords, v = NULL, params)
}
\description{
plot function for igraph vertex shape ellipse
}
\details{
This function defines the plotting function for custom igraph vertex
shape ellipse.
}
\seealso{
Other jam igraph functions: \code{\link{cnet2df}},
  \code{\link{cnet2im}}, \code{\link{cnetplotJam}},
  \code{\link{cnetplot_internalJam}},
  \code{\link{drawEllipse}}, \code{\link{igraph2pieGraph}},
  \code{\link{layout_with_qfr}},
  \code{\link{shape.coloredrectangle.plot}},
  \code{\link{xyAngle}}
}
\concept{jam igraph functions}
