% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraphshapes.R
\name{shape.jampie.plot}
\alias{shape.jampie.plot}
\title{plot function for igraph vertex shape jampie}
\usage{
shape.jampie.plot(coords, v = NULL, params)
}
\description{
plot function for igraph vertex shape jampie
}
\details{
This function is a vectorized replacement for plotting
vertex shape \code{"pie"} in much more efficient manner.
}
\seealso{
Other jam igraph shapes: 
\code{\link{jam_mypie}()},
\code{\link{shape.coloredrectangle.plot}()},
\code{\link{shape.ellipse.plot}()}
}
\concept{jam igraph shapes}
