% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycat-ui-launch.R
\name{shinycat_ui}
\alias{shinycat_ui}
\title{R-shiny ui function for shinycat}
\usage{
shinycat_ui(...)
}
\value{
A UI definition that can be passed to \code{shiny::shinyApp()}
as argument 'ui'.
}
\description{
R-shiny ui function for shinycat, internal use by launch_shinycat()
}
\details{
Internal use by \code{launch_shinycat()}, this function provides the
'ui' component for the R-shiny app.
}
\seealso{
Other jam shiny functions: 
\code{\link{launch_shinycat}()},
\code{\link{shinycat_server}()}
}
\concept{jam shiny functions}
