% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{subgraph_jam}
\alias{subgraph_jam}
\title{Subgraph using Jam extended logic}
\usage{
subgraph_jam(graph, v)
}
\arguments{
\item{graph}{\code{igraph} object}

\item{v}{\code{integer} or \code{logical} vector indicating the nodes to
retain in the final \code{igraph} object.}
}
\description{
Subgraph using Jam extended logic
}
\details{
This function extends the \code{igraph::subgraph()} function
to include proper subset of the graph attribute \code{"layout"},
which for some unknown reason does not subset the layout
matrix consistent with the subset of \code{igraph} nodes.
}
\seealso{
Other jam igraph utilities: 
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{get_cnet_nodeset}()},
\code{\link{jam_igraph}()},
\code{\link{launch_shinycat}()}
}
\concept{jam igraph utilities}
