% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{subsetCnetIgraph}
\alias{subsetCnetIgraph}
\title{Subset Cnet igraph}
\usage{
subsetCnetIgraph(gCnet, includeSets = NULL, includeGenes = NULL,
  removeSinglets = TRUE, minSetDegree = 1, minGeneDegree = 1,
  verbose = FALSE, ...)
}
\arguments{
\item{gCnet}{igraph object representing Cnet concept network data}

\item{includeSets}{character vector, or NULL, containing the set
names or labels to retain.}

\item{includeGenes}{character vector, or NULL, containing the gene
names or labels to retain.}

\item{removeSinglets}{logical whether to remove singlet graph nodes,
which are nodes that have no remaining edges.}

\item{minSetDegree}{integer value indicating the minimum number
of edges each Set node must have to be retained in the resulting
igraph. Use `minSetDegree=2` to retain only Set nodes that
have multiple connected Gene nodes.}

\item{minGeneDegree}{integer value indicating the minimum number
of edges each Gene node must have to be retained in the resulting
igraph. Use `minGeneDegree=2` to retain only Gene nodes that
connect to multiple Set nodes.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\description{
Subset Cnet igraph
}
\details{
This function produces a subset of a Cnet igraph based upon supplied
set names or gene names. This function is intended to be a convenient
method of filtering a Cnet igraph to a pre-defined set of "Set"
names.

The function assumes graph nodes have an attribute `"nodeType"` with
values either `"Set"` or `"Gene"` to indicate the type of node.

When `includeSets` is supplied, the graph is subsetted to include
only nodes with `nodeType="Set"` with matching `V(gCnet)$name` or
`V(gCnet)$label`. Then only neighboring nodes are retained, thus
removing any nodes with `nodeType="Gene"` that do not connect to
any of the given Set nodes.
The result is a proper Cnet igraph that only contains
Gene nodes connected to the subset of Set nodes.

If `includeGenes` is supplied, the graph is subsetted to include
only nodes with `nodeType="Gene"` with matching `V(gCnet)$name` or
`V(gCnet)$label`.

When `removeSinglets=TRUE` then any nodes that have no remaining
edges are removed. Especially when supplying `includeGenes`, this
option is useful to hide any Set nodes that have no connected Gene
nodes.
}
