% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-summarize-node-spacing.R
\name{summarize_node_spacing}
\alias{summarize_node_spacing}
\title{Summarize spacing between igraph nodes given a layout}
\usage{
summarize_node_spacing(
  g,
  layout = NULL,
  nodes = NULL,
  node_groups = NULL,
  each_group = FALSE,
  scaled = FALSE,
  dist_type = c("nearest_node", "all_nodes"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{g}{\code{igraph} object}

\item{layout}{passed together with \code{g} to \code{get_igraph_layout()}}

\item{nodes}{\code{character} with optional node names, or \code{integer}
index of nodes in \code{g} to define a subset of nodes for which
statistics are calculated. Useful to focus on a specific subset
of nodes, for example one or two Cnet nodesets.}

\item{node_groups}{\code{list} implemented to use nodesets. The intent is to
define node groups, then calculate statistics of node spacing across
and within node groups.}

\item{each_group}{\code{logical} indicating whether to include each node group
when \code{node_groups} is also supplied.}

\item{scaled}{\code{logical} indicating whether to report spacing
relative to the max x-axis/y-axis range, similar to the \code{min_percent}
and \code{percent_spacing} argument units in other node spacing functions.}

\item{dist_type}{\code{character} string indicating the type of distance to
summarize:
\itemize{
\item \code{"nearest_node"} only uses the nearest node to each node, which is
helpful when trying to ensure all nodes have a minimum distance
from other nodes.
\item \code{"all_nodes"} uses all node distances from each node, which is
helpful when assessing the overall spacing between nodes.
}}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to internal functions.}
}
\value{
\code{data.frame} with summary information about node distances
for connected nodes (only where edges connect any two nodes),
and unconnected nodes (only where two nodes are not connected
by an edge). When \code{node_groups} is defined, the summary
also includes each individual node group.
}
\description{
Summarize spacing between igraph nodes given a layout
}
\details{
This function is a simple wrapper to calculate typical distances
between nodes in a given network layout. It is experimental,
and intended to provide helpful information when determining
an appropriate value for \code{percent_spacing} to use for example
with \code{apply_nodeset_spacing()}. The optimal value depends upon
the number of nodes overall, also the number of nodes in
each nodeset, and the relative position of each nodeset
in layout coordinates.

The \code{node_groups} argument is intended to provide summary data
for each node group (for example Cnet \code{nodesets}) so that
individual node groups can be adjusted accordingly.
}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{call_fn_ellipsis_deprecated}()},
\code{\link{cell_fun_bivariate}()},
\code{\link{collapse_mem_clusters}()},
\code{\link{colorRamp2D}()},
\code{\link{deconcat_df2}()},
\code{\link{display_colorRamp2D}()},
\code{\link{enrichList2geneHitList}()},
\code{\link{filter_mem_genes}()},
\code{\link{filter_mem_sets}()},
\code{\link{find_colname}()},
\code{\link{get_hull_data}()},
\code{\link{get_igraph_layout}()},
\code{\link{gsubs_remove}()},
\code{\link{handle_igraph_param_list}()},
\code{\link{isColorBlank}()},
\code{\link{make_legend_bivariate}()},
\code{\link{make_point_hull}()},
\code{\link{mem_find_overlap}()},
\code{\link{order_colors}()},
\code{\link{rank_mem_clusters}()},
\code{\link{rotate_coordinates}()},
\code{\link{subgraph_jam}()},
\code{\link{subset_mem}()},
\code{\link{xyAngle}()}
}
\concept{jam utility functions}
