% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{sync_igraph_communities}
\alias{sync_igraph_communities}
\title{Sync igraph nodes and communities}
\usage{
sync_igraph_communities(g, wc, verbose = TRUE, ...)
}
\arguments{
\item{g}{\code{igraph} object}

\item{wc}{\code{communities} object, or \code{list} in form of nodegroups,
which is a \code{list} of \code{character} vectors that contain \code{igraph}
node names.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{nodegroups2communities()}
only when input \code{wc} is supplied in \code{list} nodegroups format.}
}
\value{
\code{list} with two elements:
\itemize{
\item \code{"g"} - the \code{igraph} object after subsetting to match node names
shared with \code{wc}, as necessary.
\item \verb{"wc'} - the \code{communities} object after subsetting to match
node names shared with \code{g}, as necessary. When input \code{wc} is
in \code{list} nodegroups format, that same format is returned.
}
}
\description{
Sync igraph nodes and communities
}
\details{
This function ensures that \code{igraph} nodes and corresponding
community clusters are synchronized for proper downstream use.
In particular, when using a subgraph, or when communities only
assign a subset of nodes to clusters, this function ensures the
two objects are in sync, the same order, and with the same nodes.
}
\seealso{
Other jam igraph functions: 
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{label_communities}()},
\code{\link{mem2cnet}()},
\code{\link{mem2emap}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()}
}
\concept{jam igraph functions}
