% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-communities-nodegroups.R
\name{sync_igraph_communities}
\alias{sync_igraph_communities}
\title{Sync igraph nodes and communities}
\usage{
sync_igraph_communities(g, wc, verbose = TRUE, ...)
}
\arguments{
\item{g}{\code{igraph} object}

\item{wc}{\code{communities} object, or \code{list} in form of nodegroups,
which is a \code{list} of \code{character} vectors that contain \code{igraph}
node names.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{nodegroups2communities()}
only when input \code{wc} is supplied in \code{list} nodegroups format.}
}
\value{
\code{list} with two elements:
\itemize{
\item \code{"g"} - the \code{igraph} object after subsetting to match node names
shared with \code{wc}, as necessary.
\item \verb{"wc'} - the \code{communities} object after subsetting to match
node names shared with \code{g}, as necessary. When input \code{wc} is
in \code{list} nodegroups format, that same format is returned.
}
}
\description{
Sync igraph nodes and communities
}
\details{
This function ensures that \code{igraph} nodes and corresponding
community clusters are synchronized for proper downstream use.
In particular, when using a subgraph, or when communities only
assign a subset of nodes to clusters, this function ensures the
two objects are in sync, the same order, and with the same nodes.
}
\seealso{
Other jam igraph functions: 
\code{\link{cnet2df}()},
\code{\link{cnet2im}()},
\code{\link{cnetplotJam}()},
\code{\link{cnetplot_internalJam}()},
\code{\link{color_edges_by_nodegroups}()},
\code{\link{color_edges_by_nodes}()},
\code{\link{color_edges_by_nodes_deprecated}()},
\code{\link{color_nodes_by_nodegroups}()},
\code{\link{communities2nodegroups}()},
\code{\link{drawEllipse}()},
\code{\link{edge_bundle_bipartite}()},
\code{\link{edge_bundle_nodegroups}()},
\code{\link{enrichMapJam}()},
\code{\link{fixSetLabels}()},
\code{\link{flip_edges}()},
\code{\link{get_bipartite_nodeset}()},
\code{\link{highlight_edges_by_node}()},
\code{\link{igraph2pieGraph}()},
\code{\link{jam_igraph}()},
\code{\link{jam_plot_igraph}()},
\code{\link{label_communities}()},
\code{\link{layout_with_qfr}()},
\code{\link{layout_with_qfrf}()},
\code{\link{mem2emap}()},
\code{\link{memIM2cnet}()},
\code{\link{mem_multienrichplot}()},
\code{\link{nodegroups2communities}()},
\code{\link{rectifyPiegraph}()},
\code{\link{relayout_with_qfr}()},
\code{\link{removeIgraphBlanks}()},
\code{\link{removeIgraphSinglets}()},
\code{\link{reorderIgraphNodes}()},
\code{\link{rotate_igraph_layout}()},
\code{\link{spread_igraph_labels}()},
\code{\link{subgraph_jam}()},
\code{\link{subsetCnetIgraph}()},
\code{\link{subset_igraph_components}()},
\code{\link{with_qfr}()}
}
\concept{jam igraph functions}
