% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-topenrich.R
\name{topEnrichBySource}
\alias{topEnrichBySource}
\alias{topEnrichListBySource}
\title{Subset enrichResult for top enrichment results by source}
\usage{
topEnrichBySource(
  enrichDF,
  n = 15,
  min_count = 1,
  p_cutoff = 1,
  sourceColnames = c("gs_cat", "gs_subcat"),
  sortColname = NULL,
  countColname = c("gene_count", "count", "geneHits"),
  pvalueColname = c("P.Value", "pvalue", "FDR", "adj.P.Val", "qvalue"),
  directionColname = c("direction", "NES", "activation.z.{0,1}score", "z.{0,1}score"),
  direction_cutoff = 0,
  newColname = "EnrichGroup",
  curateFrom = NULL,
  curateTo = NULL,
  sourceSubset = NULL,
  sourceSep = "_",
  subsetSets = NULL,
  descriptionColname = c("Description", "Name", "Pathway"),
  nameColname = c("ID", "Name"),
  descriptionGrep = NULL,
  nameGrep = NULL,
  verbose = FALSE,
  ...
)

topEnrichListBySource(
  enrichList,
  n = 15,
  min_count = 1,
  p_cutoff = 1,
  sourceColnames = c("gs_cat", "gs_subcat"),
  sortColname = c(pvalueColname, "P-value", "pvalue", "qvalue", "padjust", "-GeneRatio",
    "-Count", "-geneHits"),
  countColname = c("gene_count", "count", "geneHits"),
  pvalueColname = c("P.Value", "pvalue", "FDR", "adj.P.Val", "qvalue"),
  directionColname = c("direction", "NES", "activation.z.{0,1}score", "z.{0,1}score"),
  direction_cutoff = 1,
  newColname = "EnrichGroup",
  curateFrom = NULL,
  curateTo = NULL,
  sourceSubset = NULL,
  sourceSep = "_",
  subsetSets = NULL,
  descriptionColname = c("Description", "Name", "Pathway"),
  nameColname = c("ID", "Name"),
  descriptionGrep = NULL,
  nameGrep = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichDF}{\code{enrichResult} or \code{data.frame} with enrichment results.}

\item{n}{\code{integer} maximum number of pathways to retain,
after applying \code{min_count} and \code{p_cutoff} thresholds
if relevant.}

\item{min_count}{\code{integer} minimum number of genes involved
in an enrichment result to be retained, based upon values
in \code{countColname}.}

\item{p_cutoff}{\code{numeric} value indicating the enrichment
P-value threshold, pathways with enrichment P-value at
or below this threshold are retained, based upon values
in \code{pvalueColname}.}

\item{sourceColnames}{\code{character} vector of colnames in
\code{enrichDF} to consider as the \code{"Source"}. Multiple
columns will be combined using delimiter argument
\code{sourceSep}. When \code{sourceColnames} is NULL or
contains no \code{colnames(enrichDF)}, then data
is considered \code{"All"}.}

\item{sortColname}{\code{character} vector, default \code{NULL},
indicating the colnames to sort/prioritize the enrichment
data rows. Please use \code{NULL}.
\itemize{
\item Default \code{NULL} will use \code{pvalueColname} and the
reverse of \code{countColname}, to prioritize lowest P-value,
then highest gene count.
\item When \code{FALSE} it will not perform any sorting, and will
use the input data as-is.
\item When \code{character} vector is provided, its values must
exactly match the intended colnames, with optional
prefix \code{"-"} to indicate reverse sort for a particular
colname. These values are passed to \code{jamba::mixedSortDF()}
argument \code{byCols}.
}}

\item{countColname}{\code{character} vector of possible colnames
in \code{enrichDF} that should contain the \code{integer} number
of genes involved in enrichment. This vector is
passed to \code{find_colname()} to find an appropriate
matching colname in \code{enrichDF}.}

\item{pvalueColname}{\code{character} vector of possible colnames
in \code{enrichDF} that should contain the enrichment P-value
used for filtering by \code{p_cutoff}.}

\item{directionColname}{\code{character} vector of possible colnames
in \code{enrichDF} which may contain directional z-score, or
other metric used to indicate directionality. It is assumed
to be symmetric around zero, where zero indicates no
directional bias.}

\item{direction_cutoff}{\code{numeric} threshold (default \code{0}) to subset
enriched sets, filtering by magnitude of the absolute value
of the \code{directionColname}.}

\item{newColname}{\code{character} string with new column name
when \code{sourceColname} matches multiple colnames in \code{enrichDF}.
Values for each row are combined using \code{jamba::pasteByRow()}.}

\item{curateFrom, curateTo}{\code{character} vectors with
pattern,replacement values, passed to \code{gsubs()}
to allow some editing of values. The default values
convert MSigDB canonical pathways from the prefix \code{"CP:"}
to use \code{"CP"} which has the effect of combining all
canonical pathways before selecting the top \code{n} results.}

\item{sourceSubset}{\code{character} vector with a subset of
sources to retain. If there are multiple colnames in
\code{sourceColnames}, then column values are combined
using \code{jamba::pasteByRow()} and delimiter \code{sourceSep},
prior to filtering.}

\item{sourceSep}{\code{character} string used as a delimiter
when \code{sourceColnames} contains multiple colnames.}

\item{subsetSets}{\code{character} optional set names to include
by exact match.}

\item{descriptionColname, nameColname}{character vectors
indicating the colnames to consider description and name,
as returned from \code{find_colname()}. These arguments are
used only when \code{descriptionGrep} or \code{nameGrep} are
supplied.}

\item{descriptionGrep, nameGrep}{\code{character} vector of
regular expression patterns, intended to subset pathways
to include only those matching these patterns.
The \code{descriptionGrep} argument searches only \code{descriptionColname}.
The \code{nameGrep} argument searches only \code{nameColname}.
Note that the sets are combined with OR logic, such that
any pathways matched by \code{descriptionGrep} OR \code{nameGrep}
or \code{subsetSets} will be included in the output.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}

\item{enrichList}{\code{list} of \code{enrichDF} entries, each passed
to \code{topEnrichBySource()}.}
}
\value{
\code{data.frame} subset up to \code{topEnrichN} rows, after
applying optional \code{min_count} and \code{p_cutoff} filters.
}
\description{
Subset enrichResult for top enrichment results by source

Subset list of enrichResult for top enrichment results by source
}
\details{
This function takes one \code{enrichResult} object, or
a \code{data.frame} of enrichment results, and determines the
top \code{n} number of pathways sorted by P-values, within
each pathway source. This function may optionally require
\code{min_count} genes in each pathway, and \code{p_cutoff} maximum
enrichment P-value, prior to taking the top \code{topEnrichN}
entries. The default arguments do not apply filters
to \code{min_count} and \code{p_cutoff}.

When the enrichment data represents pathways from multiple
sources, the filtering and sorting is applied to each source
independently. The intent is to retain the top entries from
each source, as a method of representing each source
consistently even when one source may contain many more
pathways, and importantly where the range of enrichment P-values
may be very different for each source. For example, a database
of small canonical pathways would generally provide less
statistically significant P-values than a database of
dysregulated genes from gene expression experiments, where
each set contains a large number of genes.

This function can optionally apply basic curation of pathway
source names, and can optionally be applied to multiple
source columns. This feature is intended for sources like
MSigDB (see http://software.broadinstitute.org/gsea/msigdb/index.jsp)
which contains columns \code{"Source"} and \code{"Category"},
and where canonical pathways are either represented with \code{"CP"}
or a prefix \code{"CP:"}. The default parameters recognize this
case and curates all prefix \code{"CP:.*"} down to just \code{"CP"}
so that all canonical pathways are considered to be the
same source. For MSigDB there are also numerous other sources,
which are each independently filtered and sorted to the
top \code{topEnrichN} entries.

Finally, this function is useful to subset enrichment results
by name, using \code{descriptionGrep}, \code{nameGrep}, or \code{subsetSets}.

\code{topEnrichListBySource()} extends \code{topEnrichBySource()} by applying
filters to each \code{enrichList} entry, then keeping pathways
across all \code{enrichList} that match the filter criteria in any
one \code{enrichList}. It is most useful in the context of
\code{multiEnrichMap()} where a pathway must meet all criteria
in at least one enrichment, and that pathway should then
be included for all enrichments for the purpose of
comparative analysis.
}
\seealso{
Other jam enrichment functions: 
\code{\link{add_pathway_direction}()}

Other jam enrichment functions: 
\code{\link{add_pathway_direction}()}
}
\concept{jam enrichment functions}
