% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-base.r
\name{topEnrichListBySource}
\alias{topEnrichListBySource}
\title{Subset enrichList for top enrichment results by source}
\usage{
topEnrichListBySource(
  enrichList,
  n = 15,
  sourceColnames = c("Category", "Source"),
  sortColname = c("P-value", "pvalue", "padjust", "-GeneRatio", "-geneHits"),
  newColname = "EnrichGroup",
  curateFrom = c("CP:.*"),
  curateTo = c("CP"),
  sourceSubset = NULL,
  sourceSep = "_",
  descriptionColname = c("Description", "Name", "Pathway"),
  nameColname = c("ID", "Name"),
  descriptionGrep = NULL,
  nameGrep = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{enrichList}{\code{list} of \code{enrichDF} entries, each passed
to \code{topEnrichBySource()}.}

\item{sourceColnames}{character vector of colnames in
\code{enrichDF} to consider as the \code{"Source"}. Multiple
columns will be combined using delimiter argument
\code{sourceSep}. When \code{sourceColnames} is NULL or
contains no \code{colnames(enrichDF)}, then data
is considered \code{"All"}.}

\item{sortColname}{character vector indicating the colnames
to use to sort data, prior to selecting the top \code{n}
results by source. This argument is passed to
\code{jamba::mixedSortDF(x, byCols=sortColname)}. Columns
can be sorted in reverse order by using the prefix \code{"-"},
as described in \code{jamba::mixedSortDF()}.}

\item{newColname}{new column name to use when \code{sourceColname}
contains multiple values.}

\item{curateFrom}{character vectors with
pattern,replacement values, passed to \code{gsubs()}
to allow some editing of values. The default values
convert MSigDB canonical pathways from the prefix \code{"CP:"}
to use \code{"CP"} which has the effect of combining all
canonical pathways before selecting the top \code{n} results.}

\item{curateTo}{character vectors with
pattern,replacement values, passed to \code{gsubs()}
to allow some editing of values. The default values
convert MSigDB canonical pathways from the prefix \code{"CP:"}
to use \code{"CP"} which has the effect of combining all
canonical pathways before selecting the top \code{n} results.}

\item{sourceSubset}{character vector with a subset of
sources to retain. This filter is applied after
\code{sourceColname} values are combined with delimiter
\code{sourceSep} if there are multiple colnames in \code{sourceColname}.}

\item{sourceSep}{character string used as a delimiter
when \code{sourceColname} contains multiple colnames.}

\item{descriptionColname}{character vectors
indicating the colnames to consider description and name,
as returned from \code{find_colname()}. These arguments are
used only when \code{descriptionGrep} or \code{nameGrep} are
supplied.}

\item{nameColname}{character vectors
indicating the colnames to consider description and name,
as returned from \code{find_colname()}. These arguments are
used only when \code{descriptionGrep} or \code{nameGrep} are
supplied.}

\item{descriptionGrep}{character vector of patterns, used
to filter pathways to those matching one or more patterns.
This argument is used to help extract a specific subset
of pathways of interest using keywords.
The \code{descriptionGrep} argument searches only colname \code{"Description"}.
The \code{nameGrep} argument searches only colname \code{"Name"}.}

\item{nameGrep}{character vector of patterns, used
to filter pathways to those matching one or more patterns.
This argument is used to help extract a specific subset
of pathways of interest using keywords.
The \code{descriptionGrep} argument searches only colname \code{"Description"}.
The \code{nameGrep} argument searches only colname \code{"Name"}.}

\item{verbose}{logical indicating whether to print verbose output.}
}
\description{
Subset enrichList for top enrichment results by source
}
\details{
This function extends \code{topEnrichBySource()} by applying
consistent filters to each \code{enrichList} entry. The purpose
of this function is to generate an overall set of pathways
to return from any one or more enrichment results. Then
apply that filter consistently across all enrichment results.

For example, if \code{"Pathway A"} is the top \code{n} enriched
pathway in \code{enrichList[[1]]}, then \code{"Pathway A"} will be included
for every entry in \code{enrichList}, even when \code{"Pathway A"} is
not significant in \code{enrichList[[2]]}. The purpose is to allow
comparison of enrichment results when a pathway meets the
filter criteria in any one entry in \code{enrichList}.
}
\seealso{
Other jam enrichment functions: 
\code{\link{multiEnrichMap}()},
\code{\link{topEnrichBySource}()}
}
\concept{jam enrichment functions}
