% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{xyAngle}
\alias{xyAngle}
\title{Get angle from origin to vector of x,y coordinates}
\usage{
xyAngle(
  x,
  y = NULL,
  directed = FALSE,
  deg = TRUE,
  origin.x = 0,
  origin.y = 0,
  ...
)
}
\arguments{
\item{x}{numeric vector or two-column matrix with columns
representing x,y coordinates when y is \code{NULL}.}

\item{y}{numeric vector or \code{NULL}.}

\item{directed}{logical indicating whether to return angles around
the full circle, or only half circle. For example, in degrees
\code{c(1,1)} indicates 45 degrees, \code{c(-1,-1)} indicates -135 degrees.
When \code{directed=FALSE} then \code{c(-1,-1)} indicates 45 degrees.}

\item{deg}{logical indicating whether to return degrees, or when
\code{deg=FALSE} it returns radians.}

\item{origin.x, origin.y}{numeric input defining the coordinates
to use as the origin. When non-zero it implies the first point
of each segment.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} vector of angles in radians.
}
\description{
Get angle from origin to vector of x,y coordinates
}
\details{
This function gets the angle from origin to x,y coordinates,
allowing for vectorized input and output.
}
\examples{
# by default output is in degrees
xyAngle(1, 1);

# output in radians
xyAngle(1, 1, deg=FALSE);

# optionally different origin
xyAngle(1, 1, origin.x=1, origin.y=0);

}
\seealso{
Other jam utility functions:
\code{\link[=ashape]{ashape()}},
\code{\link[=avg_angles]{avg_angles()}},
\code{\link[=avg_colors_by_list]{avg_colors_by_list()}},
\code{\link[=cell_fun_bivariate]{cell_fun_bivariate()}},
\code{\link[=collapse_mem_clusters]{collapse_mem_clusters()}},
\code{\link[=colorRamp2D]{colorRamp2D()}},
\code{\link[=curateIPAcolnames]{curateIPAcolnames()}},
\code{\link[=deconcat_df2]{deconcat_df2()}},
\code{\link[=display_colorRamp2D]{display_colorRamp2D()}},
\code{\link[=enrichList2geneHitList]{enrichList2geneHitList()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=find_enrich_colnames]{find_enrich_colnames()}},
\code{\link[=get_hull_data]{get_hull_data()}},
\code{\link[=get_igraph_layout]{get_igraph_layout()}},
\code{\link[=gsubs_remove]{gsubs_remove()}},
\code{\link[=handle_igraph_param_list]{handle_igraph_param_list()}},
\code{\link[=isColorBlank]{isColorBlank()}},
\code{\link[=make_legend_bivariate]{make_legend_bivariate()}},
\code{\link[=make_point_hull]{make_point_hull()}},
\code{\link[=mem_find_overlap]{mem_find_overlap()}},
\code{\link[=order_colors]{order_colors()}},
\code{\link[=rank_mem_clusters]{rank_mem_clusters()}},
\code{\link[=rotate_coordinates]{rotate_coordinates()}},
\code{\link[=summarize_node_spacing]{summarize_node_spacing()}},
\code{\link[=with_ht_opts]{with_ht_opts()}}
}
\concept{jam utility functions}
