% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamenrich-igraph.r
\name{xyAngle}
\alias{xyAngle}
\title{Get angle from origin to vector of x,y coordinates}
\usage{
xyAngle(
  x,
  y = NULL,
  directed = FALSE,
  deg = TRUE,
  origin.x = 0,
  origin.y = 0,
  ...
)
}
\arguments{
\item{x}{numeric vector or two-column matrix with columns
representing x,y coordinates when y is \code{NULL}.}

\item{y}{numeric vector or \code{NULL}.}

\item{directed}{logical indicating whether to return angles around
the full circle, or only half circle. For example, in degrees
\code{c(1,1)} indicates 45 degrees, \code{c(-1,-1)} indicates -135 degrees.
When \code{directed=FALSE} then \code{c(-1,-1)} indicates 45 degrees.}

\item{deg}{logical indicating whether to return degrees, or when
\code{deg=FALSE} it returns radians.}

\item{origin.x, origin.y}{numeric input defining the coordinates
to use as the origin. When non-zero it implies the first point
of each segment.}

\item{...}{additional arguments are ignored.}
}
\description{
Get angle from origin to vector of x,y coordinates
}
\details{
This function gets the angle from origin to x,y coordinates,
allowing for vectorized input and output.
}
\examples{
# by default output is in degrees
xyAngle(1, 1);

# output in radians
xyAngle(1, 1, deg=FALSE);

# optionally different origin
xyAngle(1, 1, origin.x=1, origin.y=0);

}
\seealso{
Other jam utility functions: 
\code{\link{avg_angles}()},
\code{\link{avg_colors_by_list}()},
\code{\link{find_colname}()},
\code{\link{gsubs}()},
\code{\link{isColorBlank}()}
}
\concept{jam utility functions}
